LR.sarlm {spdep}	R Documentation

Likelihood ratio test

Description

The LR.sarlm() function provides a likelihood ratio test for objects for which a logLik() function exists for their class, or for objects of class logLik. LR1.sarlm() and Wald1.sarlm() are used internally in summary.sarlm(), but may be accessed directly; they report the values respectively of LR and Wald tests for the absence of spatial dependence in spatial lag or error models. The spatial Hausman test is available for models fitted with errorsarlm and GMerrorsar.

Usage

LR.sarlm(x, y)
## S3 method for class 'sarlm'
logLik(object, ...)
LR1.sarlm(object)
Wald1.sarlm(object)
## S3 method for class 'sarlm'
Hausman.test(object, ..., tol=NULL)
## S3 method for class 'gmsar'
Hausman.test(object, ..., tol=NULL)

Arguments

x
a logLik object or an object for which a logLik() function exists

y
a logLik object or an object for which a logLik() function exists

object
a sarlm object from lagsarlm() or errorsarlm()

...
further arguments passed to or from other methods

tol
tol argument passed to solve, default NULL

Value

The tests return objects of class htest with:

statistic
value of statistic

parameter
degrees of freedom

p.value
Probability value

estimate
varies with test

method
description of test method

logLik.sarlm() returns an object of class logLik LR1.sarlm, Hausman.sarlm and Wald1.sarlm returm objects of class htest

Note

The numbers of degrees of freedom returned by logLik.sarlm() include nuisance parameters, that is the number of regression coefficients, plus sigma, plus spatial parameter esitmate(s).

Author(s)

Roger Bivand Roger.Bivand@nhh.no

References

LeSage J and RK Pace (2009) Introduction to Spatial Econometrics. CRC Press, Boca Raton, pp. 61–63; Pace RK and LeSage J (2008) A spatial Hausman test. Economics Letters 101, 282–284.

See Also

logLik.lm, anova.sarlm

Examples

example(columbus)
mixed <- lagsarlm(CRIME ~ HOVAL + INC, data=columbus, nb2listw(col.gal.nb),
  type="mixed")
error <- errorsarlm(CRIME ~ HOVAL + INC, data=columbus, nb2listw(col.gal.nb))
LR.sarlm(mixed, error)
Hausman.test(error)

[Package spdep version 0.6-6 Index]