NY_data {spdep}	R Documentation

New York leukemia data

Description

New York leukemia data taken from the data sets supporting Waller and Gotway 2004 (the data should be loaded by running example(NY_data) to demonstrate spatial data import techniques).

Usage

data(NY_data)

Format

A data frame with 281 observations on the following 12 variables, and the binary coded spatial weights used in the source.

AREANAME

name of census tract

AREAKEY

unique FIPS code for each tract

X

x-coordinate of tract centroid (in km)

Y

y-coordinate of tract centroid (in km)

POP8

population size (1980 U.S. Census)

TRACTCAS

number of cases 1978-1982

PROPCAS

proportion of cases per tract

PCTOWNHOME

percentage of people in each tract owning their own home

PCTAGE65P

percentage of people in each tract aged 65 or more

Z

ransformed propoprtions

AVGIDIST

average distance between centroid and TCE sites

PEXPOSURE

"exposure potential": inverse distance between each census tract centroid and the nearest TCE site, IDIST, transformed via log(100*IDIST)

Details

The examples section shows how the DBF files from the book website for Chapter 9 were converted into the nydata data frame and the listw_NY spatial weights list.

Source

http://web1.sph.emory.edu/users/lwaller/ch9index.htm

References

Waller, L. and C. Gotway (2004) Applied Spatial Statistics for Public Health Data. New York: John Wiley and Sons.

Examples

## NY leukemia
library(foreign)
nydata <- read.dbf(system.file("etc/misc/nydata.dbf", package="spdep")[1])
coordinates(nydata) <- c("X", "Y")
nyadjmat <- as.matrix(read.dbf(system.file("etc/misc/nyadjwts.dbf",
 package="spdep")[1])[-1])
ID <- as.character(names(read.dbf(system.file("etc/misc/nyadjwts.dbf",
 package="spdep")[1]))[-1])
identical(substring(ID, 2, 10), substring(as.character(nydata$AREAKEY), 2, 10))
nyadjlw <- mat2listw(nyadjmat, as.character(nydata$AREAKEY))
listw_NY <- nb2listw(nyadjlw$neighbours, style="B")

[Package spdep version 0.6-6 Index]