airdist {spdep}	R Documentation

Measure distance from plot

Description

Measure a distance between two points on a plot using locator; the function checks par("plt") and par("usr") to try to ensure that the aspect ratio y/x is 1, that is that the units of measurement in both x and y are equivalent.

Usage

airdist(ann=FALSE)

Arguments

ann
annotate the plot with line measured and distance

Value

a list with members:

dist
distance measured

coords
coordinates between which distance is measured

Author(s)

Roger Bivand Roger.Bivand@nhh.no

See Also

locator

[Package spdep version 0.6-6 Index]