Data set with 4 life condition indices of Belo Horizonte region

The data are collected inthe Atlas of condition indices published by the Joao Pinheiro Foundation and UNDP.

Format

A shape polygon object with seven variables:

id

The identificator

Name

Name of city

Population

The population of city

HLCI

Health Life Condition Index

ELCI

Education Life Condition Index

CLCI

Children Life Condition Index

ELCI

Economic Life Condition Index

Examples

if (require(rgdal, quietly=TRUE)) {
bh <- readOGR(system.file("etc/shapes/bhicv.shp",
      package="spdep")[1])
}

#> OGR data source with driver: ESRI Shapefile 
#> Source: "/Users/angela/Desktop/Spatial Data Science/spdep/inst/etc/shapes/bhicv.shp", layer: "bhicv"
#> with 98 features
#> It has 8 fields