Spatial simultaneous autoregressive error model estimation

Maximum likelihood estimation of spatial simultaneous autoregressive error models of the form:

$$y = X \beta + u, u = \lambda W u + \varepsilon$$

where \(\lambda\) is found by optimize() first, and \(\beta\) and other parameters by generalized least squares subsequently. With one of the sparse matrix methods, larger numbers of observations can be handled, but the interval= argument may need be set when the weights are not row-standardised. When etype is “emixed”, a so-called spatial Durbin error model is fitted, while lmSLX fits an lm model augmented with the spatially lagged RHS variables, including the lagged intercept when the spatial weights are not row-standardised. create_WX creates spatially lagged RHS variables, and is exposed for use in model fitting functions.

errorsarlm(formula, data=list(), listw, na.action, weights=NULL,
 Durbin, etype, method="eigen", quiet=NULL, zero.policy=NULL,
 interval = NULL, tol.solve=1.0e-10, trs=NULL, control=list())
spBreg_err(formula, data = list(), listw, na.action, Durbin, etype,
    zero.policy=NULL, control=list())
lmSLX(formula, data = list(), listw, na.action, weights=NULL,
 Durbin=TRUE, zero.policy=NULL)
create_WX(x, listw, zero.policy=NULL, prefix="")

Arguments

formula
a symbolic description of the model to be fit. The details of model specification are given for lm()

data
an optional data frame containing the variables in the model. By default the variables are taken from the environment which the function is called.

listw
a listw object created for example by nb2listw

na.action
a function (default options("na.action")), can also be na.omit or na.exclude with consequences for residuals and fitted values - in these cases the spatial weights list will be subsetted to remove NAs in the data. It may be necessary to set zero.policy to TRUE because this subsetting may create no-neighbour observations. Note that only weights lists created without using the glist argument to nb2listw may be subsetted.

weights
an optional vector of weights to be used in the fitting process. Non-NULL weights can be used to indicate that different observations have different variances (with the values in weights being inversely proportional to the variances); or equivalently, when the elements of weights are positive integers w_i, that each response y_i is the mean of w_i unit-weight observations (including the case that there are w_i observations equal to y_i and the data have been summarized) - lm

Durbin
default FALSE for errorsarlm, TRUE for lmSLX (Durbin model including WX); if TRUE, full spatial Durbin model; if a formula object, the subset of explanatory variables to lag

etype
(use the ‘Durbin=’ argument - retained for backwards compatibility only) default "error", may be set to "emixed" to include the spatially lagged independent variables added to X; when "emixed", the lagged intercept is dropped for spatial weights style "W", that is row-standardised weights, but otherwise included

method
"eigen" (default) - the Jacobian is computed as the product of (1 - rho*eigenvalue) using eigenw, and "spam" or "Matrix_J" for strictly symmetric weights lists of styles "B" and "C", or made symmetric by similarity (Ord, 1975, Appendix C) if possible for styles "W" and "S", using code from the spam package or Matrix package to calculate the determinant; “Matrix” and “spam_update” provide updating Cholesky decomposition methods; "LU" provides an alternative sparse matrix decomposition approach. In addition, there are "Chebyshev" and Monte Carlo "MC" approximate log-determinant methods; the Smirnov/Anselin (2009) trace approximation is available as "moments". Three methods: "SE_classic", "SE_whichMin", and "SE_interp" are provided experimentally, the first to attempt to emulate the behaviour of Spatial Econometrics toolbox ML fitting functions. All use grids of log determinant values, and the latter two attempt to ameliorate some features of "SE_classic".

quiet
default NULL, use !verbose global option value; if FALSE, reports function values during optimization.

zero.policy
default NULL, use global option value; if TRUE assign zero to the lagged value of zones without neighbours, if FALSE assign NA - causing errorsarlm() to terminate with an error

interval
default is NULL, search interval for autoregressive parameter

tol.solve
the tolerance for detecting linear dependencies in the columns of matrices to be inverted - passed to solve() (default=1.0e-10). This may be used if necessary to extract coefficient standard errors (for instance lowering to 1e-12), but errors in solve() may constitute indications of poorly scaled variables: if the variables have scales differing much from the autoregressive coefficient, the values in this matrix may be very different in scale, and inverting such a matrix is analytically possible by definition, but numerically unstable; rescaling the RHS variables alleviates this better than setting tol.solve to a very small value

trs
default NULL, if given, a vector of powered spatial weights matrix traces output by trW; when given, insert the asymptotic analytical values into the numerical Hessian instead of the approximated values; may be used to get around some problems raised when the numerical Hessian is poorly conditioned, generating NaNs in subsequent operations. When using the numerical Hessian to get the standard error of lambda, it is very strongly advised that trs be given, as the parts of fdHess corresponding to the regression coefficients are badly approximated, affecting the standard error of lambda; the coefficient correlation matrix is unusable

control
list of extra control arguments - see section below

x
model matrix to be lagged

prefix
default empty string, may be “lag” in some cases

Details

The asymptotic standard error of \(\lambda\) is only computed when method=eigen, because the full matrix operations involved would be costly for large n typically associated with the choice of method="spam" or "Matrix". The same applies to the coefficient covariance matrix. Taken as the asymptotic matrix from the literature, it is typically badly scaled, being block-diagonal, and with the elements involving \(\lambda\) being very small, while other parts of the matrix can be very large (often many orders of magnitude in difference). It often happens that the tol.solve argument needs to be set to a smaller value than the default, or the RHS variables can be centred or reduced in range.

Note that the fitted() function for the output object assumes that the response variable may be reconstructed as the sum of the trend, the signal, and the noise (residuals). Since the values of the response variable are known, their spatial lags are used to calculate signal components (Cressie 1993, p. 564). This differs from other software, including GeoDa, which does not use knowledge of the response variable in making predictions for the fitting data.

Control arguments

tol.opt:

the desired accuracy of the optimization - passed to optimize() (default=square root of double precision machine tolerance, a larger root may be used needed, see help(boston) for an example)

returnHcov:

default TRUE, return the Vo matrix for a spatial Hausman test

pWOrder:

default 250, if returnHcov=TRUE and the method is not “eigen”, pass this order to powerWeights as the power series maximum limit

fdHess:

default NULL, then set to (method != "eigen") internally; use fdHess to compute an approximate Hessian using finite differences when using sparse matrix methods; used to make a coefficient covariance matrix when the number of observations is large; may be turned off to save resources if need be

optimHess:

default FALSE, use fdHess from nlme, if TRUE, use optim to calculate Hessian at optimum

optimHessMethod:

default “optimHess”, may be “nlm” or one of the optim methods

LAPACK:

default FALSE; logical value passed to qr in the SSE log likelihood function

compiled_sse:

default FALSE; logical value used in the log likelihood function to choose compiled code for computing SSE

Imult:

default 2; used for preparing the Cholesky decompositions for updating in the Jacobian function

super:

if NULL (default), set to FALSE to use a simplicial decomposition for the sparse Cholesky decomposition and method “Matrix_J”, set to as.logical(NA) for method “Matrix”, if TRUE, use a supernodal decomposition

cheb_q:

default 5; highest power of the approximating polynomial for the Chebyshev approximation

MC_p:

default 16; number of random variates

MC_m:

default 30; number of products of random variates matrix and spatial weights matrix

spamPivot:

default “MMD”, alternative “RCM”

in_coef

default 0.1, coefficient value for initial Cholesky decomposition in “spam_update”

type

default “MC”, used with method “moments”; alternatives “mult” and “moments”, for use if trs is missing, trW

correct

default TRUE, used with method “moments” to compute the Smirnov/Anselin correction term

trunc

default TRUE, used with method “moments” to truncate the Smirnov/Anselin correction term

SE_method

default “LU”, may be “MC”

nrho

default 200, as in SE toolbox; the size of the first stage lndet grid; it may be reduced to for example 40

interpn

default 2000, as in SE toolbox; the size of the second stage lndet grid

small_asy

default TRUE; if the method is not “eigen”, use asymmetric covariances rather than numerical Hessian ones if n <= small

small

default 1500; threshold number of observations for asymmetric covariances when the method is not “eigen”

SElndet

default NULL, may be used to pass a pre-computed SE toolbox style matrix of coefficients and their lndet values to the "SE_classic" and "SE_whichMin" methods

LU_order

default FALSE; used in “LU_prepermutate”, note warnings given for lu method

pre_eig

default NULL; may be used to pass a pre-computed vector of eigenvalues

Extra Bayesian control arguments

ldet_method

default “SE_classic”; equivalent to the method argument in lagsarlm

interval

default c(-1, 1); used unmodified or set internally by jacobianSetup

ndraw

default 2500L; integer total number of draws

nomit

default 500L; integer total number of omitted burn-in draws

thin

default 1L; integer thinning proportion

verbose

default FALSE; inverse of quiet argument in lagsarlm

detval

default NULL; not yet in use, precomputed matrix of log determinants

prior

a list with the following components:

lambdaMH

default FALSE; if TRUE use Metropolis or FALSE griddy Gibbs

Tbeta

default NULL; values of the betas variance-covariance matrix, set to diag(k)*1e+12 if NULL

c_beta

default NULL; values of the betas set to 0 if NULL

rho

default 0.5; value of the autoregressive coefficient

sige

default 1; value of the residual variance

nu

default 0; informative Gamma(nu,d0) prior on sige

d0

default 0; informative Gamma(nu,d0) prior on sige

a1

default 1.01; parameter for beta(a1,a2) prior on rho

a2

default 1.01; parameter for beta(a1,a2) prior on rho

cc

default 0.2; initial tuning parameter for M-H sampling

gG_sige

default TRUE; include sige in lambda griddy Gibbs update

Value

A list object of class sarlm

type

"error"

lambda

simultaneous autoregressive error coefficient

coefficients

GLS coefficient estimates

rest.se

GLS coefficient standard errors (are equal to asymptotic standard errors)

LL

log likelihood value at computed optimum

s2

GLS residual variance

SSE

sum of squared GLS errors

parameters

number of parameters estimated

logLik_lm.model

Log likelihood of the linear model for \(\lambda=0\)

AIC_lm.model

AIC of the linear model for \(\lambda=0\)

% \item{lm.model}{the \code{lm} object returned when estimating for \eqn{\lambda=0}{lambda=0}}
coef_lm.model

coefficients of the linear model for \(\lambda=0\)

tarX

model matrix of the GLS model

tary

response of the GLS model

y

response of the linear model for \(\lambda=0\)

X

model matrix of the linear model for \(\lambda=0\)

method

the method used to calculate the Jacobian

call

the call used to create this object

residuals

GLS residuals

% \item{lm.target}{the \code{lm} object returned for the GLS fit}
opt

object returned from numerical optimisation

fitted.values

Difference between residuals and response variable

ase

TRUE if method=eigen

% \item{formula}{model formula}
se.fit

Not used yet

lambda.se

if ase=TRUE, the asymptotic standard error of \(\lambda\)

LMtest

NULL for this model

aliased

if not NULL, details of aliased variables

LLNullLlm

Log-likelihood of the null linear model

Hcov

Spatial DGP covariance matrix for Hausman test if available

interval

line search interval

fdHess

finite difference Hessian

optimHess

optim or fdHess used

insert

logical; is TRUE, asymptotic values inserted in fdHess where feasible

timings

processing timings

f_calls

number of calls to the log likelihood function during optimization

hf_calls

number of calls to the log likelihood function during numerical Hessian computation

intern_classic

a data frame of detval matrix row choices used by the SE toolbox classic method

zero.policy

zero.policy for this model

na.action

(possibly) named vector of excluded or omitted observations if non-default na.action argument used

weights

weights used in model fitting

emixedImps

for “emixed” models, a list of three impact matrixes (impacts and standard errors) for direct, indirect and total impacts; total impacts calculated using gmodels::estimable

The internal sar.error.* functions return the value of the log likelihood function at \lambdalambda. The lmSLX function returns an lm object with a mixedImps list of three impact matrixes (impacts and standard errors) for direct, indirect and total impacts; total impacts calculated using gmodels::estimable.
References

Cliff, A. D., Ord, J. K. 1981 Spatial processes, Pion; Ord, J. K. 1975 Estimation methods for models of spatial interaction, Journal of the American Statistical Association, 70, 120-126; Anselin, L. 1988 Spatial econometrics: methods and models. (Dordrecht: Kluwer); Anselin, L. 1995 SpaceStat, a software program for the analysis of spatial data, version 1.80. Regional Research Institute, West Virginia University, Morgantown, WV; Anselin L, Bera AK (1998) Spatial dependence in linear regression models with an introduction to spatial econometrics. In: Ullah A, Giles DEA (eds) Handbook of applied economic statistics. Marcel Dekker, New York, pp. 237-289; Cressie, N. A. C. 1993 Statistics for spatial data, Wiley, New York; LeSage J and RK Pace (2009) Introduction to Spatial Econometrics. CRC Press, Boca Raton.

Roger Bivand, Gianfranco Piras (2015). Comparing Implementations of Estimation Methods for Spatial Econometrics. Journal of Statistical Software, 63(18), 1-36. https://www.jstatsoft.org/v63/i18/.

Bivand, R. S., Hauke, J., and Kossowski, T. (2013). Computing the Jacobian in Gaussian spatial autoregressive models: An illustrated comparison of available methods. Geographical Analysis, 45(2), 150-179.

See also

Examples

data(oldcol)
lw <- nb2listw(COL.nb, style="W")
ev <- eigenw(similar.listw(lw))
COL.errW.eig <- errorsarlm(CRIME ~ INC + HOVAL, data=COL.OLD,
 lw, quiet=FALSE, control=list(pre_eig=ev))

#> 
#> Spatial autoregressive error model
#> 
#> Jacobian calculated using neighbourhood matrix eigenvalues
#> 
#> lambda: -0.5674437  function: -195.8051  Jacobian: -1.636549  SSE: 7936.201 
#> lambda: 0.03126655  function: -187.0219  Jacobian: -0.005318373  SSE: 5927.009 
#> lambda: 0.4012898  function: -183.8422  Jacobian: -0.9953987  SSE: 4999.419 
#> lambda: 0.6299767  function: -183.4895  Jacobian: -2.818134  SSE: 4574.641 
#> lambda: 0.5811116  function: -183.3887  Jacobian: -2.314073  SSE: 4650.566 
#> lambda: 0.554104  function: -183.3817  Jacobian: -2.066354  SSE: 4696.49 
#> lambda: 0.5621834  function: -183.3805  Jacobian: -2.138326  SSE: 4682.474 
#> lambda: 0.5617028  function: -183.3805  Jacobian: -2.133995  SSE: 4683.301 
#> lambda: 0.5617888  function: -183.3805  Jacobian: -2.134769  SSE: 4683.153 
#> lambda: 0.5617902  function: -183.3805  Jacobian: -2.134782  SSE: 4683.151 
#> lambda: 0.5617903  function: -183.3805  Jacobian: -2.134782  SSE: 4683.151 
#> lambda: 0.5617903  function: -183.3805  Jacobian: -2.134783  SSE: 4683.15 
#> lambda: 0.5617903  function: -183.3805  Jacobian: -2.134783  SSE: 4683.151 
#> lambda: 0.5617903  function: -183.3805  Jacobian: -2.134782  SSE: 4683.151 

summary(COL.errW.eig)

#> 
#> Call:errorsarlm(formula = CRIME ~ INC + HOVAL, data = COL.OLD, listw = lw, 
#>     quiet = FALSE, control = list(pre_eig = ev))
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -34.81174  -6.44031  -0.72142   7.61476  23.33626 
#> 
#> Type: error 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) 59.893219   5.366163 11.1613 < 2.2e-16
#> INC         -0.941312   0.330569 -2.8476 0.0044057
#> HOVAL       -0.302250   0.090476 -3.3407 0.0008358
#> 
#> Lambda: 0.56179, LR test value: 7.9935, p-value: 0.0046945
#> Asymptotic standard error: 0.13387
#>     z-value: 4.1966, p-value: 2.7098e-05
#> Wald statistic: 17.611, p-value: 2.7098e-05
#> 
#> Log likelihood: -183.3805 for error model
#> ML residual variance (sigma squared): 95.575, (sigma: 9.7762)
#> Number of observations: 49 
#> Number of parameters estimated: 5 
#> AIC: 376.76, (AIC for lm: 382.75)
#> 

COL.errW.eig_ev <- errorsarlm(CRIME ~ INC + HOVAL, data=COL.OLD,
 lw, control=list(pre_eig=ev))
all.equal(coefficients(COL.errW.eig), coefficients(COL.errW.eig_ev))

#> [1] TRUE

COL.errB.eig <- errorsarlm(CRIME ~ INC + HOVAL, data=COL.OLD,
 nb2listw(COL.nb, style="B"))
summary(COL.errB.eig)

#> 
#> Call:
#> errorsarlm(formula = CRIME ~ INC + HOVAL, data = COL.OLD, listw = nb2listw(COL.nb, 
#>     style = "B"))
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -32.19010  -5.22646  -0.69952   7.92588  24.23511 
#> 
#> Type: error 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) 55.383119   5.449775 10.1625 < 2.2e-16
#> INC         -0.936595   0.319355 -2.9328 0.0033596
#> HOVAL       -0.299857   0.088678 -3.3814 0.0007212
#> 
#> Lambda: 0.12686, LR test value: 10.654, p-value: 0.0010983
#> Asymptotic standard error: 0.021745
#>     z-value: 5.8342, p-value: 5.4044e-09
#> Wald statistic: 34.038, p-value: 5.4044e-09
#> 
#> Log likelihood: -182.0502 for error model
#> ML residual variance (sigma squared): 88.744, (sigma: 9.4204)
#> Number of observations: 49 
#> Number of parameters estimated: 5 
#> AIC: 374.1, (AIC for lm: 382.75)
#> 

W <- as(nb2listw(COL.nb), "CsparseMatrix")
trMatc <- trW(W, type="mult")
COL.errW.M <- errorsarlm(CRIME ~ INC + HOVAL, data=COL.OLD,
 lw, method="Matrix", quiet=FALSE, trs=trMatc)

#> 
#> Spatial autoregressive error model
#> 
#> Jacobian calculated using sparse matrix Cholesky decomposition
#> lambda: -0.2364499  function: -190.4287  Jacobian: -0.2899343  SSE: 6732.556 
#> lambda: 0.2354499  function: -185.0024  Jacobian: -0.3201148  SSE: 5388.346 
#> lambda: 0.5271001  function: -183.4053  Jacobian: -1.838306  SSE: 4744.975 
#> lambda: 0.728364  function: -184.1244  Jacobian: -4.106761  SSE: 4454.203 
#> lambda: 0.5304163  function: -183.4009  Jacobian: -1.865308  SSE: 4738.889 
#> lambda: 0.5557478  function: -183.3812  Jacobian: -2.080853  SSE: 4693.62 
#> lambda: 0.6216813  function: -183.4637  Jacobian: -2.727084  SSE: 4586.84 
#> lambda: 0.5627129  function: -183.3805  Jacobian: -2.143105  SSE: 4681.563 
#> lambda: 0.5618852  function: -183.3805  Jacobian: -2.135638  SSE: 4682.987 
#> lambda: 0.5617866  function: -183.3805  Jacobian: -2.134749  SSE: 4683.157 
#> lambda: 0.5617903  function: -183.3805  Jacobian: -2.134783  SSE: 4683.15 
#> lambda: 0.5617903  function: -183.3805  Jacobian: -2.134782  SSE: 4683.151 
#> lambda: 0.5617902  function: -183.3805  Jacobian: -2.134782  SSE: 4683.151 
#> lambda: 0.5617903  function: -183.3805  Jacobian: -2.134782  SSE: 4683.151 
#> lambda: 0.5617903  function: -183.3805  Jacobian: -2.134783  SSE: 4683.151 
#> lambda: 0.5617903  function: -183.3805  Jacobian: -2.134782  SSE: 4683.151 

summary(COL.errW.M)

#> 
#> Call:errorsarlm(formula = CRIME ~ INC + HOVAL, data = COL.OLD, listw = lw, 
#>     method = "Matrix", quiet = FALSE, trs = trMatc)
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -34.81174  -6.44031  -0.72142   7.61476  23.33626 
#> 
#> Type: error 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) 59.893219   5.366163 11.1613 < 2.2e-16
#> INC         -0.941312   0.330569 -2.8476 0.0044057
#> HOVAL       -0.302250   0.090476 -3.3407 0.0008358
#> 
#> Lambda: 0.56179, LR test value: 7.9935, p-value: 0.0046945
#> Asymptotic standard error: 0.13387
#>     z-value: 4.1966, p-value: 2.7098e-05
#> Wald statistic: 17.611, p-value: 2.7098e-05
#> 
#> Log likelihood: -183.3805 for error model
#> ML residual variance (sigma squared): 95.575, (sigma: 9.7762)
#> Number of observations: 49 
#> Number of parameters estimated: 5 
#> AIC: 376.76, (AIC for lm: 382.75)
#> 

COL.SDEM.eig <- errorsarlm(CRIME ~ INC + HOVAL, data=COL.OLD,
 lw, etype="emixed", control=list(pre_eig=ev))
summary(COL.SDEM.eig)

#> 
#> Call:errorsarlm(formula = CRIME ~ INC + HOVAL, data = COL.OLD, listw = lw, 
#>     etype = "emixed", control = list(pre_eig = ev))
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -37.31635  -6.54376  -0.22212   6.44591  23.15801 
#> 
#> Type: error 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) 73.545133   8.783543  8.3731 < 2.2e-16
#> INC         -1.051673   0.319514 -3.2915 0.0009966
#> HOVAL       -0.275608   0.091151 -3.0236 0.0024976
#> lag.INC     -1.156711   0.578629 -1.9991 0.0456024
#> lag.HOVAL    0.111691   0.198993  0.5613 0.5746048
#> 
#> Lambda: 0.4254, LR test value: 4.9871, p-value: 0.025537
#> Asymptotic standard error: 0.15842
#>     z-value: 2.6852, p-value: 0.0072485
#> Wald statistic: 7.2103, p-value: 0.0072485
#> 
#> Log likelihood: -181.5846 for error model
#> ML residual variance (sigma squared): 92.531, (sigma: 9.6193)
#> Number of observations: 49 
#> Number of parameters estimated: 7 
#> AIC: 377.17, (AIC for lm: 380.16)
#> 

COL.SDEM.eig <- errorsarlm(CRIME ~ INC + HOVAL, data=COL.OLD,
 lw, Durbin=TRUE, control=list(pre_eig=ev))
summary(COL.SDEM.eig)

#> 
#> Call:errorsarlm(formula = CRIME ~ INC + HOVAL, data = COL.OLD, listw = lw, 
#>     Durbin = TRUE, control = list(pre_eig = ev))
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -37.31635  -6.54376  -0.22212   6.44591  23.15801 
#> 
#> Type: error 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) 73.545133   8.783543  8.3731 < 2.2e-16
#> INC         -1.051673   0.319514 -3.2915 0.0009966
#> HOVAL       -0.275608   0.091151 -3.0236 0.0024976
#> lag.INC     -1.156711   0.578629 -1.9991 0.0456024
#> lag.HOVAL    0.111691   0.198993  0.5613 0.5746048
#> 
#> Lambda: 0.4254, LR test value: 4.9871, p-value: 0.025537
#> Asymptotic standard error: 0.15842
#>     z-value: 2.6852, p-value: 0.0072485
#> Wald statistic: 7.2103, p-value: 0.0072485
#> 
#> Log likelihood: -181.5846 for error model
#> ML residual variance (sigma squared): 92.531, (sigma: 9.6193)
#> Number of observations: 49 
#> Number of parameters estimated: 7 
#> AIC: 377.17, (AIC for lm: 380.16)
#> 

COL.SDEM.eig <- errorsarlm(CRIME ~ DISCBD + INC + HOVAL, data=COL.OLD,
 lw, Durbin=~INC, control=list(pre_eig=ev))
summary(COL.SDEM.eig)

#> 
#> Call:errorsarlm(formula = CRIME ~ DISCBD + INC + HOVAL, data = COL.OLD, 
#>     listw = lw, Durbin = ~INC, control = list(pre_eig = ev))
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -34.61867  -7.31993   0.82879   5.92877  17.82211 
#> 
#> Type: error 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) 68.961912   6.985784  9.8717 < 2.2e-16
#> DISCBD      -5.412936   2.009281 -2.6940  0.007061
#> INC         -0.899425   0.315174 -2.8537  0.004321
#> HOVAL       -0.202846   0.090125 -2.2507  0.024403
#> lag.INC      0.161858   0.603859  0.2680  0.788669
#> 
#> Lambda: 0.24524, LR test value: 1.2719, p-value: 0.25942
#> Asymptotic standard error: 0.18393
#>     z-value: 1.3334, p-value: 0.18241
#> Wald statistic: 1.7778, p-value: 0.18241
#> 
#> Log likelihood: -178.9895 for error model
#> ML residual variance (sigma squared): 85.935, (sigma: 9.2701)
#> Number of observations: 49 
#> Number of parameters estimated: 7 
#> AIC: 371.98, (AIC for lm: 371.25)
#> 

summary(impacts(COL.SDEM.eig))

#> Impact measures (SDEM, estimable, n):
#>            Direct  Indirect      Total
#> DISCBD -5.4129365        NA -5.4129365
#> INC    -0.8994251 0.1618581 -0.7375670
#> HOVAL  -0.2028457        NA -0.2028457
#> ========================================================
#> Standard errors:
#>            Direct  Indirect      Total
#> DISCBD 2.00928139        NA 2.00928139
#> INC    0.31517363 0.6038588 0.67559401
#> HOVAL  0.09012493        NA 0.09012493
#> ========================================================
#> Z-values:
#>           Direct  Indirect     Total
#> DISCBD -2.693966        NA -2.693966
#> INC    -2.853745 0.2680396 -1.091731
#> HOVAL  -2.250717        NA -2.250717
#> 
#> p-values:
#>        Direct    Indirect Total    
#> DISCBD 0.0070607 NA       0.0070607
#> INC    0.0043207 0.78867  0.2749513
#> HOVAL  0.0244035 NA       0.0244035
#> 

COL.SLX <- lmSLX(CRIME ~ INC + HOVAL, data=COL.OLD, listw=lw)
summary(COL.SLX)

#> 
#> Call:
#> lm(formula = formula(paste("y ~ ", paste(colnames(x)[-1], collapse = "+"))), 
#>     data = as.data.frame(x), weights = weights)
#> 
#> Residuals:
#>     Min      1Q  Median      3Q     Max 
#> -36.536  -7.835   0.474   8.349  25.594 
#> 
#> Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)  75.0287     6.6260  11.323 1.26e-14 ***
#> INC          -1.1089     0.3738  -2.967  0.00485 ** 
#> HOVAL        -0.2897     0.1014  -2.858  0.00649 ** 
#> lag.INC      -1.3710     0.5613  -2.443  0.01867 *  
#> lag.HOVAL     0.1918     0.2003   0.957  0.34369    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error: 10.93 on 44 degrees of freedom
#> Multiple R-squared:  0.6088,	Adjusted R-squared:  0.5732 
#> F-statistic: 17.12 on 4 and 44 DF,  p-value: 1.553e-08
#> 

summary(impacts(COL.SLX))

#> Impact measures (SLX, estimable, n-k):
#>           Direct   Indirect       Total
#> INC   -1.1089293 -1.3709725 -2.47990173
#> HOVAL -0.2897283  0.1917608 -0.09796753
#> ========================================================
#> Standard errors:
#>          Direct  Indirect     Total
#> INC   0.3738129 0.5612771 0.4965456
#> HOVAL 0.1013673 0.2003335 0.2028016
#> ========================================================
#> Z-values:
#>          Direct   Indirect      Total
#> INC   -2.966535 -2.4425945 -4.9943086
#> HOVAL -2.858202  0.9572079 -0.4830709
#> 
#> p-values:
#>       Direct    Indirect Total     
#> INC   0.0030118 0.014582 5.9047e-07
#> HOVAL 0.0042605 0.338462 0.62905   
#> 

COL.SLX <- lmSLX(CRIME ~ INC + HOVAL + I(HOVAL^2), data=COL.OLD, listw=lw, Durbin=TRUE)
summary(impacts(COL.SLX))

#> Impact measures (SLX, estimable, n-k):
#>                  Direct     Indirect       Total
#> INC        -0.947594274 -1.275338647 -2.22293292
#> HOVAL      -0.777427839 -0.355048446 -1.13247628
#> I(HOVAL^2)  0.004639919  0.005608104  0.01024802
#> ========================================================
#> Standard errors:
#>                 Direct   Indirect      Total
#> INC        0.398832844 0.59687399 0.56325440
#> HOVAL      0.464456540 0.98213200 1.07348981
#> I(HOVAL^2) 0.004226259 0.00908385 0.01025212
#> ========================================================
#> Z-values:
#>               Direct   Indirect      Total
#> INC        -2.375918 -2.1366966 -3.9465877
#> HOVAL      -1.673844 -0.3615079 -1.0549483
#> I(HOVAL^2)  1.097879  0.6173709  0.9996008
#> 
#> p-values:
#>            Direct   Indirect Total     
#> INC        0.017505 0.032623 7.9273e-05
#> HOVAL      0.094161 0.717720 0.29145   
#> I(HOVAL^2) 0.272258 0.536990 0.31750   
#> 

summary(COL.SLX)

#> 
#> Call:
#> lm(formula = formula(paste("y ~ ", paste(colnames(x)[-1], collapse = "+"))), 
#>     data = as.data.frame(x), weights = weights)
#> 
#> Residuals:
#>     Min      1Q  Median      3Q     Max 
#> -33.974  -7.764   0.907   6.820  24.395 
#> 
#> Coefficients:
#>                 Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)    92.459835  19.278187   4.796 2.06e-05 ***
#> INC            -0.947594   0.398833  -2.376   0.0221 *  
#> HOVAL          -0.777428   0.464457  -1.674   0.1016    
#> I.HOVAL.2.      0.004640   0.004226   1.098   0.2785    
#> lag.INC        -1.275339   0.596874  -2.137   0.0385 *  
#> lag.HOVAL      -0.355048   0.982132  -0.362   0.7195    
#> lag.I.HOVAL.2.  0.005608   0.009084   0.617   0.5403    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error: 10.99 on 42 degrees of freedom
#> Multiple R-squared:  0.6224,	Adjusted R-squared:  0.5684 
#> F-statistic: 11.54 on 6 and 42 DF,  p-value: 1.361e-07
#> 

COL.SLX <- lmSLX(CRIME ~ INC + HOVAL + I(HOVAL^2), data=COL.OLD, listw=lw, Durbin=~INC)
summary(impacts(COL.SLX))

#> Impact measures (SLX, estimable, n-k):
#>                  Direct  Indirect        Total
#> INC        -1.079064628 -1.010896 -2.089960575
#> HOVAL      -0.634518755        NA -0.634518755
#> I(HOVAL^2)  0.003455273        NA  0.003455273
#> ========================================================
#> Standard errors:
#>                Direct  Indirect      Total
#> INC        0.38471071 0.4552667 0.44650193
#> HOVAL      0.44760078        NA 0.44760078
#> I(HOVAL^2) 0.00411036        NA 0.00411036
#> ========================================================
#> Z-values:
#>                Direct  Indirect      Total
#> INC        -2.8048728 -2.220448 -4.6807425
#> HOVAL      -1.4175997        NA -1.4175997
#> I(HOVAL^2)  0.8406254        NA  0.8406254
#> 
#> p-values:
#>            Direct    Indirect Total     
#> INC        0.0050336 0.026388 2.8584e-06
#> HOVAL      0.1563077 NA       0.15631   
#> I(HOVAL^2) 0.4005579 NA       0.40056   
#> 

summary(COL.SLX)

#> 
#> Call:
#> lm(formula = formula(paste("y ~ ", paste(colnames(x)[-1], collapse = "+"))), 
#>     data = as.data.frame(x), weights = weights)
#> 
#> Residuals:
#>     Min      1Q  Median      3Q     Max 
#> -35.908  -7.085   0.770   7.147  24.684 
#> 
#> Coefficients:
#>              Estimate Std. Error t value Pr(>|t|)    
#> (Intercept) 83.679968   9.264846   9.032  1.4e-11 ***
#> INC         -1.079065   0.384711  -2.805  0.00747 ** 
#> HOVAL       -0.634519   0.447601  -1.418  0.16335    
#> I.HOVAL.2.   0.003455   0.004110   0.841  0.40510    
#> lag.INC     -1.010896   0.455267  -2.220  0.03159 *  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error: 10.96 on 44 degrees of freedom
#> Multiple R-squared:  0.607,	Adjusted R-squared:  0.5712 
#> F-statistic: 16.99 on 4 and 44 DF,  p-value: 1.716e-08
#> 

COL.SLX <- lmSLX(CRIME ~ INC, data=COL.OLD, listw=lw)
summary(COL.SLX)

#> 
#> Call:
#> lm(formula = formula(paste("y ~ ", paste(colnames(x)[-1], collapse = "+"))), 
#>     data = as.data.frame(x), weights = weights)
#> 
#> Residuals:
#>     Min      1Q  Median      3Q     Max 
#> -49.297  -6.658   0.669   7.315  29.826 
#> 
#> Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)  73.9849     6.2081  11.918 1.16e-15 ***
#> INC          -1.5889     0.3564  -4.458 5.28e-05 ***
#> lag.INC      -1.0859     0.4812  -2.257   0.0288 *  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error: 11.65 on 46 degrees of freedom
#> Multiple R-squared:  0.5353,	Adjusted R-squared:  0.5151 
#> F-statistic: 26.49 on 2 and 46 DF,  p-value: 2.214e-08
#> 

summary(impacts(COL.SLX))

#> Impact measures (SLX, estimable, n-k):
#>        Direct  Indirect     Total
#> INC -1.588901 -1.085867 -2.674768
#> ========================================================
#> Standard errors:
#>        Direct  Indirect    Total
#> INC 0.3564039 0.4811809 0.407313
#> ========================================================
#> Z-values:
#>        Direct  Indirect     Total
#> INC -4.458147 -2.256671 -6.566861
#> 
#> p-values:
#>     Direct     Indirect Total     
#> INC 8.2671e-06 0.024029 5.1387e-11
#> 

crds <- cbind(COL.OLD$X, COL.OLD$Y)
mdist <- sqrt(sum(diff(apply(crds, 2, range))^2))
dnb <- dnearneigh(crds, 0, mdist)
dists <- nbdists(dnb, crds)
f <- function(x, form, data, dnb, dists, verbose) {
  glst <- lapply(dists, function(d) 1/(d^x))
  lw <- nb2listw(dnb, glist=glst, style="B")
  res <- logLik(lmSLX(form=form, data=data, listw=lw))
  if (verbose) cat("power:", x, "logLik:", res, "\n")
  res
}
opt <- optimize(f, interval=c(0.1, 4), form=CRIME ~ INC + HOVAL,
 data=COL.OLD, dnb=dnb, dists=dists, verbose=TRUE, maximum=TRUE)

#> power: 1.589667 logLik: -172.6864 
#> power: 2.510333 logLik: -177.741 
#> power: 1.020665 logLik: -171.5379 
#> power: 0.8721475 logLik: -171.7979 
#> power: 1.11302 logLik: -171.4973 
#> power: 1.107329 logLik: -171.497 
#> power: 1.105705 logLik: -171.497 
#> power: 1.105746 logLik: -171.497 
#> power: 1.105664 logLik: -171.497 
#> power: 1.105705 logLik: -171.497 

glst <- lapply(dists, function(d) 1/(d^opt$maximum))
lw <- nb2listw(dnb, glist=glst, style="B")
SLX <- lmSLX(CRIME ~ INC + HOVAL, data=COL.OLD, listw=lw)
summary(SLX)

#> 
#> Call:
#> lm(formula = formula(paste("y ~ ", paste(colnames(x)[-1], collapse = "+"))), 
#>     data = as.data.frame(x), weights = weights)
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -25.1535  -4.7529  -0.9819   4.8053  22.2274 
#> 
#> Coefficients:
#>                 Estimate Std. Error t value Pr(>|t|)  
#> (Intercept)     18.78498   11.23795   1.672   0.1019  
#> INC             -0.65428    0.29463  -2.221   0.0317 *
#> HOVAL           -0.18244    0.08138  -2.242   0.0302 *
#> lag..Intercept.  6.27790    4.44284   1.413   0.1648  
#> lag.INC         -0.12958    0.28794  -0.450   0.6550  
#> lag.HOVAL        0.02668    0.11273   0.237   0.8141  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error: 8.553 on 43 degrees of freedom
#> Multiple R-squared:  0.7659,	Adjusted R-squared:  0.7387 
#> F-statistic: 28.14 on 5 and 43 DF,  p-value: 1.544e-12
#> 

summary(impacts(SLX))

#> Impact measures (SLX, estimable, n-k):
#>           Direct   Indirect      Total
#> INC   -0.6542760  6.2778972  5.6236212
#> HOVAL -0.1824383 -0.1295774 -0.3120157
#> ========================================================
#> Standard errors:
#>           Direct  Indirect     Total
#> INC   0.29463342 4.4428388 4.5116967
#> HOVAL 0.08138257 0.2879383 0.3178755
#> ========================================================
#> Z-values:
#>          Direct   Indirect      Total
#> INC   -2.220644  1.4130374  1.2464537
#> HOVAL -2.241736 -0.4500179 -0.9815656
#> 
#> p-values:
#>       Direct   Indirect Total  
#> INC   0.026375 0.15764  0.21260
#> HOVAL 0.024978 0.65270  0.32631
#> 

NA.COL.OLD <- COL.OLD
NA.COL.OLD$CRIME[20:25] <- NA
COL.err.NA <- errorsarlm(CRIME ~ INC + HOVAL, data=NA.COL.OLD,
 nb2listw(COL.nb), na.action=na.exclude)
COL.err.NA$na.action

#> 1020 1021 1022 1023 1024 1025 
#>   20   21   22   23   24   25 
#> attr(,"class")
#> [1] "exclude"

COL.err.NA

#> 
#> Call:
#> errorsarlm(formula = CRIME ~ INC + HOVAL, data = NA.COL.OLD, 
#>     listw = nb2listw(COL.nb), na.action = na.exclude)
#> Type: error 
#> 
#> Coefficients:
#>      lambda (Intercept)         INC       HOVAL 
#>   0.5748430  58.2460528  -0.8473028  -0.3024909 
#> 
#> Log likelihood: -161.8763 

resid(COL.err.NA)

#>         1001         1002         1003         1004         1005         1006 
#>  -4.18270830 -11.44133842   0.31874928 -34.47163074   2.42244758  -4.32095072 
#>         1007         1008         1009         1010         1011         1012 
#>   8.66744165 -13.38669935  -1.92276585  17.85753951  -1.11484596  -2.30434792 
#>         1013         1014         1015         1016         1017         1018 
#>  -8.16935116  -5.80500231   0.14973721   5.93191444  -7.03028271   2.39112829 
#>         1019         1020         1021         1022         1023         1024 
#>  -8.95099917           NA           NA           NA           NA           NA 
#>         1025         1026         1027         1028         1029         1030 
#>           NA  -2.52940711  -9.60025384  -6.95635586  -0.43630579   5.98493664 
#>         1031         1032         1033         1034         1035         1036 
#>   6.25882669   7.75527031  10.83413235  23.23927250  -0.05594958   1.43808573 
#>         1037         1038         1039         1040         1041         1042 
#>   9.51995265  12.18295373   8.31031724  17.06834503   7.04418392   7.50088924 
#>         1043         1044         1045         1046         1047         1048 
#>  -7.78485976  -6.79207446  -7.94977534 -11.25362117  -5.68994630   5.04837399 
#>         1049 
#>   2.22497381 

lw <- nb2listw(COL.nb, style="W")
print(system.time(ev <- eigenw(similar.listw(lw))))

#>    user  system elapsed 
#>   0.001   0.000   0.001 

print(system.time(COL.errW.eig <- errorsarlm(CRIME ~ INC + HOVAL, data=COL.OLD,
 lw, method="eigen", control=list(pre_eig=ev))))

#>    user  system elapsed 
#>   0.166   0.000   0.167 

ocoef <- coefficients(COL.errW.eig)
print(system.time(COL.errW.eig <- errorsarlm(CRIME ~ INC + HOVAL, data=COL.OLD,
 lw, method="eigen", control=list(pre_eig=ev, LAPACK=FALSE))))

#>    user  system elapsed 
#>   0.181   0.000   0.183 

print(all.equal(ocoef, coefficients(COL.errW.eig)))

#> [1] TRUE

print(system.time(COL.errW.eig <- errorsarlm(CRIME ~ INC + HOVAL, data=COL.OLD,
 lw, method="eigen", control=list(pre_eig=ev, compiled_sse=TRUE))))

#>    user  system elapsed 
#>   0.175   0.000   0.180 

print(all.equal(ocoef, coefficients(COL.errW.eig)))

#> [1] TRUE

print(system.time(COL.errW.eig <- errorsarlm(CRIME ~ INC + HOVAL, data=COL.OLD,
 lw, method="Matrix_J", control=list(super=TRUE))))

#>    user  system elapsed 
#>   0.204   0.000   0.205 

print(all.equal(ocoef, coefficients(COL.errW.eig)))

#> [1] TRUE

print(system.time(COL.errW.eig <- errorsarlm(CRIME ~ INC + HOVAL, data=COL.OLD,
 lw, method="Matrix_J", control=list(super=FALSE))))

#>    user  system elapsed 
#>   0.208   0.000   0.209 

print(all.equal(ocoef, coefficients(COL.errW.eig)))

#> [1] TRUE

print(system.time(COL.errW.eig <- errorsarlm(CRIME ~ INC + HOVAL, data=COL.OLD,
 lw, method="Matrix_J", control=list(super=as.logical(NA)))))

#>    user  system elapsed 
#>   0.203   0.000   0.212 

print(all.equal(ocoef, coefficients(COL.errW.eig)))

#> [1] TRUE

print(system.time(COL.errW.eig <- errorsarlm(CRIME ~ INC + HOVAL, data=COL.OLD,
 lw, method="Matrix", control=list(super=TRUE))))

#>    user  system elapsed 
#>   0.180   0.000   0.186 

print(all.equal(ocoef, coefficients(COL.errW.eig)))

#> [1] TRUE

print(system.time(COL.errW.eig <- errorsarlm(CRIME ~ INC + HOVAL, data=COL.OLD,
 lw, method="Matrix", control=list(super=FALSE))))

#>    user  system elapsed 
#>   0.192   0.002   0.219 

print(all.equal(ocoef, coefficients(COL.errW.eig)))

#> [1] TRUE

print(system.time(COL.errW.eig <- errorsarlm(CRIME ~ INC + HOVAL, data=COL.OLD,
 lw, method="Matrix", control=list(super=as.logical(NA)))))

#>    user  system elapsed 
#>   0.180   0.000   0.183 

print(all.equal(ocoef, coefficients(COL.errW.eig)))

#> [1] TRUE

print(system.time(COL.errW.eig <- errorsarlm(CRIME ~ INC + HOVAL, data=COL.OLD,
 lw, method="spam", control=list(spamPivot="MMD"))))

#>    user  system elapsed 
#>   0.213   0.001   0.218 

print(all.equal(ocoef, coefficients(COL.errW.eig)))

#> [1] TRUE

print(system.time(COL.errW.eig <- errorsarlm(CRIME ~ INC + HOVAL, data=COL.OLD,
 lw, method="spam", control=list(spamPivot="RCM"))))

#>    user  system elapsed 
#>   0.185   0.000   0.188 

print(all.equal(ocoef, coefficients(COL.errW.eig)))

#> [1] TRUE

print(system.time(COL.errW.eig <- errorsarlm(CRIME ~ INC + HOVAL, data=COL.OLD,
 lw, method="spam_update", control=list(spamPivot="MMD"))))

#>    user  system elapsed 
#>   0.177   0.000   0.179 

print(all.equal(ocoef, coefficients(COL.errW.eig)))

#> [1] TRUE

print(system.time(COL.errW.eig <- errorsarlm(CRIME ~ INC + HOVAL, data=COL.OLD,
 lw, method="spam_update", control=list(spamPivot="RCM"))))

#>    user  system elapsed 
#>   0.196   0.000   0.217 

print(all.equal(ocoef, coefficients(COL.errW.eig)))

#> [1] TRUE

if (require(coda, quietly=TRUE)) {
set.seed(1)
COL.err.Bayes <- spBreg_err(CRIME ~ INC + HOVAL, data=COL.OLD, listw=lw)
print(summary(COL.err.Bayes))
print(raftery.diag(COL.err.Bayes, r=0.01))
set.seed(1)
COL.err.Bayes <- spBreg_err(CRIME ~ INC + HOVAL, data=COL.OLD, listw=lw,
 control=list(prior=list(lambdaMH=TRUE)))
print(summary(COL.err.Bayes))
print(raftery.diag(COL.err.Bayes, r=0.01))
set.seed(1)
COL.err.Bayes <- spBreg_err(CRIME ~ INC + HOVAL, data=COL.OLD, listw=lw,
 Durbin=TRUE)
print(summary(COL.err.Bayes))
print(summary(impacts(COL.err.Bayes)))
print(raftery.diag(COL.err.Bayes, r=0.01))
set.seed(1)
COL.err.Bayes <- spBreg_err(CRIME ~ INC + HOVAL, data=COL.OLD, listw=lw,
 Durbin=TRUE, control=list(prior=list(lambdaMH=TRUE)))
print(summary(COL.err.Bayes))
print(summary(impacts(COL.err.Bayes)))
print(raftery.diag(COL.err.Bayes, r=0.01))
set.seed(1)
COL.err.Bayes <- spBreg_err(CRIME ~ INC + HOVAL, data=COL.OLD, listw=lw,
 Durbin=~INC)
print(summary(COL.err.Bayes))
print(summary(impacts(COL.err.Bayes)))
print(raftery.diag(COL.err.Bayes, r=0.01))
set.seed(1)
COL.err.Bayes <- spBreg_err(CRIME ~ INC + HOVAL, data=COL.OLD, listw=lw,
 Durbin=~INC, control=list(prior=list(lambdaMH=TRUE)))
print(summary(COL.err.Bayes))
print(summary(impacts(COL.err.Bayes)))
print(raftery.diag(COL.err.Bayes, r=0.01))
}

#> 
#> Iterations = 501:2500
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean       SD Naive SE Time-series SE
#> (Intercept)  59.8221  7.41100 0.165715       0.188495
#> INC          -0.9452  0.40434 0.009041       0.010116
#> HOVAL        -0.3019  0.09875 0.002208       0.002249
#> lambda        0.5690  0.16030 0.003585       0.003893
#> sige        116.5436 26.82453 0.599815       0.649112
#> 
#> 2. Quantiles for each variable:
#> 
#>                2.5%     25%      50%      75%    97.5%
#> (Intercept) 45.2995 55.5884  59.9721  64.5786  72.6933
#> INC         -1.7796 -1.1999  -0.9416  -0.6777  -0.1594
#> HOVAL       -0.4881 -0.3695  -0.3009  -0.2357  -0.1135
#> lambda       0.2156  0.4707   0.5828   0.6798   0.8419
#> sige        74.4166 97.6479 113.0769 131.2469 178.4355
#> 
#> 
#> Quantile (q) = 0.025
#> Accuracy (r) = +/- 0.01
#> Probability (s) = 0.95 
#>                                                    
#>              Burn-in  Total Lower bound  Dependence
#>              (M)      (N)   (Nmin)       factor (I)
#>  (Intercept) 3        1052  937          1.120     
#>  INC         3        1010  937          1.080     
#>  HOVAL       2        930   937          0.993     
#>  lambda      2        946   937          1.010     
#>  sige        2        930   937          0.993     
#> 
#> 
#> Iterations = 501:2500
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD Naive SE Time-series SE
#> (Intercept)  58.7696  8.9422 0.199953       0.249755
#> INC          -0.8978  0.3975 0.008887       0.011839
#> HOVAL        -0.3022  0.1012 0.002263       0.002424
#> lambda        0.6154  0.1553 0.003473       0.008082
#> sige        117.6540 27.4230 0.613198       0.703192
#> 
#> 2. Quantiles for each variable:
#> 
#>                2.5%     25%      50%      75%    97.5%
#> (Intercept) 42.3897 54.4585  59.4154  63.7505  72.8358
#> INC         -1.6953 -1.1702  -0.8836  -0.6288  -0.1264
#> HOVAL       -0.4957 -0.3710  -0.3026  -0.2332  -0.1022
#> lambda       0.3128  0.5079   0.6183   0.7280   0.8923
#> sige        77.3438 97.7834 113.2635 133.1907 180.1621
#> 
#> 
#> Quantile (q) = 0.025
#> Accuracy (r) = +/- 0.01
#> Probability (s) = 0.95 
#>                                                    
#>              Burn-in  Total Lower bound  Dependence
#>              (M)      (N)   (Nmin)       factor (I)
#>  (Intercept) 4        1242  937          1.330     
#>  INC         2        930   937          0.993     
#>  HOVAL       2        969   937          1.030     
#>  lambda      10       2727  937          2.910     
#>  sige        2        892   937          0.952     
#> 
#> 
#> Iterations = 501:2500
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                  Mean      SD Naive SE Time-series SE
#> (Intercept)  72.68153 11.6048 0.259491       0.259491
#> INC          -1.02939  0.3745 0.008375       0.007390
#> HOVAL        -0.28159  0.1050 0.002348       0.002101
#> lag.INC      -1.04431  0.7217 0.016137       0.017157
#> lag.HOVAL     0.08579  0.2324 0.005196       0.005196
#> lambda        0.48345  0.1753 0.003920       0.003920
#> sige        113.22054 27.0946 0.605854       0.768548
#> 
#> 2. Quantiles for each variable:
#> 
#>                2.5%      25%       50%      75%     97.5%
#> (Intercept) 48.0258 65.77791  73.31319  80.1212  94.01866
#> INC         -1.7440 -1.28111  -1.03539  -0.7763  -0.29753
#> HOVAL       -0.4850 -0.34902  -0.28155  -0.2134  -0.07358
#> lag.INC     -2.3746 -1.51700  -1.06007  -0.6015   0.44654
#> lag.HOVAL   -0.3856 -0.06811   0.08553   0.2446   0.52837
#> lambda       0.1145  0.36368   0.49625   0.6078   0.79895
#> sige        73.3073 93.88527 108.49776 127.8573 177.04625
#> 
#> Impact measures (SDEM, MCMC, n):
#>           Direct    Indirect      Total
#> INC   -1.0293867 -1.04431275 -2.0736995
#> HOVAL -0.2815921  0.08578589 -0.1958062
#> ========================================================
#> Standard errors:
#>          Direct  Indirect     Total
#> INC   0.3549163 0.6838673 0.8165193
#> HOVAL 0.0994986 0.2201879 0.2645537
#> ========================================================
#> Z-values:
#>          Direct   Indirect     Total
#> INC   -2.900364 -1.5270694 -2.539682
#> HOVAL -2.830111  0.3896031 -0.740138
#> 
#> p-values:
#>       Direct    Indirect Total   
#> INC   0.0037273 0.12674  0.011095
#> HOVAL 0.0046532 0.69683  0.459216
#> 
#> 
#> Quantile (q) = 0.025
#> Accuracy (r) = +/- 0.01
#> Probability (s) = 0.95 
#>                                                    
#>              Burn-in  Total Lower bound  Dependence
#>              (M)      (N)   (Nmin)       factor (I)
#>  (Intercept) 2        930   937          0.993     
#>  INC         2        930   937          0.993     
#>  HOVAL       3        1010  937          1.080     
#>  lag.INC     2        930   937          0.993     
#>  lag.HOVAL   3        1052  937          1.120     
#>  lambda      2        892   937          0.952     
#>  sige        2        913   937          0.974     
#> 
#> 
#> Iterations = 501:2500
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                  Mean        SD  Naive SE Time-series SE
#> (Intercept) -31.39619 2791.7919 62.426364      55.399688
#> INC          -0.92744    0.4643  0.010381       0.021587
#> HOVAL        -0.28730    0.1272  0.002844       0.002844
#> lag.INC      -0.80036    1.0203  0.022815       0.052169
#> lag.HOVAL     0.05084    0.2855  0.006384       0.008611
#> lambda        0.68468    0.2072  0.004633       0.034485
#> sige        127.02847   34.2077  0.764907       3.123875
#> 
#> 2. Quantiles for each variable:
#> 
#>                   2.5%      25%       50%      75%      97.5%
#> (Intercept) -7462.6706  58.6130  71.02612  81.4215 6176.51971
#> INC            -1.7649  -1.2222  -0.95846  -0.6641    0.11485
#> HOVAL          -0.5582  -0.3632  -0.28227  -0.2076   -0.04721
#> lag.INC        -2.7193  -1.4753  -0.86419  -0.2074    1.43187
#> lag.HOVAL      -0.5424  -0.1312   0.06131   0.2385    0.61179
#> lambda          0.2821   0.5503   0.66750   0.8323    0.99972
#> sige           77.2768 101.7843 121.19261 144.6894  211.36169
#> 
#> Impact measures (SDEM, MCMC, n):
#>           Direct    Indirect      Total
#> INC   -0.9274355 -0.80035884 -1.7277943
#> HOVAL -0.2873048  0.05083708 -0.2364677
#> ========================================================
#> Standard errors:
#>          Direct  Indirect     Total
#> INC   0.4399307 0.9668633 1.2459420
#> HOVAL 0.1205389 0.2705534 0.3500774
#> ========================================================
#> Z-values:
#>          Direct   Indirect      Total
#> INC   -2.108140 -0.8277890 -1.3867374
#> HOVAL -2.383502  0.1879003 -0.6754725
#> 
#> p-values:
#>       Direct   Indirect Total  
#> INC   0.035019 0.40779  0.16552
#> HOVAL 0.017149 0.85095  0.49938
#> 
#> 
#> Quantile (q) = 0.025
#> Accuracy (r) = +/- 0.01
#> Probability (s) = 0.95 
#>                                                    
#>              Burn-in  Total Lower bound  Dependence
#>              (M)      (N)   (Nmin)       factor (I)
#>  (Intercept) 4        1242  937          1.330     
#>  INC         3        1010  937          1.080     
#>  HOVAL       3        1010  937          1.080     
#>  lag.INC     2        969   937          1.030     
#>  lag.HOVAL   2        892   937          0.952     
#>  lambda      11       3012  937          3.210     
#>  sige        2        950   937          1.010     
#> 
#> 
#> Iterations = 501:2500
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD Naive SE Time-series SE
#> (Intercept)  74.0524 10.9290 0.244380       0.244380
#> INC          -1.0151  0.3685 0.008240       0.008240
#> HOVAL        -0.2881  0.1010 0.002258       0.002258
#> lag.INC      -0.9149  0.5843 0.013065       0.012552
#> lambda        0.4887  0.1759 0.003934       0.004111
#> sige        112.2781 26.3936 0.590178       0.642304
#> 
#> 2. Quantiles for each variable:
#> 
#>                2.5%     25%      50%      75%     97.5%
#> (Intercept) 50.3754 67.8961  74.6926  81.4101  93.34744
#> INC         -1.7080 -1.2764  -1.0174  -0.7598  -0.31032
#> HOVAL       -0.4888 -0.3557  -0.2887  -0.2237  -0.08677
#> lag.INC     -2.0138 -1.2858  -0.9469  -0.5476   0.28401
#> lambda       0.1115  0.3757   0.4987   0.6151   0.78892
#> sige        72.7390 93.1156 108.5935 126.4337 177.50915
#> 
#> Impact measures (SDEM, MCMC, n):
#>           Direct   Indirect      Total
#> INC   -1.0151369 -0.9149477 -1.9300846
#> HOVAL -0.2881292         NA -0.2881292
#> ========================================================
#> Standard errors:
#>           Direct  Indirect      Total
#> INC   0.35314480 0.5599113 0.67653043
#> HOVAL 0.09675265        NA 0.09675265
#> ========================================================
#> Z-values:
#>          Direct  Indirect     Total
#> INC   -2.874563 -1.634094 -2.852916
#> HOVAL -2.977998        NA -2.977998
#> 
#> p-values:
#>       Direct    Indirect Total    
#> INC   0.0040459 0.10224  0.0043320
#> HOVAL 0.0029014 NA       0.0029014
#> 
#> 
#> Quantile (q) = 0.025
#> Accuracy (r) = +/- 0.01
#> Probability (s) = 0.95 
#>                                                    
#>              Burn-in  Total Lower bound  Dependence
#>              (M)      (N)   (Nmin)       factor (I)
#>  (Intercept) 2        969   937          1.030     
#>  INC         2        930   937          0.993     
#>  HOVAL       3        1096  937          1.170     
#>  lag.INC     2        930   937          0.993     
#>  lambda      2        892   937          0.952     
#>  sige        3        1010  937          1.080     
#> 
#> 
#> Iterations = 501:2500
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean       SD Naive SE Time-series SE
#> (Intercept)  75.2122 280.0161 6.261351       2.317368
#> INC          -0.9792   0.3958 0.008851       0.011485
#> HOVAL        -0.2875   0.0988 0.002209       0.002045
#> lag.INC      -0.8546   0.7076 0.015823       0.017724
#> lambda        0.5734   0.1742 0.003896       0.010450
#> sige        115.0232  27.9503 0.624987       0.814128
#> 
#> 2. Quantiles for each variable:
#> 
#>                2.5%     25%      50%      75%     97.5%
#> (Intercept) 41.0609 65.7069  73.5908  81.2756  98.58537
#> INC         -1.7566 -1.2427  -0.9817  -0.7088  -0.20021
#> HOVAL       -0.4834 -0.3533  -0.2866  -0.2229  -0.09669
#> lag.INC     -2.1066 -1.3221  -0.8901  -0.4221   0.67329
#> lambda       0.2167  0.4514   0.5784   0.6934   0.93900
#> sige        73.4904 95.0488 110.0353 129.7835 181.67444
#> 
#> Impact measures (SDEM, MCMC, n):
#>           Direct   Indirect      Total
#> INC   -0.9791842 -0.8545618 -1.8337459
#> HOVAL -0.2874605         NA -0.2874605
#> ========================================================
#> Standard errors:
#>           Direct  Indirect      Total
#> INC   0.37933568 0.6781451 0.85399103
#> HOVAL 0.09468092        NA 0.09468092
#> ========================================================
#> Z-values:
#>          Direct  Indirect     Total
#> INC   -2.581313 -1.260146 -2.147266
#> HOVAL -3.036097        NA -3.036097
#> 
#> p-values:
#>       Direct    Indirect Total    
#> INC   0.0098425 0.20762  0.0317721
#> HOVAL 0.0023966 NA       0.0023966
#> 
#> 
#> Quantile (q) = 0.025
#> Accuracy (r) = +/- 0.01
#> Probability (s) = 0.95 
#>                                                    
#>              Burn-in  Total Lower bound  Dependence
#>              (M)      (N)   (Nmin)       factor (I)
#>  (Intercept) 5        1472  937          1.570     
#>  INC         2        930   937          0.993     
#>  HOVAL       2        969   937          1.030     
#>  lag.INC     3        1010  937          1.080     
#>  lambda      11       3046  937          3.250     
#>  sige        2        930   937          0.993     
#>