house {spdep}	R Documentation

Lucas county OH housing

Description

Data on 25,357 single family homes sold in Lucas County, Ohio, 1993-1998 from the county auditor, together with an nb neighbour object constructed as a sphere of influence graph from projected coordinates.

Usage

data(house)

Format

The format is: Formal class 'SpatialPointsDataFrame' [package "sp"] with 5 slots. The data slot is a data frame with 25357 observations on the following 24 variables.

price

a numeric vector

yrbuilt

a numeric vector

stories

a factor with levels one bilevel multilvl one+half two two+half three

TLA

a numeric vector

wall

a factor with levels stucdrvt ccbtile metlvnyl brick stone wood partbrk

beds

a numeric vector

baths

a numeric vector

halfbaths

a numeric vector

frontage

a numeric vector

depth

a numeric vector

garage

a factor with levels no garage basement attached detached carport

garagesqft

a numeric vector

rooms

a numeric vector

lotsize

a numeric vector

sdate

a numeric vector

avalue

a numeric vector

s1993

a numeric vector

s1994

a numeric vector

s1995

a numeric vector

s1996

a numeric vector

s1997

a numeric vector

s1998

a numeric vector

syear

a factor with levels 1993 1994 1995 1996 1997 1998

age

a numeric vector

Its projection is CRS(+init=epsg:2834), the Ohio North State Plane.

Source

Dataset included in the Spatial Econometrics toolbox for Matlab, http://www.spatial-econometrics.com/html/jplv7.zip.

Examples

## Not run: 
# house <- read.table("house.dat", header=FALSE)
# names(house) <- c("price", "yrbuilt", "stories", "TLA", "wall", "beds",
#   "baths", "halfbaths", "frontage", "depth", "garage", "garagesqft", "rooms",
#   "lotsize", "sdate", "avalue", "long", "lat", "s1993", "s1994", "s1995",
#   "s1996", "s1997", "s1998")
# house$syear <- 1992 + house$s1993 + 2*house$s1994 + 3*house$s1995 +
# 4*house$s1996 + 5*house$s1997 + 6*house$s1998
# house$syear <- factor(house$syear)
# house$age <- (1999 - house$yrbuilt)/100
# house$stories <- factor(house$stories, levels=1:7, labels=c("one",
#  "bilevel", "multilvl", "one+half", "two", "two+half", "three"))
# house$wall <- factor(house$wall, levels=1:7, labels=c("stucdrvt",
#  "ccbtile", "metlvnyl", "brick", "stone", "wood", "partbrk"))
# house$garage <- factor(house$garage, levels=0:4, labels=c("no garage",
#  "basement", "attached", "detached", "carport"))
# library(sp)
# coordinates(house) <- c("long", "lat")
# proj4string(house) <- CRS("+proj=longlat +ellps=GRS80")
# library(rgdal)
# house <- spTransform(house, CRS("+init=epsg:2834"))
# library(spdep)
# LO_nb <- graph2nb(soi.graph(tri2nb(coordinates(house)), coordinates(house)))
# W <- as(nb2listw(LO_nb), "CsparseMatrix")
# trMat <- trW(W, type="mult")

## End(Not run)
data(house)
## maybe str(house) ; plot(house) ...

[Package spdep version 0.6-6 Index]