Impacts in spatial lag models

The calculation of impacts for spatial lag and spatial Durbin models is needed in order to interpret the regression coefficients correctly, because of the spillovers between the terms in these data generation processes (unlike the spatial error model). Methods for “SLX” and Bayesian fitted models are also provided, the former do not need MC simulations, while the latter pass through MCMC draws.

# S3 method for sarlm
impacts(obj, …, tr, R = NULL, listw = NULL, evalues=NULL,
 useHESS = NULL, tol = 1e-06, empirical = FALSE, Q=NULL)
# S3 method for stsls
impacts(obj, …, tr, R = NULL, listw = NULL, evalues=NULL,
 tol = 1e-06, empirical = FALSE, Q=NULL)
# S3 method for gmsar
impacts(obj, …, n = NULL, tr = NULL, R = NULL,
 listw = NULL, evalues=NULL, tol = 1e-06, empirical = FALSE, Q=NULL)
# S3 method for lagmess
impacts(obj, ..., R=NULL, listw=NULL, tol=1e-6,
 empirical=FALSE)
# S3 method for SLX
impacts(obj, ...)
# S3 method for MCMC_sar_g
impacts(obj, ..., tr=NULL, listw=NULL, evalues=NULL, Q=NULL)
# S3 method for MCMC_sem_g
impacts(obj, ..., tr=NULL, listw=NULL, evalues=NULL, Q=NULL)
# S3 method for MCMC_sac_g
impacts(obj, ..., tr=NULL, listw=NULL, evalues=NULL, Q=NULL)
# S3 method for lagImpact
plot(x, ..., choice="direct", trace=FALSE, density=TRUE)
# S3 method for lagImpact
print(x, ..., reportQ=NULL)
# S3 method for lagImpact
summary(object, ..., zstats=FALSE, short=FALSE, reportQ=NULL)
# S3 method for WXImpact
print(x, ...)
# S3 method for WXImpact
summary(object, ..., adjust_k=(attr(object, "type") == "SDEM"))
# S3 method for lagImpact
HPDinterval(obj, prob = 0.95, ..., choice="direct")
intImpacts(rho, beta, P, n, mu, Sigma, irho, drop2beta, bnames, interval,
 type, tr, R, listw, evalues, tol, empirical, Q, icept, iicept, p, mess=FALSE,
 samples=NULL, zero_fill = NULL, dvars = NULL)

Arguments

obj
A spatial regression object created by lagsarlm, lagmess or by lmSLX; in HPDinterval.lagImpact, a lagImpact object

…
Arguments passed through to methods in the coda package

tr
A vector of traces of powers of the spatial weights matrix created using trW, for approximate impact measures; if not given, listw must be given for exact measures (for small to moderate spatial weights matrices); the traces must be for the same spatial weights as were used in fitting the spatial regression, and must be row-standardised

listw
If tr is not given, a spatial weights object as created by nb2listw; they must be the same spatial weights as were used in fitting the spatial regression, but do not have to be row-standardised

evalues
vector of eigenvalues of spatial weights matrix for impacts calculations

n
defaults to length(obj$residuals); in the method for gmsar objects it may be used in panel settings to compute the impacts for cross-sectional weights only, suggested by Angela Parenti

R
If given, simulations are used to compute distributions for the impact measures, returned as mcmc objects; the objects are used for convenience but are not output by an MCMC process

useHESS
Use the Hessian approximation (if available) even if the asymptotic coefficient covariance matrix is available; used for comparing methods

tol
Argument passed to mvrnorm: tolerance (relative to largest variance) for numerical lack of positive-definiteness in the coefficient covariance matrix

empirical
Argument passed to mvrnorm (default FALSE): if true, the coefficients and their covariance matrix specify the empirical not population mean and covariance matrix

Q
default NULL, else an integer number of cumulative power series impacts to calculate if tr is given

reportQ
default NULL; if TRUE and Q given as an argument to impacts, report impact components

x, object
lagImpact objects created by impacts methods

zstats
default FALSE, if TRUE, also return z-values and p-values for the impacts based on the simulations

short
default FALSE, if TRUE passed to the print summary method to omit printing of the mcmc summaries

choice
One of three impacts: direct, indirect, or total

trace
Argument passed to plot.mcmc: plot trace plots

density
Argument passed to plot.mcmc: plot density plots

prob
Argument passed to HPDinterval.mcmc: a numeric scalar in the interval (0,1) giving the target probability content of the intervals

adjust_k
default TRUE if SDEM else FALSE, adjust internal OLS SDEM standard errors by dividing by n rather than (n-k) (default changed and bug fixed after 0.7-8; standard errors now ML in SDEM summary and impacts summary and identical - for SLX use FALSE)

rho, beta, P, mu, Sigma, irho, drop2beta, bnames, interval, type, icept, iicept, p, mess, samples, zero_fill, dvars
internal arguments shared inside impacts methods

Details

If called without R being set, the method returns the direct, indirect and total impacts for the variables in the model, for the variables themselves in tha spatial lag model case, for the variables and their spatial lags in the spatial Durbin (mixed) model case. The spatial lag impact measures are computed using eq. 2.46 (LeSage and Pace, 2009, p. 38), either using the exact dense matrix (when listw is given), or traces of powers of the weights matrix (when tr is given). When the traces are created by powering sparse matrices, the exact and the trace methods should give very similar results, unless the number of powers used is very small, or the spatial coefficient is close to its bounds.

If R is given, simulations will be used to create distributions for the impact measures, provided that the fitted model object contains a coefficient covariance matrix. The simulations are made using mvrnorm with the coefficients and their covariance matrix from the fitted model.

The simulations are stored as mcmc objects as defined in the coda package; the objects are used for convenience but are not output by an MCMC process. The simulated values of the coefficients are checked to see that the spatial coefficient remains within its valid interval --- draws outside the interval are discarded.

If a model is fitted with the “Durbin=” set to a formula subsetting the explanatory variables, the impacts object returned reports Durbin impacts for variables included in the formula and lag impacts for the other variables.

When Q and tr are given, addition impact component results are provided for each step in the traces of powers of the weights matrix up to and including the Q'th power. This increases computing time because the output object is substantially increased in size in proportion to the size of Q.

The method for gmsar objects is only for those of type SARAR output by gstsls, and assume that the spatial error coefficient is fixed, and thus omitted from the coefficients and covariance matrix used for simulation.

Value

An object of class lagImpact.

If no simulation is carried out, the object returned is a list with:

direct

numeric vector

indirect

numeric vector

total

numeric vector

and a matching Qres list attribute if Q was given. If simulation is carried out, the object returned is a list with:
res

a list with three components as for the non-simulation case, with a matching Qres list attribute if Q was given

sres

a list with three mcmc matrices, for the direct, indirect and total impacts with a matching Qmcmc list attribute if Q was given

References

LeSage J and RK Pace (2009) Introduction to Spatial Econometrics. CRC Press, Boca Raton, pp. 33--42, 114--115; LeSage J and MM Fischer (2008) Spatial growth regressions: model specification, estimation and interpretation. Spatial Economic Analysis 3 (3), pp. 275--304.

Roger Bivand, Gianfranco Piras (2015). Comparing Implementations of Estimation Methods for Spatial Econometrics. Journal of Statistical Software, 63(18), 1-36. https://www.jstatsoft.org/v63/i18/.

See also

Examples

if (require(rgdal, quietly=TRUE)) {
example(columbus, package="spData")
listw <- nb2listw(col.gal.nb)
ev <- eigenw(listw)
lobj <- lagsarlm(CRIME ~ INC + HOVAL, columbus, listw,
 control=list(pre_eig=ev))
summary(lobj)
mobj <- lagsarlm(CRIME ~ INC + HOVAL, columbus, listw, Durbin=TRUE,
 control=list(pre_eig=ev))
summary(mobj)
mobj1 <- lagsarlm(CRIME ~ INC + HOVAL, columbus, listw, Durbin= ~ INC,
 control=list(pre_eig=ev))
summary(mobj1)
W <- as(listw, "CsparseMatrix")
trMatc <- trW(W, type="mult")
trMC <- trW(W, type="MC")
set.seed(1)
impacts(lobj, listw=listw)
impacts(lobj, tr=trMatc)
impacts(lobj, tr=trMC)
impacts(lobj, evalues=ev)
lobj1 <- stsls(CRIME ~ INC + HOVAL, columbus, listw)
loobj1 <- impacts(lobj1, R=200, tr=trMatc)
summary(loobj1, zstats=TRUE, short=TRUE)
loobj2 <- impacts(lobj1, R=200, evalues=ev)
summary(loobj2, zstats=TRUE, short=TRUE)
library(coda)
HPDinterval(loobj1)
lobj1r <- stsls(CRIME ~ INC + HOVAL, columbus, listw, robust=TRUE)
loobj1r <- impacts(lobj1r, tr=trMatc, R=200)
summary(loobj1r, zstats=TRUE, short=TRUE)
lobjIQ5 <- impacts(lobj, tr=trMatc, R=200, Q=5)
summary(lobjIQ5, zstats=TRUE, short=TRUE)
summary(lobjIQ5, zstats=TRUE, short=TRUE, reportQ=TRUE)
impacts(mobj, listw=listw)
impacts(mobj, tr=trMatc)
impacts(mobj, tr=trMC)
impacts(mobj1, tr=trMatc)
impacts(mobj1, listw=listw)
try(impacts(mobj, evalues=ev))
summary(impacts(mobj, tr=trMatc, R=200), short=TRUE, zstats=TRUE)
summary(impacts(mobj1, tr=trMatc, R=200), short=TRUE, zstats=TRUE)
xobj <- lmSLX(CRIME ~ INC + HOVAL, columbus, listw)
summary(impacts(xobj))
eobj <- errorsarlm(CRIME ~ INC + HOVAL, columbus, listw, etype="emixed")
summary(impacts(eobj), adjust_k=TRUE)

#> Error: <text>:48:0: unexpected end of input
#> 46: summary(impacts(eobj), adjust_k=TRUE)
#> 47: 
#>    ^

mobj1 <- lagsarlm(CRIME ~ INC + HOVAL, columbus, listw, type="mixed",
method="Matrix", control=list(fdHess=TRUE))

#> Error in lagsarlm(CRIME ~ INC + HOVAL, columbus, listw, type = "mixed",     method = "Matrix", control = list(fdHess = TRUE)): object 'listw' not found

summary(mobj1)

#> Error in summary(mobj1): object 'mobj1' not found

set.seed(1)
summary(impacts(mobj1, tr=trMatc, R=1000), zstats=TRUE, short=TRUE)

#> Error in impacts(mobj1, tr = trMatc, R = 1000): object 'mobj1' not found

summary(impacts(mobj, tr=trMatc, R=1000), zstats=TRUE, short=TRUE)

#> Error in impacts(mobj, tr = trMatc, R = 1000): object 'mobj' not found

mobj2 <- lagsarlm(CRIME ~ INC + HOVAL, columbus, listw, type="mixed",
method="Matrix", control=list(fdHess=TRUE, optimHess=TRUE))

#> Error in lagsarlm(CRIME ~ INC + HOVAL, columbus, listw, type = "mixed",     method = "Matrix", control = list(fdHess = TRUE, optimHess = TRUE)): object 'listw' not found

summary(impacts(mobj2, tr=trMatc, R=1000), zstats=TRUE, short=TRUE)

#> Error in impacts(mobj2, tr = trMatc, R = 1000): object 'mobj2' not found

mobj3 <- lagsarlm(CRIME ~ INC + HOVAL, columbus, listw, type="mixed",
method="spam", control=list(fdHess=TRUE))

#> Error in lagsarlm(CRIME ~ INC + HOVAL, columbus, listw, type = "mixed",     method = "spam", control = list(fdHess = TRUE)): object 'listw' not found

summary(impacts(mobj3, tr=trMatc, R=1000), zstats=TRUE, short=TRUE)

#> Error in impacts(mobj3, tr = trMatc, R = 1000): object 'mobj3' not found

data(boston, package="spData")
Wb <- as(nb2listw(boston.soi), "CsparseMatrix")
trMatb <- trW(Wb, type="mult")
gp2mMi <- lagsarlm(log(CMEDV) ~ CRIM + ZN + INDUS + CHAS + I(NOX^2) +
I(RM^2) +  AGE + log(DIS) + log(RAD) + TAX + PTRATIO + B + log(LSTAT),
data=boston.c, nb2listw(boston.soi), type="mixed", method="Matrix",
control=list(fdHess=TRUE), trs=trMatb)
summary(gp2mMi)

#> 
#> Call:lagsarlm(formula = log(CMEDV) ~ CRIM + ZN + INDUS + CHAS + I(NOX^2) + 
#>     I(RM^2) + AGE + log(DIS) + log(RAD) + TAX + PTRATIO + B + 
#>     log(LSTAT), data = boston.c, listw = nb2listw(boston.soi), 
#>     type = "mixed", method = "Matrix", trs = trMatb, control = list(fdHess = TRUE))
#> 
#> Residuals:
#>        Min         1Q     Median         3Q        Max 
#> -0.6316833 -0.0629790 -0.0090776  0.0682421  0.6991072 
#> 
#> Type: mixed 
#> Coefficients: (asymptotic standard errors) 
#>                   Estimate  Std. Error  z value  Pr(>|z|)
#> (Intercept)     1.89816218  0.24400182   7.7793 7.327e-15
#> CRIM           -0.00571021  0.00093857  -6.0839 1.173e-09
#> ZN              0.00069091  0.00051874   1.3319 0.1828931
#> INDUS          -0.00111343  0.00307380  -0.3622 0.7171778
#> CHAS1          -0.04163225  0.02738839  -1.5201 0.1284937
#> I(NOX^2)       -0.01034950  0.19358633  -0.0535 0.9573639
#> I(RM^2)         0.00794979  0.00102109   7.7856 6.883e-15
#> AGE            -0.00128789  0.00048929  -2.6322 0.0084838
#> log(DIS)       -0.12404108  0.09510145  -1.3043 0.1921304
#> log(RAD)        0.05863502  0.02257529   2.5973 0.0093957
#> TAX            -0.00049084  0.00012146  -4.0412 5.317e-05
#> PTRATIO        -0.01319853  0.00595331  -2.2170 0.0266227
#> B               0.00056383  0.00011084   5.0867 3.643e-07
#> log(LSTAT)     -0.24724454  0.02265149 -10.9152 < 2.2e-16
#> lag.CRIM       -0.00464215  0.00173900  -2.6694 0.0075978
#> lag.ZN         -0.00037937  0.00070703  -0.5366 0.5915659
#> lag.INDUS       0.00025064  0.00385911   0.0649 0.9482165
#> lag.CHAS1       0.12518252  0.04083949   3.0652 0.0021750
#> lag.I(NOX^2)   -0.38640401  0.22253428  -1.7364 0.0824967
#> lag.I(RM^2)    -0.00451252  0.00148919  -3.0302 0.0024440
#> lag.AGE         0.00149678  0.00068470   2.1860 0.0288128
#> lag.log(DIS)   -0.00453785  0.10056961  -0.0451 0.9640105
#> lag.log(RAD)   -0.00940702  0.03127787  -0.3008 0.7636001
#> lag.TAX         0.00041083  0.00017859   2.3004 0.0214237
#> lag.PTRATIO     0.00060355  0.00789994   0.0764 0.9391011
#> lag.B          -0.00050781  0.00014107  -3.5996 0.0003187
#> lag.log(LSTAT)  0.09846781  0.03574182   2.7550 0.0058697
#> 
#> Rho: 0.59578, LR test value: 181.68, p-value: < 2.22e-16
#> Asymptotic standard error: 0.037474
#>     z-value: 15.899, p-value: < 2.22e-16
#> Wald statistic: 252.76, p-value: < 2.22e-16
#> 
#> Log likelihood: 300.6131 for mixed model
#> ML residual variance (sigma squared): 0.016011, (sigma: 0.12654)
#> Number of observations: 506 
#> Number of parameters estimated: 29 
#> AIC: -543.23, (AIC for lm: -363.55)
#> LM test for residual autocorrelation
#> test value: 29.772, p-value: 4.8604e-08
#> 

summary(impacts(gp2mMi, tr=trMatb, R=1000), zstats=TRUE, short=TRUE)

#> Impact measures (mixed, trace):
#>                   Direct      Indirect         Total
#> CRIM       -0.0074555753 -1.815485e-02 -0.0256104225
#> ZN          0.0006979073  7.279849e-05  0.0007707058
#> INDUS      -0.0012029822 -9.314672e-04 -0.0021344494
#> CHAS1      -0.0198526436  2.265454e-01  0.2066927455
#> I(NOX^2)   -0.0955268264 -8.859909e-01 -0.9815177559
#> I(RM^2)     0.0079983430  5.050177e-04  0.0085033608
#> AGE        -0.0011296134  1.646365e-03  0.0005167515
#> log(DIS)   -0.1410601687 -1.770278e-01 -0.3180879407
#> log(RAD)    0.0641735548  5.761026e-02  0.1217838138
#> TAX        -0.0004651543  2.672120e-04 -0.0001979424
#> PTRATIO    -0.0147737151 -1.638465e-02 -0.0311583673
#> B           0.0005265343 -3.879424e-04  0.0001385920
#> log(LSTAT) -0.2578403213 -1.102144e-01 -0.3680547068
#> ========================================================
#> Simulation results (asymptotic variance matrix):
#> ========================================================
#> Simulated standard errors
#>                  Direct     Indirect        Total
#> CRIM       0.0009825947 0.0035673793 0.0039941079
#> ZN         0.0004893718 0.0012186281 0.0012429895
#> INDUS      0.0028334621 0.0055973954 0.0053186322
#> CHAS1      0.0283708321 0.0812423795 0.0913355496
#> I(NOX^2)   0.1855060888 0.2921285220 0.2436117326
#> I(RM^2)    0.0010567046 0.0031433490 0.0035619449
#> AGE        0.0004659277 0.0012359580 0.0013035531
#> log(DIS)   0.0875172480 0.1169983912 0.0754055941
#> log(RAD)   0.0206253904 0.0508355125 0.0495616064
#> TAX        0.0001121870 0.0003046755 0.0003189050
#> PTRATIO    0.0054607955 0.0123210949 0.0118526404
#> B          0.0001033402 0.0002298688 0.0002317636
#> log(LSTAT) 0.0212790458 0.0619329522 0.0649277399
#> 
#> Simulated z-values:
#>                 Direct    Indirect      Total
#> CRIM        -7.5687958 -5.12111000 -6.4359805
#> ZN           1.4916035  0.04855513  0.6348559
#> INDUS       -0.4103920 -0.15456513 -0.3812996
#> CHAS1       -0.7026843  2.77932883  2.2539258
#> I(NOX^2)    -0.5037322 -3.05392436 -4.0457157
#> I(RM^2)      7.5459998  0.18384377  2.4008731
#> AGE         -2.3773644  1.29406030  0.3772184
#> log(DIS)    -1.5975553 -1.53421335 -4.2346214
#> log(RAD)     3.1315355  1.16521418  2.4983734
#> TAX         -4.1773159  0.83787480 -0.6690406
#> PTRATIO     -2.6985674 -1.30214929 -2.5969091
#> B            5.0430369 -1.71412155  0.5485124
#> log(LSTAT) -12.1251369 -1.75005587 -5.6431576
#> 
#> Simulated p-values:
#>            Direct     Indirect   Total     
#> CRIM       3.7748e-14 3.0374e-07 1.2268e-10
#> ZN         0.1358031  0.9612738  0.5255224 
#> INDUS      0.6815184  0.8771642  0.7029810 
#> CHAS1      0.4822525  0.0054471  0.0242008 
#> I(NOX^2)   0.6144496  0.0022587  5.2164e-05
#> I(RM^2)    4.4853e-14 0.8541360  0.0163560 
#> AGE        0.0174369  0.1956446  0.7060113 
#> log(DIS)   0.1101420  0.1249772  2.2894e-05
#> log(RAD)   0.0017389  0.2439323  0.0124765 
#> TAX        2.9497e-05 0.4021010  0.5034696 
#> PTRATIO    0.0069639  0.1928654  0.0094067 
#> B          4.5820e-07 0.0865064  0.5833401 
#> log(LSTAT) < 2.22e-16 0.0801087  1.6696e-08

data(house, package="spData")
lw <- nb2listw(LO_nb)
form <- formula(log(price) ~ age + I(age^2) + I(age^3) + log(lotsize) +
   rooms + log(TLA) + beds + syear)
lobj <- lagsarlm(form, house, lw, method="Matrix",
 control=list(fdHess=TRUE), trs=trMat)
summary(lobj)

#> 
#> Call:lagsarlm(formula = form, data = house, listw = lw, method = "Matrix", 
#>     trs = trMat, control = list(fdHess = TRUE))
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -2.488174 -0.128732  0.030689  0.174778  2.845118 
#> 
#> Type: lag 
#> Coefficients: (numerical Hessian approximate standard errors) 
#>                Estimate Std. Error  z value  Pr(>|z|)
#> (Intercept)   0.2583272  0.0656170   3.9369 8.254e-05
#> age           1.3084686  0.0563647  23.2143 < 2.2e-16
#> I(age^2)     -2.3213256  0.1024831 -22.6508 < 2.2e-16
#> I(age^3)      0.6548946  0.0551842  11.8674 < 2.2e-16
#> log(lotsize)  0.0729753  0.0030642  23.8153 < 2.2e-16
#> rooms        -0.0025340  0.0030414  -0.8332 0.4047438
#> log(TLA)      0.5778330  0.0101653  56.8437 < 2.2e-16
#> beds          0.0156215  0.0045220   3.4545 0.0005512
#> syear1994     0.0444752  0.0073880   6.0199 1.745e-09
#> syear1995     0.0860740  0.0072149  11.9300 < 2.2e-16
#> syear1996     0.1059371  0.0069873  15.1613 < 2.2e-16
#> syear1997     0.1473471  0.0069302  21.2615 < 2.2e-16
#> syear1998     0.2007216  0.0071253  28.1703 < 2.2e-16
#> 
#> Rho: 0.52281, LR test value: 12594, p-value: < 2.22e-16
#> Approximate (numerical Hessian) standard error: 0.0038416
#>     z-value: 136.09, p-value: < 2.22e-16
#> Wald statistic: 18521, p-value: < 2.22e-16
#> 
#> Log likelihood: -7670.362 for lag model
#> ML residual variance (sigma squared): 0.094786, (sigma: 0.30787)
#> Number of observations: 25357 
#> Number of parameters estimated: 15 
#> AIC: 15371, (AIC for lm: 27962)
#> 

loobj <- impacts(lobj, tr=trMat, R=1000)
summary(loobj, zstats=TRUE, short=TRUE)

#> Impact measures (lag, trace):
#>                    Direct     Indirect        Total
#> age           1.495027201  1.247025340  2.742052541
#> I(age^2)     -2.652295154 -2.212320461 -4.864615615
#> I(age^3)      0.748268069  0.624141983  1.372410052
#> log(lotsize)  0.083379995  0.069548545  0.152928539
#> rooms        -0.002895346 -0.002415053 -0.005310398
#> log(TLA)      0.660219207  0.550699064  1.210918271
#> beds          0.017848752  0.014887921  0.032736673
#> syear1994     0.050816402  0.042386748  0.093203150
#> syear1995     0.098346272  0.082032148  0.180378420
#> syear1996     0.121041417  0.100962520  0.222003938
#> syear1997     0.168355573  0.140427991  0.308783564
#> syear1998     0.229340074  0.191296107  0.420636180
#> ========================================================
#> Simulation results (mixed Hessian approximation variance matrix):
#> ========================================================
#> Simulated standard errors
#>                   Direct    Indirect       Total
#> age          0.065159670 0.056074844 0.120047960
#> I(age^2)     0.118101828 0.100445846 0.216465436
#> I(age^3)     0.064231590 0.053764246 0.117689824
#> log(lotsize) 0.003344771 0.002753924 0.006023682
#> rooms        0.003346013 0.002789327 0.006135140
#> log(TLA)     0.011081825 0.010909583 0.020741960
#> beds         0.004880528 0.004087197 0.008965259
#> syear1994    0.008302791 0.006969403 0.015261051
#> syear1995    0.008371747 0.007112559 0.015443796
#> syear1996    0.008045969 0.006889221 0.014872608
#> syear1997    0.008179808 0.007187624 0.015252035
#> syear1998    0.008406694 0.007550248 0.015751225
#> 
#> Simulated z-values:
#>                   Direct    Indirect       Total
#> age           22.9266686  22.2072306  22.8172236
#> I(age^2)     -22.4311583 -21.9841384 -22.4395001
#> I(age^3)      11.6129151  11.5643840  11.6209400
#> log(lotsize)  24.9485184  25.2561836  25.3998603
#> rooms         -0.8151298  -0.8150624  -0.8151258
#> log(TLA)      59.5527853  50.4250112  58.3392000
#> beds           3.6285293   3.6124764   3.6222089
#> syear1994      6.1088642   6.0671451   6.0942726
#> syear1995     11.7402118  11.5197415  11.6694706
#> syear1996     15.0114911  14.6150645  14.8910278
#> syear1997     20.5610088  19.5063150  20.2195422
#> syear1998     27.3003157  25.3393120  26.7168753
#> 
#> Simulated p-values:
#>              Direct     Indirect   Total     
#> age          < 2.22e-16 < 2.22e-16 < 2.22e-16
#> I(age^2)     < 2.22e-16 < 2.22e-16 < 2.22e-16
#> I(age^3)     < 2.22e-16 < 2.22e-16 < 2.22e-16
#> log(lotsize) < 2.22e-16 < 2.22e-16 < 2.22e-16
#> rooms        0.41499802 0.41503662 0.4150003 
#> log(TLA)     < 2.22e-16 < 2.22e-16 < 2.22e-16
#> beds         0.00028504 0.00030329 0.0002921 
#> syear1994    1.0034e-09 1.302e-09  1.0994e-09
#> syear1995    < 2.22e-16 < 2.22e-16 < 2.22e-16
#> syear1996    < 2.22e-16 < 2.22e-16 < 2.22e-16
#> syear1997    < 2.22e-16 < 2.22e-16 < 2.22e-16
#> syear1998    < 2.22e-16 < 2.22e-16 < 2.22e-16

lobj1 <- stsls(form, house, lw)
loobj1 <- impacts(lobj1, tr=trMat, R=1000)
summary(loobj1, zstats=TRUE, short=TRUE)

#> Impact measures (lag, trace):
#>                    Direct     Indirect       Total
#> age           1.492873357  1.265395569  2.75826893
#> I(age^2)     -2.642549749 -2.239889088 -4.88243884
#> I(age^3)      0.744861040  0.631362235  1.37622327
#> log(lotsize)  0.082511049  0.069938361  0.15244941
#> rooms        -0.003035731 -0.002573159 -0.00560889
#> log(TLA)      0.658778386  0.558396495  1.21717488
#> beds          0.018256951  0.015475033  0.03373198
#> syear1994     0.050971955  0.043205062  0.09417702
#> syear1995     0.098646802  0.083615416  0.18226222
#> syear1996     0.121389057  0.102892301  0.22428136
#> syear1997     0.168907859  0.143170387  0.31207825
#> syear1998     0.230052473  0.194998041  0.42505051
#> ========================================================
#> Simulation results (asymptotic IV variance matrix):
#> ========================================================
#> Simulated standard errors
#>                   Direct    Indirect       Total
#> age          0.066761364 0.062449803 0.126307377
#> I(age^2)     0.121471394 0.110471871 0.226738247
#> I(age^3)     0.065395151 0.056646363 0.121243929
#> log(lotsize) 0.003587125 0.002998270 0.006400355
#> rooms        0.003537495 0.002997679 0.006534655
#> log(TLA)     0.012207271 0.013892976 0.023317198
#> beds         0.005265662 0.004501404 0.009760991
#> syear1994    0.008556363 0.007300904 0.015826967
#> syear1995    0.008102610 0.007105658 0.015096919
#> syear1996    0.007980096 0.007197483 0.015012935
#> syear1997    0.007930940 0.007518526 0.015147369
#> syear1998    0.008244868 0.008348955 0.016092252
#> 
#> Simulated z-values:
#>                   Direct   Indirect       Total
#> age           22.3544113  20.239745  21.8227879
#> I(age^2)     -21.7559867 -20.259260 -21.5261802
#> I(age^3)      11.3907499  11.136673  11.3469749
#> log(lotsize)  23.0219871  23.321697  23.8279742
#> rooms         -0.7842976  -0.784369  -0.7843926
#> log(TLA)      53.9206722  40.122788  52.1353040
#> beds           3.4164154   3.386731   3.4048522
#> syear1994      5.9485389   5.904262   5.9395024
#> syear1995     12.1619428  11.745851  12.0558028
#> syear1996     15.2148517  14.288072  14.9373950
#> syear1997     21.2800989  19.012543  20.5789844
#> syear1998     27.8828585  23.321737  26.3855323
#> 
#> Simulated p-values:
#>              Direct     Indirect   Total     
#> age          < 2.22e-16 < 2.22e-16 < 2.22e-16
#> I(age^2)     < 2.22e-16 < 2.22e-16 < 2.22e-16
#> I(age^3)     < 2.22e-16 < 2.22e-16 < 2.22e-16
#> log(lotsize) < 2.22e-16 < 2.22e-16 < 2.22e-16
#> rooms        0.43286553 0.43282365 0.432810  
#> log(TLA)     < 2.22e-16 < 2.22e-16 < 2.22e-16
#> beds         0.00063451 0.00070731 0.000662  
#> syear1994    2.7055e-09 3.5423e-09 2.8589e-09
#> syear1995    < 2.22e-16 < 2.22e-16 < 2.22e-16
#> syear1996    < 2.22e-16 < 2.22e-16 < 2.22e-16
#> syear1997    < 2.22e-16 < 2.22e-16 < 2.22e-16
#> syear1998    < 2.22e-16 < 2.22e-16 < 2.22e-16

mobj <- lagsarlm(form, house, lw, type="mixed",
 method="Matrix", control=list(fdHess=TRUE), trs=trMat)
summary(mobj)

#> 
#> Call:lagsarlm(formula = form, data = house, listw = lw, type = "mixed", 
#>     method = "Matrix", trs = trMat, control = list(fdHess = TRUE))
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -2.421437 -0.126022  0.029935  0.167534  2.895359 
#> 
#> Type: mixed 
#> Coefficients: (numerical Hessian approximate standard errors) 
#>                     Estimate  Std. Error  z value  Pr(>|z|)
#> (Intercept)       0.67918314  0.08721235   7.7877 6.883e-15
#> age               0.97905382  0.08453611  11.5815 < 2.2e-16
#> I(age^2)         -1.87856579  0.13822485 -13.5907 < 2.2e-16
#> I(age^3)          0.55454470  0.06800288   8.1547 4.441e-16
#> log(lotsize)      0.10944213  0.00522190  20.9583 < 2.2e-16
#> rooms             0.00136728  0.00302395   0.4521  0.651162
#> log(TLA)          0.62148335  0.01060003  58.6303 < 2.2e-16
#> beds              0.01294189  0.00448716   2.8842  0.003924
#> syear1994         0.04031417  0.00725573   5.5562 2.757e-08
#> syear1995         0.08394251  0.00708737  11.8440 < 2.2e-16
#> syear1996         0.10400813  0.00686112  15.1591 < 2.2e-16
#> syear1997         0.14696274  0.00680970  21.5814 < 2.2e-16
#> syear1998         0.20021074  0.00699781  28.6105 < 2.2e-16
#> lag.age          -0.04653535  0.09647336  -0.4824  0.629547
#> lag.I(age^2)      0.27074756  0.16312167   1.6598  0.096957
#> lag.I(age^3)     -0.32908972  0.08295047  -3.9673 7.269e-05
#> lag.log(lotsize) -0.05072331  0.00577489  -8.7834 < 2.2e-16
#> lag.rooms         0.00050653  0.00441843   0.1146  0.908731
#> lag.log(TLA)     -0.07588857  0.01502932  -5.0494 4.433e-07
#> lag.beds         -0.04306553  0.00657416  -6.5507 5.726e-11
#> lag.syear1994    -0.02428053  0.01111700  -2.1841  0.028956
#> lag.syear1995    -0.04415580  0.01084657  -4.0709 4.682e-05
#> lag.syear1996    -0.05712325  0.01050600  -5.4372 5.412e-08
#> lag.syear1997    -0.09260271  0.01042633  -8.8816 < 2.2e-16
#> lag.syear1998    -0.09696080  0.01078352  -8.9916 < 2.2e-16
#> 
#> Rho: 0.53828, LR test value: 9593.4, p-value: < 2.22e-16
#> Approximate (numerical Hessian) standard error: 0.0046668
#>     z-value: 115.34, p-value: < 2.22e-16
#> Wald statistic: 13304, p-value: < 2.22e-16
#> 
#> Log likelihood: -7307.507 for mixed model
#> ML residual variance (sigma squared): 0.091319, (sigma: 0.30219)
#> Number of observations: 25357 
#> Number of parameters estimated: 27 
#> AIC: 14669, (AIC for lm: 24260)
#> 

moobj <- impacts(mobj, tr=trMat, R=1000)
summary(moobj, zstats=TRUE, short=TRUE)

#> Impact measures (mixed, trace):
#>                    Direct     Indirect        Total
#> age           1.116551228  0.903101121  2.019652349
#> I(age^2)     -2.090461358 -1.391758011 -3.482219369
#> I(age^3)      0.545790447 -0.057499114  0.488291333
#> log(lotsize)  0.111785027  0.015388439  0.127173466
#> rooms         0.001722848  0.002335444  0.004058292
#> log(TLA)      0.695500044  0.486151389  1.181651434
#> beds          0.002611230 -0.067853136 -0.065241906
#> syear1994     0.039575758 -0.004850020  0.034725738
#> syear1995     0.084236861  0.001933350  0.086170211
#> syear1996     0.103682464 -0.002139009  0.101543455
#> syear1997     0.143100555 -0.025367367  0.117733187
#> syear1998     0.203303767  0.020315386  0.223619153
#> ========================================================
#> Simulation results (mixed Hessian approximation variance matrix):
#> ========================================================
#> Simulated standard errors
#>                   Direct    Indirect       Total
#> age          0.077925351 0.133375095 0.146815832
#> I(age^2)     0.134517146 0.238296829 0.278571515
#> I(age^3)     0.068830959 0.128769819 0.156600948
#> log(lotsize) 0.004766226 0.007062475 0.007356136
#> rooms        0.003622726 0.008606519 0.011067116
#> log(TLA)     0.011764634 0.026259475 0.033060599
#> beds         0.005231272 0.012493821 0.015915221
#> syear1994    0.008942966 0.021845951 0.028344722
#> syear1995    0.008524353 0.021045048 0.027231971
#> syear1996    0.008205057 0.020386026 0.026295470
#> syear1997    0.008377581 0.020606130 0.026538431
#> syear1998    0.008553225 0.021404295 0.027588655
#> 
#> Simulated z-values:
#>                   Direct    Indirect       Total
#> age           14.3330903  6.79435482  13.7799090
#> I(age^2)     -15.5339855 -5.86851877 -12.5211539
#> I(age^3)       7.9135955 -0.41523779   3.1368282
#> log(lotsize)  23.4882813  2.14469543  17.2777285
#> rooms          0.4927074  0.29191530   0.3882961
#> log(TLA)      59.0826292 18.50393093  35.7219484
#> beds           0.5170681 -5.44045245  -4.1009244
#> syear1994      4.4495327 -0.24052054   1.2184851
#> syear1995      9.9016123  0.09650671   3.1740571
#> syear1996     12.6998683 -0.07544440   3.9042897
#> syear1997     17.1458327 -1.20850478   4.4741905
#> syear1998     23.7694725  0.91368089   8.0780429
#> 
#> Simulated p-values:
#>              Direct     Indirect   Total     
#> age          < 2.22e-16 1.0880e-11 < 2.22e-16
#> I(age^2)     < 2.22e-16 4.3971e-09 < 2.22e-16
#> I(age^3)     2.4425e-15 0.677968   0.0017079 
#> log(lotsize) < 2.22e-16 0.031977   < 2.22e-16
#> rooms        0.62222    0.770351   0.6977969 
#> log(TLA)     < 2.22e-16 < 2.22e-16 < 2.22e-16
#> beds         0.60511    5.3145e-08 4.1150e-05
#> syear1994    8.6057e-06 0.809927   0.2230397 
#> syear1995    < 2.22e-16 0.923118   0.0015032 
#> syear1996    < 2.22e-16 0.939861   9.4503e-05
#> syear1997    < 2.22e-16 0.226853   7.6701e-06
#> syear1998    < 2.22e-16 0.360885   6.6613e-16

}

#> Error: <text>:1:1: unexpected '}'
#> 1: }
#>     ^