Matrix exponential spatial lag model

The function fits a matrix exponential spatial lag model, using optim to find the value of alpha, the spatial coefficient.

lagmess(formula, data = list(), listw, zero.policy = NULL, na.action = na.fail,
 q = 10, start = -2.5, control=list(), method="BFGS", verbose=NULL,
 use_expm=FALSE)
# S3 method for lagmess
summary(object, ...)
# S3 method for lagmess
print(x, ...)
# S3 method for summary.lagmess
print(x, digits = max(5, .Options$digits - 3),
    signif.stars = FALSE, ...)
# S3 method for lagmess
residuals(object, ...)
# S3 method for lagmess
deviance(object, ...)
# S3 method for lagmess
coef(object, ...)
# S3 method for lagmess
fitted(object, ...)
# S3 method for lagmess
logLik(object, ...)

Arguments

formula
a symbolic description of the model to be fit. The details of model specification are given for lm()

data
an optional data frame containing the variables in the model. By default the variables are taken from the environment which the function is called.

listw
a listw object created for example by nb2listw

zero.policy
default NULL, use global option value; if TRUE assign zero to the lagged value of zones without neighbours, if FALSE assign NA - causing lagmess() to terminate with an error

na.action
a function (default options("na.action")), can also be na.omit or na.exclude with consequences for residuals and fitted values - in these cases the weights list will be subsetted to remove NAs in the data. It may be necessary to set zero.policy to TRUE because this subsetting may create no-neighbour observations. Note that only weights lists created without using the glist argument to nb2listw may be subsetted.

q
default 10; number of powers of the spatial weights to use

start
starting value for numerical optimization, should be a small negative number

control
control parameters passed to optim

method
default BFGS, method passed to optim

verbose
default NULL, use global option value; if TRUE report function values during optimization

use_expm
default FALSE; if TRUE use expm::expAtv instead of a truncated power series of W

x,object
Objects of classes lagmess or summary.lagmess to be passed to methods

digits
the number of significant digits to use when printing

signif.stars
logical. If TRUE, "significance stars" are printed for each coefficient.

…
further arguments passed to or from other methods

Details

The underlying spatial lag model:

$$y = \rho W y + X \beta + \varepsilon$$

where \(\rho\) is the spatial parameter may be fitted by maximum likelihood. In that case, the log likelihood function includes the logartithm of cumbersome Jacobian term \(|I - \rho W|\). If we rewrite the model as:

$$S y = X \beta + \varepsilon$$

we see that in the ML case \(S y = (I - \rho W) y\). If W is row-stochastic, S may be expressed as a linear combination of row-stochastic matrices. By pre-computing the matrix \([y Wy, W^2y, ..., W^{q-1}y]\), the term \(S y (\alpha)\) can readily be found by numerical optimization using the matrix exponential approach. \(\alpha\) and \(\rho\) are related as \(\rho = 1 - \exp{\alpha}\), conditional on the number of matrix power terms taken q.

Value

The function returns an object of class lagmess with components:

lmobj

the lm object returned after fitting alpha

alpha

the spatial coefficient

alphase

the standard error of the spatial coefficient using the numerical Hessian

rho

the value of rho implied by alpha

bestmess

the object returned by optim

q

the number of powers of the spatial weights used

start

the starting value for numerical optimization used

na.action

(possibly) named vector of excluded or omitted observations if non-default na.action argument used

nullLL

the log likelihood of the aspatial model for the same data

References

J. P. LeSage and R. K. Pace (2007) A matrix exponential specification. Journal of Econometrics, 140, 190-214; J. P. LeSage and R. K. Pace (2009) Introduction to Spatial Econometrics. CRC Press, Chapter 9.

See also

Examples

data(baltimore, package="spData")
baltimore$AGE <- ifelse(baltimore$AGE < 1, 1, baltimore$AGE)
lw <- nb2listw(knn2nb(knearneigh(cbind(baltimore$X, baltimore$Y), k=7)))
obj1 <- lm(log(PRICE) ~ PATIO + log(AGE) + log(SQFT),
 data=baltimore)
lm.morantest(obj1, lw)

#> 
#> 	Global Moran I for regression residuals
#> 
#> data:  
#> model: lm(formula = log(PRICE) ~ PATIO + log(AGE) + log(SQFT), data =
#> baltimore)
#> weights: lw
#> 
#> Moran I statistic standard deviate = 7.4782, p-value = 3.769e-14
#> alternative hypothesis: greater
#> sample estimates:
#> Observed Moran I      Expectation         Variance 
#>      0.245497850     -0.007859541      0.001147832 
#> 

lm.LMtests(obj1, lw, test="all")

#> 
#> 	Lagrange multiplier diagnostics for spatial dependence
#> 
#> data:  
#> model: lm(formula = log(PRICE) ~ PATIO + log(AGE) + log(SQFT), data =
#> baltimore)
#> weights: lw
#> 
#> LMerr = 48.823, df = 1, p-value = 2.802e-12
#> 
#> 
#> 	Lagrange multiplier diagnostics for spatial dependence
#> 
#> data:  
#> model: lm(formula = log(PRICE) ~ PATIO + log(AGE) + log(SQFT), data =
#> baltimore)
#> weights: lw
#> 
#> LMlag = 83.434, df = 1, p-value < 2.2e-16
#> 
#> 
#> 	Lagrange multiplier diagnostics for spatial dependence
#> 
#> data:  
#> model: lm(formula = log(PRICE) ~ PATIO + log(AGE) + log(SQFT), data =
#> baltimore)
#> weights: lw
#> 
#> RLMerr = 1.2309, df = 1, p-value = 0.2672
#> 
#> 
#> 	Lagrange multiplier diagnostics for spatial dependence
#> 
#> data:  
#> model: lm(formula = log(PRICE) ~ PATIO + log(AGE) + log(SQFT), data =
#> baltimore)
#> weights: lw
#> 
#> RLMlag = 35.843, df = 1, p-value = 2.139e-09
#> 
#> 
#> 	Lagrange multiplier diagnostics for spatial dependence
#> 
#> data:  
#> model: lm(formula = log(PRICE) ~ PATIO + log(AGE) + log(SQFT), data =
#> baltimore)
#> weights: lw
#> 
#> SARMA = 84.665, df = 2, p-value < 2.2e-16
#> 

system.time(obj2 <- lagmess(log(PRICE) ~ PATIO + log(AGE) + log(SQFT), data=baltimore, listw=lw))

#>    user  system elapsed 
#>   0.051   0.000   0.052 

summary(obj2)

#> Matrix exponential spatial lag model:
#> 
#> Call:
#> lagmess(formula = log(PRICE) ~ PATIO + log(AGE) + log(SQFT), 
#>     data = baltimore, listw = lw)
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -2.026015 -0.141654  0.050901  0.224067  1.072942 
#> 
#> Coefficients:
#>              Estimate Std. Error t value  Pr(>|t|)
#> (Intercept)  1.549087   0.214465  7.2230 9.553e-12
#> PATIO        0.259322   0.086871  2.9851  0.003176
#> log(AGE)    -0.148169   0.035244 -4.2041 3.899e-05
#> log(SQFT)    0.300203   0.071582  4.1939 4.066e-05
#> 
#> Residual standard error: 0.41648 on 207 degrees of freedom
#> Multiple R-squared:  0.22384,	Adjusted R-squared:  0.21259 
#> F-statistic: 19.899 on 3 and 207 DF,  p-value: 2.2574e-11
#> 
#> Alpha: -0.64248, standard error: 0.10411
#>     z-value: -6.1711, p-value: 6.7823e-10
#> LR test value: 48.392, p-value: 3.4905e-12
#> Implied rho: 0.474015 
#> 

system.time(obj2a <- lagmess(log(PRICE) ~ PATIO + log(AGE) + log(SQFT), data=baltimore, listw=lw,
 use_expm=TRUE))

#>    user  system elapsed 
#>   0.647   0.000   0.650 

summary(obj2a)

#> Matrix exponential spatial lag model:
#> (calculated with expm)
#> 
#> Call:
#> lagmess(formula = log(PRICE) ~ PATIO + log(AGE) + log(SQFT), 
#>     data = baltimore, listw = lw, use_expm = TRUE)
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -2.026015 -0.141654  0.050901  0.224067  1.072942 
#> 
#> Coefficients:
#>              Estimate Std. Error t value  Pr(>|t|)
#> (Intercept)  1.549087   0.214465  7.2230 9.553e-12
#> PATIO        0.259322   0.086871  2.9851  0.003176
#> log(AGE)    -0.148169   0.035244 -4.2041 3.899e-05
#> log(SQFT)    0.300203   0.071582  4.1939 4.066e-05
#> 
#> Residual standard error: 0.41648 on 207 degrees of freedom
#> Multiple R-squared:  0.22384,	Adjusted R-squared:  0.21259 
#> F-statistic: 19.899 on 3 and 207 DF,  p-value: 2.2574e-11
#> 
#> Alpha: -0.64248, standard error: 0.10411
#>     z-value: -6.1711, p-value: 6.7823e-10
#> LR test value: 48.392, p-value: 3.4905e-12
#> Implied rho: 0.474015 
#> 

obj3 <- lagsarlm(log(PRICE) ~ PATIO + log(AGE) + log(SQFT), data=baltimore, listw=lw)
summary(obj3)

#> 
#> Call:lagsarlm(formula = log(PRICE) ~ PATIO + log(AGE) + log(SQFT), 
#>     data = baltimore, listw = lw)
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.968546 -0.147936  0.042912  0.199676  1.076176 
#> 
#> Type: lag 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept)  1.257955   0.320454  3.9255 8.653e-05
#> PATIO        0.244879   0.083432  2.9351 0.0033348
#> log(AGE)    -0.131946   0.034502 -3.8243 0.0001311
#> log(SQFT)    0.278307   0.070246  3.9619 7.437e-05
#> 
#> Rho: 0.5574, LR test value: 52.786, p-value: 3.7192e-13
#> Asymptotic standard error: 0.072691
#>     z-value: 7.6681, p-value: 1.7542e-14
#> Wald statistic: 58.8, p-value: 1.7431e-14
#> 
#> Log likelihood: -110.3622 for lag model
#> ML residual variance (sigma squared): 0.15883, (sigma: 0.39853)
#> Number of observations: 211 
#> Number of parameters estimated: 6 
#> AIC: 232.72, (AIC for lm: 283.51)
#> LM test for residual autocorrelation
#> test value: 8.8042, p-value: 0.0030054
#> 

data(boston, package="spData")
lw <- nb2listw(boston.soi)
gp2 <- lagsarlm(log(CMEDV) ~ CRIM + ZN + INDUS + CHAS + I(NOX^2) + I(RM^2)
 +  AGE + log(DIS) + log(RAD) + TAX + PTRATIO + B + log(LSTAT),
 data=boston.c, lw, method="Matrix")
summary(gp2)

#> 
#> Call:lagsarlm(formula = log(CMEDV) ~ CRIM + ZN + INDUS + CHAS + I(NOX^2) + 
#>     I(RM^2) + AGE + log(DIS) + log(RAD) + TAX + PTRATIO + B + 
#>     log(LSTAT), data = boston.c, listw = lw, method = "Matrix")
#> 
#> Residuals:
#>        Min         1Q     Median         3Q        Max 
#> -0.5262308 -0.0749699 -0.0044237  0.0713409  0.7122121 
#> 
#> Type: lag 
#> Coefficients: (asymptotic standard errors) 
#>                Estimate  Std. Error  z value  Pr(>|z|)
#> (Intercept)  2.2796e+00  1.7495e-01  13.0302 < 2.2e-16
#> CRIM        -7.1045e-03  9.6236e-04  -7.3824 1.554e-13
#> ZN           3.7985e-04  3.8510e-04   0.9864 0.3239507
#> INDUS        1.2572e-03  1.7986e-03   0.6990 0.4845472
#> CHAS1        7.3677e-03  2.5416e-02   0.2899 0.7719059
#> I(NOX^2)    -2.6892e-01  8.8026e-02  -3.0550 0.0022508
#> I(RM^2)      6.7243e-03  1.0039e-03   6.6985 2.106e-11
#> AGE         -2.7682e-04  4.0062e-04  -0.6910 0.4895829
#> log(DIS)    -1.5830e-01  2.5554e-02  -6.1947 5.841e-10
#> log(RAD)     7.0689e-02  1.4616e-02   4.8363 1.323e-06
#> TAX         -3.6569e-04  9.3744e-05  -3.9009 9.582e-05
#> PTRATIO     -1.2011e-02  3.9599e-03  -3.0330 0.0024211
#> B            2.8432e-04  7.9402e-05   3.5807 0.0003427
#> log(LSTAT)  -2.3216e-01  2.0425e-02 -11.3663 < 2.2e-16
#> 
#> Rho: 0.48537, LR test value: 214.06, p-value: < 2.22e-16
#> Asymptotic standard error: 0.029426
#>     z-value: 16.494, p-value: < 2.22e-16
#> Wald statistic: 272.06, p-value: < 2.22e-16
#> 
#> Log likelihood: 264.0089 for lag model
#> ML residual variance (sigma squared): 0.019276, (sigma: 0.13884)
#> Number of observations: 506 
#> Number of parameters estimated: 16 
#> AIC: -496.02, (AIC for lm: -283.96)
#> LM test for residual autocorrelation
#> test value: 10.74, p-value: 0.0010486
#> 

gp2a <- lagmess(CMEDV ~ CRIM + ZN + INDUS + CHAS + I(NOX^2) + I(RM^2)
 +  AGE + log(DIS) + log(RAD) + TAX + PTRATIO + B + log(LSTAT),
 data=boston.c, lw)
summary(gp2a)

#> Matrix exponential spatial lag model:
#> 
#> Call:
#> lagmess(formula = CMEDV ~ CRIM + ZN + INDUS + CHAS + I(NOX^2) + 
#>     I(RM^2) + AGE + log(DIS) + log(RAD) + TAX + PTRATIO + B + 
#>     log(LSTAT), data = boston.c, listw = lw)
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -17.03755  -2.05386  -0.30295   1.67710  21.82120 
#> 
#> Coefficients:
#>               Estimate Std. Error  t value  Pr(>|t|)
#> (Intercept) 35.3206850  3.0128759  11.7232 < 2.2e-16
#> CRIM        -0.0986056  0.0242819  -4.0609 5.688e-05
#> ZN           0.0198500  0.0098638   2.0124 0.0447222
#> INDUS        0.0071211  0.0461104   0.1544 0.8773301
#> CHAS1        0.8158059  0.6477224   1.2595 0.2084473
#> I(NOX^2)    -9.2592911  2.2072427  -4.1950 3.238e-05
#> I(RM^2)      0.2434745  0.0256001   9.5107 < 2.2e-16
#> AGE         -0.0040683  0.0102667  -0.3963 0.6920816
#> log(DIS)    -5.3974116  0.6514323  -8.2855 1.125e-15
#> log(RAD)     1.7142905  0.3732772   4.5925 5.569e-06
#> TAX         -0.0087053  0.0023933  -3.6373 0.0003046
#> PTRATIO     -0.4118524  0.0977997  -4.2112 3.021e-05
#> B            0.0056141  0.0020116   2.7908 0.0054614
#> log(LSTAT)  -6.1484203  0.4878957 -12.6019 < 2.2e-16
#> 
#> Residual standard error: 3.5594 on 492 degrees of freedom
#> Multiple R-squared:  0.76221,	Adjusted R-squared:  0.75593 
#> F-statistic: 121.31 on 13 and 492 DF,  p-value: < 2.22e-16
#> 
#> Alpha: -0.41361, standard error: 0.038521
#>     z-value: -10.737, p-value: < 2.22e-16
#> LR test value: 121.4, p-value: < 2.22e-16
#> Implied rho: 0.3387434 
#>