Spatial simultaneous autoregressive lag model estimation

The lagsarlm function provides Maximum likelihood estimation of spatial simultaneous autoregressive lag and spatial Durbin (mixed) models of the form:

$$y = \rho W y + X \beta + \varepsilon$$

where \(\rho\) is found by optimize() first, and \(\beta\) and other parameters by generalized least squares subsequently (one-dimensional search using optim performs badly on some platforms). In the spatial Durbin (mixed) model, the spatially lagged independent variables are added to X. Note that interpretation of the fitted coefficients should use impact measures, because of the feedback loops induced by the data generation process for this model. With one of the sparse matrix methods, larger numbers of observations can be handled, but the interval= argument may need be set when the weights are not row-standardised.

The spBreg_lag function is an early-release version of the Matlab Spatial Econometrics Toolbox function sar_g.m, using drawing by inversion, and not accommodating heteroskedastic disturbances.

lagsarlm(formula, data = list(), listw,
        na.action, Durbin, type, method="eigen", quiet=NULL,
        zero.policy=NULL, interval=NULL, tol.solve=1.0e-10, trs=NULL,
        control=list())
spBreg_lag(formula, data = list(), listw, na.action, Durbin, type,
    zero.policy=NULL, control=list())

Arguments

formula
a symbolic description of the model to be fit. The details of model specification are given for lm()

data
an optional data frame containing the variables in the model. By default the variables are taken from the environment which the function is called.

listw
a listw object created for example by nb2listw

na.action
a function (default options("na.action")), can also be na.omit or na.exclude with consequences for residuals and fitted values - in these cases the weights list will be subsetted to remove NAs in the data. It may be necessary to set zero.policy to TRUE because this subsetting may create no-neighbour observations. Note that only weights lists created without using the glist argument to nb2listw may be subsetted.

Durbin
default FALSE (spatial lag model); if TRUE, full spatial Durbin model; if a formula object, the subset of explanatory variables to lag

type
(use the ‘Durbin=’ argument - retained for backwards compatibility only) default "lag", may be set to "mixed"; when "mixed", the lagged intercept is dropped for spatial weights style "W", that is row-standardised weights, but otherwise included; “Durbin” may be used instead of “mixed”

method
"eigen" (default) - the Jacobian is computed as the product of (1 - rho*eigenvalue) using eigenw, and "spam" or "Matrix_J" for strictly symmetric weights lists of styles "B" and "C", or made symmetric by similarity (Ord, 1975, Appendix C) if possible for styles "W" and "S", using code from the spam or Matrix packages to calculate the determinant; “Matrix” and “spam_update” provide updating Cholesky decomposition methods; "LU" provides an alternative sparse matrix decomposition approach. In addition, there are "Chebyshev" and Monte Carlo "MC" approximate log-determinant methods; the Smirnov/Anselin (2009) trace approximation is available as "moments". Three methods: "SE_classic", "SE_whichMin", and "SE_interp" are provided experimentally, the first to attempt to emulate the behaviour of Spatial Econometrics toolbox ML fitting functions. All use grids of log determinant values, and the latter two attempt to ameliorate some features of "SE_classic".

quiet
default NULL, use !verbose global option value; if FALSE, reports function values during optimization.

zero.policy
default NULL, use global option value; if TRUE assign zero to the lagged value of zones without neighbours, if FALSE (default) assign NA - causing lagsarlm() to terminate with an error

interval
default is NULL, search interval for autoregressive parameter

tol.solve
the tolerance for detecting linear dependencies in the columns of matrices to be inverted - passed to solve() (default=1.0e-10). This may be used if necessary to extract coefficient standard errors (for instance lowering to 1e-12), but errors in solve() may constitute indications of poorly scaled variables: if the variables have scales differing much from the autoregressive coefficient, the values in this matrix may be very different in scale, and inverting such a matrix is analytically possible by definition, but numerically unstable; rescaling the RHS variables alleviates this better than setting tol.solve to a very small value

trs
default NULL, if given, a vector of powered spatial weights matrix traces output by trW; when given, insert the asymptotic analytical values into the numerical Hessian instead of the approximated values; may be used to get around some problems raised when the numerical Hessian is poorly conditioned, generating NaNs in subsequent operations; the use of trs is recommended

control
list of extra control arguments - see section below

Details

The asymptotic standard error of \(\rho\) is only computed when method=eigen, because the full matrix operations involved would be costly for large n typically associated with the choice of method="spam" or "Matrix". The same applies to the coefficient covariance matrix. Taken as the asymptotic matrix from the literature, it is typically badly scaled, and with the elements involving \(\rho\) being very small, while other parts of the matrix can be very large (often many orders of magnitude in difference). It often happens that the tol.solve argument needs to be set to a smaller value than the default, or the RHS variables can be centred or reduced in range.

Versions of the package from 0.4-38 include numerical Hessian values where asymptotic standard errors are not available. This change has been introduced to permit the simulation of distributions for impact measures. The warnings made above with regard to variable scaling also apply in this case.

Note that the fitted() function for the output object assumes that the response variable may be reconstructed as the sum of the trend, the signal, and the noise (residuals). Since the values of the response variable are known, their spatial lags are used to calculate signal components (Cressie 1993, p. 564). This differs from other software, including GeoDa, which does not use knowledge of the response variable in making predictions for the fitting data.

Control arguments

tol.opt:

the desired accuracy of the optimization - passed to optimize() (default=square root of double precision machine tolerance, a larger root may be used needed, see help(boston) for an example)

fdHess:

default NULL, then set to (method != "eigen") internally; use fdHess to compute an approximate Hessian using finite differences when using sparse matrix methods; used to make a coefficient covariance matrix when the number of observations is large; may be turned off to save resources if need be

optimHess:

default FALSE, use fdHess from nlme, if TRUE, use optim to calculate Hessian at optimum

optimHessMethod:

default “optimHess”, may be “nlm” or one of the optim methods

compiled_sse:

default FALSE; logical value used in the log likelihood function to choose compiled code for computing SSE

Imult:

default 2; used for preparing the Cholesky decompositions for updating in the Jacobian function

super:

if NULL (default), set to FALSE to use a simplicial decomposition for the sparse Cholesky decomposition and method “Matrix_J”, set to as.logical(NA) for method “Matrix”, if TRUE, use a supernodal decomposition

cheb_q:

default 5; highest power of the approximating polynomial for the Chebyshev approximation

MC_p:

default 16; number of random variates

MC_m:

default 30; number of products of random variates matrix and spatial weights matrix

spamPivot:

default “MMD”, alternative “RCM”

in_coef

default 0.1, coefficient value for initial Cholesky decomposition in “spam_update”

type

default “MC”, used with method “moments”; alternatives “mult” and “moments”, for use if trs is missing, trW

correct

default TRUE, used with method “moments” to compute the Smirnov/Anselin correction term

trunc

default TRUE, used with method “moments” to truncate the Smirnov/Anselin correction term

SE_method

default “LU”, may be “MC”

nrho

default 200, as in SE toolbox; the size of the first stage lndet grid; it may be reduced to for example 40

interpn

default 2000, as in SE toolbox; the size of the second stage lndet grid

small_asy

default TRUE; if the method is not “eigen”, use asymmetric covariances rather than numerical Hessian ones if n <= small

small

default 1500; threshold number of observations for asymmetric covariances when the method is not “eigen”

SElndet

default NULL, may be used to pass a pre-computed SE toolbox style matrix of coefficients and their lndet values to the "SE_classic" and "SE_whichMin" methods

LU_order

default FALSE; used in “LU_prepermutate”, note warnings given for lu method

pre_eig

default NULL; may be used to pass a pre-computed vector of eigenvalues

OrdVsign

default 1; used to set the sign of the final component to negative if -1 (alpha times ((sigma squared) squared) in Ord (1975) equation B.1).

Extra Bayesian control arguments

ldet_method

default “SE_classic”; equivalent to the method argument in lagsarlm

interval

default c(-1, 1); used unmodified or set internally by jacobianSetup

ndraw

default 2500L; integer total number of draws

nomit

default 500L; integer total number of omitted burn-in draws

thin

default 1L; integer thinning proportion

verbose

default FALSE; inverse of quiet argument in lagsarlm

detval

default NULL; not yet in use, precomputed matrix of log determinants

prior

a list with the following components:

rhoMH

default FALSE; use Metropolis or griddy Gibbs

Tbeta

default NULL; values of the betas variance-covariance matrix, set to diag(k)*1e+12 if NULL

c_beta

default NULL; values of the betas set to 0 if NULL

rho

default 0.5; value of the autoregressive coefficient

sige

default 1; value of the residual variance

nu

default 0; informative Gamma(nu,d0) prior on sige

d0

default 0; informative Gamma(nu,d0) prior on sige

a1

default 1.01; parameter for beta(a1,a2) prior on rho

a2

default 1.01; parameter for beta(a1,a2) prior on rho

Value

A list object of class sarlm

type

"lag" or "mixed"

dvars

vector of length 2, numbers of columns in X and WX; if Durbin is given as a formula, the formula as attribute “f”, the indices of the included Wx as “inds”, and indices of added zero Wx coefficients as “zero_fill”

rho

simultaneous autoregressive lag coefficient

coefficients

GLS coefficient estimates

rest.se

asymptotic standard errors if ase=TRUE, otherwise approximate numeriacal Hessian-based values

LL

log likelihood value at computed optimum

s2

GLS residual variance

SSE

sum of squared GLS errors

parameters

number of parameters estimated

% \item{lm.model}{the \code{lm} object returned when estimating for \eqn{\rho=0}{rho=0}}
logLik_lm.model

Log likelihood of the linear model for \(\rho=0\)

AIC_lm.model

AIC of the linear model for \(\rho=0\)

method

the method used to calculate the Jacobian

call

the call used to create this object

residuals

GLS residuals

tarX

model matrix of the GLS model

tary

response of the GLS model

y

response of the linear model for \(\rho=0\)

X

model matrix of the linear model for \(\rho=0\)

% \item{lm.target}{the \code{lm} object returned for the GLS fit}
opt

object returned from numerical optimisation

fitted.values

Difference between residuals and response variable

se.fit

Not used yet

% \item{formula}{model formula}
ase

TRUE if method=eigen

rho.se

if ase=TRUE, the asymptotic standard error of \(\rho\), otherwise approximate numeriacal Hessian-based value

LMtest

if ase=TRUE, the Lagrange Multiplier test for the absence of spatial autocorrelation in the lag model residuals

resvar

the asymptotic coefficient covariance matrix for (s2, rho, B)

zero.policy

zero.policy for this model

aliased

the aliased explanatory variables (if any)

listw_style

the style of the spatial weights used

interval

the line search interval used to find \(\rho\)

fdHess

the numerical Hessian-based coefficient covariance matrix for (rho, B) if computed

optimHess

if TRUE and fdHess returned, optim used to calculate Hessian at optimum

insert

if TRUE and fdHess returned, the asymptotic analytical values are inserted into the numerical Hessian instead of the approximated values, and its size increased to include the first row/column for sigma2

LLNullLlm

Log-likelihood of the null linear model

timings

processing timings

f_calls

number of calls to the log likelihood function during optimization

hf_calls

number of calls to the log likelihood function during numerical Hessian computation

intern_classic

a data frame of detval matrix row choices used by the SE toolbox classic method

na.action

(possibly) named vector of excluded or omitted observations if non-default na.action argument used

The internal sar.lag.mixed.* functions return the value of the log likelihood function at \rhorho.
References

Cliff, A. D., Ord, J. K. 1981 Spatial processes, Pion; Ord, J. K. 1975 Estimation methods for models of spatial interaction, Journal of the American Statistical Association, 70, 120-126; Anselin, L. 1988 Spatial econometrics: methods and models. (Dordrecht: Kluwer); Anselin, L. 1995 SpaceStat, a software program for the analysis of spatial data, version 1.80. Regional Research Institute, West Virginia University, Morgantown, WV; Anselin L, Bera AK (1998) Spatial dependence in linear regression models with an introduction to spatial econometrics. In: Ullah A, Giles DEA (eds) Handbook of applied economic statistics. Marcel Dekker, New York, pp. 237-289; Cressie, N. A. C. 1993 Statistics for spatial data, Wiley, New York; LeSage J and RK Pace (2009) Introduction to Spatial Econometrics. CRC Press, Boca Raton.

Roger Bivand, Gianfranco Piras (2015). Comparing Implementations of Estimation Methods for Spatial Econometrics. Journal of Statistical Software, 63(18), 1-36. https://www.jstatsoft.org/v63/i18/.

Bivand, R. S., Hauke, J., and Kossowski, T. (2013). Computing the Jacobian in Gaussian spatial autoregressive models: An illustrated comparison of available methods. Geographical Analysis, 45(2), 150-179.

See also

Examples

data(oldcol)
listw <- nb2listw(COL.nb, style="W")
ev <- eigenw(listw)
W <- as(listw, "CsparseMatrix")
trMatc <- trW(W, type="mult")
COL.lag.eig <- lagsarlm(CRIME ~ INC + HOVAL, data=COL.OLD, listw=listw,
 method="eigen", quiet=FALSE, control=list(pre_eig=ev, OrdVsign=1))

#> 
#> Spatial lag model
#> Jacobian calculated using neighbourhood matrix eigenvalues
#> 
#> rho:	 -0.5674437 	function value:	 -202.2909 
#> rho:	 0.03126655 	function value:	 -186.749 
#> rho:	 0.4012898 	function value:	 -182.419 
#> rho:	 0.6138418 	function value:	 -183.5636 
#> rho:	 0.4157662 	function value:	 -182.398 
#> rho:	 0.4295565 	function value:	 -182.3905 
#> rho:	 0.4311288 	function value:	 -182.3904 
#> rho:	 0.4310273 	function value:	 -182.3904 
#> rho:	 0.4310232 	function value:	 -182.3904 
#> rho:	 0.4310232 	function value:	 -182.3904 
#> rho:	 0.4310232 	function value:	 -182.3904 
#> rho:	 0.4310232 	function value:	 -182.3904 

summary(COL.lag.eig, correlation=TRUE)

#> 
#> Call:lagsarlm(formula = CRIME ~ INC + HOVAL, data = COL.OLD, listw = listw, 
#>     method = "eigen", quiet = FALSE, control = list(pre_eig = ev, 
#>         OrdVsign = 1))
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -37.68585  -5.35636   0.05421   6.02013  23.20555 
#> 
#> Type: lag 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) 45.079251   7.177347  6.2808 3.369e-10
#> INC         -1.031616   0.305143 -3.3808 0.0007229
#> HOVAL       -0.265926   0.088499 -3.0049 0.0026570
#> 
#> Rho: 0.43102, LR test value: 9.9736, p-value: 0.001588
#> Asymptotic standard error: 0.11768
#>     z-value: 3.6626, p-value: 0.00024962
#> Wald statistic: 13.415, p-value: 0.00024962
#> 
#> Log likelihood: -182.3904 for lag model
#> ML residual variance (sigma squared): 95.494, (sigma: 9.7721)
#> Number of observations: 49 
#> Number of parameters estimated: 5 
#> AIC: 374.78, (AIC for lm: 382.75)
#> LM test for residual autocorrelation
#> test value: 0.31955, p-value: 0.57188
#> 
#>  Correlation of coefficients 
#>             sigma rho   (Intercept) INC  
#> rho         -0.14                        
#> (Intercept)  0.12 -0.83                  
#> INC         -0.05  0.35 -0.61            
#> HOVAL       -0.01  0.08 -0.25       -0.44
#> 

COL.lag.eig$fdHess

#> [1] FALSE

COL.lag.eig$resvar

#>                    sigma           rho (Intercept)         INC         HOVAL
#> sigma       379.77510361 -0.3236306318  16.3015079 -0.29590801 -0.0202478459
#> rho          -0.32363063  0.0138487533  -0.6975717  0.01266245  0.0008664428
#> (Intercept)  16.30150788 -0.6975716738  51.5143034 -1.32602704 -0.1616379856
#> INC          -0.29590801  0.0126624511  -1.3260270  0.09311223 -0.0117959715
#> HOVAL        -0.02024785  0.0008664428  -0.1616380 -0.01179597  0.0078320057

# using the apparent sign in Ord (1975, equation B.1) 
COL.lag.eigb <- lagsarlm(CRIME ~ INC + HOVAL, data=COL.OLD, listw=listw,
 method="eigen", control=list(pre_eig=ev, OrdVsign=-1))
summary(COL.lag.eigb)

#> 
#> Call:lagsarlm(formula = CRIME ~ INC + HOVAL, data = COL.OLD, listw = listw, 
#>     method = "eigen", control = list(pre_eig = ev, OrdVsign = -1))
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -37.68585  -5.35636   0.05421   6.02013  23.20555 
#> 
#> Type: lag 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) 45.079251   9.617835  4.6870 2.772e-06
#> INC         -1.031616   0.326524 -3.1594  0.001581
#> HOVAL       -0.265926   0.088855 -2.9928  0.002764
#> 
#> Rho: 0.43102, LR test value: 9.9736, p-value: 0.001588
#> Asymptotic standard error: 0.17322
#>     z-value: 2.4884, p-value: 0.012833
#> Wald statistic: 6.1919, p-value: 0.012833
#> 
#> Log likelihood: -182.3904 for lag model
#> ML residual variance (sigma squared): 95.494, (sigma: 9.7721)
#> Number of observations: 49 
#> Number of parameters estimated: 5 
#> AIC: 374.78, (AIC for lm: 382.75)
#> LM test for residual autocorrelation
#> test value: -0.93825, p-value: 1
#> 

COL.lag.eigb$fdHess

#> [1] FALSE

COL.lag.eigb$resvar

#>                    sigma          rho (Intercept)         INC        HOVAL
#> sigma       388.59742931 -0.701154266  35.3176451 -0.64109248 -0.043867490
#> rho          -0.70115427  0.030003688  -1.5113074  0.02743353  0.001877171
#> (Intercept)  35.31764510 -1.511307358  92.5027556 -2.07005707 -0.212549096
#> INC          -0.64109248  0.027433533  -2.0700571  0.10661800 -0.010871824
#> HOVAL        -0.04386749  0.001877171  -0.2125491 -0.01087182  0.007895242

# force numerical Hessian
COL.lag.eig1 <- lagsarlm(CRIME ~ INC + HOVAL, data=COL.OLD,
 listw=listw, method="Matrix", control=list(small=25))
summary(COL.lag.eig1)

#> 
#> Call:lagsarlm(formula = CRIME ~ INC + HOVAL, data = COL.OLD, listw = listw, 
#>     method = "Matrix", control = list(small = 25))
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -37.68585  -5.35636   0.05421   6.02013  23.20555 
#> 
#> Type: lag 
#> Coefficients: (numerical Hessian approximate standard errors) 
#>             Estimate Std. Error z value Pr(>|z|)
#> (Intercept) 45.07925    7.87089  5.7273 1.02e-08
#> INC         -1.03162    0.32842 -3.1411 0.001683
#> HOVAL       -0.26593    0.08823 -3.0140 0.002578
#> 
#> Rho: 0.43102, LR test value: 9.9736, p-value: 0.001588
#> Approximate (numerical Hessian) standard error: 0.12362
#>     z-value: 3.4868, p-value: 0.00048892
#> Wald statistic: 12.157, p-value: 0.00048892
#> 
#> Log likelihood: -182.3904 for lag model
#> ML residual variance (sigma squared): 95.494, (sigma: 9.7721)
#> Number of observations: 49 
#> Number of parameters estimated: 5 
#> AIC: 374.78, (AIC for lm: 382.75)
#> 

COL.lag.eig1$fdHess

#>                       rho (Intercept)         INC         HOVAL
#> rho          0.0152812237  -0.8345267  0.02005558  0.0002830225
#> (Intercept) -0.8345266779  61.9509427 -1.78340176 -0.1334585288
#> INC          0.0200555771  -1.7834018  0.10786140 -0.0122204417
#> HOVAL        0.0002830225  -0.1334585 -0.01222044  0.0077846055

# force LeSage & Pace (2008, p. 57) approximation 
COL.lag.eig1a <- lagsarlm(CRIME ~ INC + HOVAL, data=COL.OLD,
 listw=listw, method="Matrix", control=list(small=25), trs=trMatc)
summary(COL.lag.eig1a)

#> 
#> Call:lagsarlm(formula = CRIME ~ INC + HOVAL, data = COL.OLD, listw = listw, 
#>     method = "Matrix", trs = trMatc, control = list(small = 25))
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -37.68585  -5.35636   0.05421   6.02013  23.20555 
#> 
#> Type: lag 
#> Coefficients: (numerical Hessian approximate standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) 45.079249   7.937572  5.6792 1.353e-08
#> INC         -1.031616   0.329337 -3.1324  0.001734
#> HOVAL       -0.265926   0.088222 -3.0143  0.002576
#> 
#> Rho: 0.43102, LR test value: 9.9736, p-value: 0.001588
#> Approximate (numerical Hessian) standard error: 0.12503
#>     z-value: 3.4473, p-value: 0.00056624
#> Wald statistic: 11.884, p-value: 0.00056624
#> 
#> Log likelihood: -182.3904 for lag model
#> ML residual variance (sigma squared): 95.494, (sigma: 9.7721)
#> Number of observations: 49 
#> Number of parameters estimated: 5 
#> AIC: 374.78, (AIC for lm: 382.75)
#> 

COL.lag.eig1a$fdHess

#>                    sigma2           rho (Intercept)         INC         HOVAL
#> sigma2      380.749548407 -0.3653290756  19.9519208 -0.47947507 -0.0067851124
#> rho          -0.365329076  0.0156331058  -0.8537795  0.02051762  0.0002903475
#> (Intercept)  19.951920812 -0.8537795385  63.0050547 -1.80875071 -0.1338515483
#> INC          -0.479475072  0.0205176238  -1.8087507  0.10846276 -0.0122071279
#> HOVAL        -0.006785112  0.0002903475  -0.1338515 -0.01220713  0.0077831895

COL.lag.eig$resvar[2,2]

#> [1] 0.01384875

# using the apparent sign in Ord (1975, equation B.1) 
COL.lag.eigb$resvar[2,2]

#> [1] 0.03000369

# force numerical Hessian
COL.lag.eig1$fdHess[1,1]

#> [1] 0.01528122

# force LeSage & Pace (2008, p. 57) approximation 
COL.lag.eig1a$fdHess[2,2]

#> [1] 0.01563311

system.time(COL.lag.M <- lagsarlm(CRIME ~ INC + HOVAL, data=COL.OLD,
 nb2listw(COL.nb), method="Matrix", quiet=FALSE))

#> 
#> Spatial lag model
#> Jacobian calculated using sparse matrix Cholesky decomposition
#> rho:	 -0.2364499 	function value:	 -192.9523 
#> rho:	 0.2354499 	function value:	 -183.542 
#> rho:	 0.5271001 	function value:	 -182.7039 
#> rho:	 0.4455543 	function value:	 -182.3974 
#> rho:	 0.4267907 	function value:	 -182.391 
#> rho:	 0.4311986 	function value:	 -182.3904 
#> rho:	 0.4310114 	function value:	 -182.3904 
#> rho:	 0.4310231 	function value:	 -182.3904 
#> rho:	 0.4310232 	function value:	 -182.3904 
#> rho:	 0.4310232 	function value:	 -182.3904 
#> rho:	 0.4310232 	function value:	 -182.3904 
#> rho:	 0.4310232 	function value:	 -182.3904 
#> Computing eigenvalues ...
#> 

#>    user  system elapsed 
#>   0.197   0.000   0.199 

summary(COL.lag.M)

#> 
#> Call:
#> lagsarlm(formula = CRIME ~ INC + HOVAL, data = COL.OLD, listw = nb2listw(COL.nb), 
#>     method = "Matrix", quiet = FALSE)
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -37.68585  -5.35636   0.05421   6.02013  23.20555 
#> 
#> Type: lag 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) 45.079249   7.177346  6.2808 3.369e-10
#> INC         -1.031616   0.305143 -3.3808 0.0007229
#> HOVAL       -0.265926   0.088499 -3.0049 0.0026570
#> 
#> Rho: 0.43102, LR test value: 9.9736, p-value: 0.001588
#> Asymptotic standard error: 0.11768
#>     z-value: 3.6626, p-value: 0.00024962
#> Wald statistic: 13.415, p-value: 0.00024962
#> 
#> Log likelihood: -182.3904 for lag model
#> ML residual variance (sigma squared): 95.494, (sigma: 9.7721)
#> Number of observations: 49 
#> Number of parameters estimated: 5 
#> AIC: 374.78, (AIC for lm: 382.75)
#> LM test for residual autocorrelation
#> test value: 0.31954, p-value: 0.57188
#> 

impacts(COL.lag.M, listw=nb2listw(COL.nb))

#> Impact measures (lag, exact):
#>           Direct   Indirect      Total
#> INC   -1.0860220 -0.7270848 -1.8131068
#> HOVAL -0.2799509 -0.1874254 -0.4673763

system.time(COL.lag.sp <- lagsarlm(CRIME ~ INC + HOVAL, data=COL.OLD,
 nb2listw(COL.nb), method="spam", quiet=FALSE))

#> 
#> Spatial lag model
#> Jacobian calculated using sparse matrix Cholesky decomposition
#> rho:	 -0.2364499 	function value:	 -192.9523 
#> rho:	 0.2354499 	function value:	 -183.542 
#> rho:	 0.5271001 	function value:	 -182.7039 
#> rho:	 0.4455543 	function value:	 -182.3974 
#> rho:	 0.4267907 	function value:	 -182.391 
#> rho:	 0.4311986 	function value:	 -182.3904 
#> rho:	 0.4310114 	function value:	 -182.3904 
#> rho:	 0.4310231 	function value:	 -182.3904 
#> rho:	 0.4310232 	function value:	 -182.3904 
#> rho:	 0.4310232 	function value:	 -182.3904 
#> rho:	 0.4310232 	function value:	 -182.3904 
#> rho:	 0.4310232 	function value:	 -182.3904 
#> Computing eigenvalues ...
#> 

#>    user  system elapsed 
#>   0.214   0.000   0.219 

summary(COL.lag.sp)

#> 
#> Call:
#> lagsarlm(formula = CRIME ~ INC + HOVAL, data = COL.OLD, listw = nb2listw(COL.nb), 
#>     method = "spam", quiet = FALSE)
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -37.68585  -5.35636   0.05421   6.02013  23.20555 
#> 
#> Type: lag 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) 45.079250   7.177347  6.2808 3.369e-10
#> INC         -1.031616   0.305143 -3.3808 0.0007229
#> HOVAL       -0.265926   0.088499 -3.0049 0.0026570
#> 
#> Rho: 0.43102, LR test value: 9.9736, p-value: 0.001588
#> Asymptotic standard error: 0.11768
#>     z-value: 3.6626, p-value: 0.00024962
#> Wald statistic: 13.415, p-value: 0.00024962
#> 
#> Log likelihood: -182.3904 for lag model
#> ML residual variance (sigma squared): 95.494, (sigma: 9.7721)
#> Number of observations: 49 
#> Number of parameters estimated: 5 
#> AIC: 374.78, (AIC for lm: 382.75)
#> LM test for residual autocorrelation
#> test value: 0.31954, p-value: 0.57188
#> 

COL.lag.B <- lagsarlm(CRIME ~ INC + HOVAL, data=COL.OLD,
 nb2listw(COL.nb, style="B"))
summary(COL.lag.B)

#> 
#> Call:
#> lagsarlm(formula = CRIME ~ INC + HOVAL, data = COL.OLD, listw = nb2listw(COL.nb, 
#>     style = "B"))
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -33.6975  -4.7996  -1.4808   5.3648  25.8922 
#> 
#> Type: lag 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) 52.403878   5.961310  8.7907 < 2.2e-16
#> INC         -1.175261   0.300473 -3.9114 9.178e-05
#> HOVAL       -0.252680   0.087613 -2.8840  0.003926
#> 
#> Rho: 0.051981, LR test value: 12.764, p-value: 0.00035336
#> Asymptotic standard error: 0.014428
#>     z-value: 3.6028, p-value: 0.00031482
#> Wald statistic: 12.98, p-value: 0.00031482
#> 
#> Log likelihood: -180.9953 for lag model
#> ML residual variance (sigma squared): 93.29, (sigma: 9.6587)
#> Number of observations: 49 
#> Number of parameters estimated: 5 
#> AIC: 371.99, (AIC for lm: 382.75)
#> LM test for residual autocorrelation
#> test value: 0.0028047, p-value: 0.95776
#> 

COL.mixed.B <- lagsarlm(CRIME ~ INC + HOVAL, data=COL.OLD,
 nb2listw(COL.nb, style="B"), type="mixed", tol.solve=1e-9,
 control=list(pre_eig=ev))
summary(COL.mixed.B)

#> 
#> Call:
#> lagsarlm(formula = CRIME ~ INC + HOVAL, data = COL.OLD, listw = nb2listw(COL.nb, 
#>     style = "B"), type = "mixed", tol.solve = 1e-09, control = list(pre_eig = ev))
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -34.8460  -4.2057  -0.1195   4.6525  21.6112 
#> 
#> Type: mixed 
#> Coefficients: (asymptotic standard errors) 
#>                    Estimate  Std. Error z value  Pr(>|z|)
#> (Intercept)      5.4215e+01  5.6639e+00  9.5719 < 2.2e-16
#> INC             -8.2386e-01  3.1643e-01 -2.6036 0.0092262
#> HOVAL           -3.0085e-01  8.5629e-02 -3.5134 0.0004424
#> lag.(Intercept) -7.5493e+00  1.7307e+00 -4.3620 1.289e-05
#> lag.INC          2.1544e-05  1.2216e-01  0.0002 0.9998593
#> lag.HOVAL        7.2458e-02  3.9007e-02  1.8576 0.0632281
#> 
#> Rho: 0.15212, LR test value: 7.435, p-value: 0.0063967
#> Asymptotic standard error: 0.015565
#>     z-value: 9.7735, p-value: < 2.22e-16
#> Wald statistic: 95.522, p-value: < 2.22e-16
#> 
#> Log likelihood: -177.7722 for mixed model
#> ML residual variance (sigma squared): 82.502, (sigma: 9.0831)
#> Number of observations: 49 
#> Number of parameters estimated: 8 
#> AIC: 371.54, (AIC for lm: 376.98)
#> LM test for residual autocorrelation
#> test value: -26.79, p-value: 1
#> 

COL.mixed.W <- lagsarlm(CRIME ~ INC + HOVAL, data=COL.OLD,
 listw, type="mixed",
 control=list(pre_eig=ev))
summary(COL.mixed.W)

#> 
#> Call:lagsarlm(formula = CRIME ~ INC + HOVAL, data = COL.OLD, listw = listw, 
#>     type = "mixed", control = list(pre_eig = ev))
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -37.47829  -6.46731  -0.33835   6.05200  22.62969 
#> 
#> Type: mixed 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) 42.822414  12.667204  3.3806 0.0007233
#> INC         -0.914223   0.331094 -2.7612 0.0057586
#> HOVAL       -0.293738   0.089212 -3.2926 0.0009927
#> lag.INC     -0.520284   0.565129 -0.9206 0.3572355
#> lag.HOVAL    0.245640   0.178917  1.3729 0.1697756
#> 
#> Rho: 0.42634, LR test value: 5.3693, p-value: 0.020494
#> Asymptotic standard error: 0.15623
#>     z-value: 2.7288, p-value: 0.0063561
#> Wald statistic: 7.4465, p-value: 0.0063561
#> 
#> Log likelihood: -181.3935 for mixed model
#> ML residual variance (sigma squared): 91.791, (sigma: 9.5808)
#> Number of observations: 49 
#> Number of parameters estimated: 7 
#> AIC: 376.79, (AIC for lm: 380.16)
#> LM test for residual autocorrelation
#> test value: 0.28919, p-value: 0.59074
#> 

COL.mixed.D00 <- lagsarlm(CRIME ~ INC + HOVAL, data=COL.OLD,
 listw, Durbin=TRUE,
 control=list(pre_eig=ev))
summary(COL.mixed.D00)

#> 
#> Call:lagsarlm(formula = CRIME ~ INC + HOVAL, data = COL.OLD, listw = listw, 
#>     Durbin = TRUE, control = list(pre_eig = ev))
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -37.47829  -6.46731  -0.33835   6.05200  22.62969 
#> 
#> Type: mixed 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) 42.822414  12.667204  3.3806 0.0007233
#> INC         -0.914223   0.331094 -2.7612 0.0057586
#> HOVAL       -0.293738   0.089212 -3.2926 0.0009927
#> lag.INC     -0.520284   0.565129 -0.9206 0.3572355
#> lag.HOVAL    0.245640   0.178917  1.3729 0.1697756
#> 
#> Rho: 0.42634, LR test value: 5.3693, p-value: 0.020494
#> Asymptotic standard error: 0.15623
#>     z-value: 2.7288, p-value: 0.0063561
#> Wald statistic: 7.4465, p-value: 0.0063561
#> 
#> Log likelihood: -181.3935 for mixed model
#> ML residual variance (sigma squared): 91.791, (sigma: 9.5808)
#> Number of observations: 49 
#> Number of parameters estimated: 7 
#> AIC: 376.79, (AIC for lm: 380.16)
#> LM test for residual autocorrelation
#> test value: 0.28919, p-value: 0.59074
#> 

COL.mixed.D01 <- lagsarlm(CRIME ~ INC + HOVAL, data=COL.OLD,
 listw, Durbin=FALSE,
 control=list(pre_eig=ev))
summary(COL.mixed.D01)

#> 
#> Call:lagsarlm(formula = CRIME ~ INC + HOVAL, data = COL.OLD, listw = listw, 
#>     Durbin = FALSE, control = list(pre_eig = ev))
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -37.68585  -5.35636   0.05421   6.02013  23.20555 
#> 
#> Type: lag 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) 45.079251   7.177347  6.2808 3.369e-10
#> INC         -1.031616   0.305143 -3.3808 0.0007229
#> HOVAL       -0.265926   0.088499 -3.0049 0.0026570
#> 
#> Rho: 0.43102, LR test value: 9.9736, p-value: 0.001588
#> Asymptotic standard error: 0.11768
#>     z-value: 3.6626, p-value: 0.00024962
#> Wald statistic: 13.415, p-value: 0.00024962
#> 
#> Log likelihood: -182.3904 for lag model
#> ML residual variance (sigma squared): 95.494, (sigma: 9.7721)
#> Number of observations: 49 
#> Number of parameters estimated: 5 
#> AIC: 374.78, (AIC for lm: 382.75)
#> LM test for residual autocorrelation
#> test value: 0.31955, p-value: 0.57188
#> 

COL.mixed.D1 <- lagsarlm(CRIME ~ INC + HOVAL, data=COL.OLD,
 listw, Durbin= ~ INC + HOVAL,
 control=list(pre_eig=ev))
summary(COL.mixed.D1)

#> 
#> Call:lagsarlm(formula = CRIME ~ INC + HOVAL, data = COL.OLD, listw = listw, 
#>     Durbin = ~INC + HOVAL, control = list(pre_eig = ev))
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -37.47829  -6.46731  -0.33835   6.05200  22.62969 
#> 
#> Type: mixed 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) 42.822414  12.667204  3.3806 0.0007233
#> INC         -0.914223   0.331094 -2.7612 0.0057586
#> HOVAL       -0.293738   0.089212 -3.2926 0.0009927
#> lag.INC     -0.520284   0.565129 -0.9206 0.3572355
#> lag.HOVAL    0.245640   0.178917  1.3729 0.1697756
#> 
#> Rho: 0.42634, LR test value: 5.3693, p-value: 0.020494
#> Asymptotic standard error: 0.15623
#>     z-value: 2.7288, p-value: 0.0063561
#> Wald statistic: 7.4465, p-value: 0.0063561
#> 
#> Log likelihood: -181.3935 for mixed model
#> ML residual variance (sigma squared): 91.791, (sigma: 9.5808)
#> Number of observations: 49 
#> Number of parameters estimated: 7 
#> AIC: 376.79, (AIC for lm: 380.16)
#> LM test for residual autocorrelation
#> test value: 0.28919, p-value: 0.59074
#> 

f <- CRIME ~ INC + HOVAL
COL.mixed.D2 <- lagsarlm(f, data=COL.OLD, listw,
 Durbin=as.formula(delete.response(terms(f))),
 control=list(pre_eig=ev))
summary(COL.mixed.D2)

#> 
#> Call:
#> lagsarlm(formula = f, data = COL.OLD, listw = listw, Durbin = as.formula(delete.response(terms(f))), 
#>     control = list(pre_eig = ev))
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -37.47829  -6.46731  -0.33835   6.05200  22.62969 
#> 
#> Type: mixed 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) 42.822414  12.667204  3.3806 0.0007233
#> INC         -0.914223   0.331094 -2.7612 0.0057586
#> HOVAL       -0.293738   0.089212 -3.2926 0.0009927
#> lag.INC     -0.520284   0.565129 -0.9206 0.3572355
#> lag.HOVAL    0.245640   0.178917  1.3729 0.1697756
#> 
#> Rho: 0.42634, LR test value: 5.3693, p-value: 0.020494
#> Asymptotic standard error: 0.15623
#>     z-value: 2.7288, p-value: 0.0063561
#> Wald statistic: 7.4465, p-value: 0.0063561
#> 
#> Log likelihood: -181.3935 for mixed model
#> ML residual variance (sigma squared): 91.791, (sigma: 9.5808)
#> Number of observations: 49 
#> Number of parameters estimated: 7 
#> AIC: 376.79, (AIC for lm: 380.16)
#> LM test for residual autocorrelation
#> test value: 0.28919, p-value: 0.59074
#> 

COL.mixed.D1a <- lagsarlm(CRIME ~ INC + HOVAL, data=COL.OLD,
 listw, Durbin= ~ INC,
 control=list(pre_eig=ev))
summary(COL.mixed.D1a)

#> 
#> Call:lagsarlm(formula = CRIME ~ INC + HOVAL, data = COL.OLD, listw = listw, 
#>     Durbin = ~INC, control = list(pre_eig = ev))
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -37.82800  -5.85207   0.12047   6.00137  23.19963 
#> 
#> Type: mixed 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) 48.814687  12.198232  4.0018 6.287e-05
#> INC         -1.006620   0.330641 -3.0445  0.002331
#> HOVAL       -0.265514   0.088768 -2.9911  0.002780
#> lag.INC     -0.186684   0.530550 -0.3519  0.724936
#> 
#> Rho: 0.39229, LR test value: 4.5007, p-value: 0.033881
#> Asymptotic standard error: 0.1561
#>     z-value: 2.513, p-value: 0.011971
#> Wald statistic: 6.3151, p-value: 0.011971
#> 
#> Log likelihood: -182.3328 for mixed model
#> ML residual variance (sigma squared): 96.122, (sigma: 9.8042)
#> Number of observations: 49 
#> Number of parameters estimated: 6 
#> AIC: 376.67, (AIC for lm: 379.17)
#> LM test for residual autocorrelation
#> test value: 2.6134, p-value: 0.10596
#> 

try(COL.mixed.D1 <- lagsarlm(CRIME ~ INC + HOVAL, data=COL.OLD,
 listw, Durbin= ~ inc + HOVAL,
 control=list(pre_eig=ev)))
try(COL.mixed.D1 <- lagsarlm(CRIME ~ INC + HOVAL, data=COL.OLD,
 listw, Durbin= ~ DISCBD + HOVAL,
 control=list(pre_eig=ev)))
NA.COL.OLD <- COL.OLD
NA.COL.OLD$CRIME[20:25] <- NA
COL.lag.NA <- lagsarlm(CRIME ~ INC + HOVAL, data=NA.COL.OLD,
 nb2listw(COL.nb), na.action=na.exclude,
 control=list(tol.opt=.Machine$double.eps^0.4))
COL.lag.NA$na.action

#> 1020 1021 1022 1023 1024 1025 
#>   20   21   22   23   24   25 
#> attr(,"class")
#> [1] "exclude"

COL.lag.NA

#> 
#> Call:
#> lagsarlm(formula = CRIME ~ INC + HOVAL, data = NA.COL.OLD, listw = nb2listw(COL.nb), 
#>     na.action = na.exclude, control = list(tol.opt = .Machine$double.eps^0.4))
#> Type: lag 
#> 
#> Coefficients:
#>         rho (Intercept)         INC       HOVAL 
#>   0.4537820  43.1054975  -0.9267352  -0.2715541 
#> 
#> Log likelihood: -160.8867 

resid(COL.lag.NA)

#>        1001        1002        1003        1004        1005        1006 
#>  -4.4352945 -13.2750948  -2.9233555 -37.3067542   1.3568011  -3.8828027 
#>        1007        1008        1009        1010        1011        1012 
#>   5.9980436 -12.8113318  -4.0482558  18.1813323   5.9714203   1.0035599 
#>        1013        1014        1015        1016        1017        1018 
#>  -1.7490666  -1.7490651   5.8456328  10.1074158  -4.3706895   3.3713771 
#>        1019        1020        1021        1022        1023        1024 
#>  -4.0823022          NA          NA          NA          NA          NA 
#>        1025        1026        1027        1028        1029        1030 
#>          NA  -6.0553296 -11.5813311  -8.0011389  -1.9047478   3.9744243 
#>        1031        1032        1033        1034        1035        1036 
#>   4.8148614   6.0673483  10.2059847  22.7419913  -2.0830998   0.1422777 
#>        1037        1038        1039        1040        1041        1042 
#>   8.6436388   8.8150864   6.4974685  15.4121789   9.4182148   5.6427603 
#>        1043        1044        1045        1046        1047        1048 
#>  -7.5727123  -7.5983733  -9.7792620 -10.0870764  -3.1242393   3.4921796 
#>        1049 
#>   0.7173251 

data(boston, package="spData")
gp2mM <- lagsarlm(log(CMEDV) ~ CRIM + ZN + INDUS + CHAS + I(NOX^2) +
I(RM^2) +  AGE + log(DIS) + log(RAD) + TAX + PTRATIO + B + log(LSTAT),
data=boston.c, nb2listw(boston.soi), type="mixed", method="Matrix")
summary(gp2mM)

#> 
#> Call:lagsarlm(formula = log(CMEDV) ~ CRIM + ZN + INDUS + CHAS + I(NOX^2) + 
#>     I(RM^2) + AGE + log(DIS) + log(RAD) + TAX + PTRATIO + B + 
#>     log(LSTAT), data = boston.c, listw = nb2listw(boston.soi), 
#>     type = "mixed", method = "Matrix")
#> 
#> Residuals:
#>        Min         1Q     Median         3Q        Max 
#> -0.6316833 -0.0629790 -0.0090776  0.0682421  0.6991072 
#> 
#> Type: mixed 
#> Coefficients: (asymptotic standard errors) 
#>                   Estimate  Std. Error  z value  Pr(>|z|)
#> (Intercept)     1.89816218  0.22759605   8.3400 < 2.2e-16
#> CRIM           -0.00571021  0.00093504  -6.1069 1.016e-09
#> ZN              0.00069091  0.00051869   1.3320  0.182851
#> INDUS          -0.00111343  0.00307354  -0.3623  0.717155
#> CHAS1          -0.04163225  0.02739364  -1.5198  0.128567
#> I(NOX^2)       -0.01034950  0.19360214  -0.0535  0.957367
#> I(RM^2)         0.00794979  0.00102063   7.7891 6.661e-15
#> AGE            -0.00128789  0.00048920  -2.6326  0.008473
#> log(DIS)       -0.12404108  0.09510940  -1.3042  0.192168
#> log(RAD)        0.05863502  0.02257078   2.5978  0.009382
#> TAX            -0.00049084  0.00012145  -4.0416 5.308e-05
#> PTRATIO        -0.01319853  0.00595352  -2.2169  0.026628
#> B               0.00056383  0.00011089   5.0847 3.682e-07
#> log(LSTAT)     -0.24724454  0.02262033 -10.9302 < 2.2e-16
#> lag.CRIM       -0.00464215  0.00172935  -2.6843  0.007267
#> lag.ZN         -0.00037937  0.00070584  -0.5375  0.590940
#> lag.INDUS       0.00025064  0.00385901   0.0649  0.948215
#> lag.CHAS1       0.12518252  0.04071559   3.0746  0.002108
#> lag.I(NOX^2)   -0.38640401  0.22157523  -1.7439  0.081177
#> lag.I(RM^2)    -0.00451252  0.00153180  -2.9459  0.003220
#> lag.AGE         0.00149678  0.00068418   2.1877  0.028693
#> lag.log(DIS)   -0.00453785  0.10046478  -0.0452  0.963973
#> lag.log(RAD)   -0.00940702  0.03104930  -0.3030  0.761912
#> lag.TAX         0.00041083  0.00017867   2.2994  0.021481
#> lag.PTRATIO     0.00060355  0.00788837   0.0765  0.939012
#> lag.B          -0.00050781  0.00014155  -3.5874  0.000334
#> lag.log(LSTAT)  0.09846781  0.03399423   2.8966  0.003772
#> 
#> Rho: 0.59578, LR test value: 181.68, p-value: < 2.22e-16
#> Asymptotic standard error: 0.038445
#>     z-value: 15.497, p-value: < 2.22e-16
#> Wald statistic: 240.16, p-value: < 2.22e-16
#> 
#> Log likelihood: 300.6131 for mixed model
#> ML residual variance (sigma squared): 0.016011, (sigma: 0.12654)
#> Number of observations: 506 
#> Number of parameters estimated: 29 
#> AIC: -543.23, (AIC for lm: -363.55)
#> LM test for residual autocorrelation
#> test value: 29.772, p-value: 4.8604e-08
#> 

W <- as(nb2listw(boston.soi), "CsparseMatrix")
trMatb <- trW(W, type="mult")
gp2mMi <- lagsarlm(log(CMEDV) ~ CRIM + ZN + INDUS + CHAS + I(NOX^2) +
I(RM^2) +  AGE + log(DIS) + log(RAD) + TAX + PTRATIO + B + log(LSTAT),
data=boston.c, nb2listw(boston.soi), type="mixed", method="Matrix",
trs=trMatb)
summary(gp2mMi)

#> 
#> Call:lagsarlm(formula = log(CMEDV) ~ CRIM + ZN + INDUS + CHAS + I(NOX^2) + 
#>     I(RM^2) + AGE + log(DIS) + log(RAD) + TAX + PTRATIO + B + 
#>     log(LSTAT), data = boston.c, listw = nb2listw(boston.soi), 
#>     type = "mixed", method = "Matrix", trs = trMatb)
#> 
#> Residuals:
#>        Min         1Q     Median         3Q        Max 
#> -0.6316833 -0.0629790 -0.0090776  0.0682421  0.6991072 
#> 
#> Type: mixed 
#> Coefficients: (asymptotic standard errors) 
#>                   Estimate  Std. Error  z value  Pr(>|z|)
#> (Intercept)     1.89816218  0.22759605   8.3400 < 2.2e-16
#> CRIM           -0.00571021  0.00093504  -6.1069 1.016e-09
#> ZN              0.00069091  0.00051869   1.3320  0.182851
#> INDUS          -0.00111343  0.00307354  -0.3623  0.717155
#> CHAS1          -0.04163225  0.02739364  -1.5198  0.128567
#> I(NOX^2)       -0.01034950  0.19360214  -0.0535  0.957367
#> I(RM^2)         0.00794979  0.00102063   7.7891 6.661e-15
#> AGE            -0.00128789  0.00048920  -2.6326  0.008473
#> log(DIS)       -0.12404108  0.09510940  -1.3042  0.192168
#> log(RAD)        0.05863502  0.02257078   2.5978  0.009382
#> TAX            -0.00049084  0.00012145  -4.0416 5.308e-05
#> PTRATIO        -0.01319853  0.00595352  -2.2169  0.026628
#> B               0.00056383  0.00011089   5.0847 3.682e-07
#> log(LSTAT)     -0.24724454  0.02262033 -10.9302 < 2.2e-16
#> lag.CRIM       -0.00464215  0.00172935  -2.6843  0.007267
#> lag.ZN         -0.00037937  0.00070584  -0.5375  0.590940
#> lag.INDUS       0.00025064  0.00385901   0.0649  0.948215
#> lag.CHAS1       0.12518252  0.04071559   3.0746  0.002108
#> lag.I(NOX^2)   -0.38640401  0.22157523  -1.7439  0.081177
#> lag.I(RM^2)    -0.00451252  0.00153180  -2.9459  0.003220
#> lag.AGE         0.00149678  0.00068418   2.1877  0.028693
#> lag.log(DIS)   -0.00453785  0.10046478  -0.0452  0.963973
#> lag.log(RAD)   -0.00940702  0.03104930  -0.3030  0.761912
#> lag.TAX         0.00041083  0.00017867   2.2994  0.021481
#> lag.PTRATIO     0.00060355  0.00788837   0.0765  0.939012
#> lag.B          -0.00050781  0.00014155  -3.5874  0.000334
#> lag.log(LSTAT)  0.09846781  0.03399423   2.8966  0.003772
#> 
#> Rho: 0.59578, LR test value: 181.68, p-value: < 2.22e-16
#> Asymptotic standard error: 0.038445
#>     z-value: 15.497, p-value: < 2.22e-16
#> Wald statistic: 240.16, p-value: < 2.22e-16
#> 
#> Log likelihood: 300.6131 for mixed model
#> ML residual variance (sigma squared): 0.016011, (sigma: 0.12654)
#> Number of observations: 506 
#> Number of parameters estimated: 29 
#> AIC: -543.23, (AIC for lm: -363.55)
#> LM test for residual autocorrelation
#> test value: 29.772, p-value: 4.8604e-08
#> 

set.seed(1)
COL.lag.Bayes <- spBreg_lag(CRIME ~ INC + HOVAL, data=COL.OLD,
 listw=listw)
summary(COL.lag.Bayes)

#> 
#> Iterations = 501:2500
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean       SD Naive SE Time-series SE
#> (Intercept)  45.8422  8.33297 0.186331       0.220103
#> INC          -1.0460  0.35199 0.007871       0.008886
#> HOVAL        -0.2675  0.09481 0.002120       0.002120
#> rho           0.4174  0.12987 0.002904       0.003147
#> sige        108.9668 24.59087 0.549869       0.579894
#> 
#> 2. Quantiles for each variable:
#> 
#>                2.5%     25%      50%      75%     97.5%
#> (Intercept) 30.1130 40.3963  45.6116  51.2449  62.52518
#> INC         -1.7618 -1.2684  -1.0386  -0.8170  -0.35564
#> HOVAL       -0.4517 -0.3332  -0.2659  -0.2043  -0.08908
#> rho          0.1536  0.3324   0.4197   0.5038   0.66183
#> sige        71.2014 91.6918 105.3263 123.1005 165.01165
#> 

summary(impacts(COL.lag.Bayes, tr=trMatc), short=TRUE, zstats=TRUE)

#> Impact measures (lag, trace):
#>           Direct   Indirect      Total
#> INC   -1.0970636 -0.6983134 -1.7953770
#> HOVAL -0.2805878 -0.1786024 -0.4591902
#> ========================================================
#> MCMC sample impact results:
#> ========================================================
#> Simulated standard errors
#>          Direct  Indirect     Total
#> INC   0.3751705 0.5471652 0.8243027
#> HOVAL 0.1015673 0.1463852 0.2238372
#> 
#> Simulated z-values:
#>          Direct  Indirect     Total
#> INC   -2.951364 -1.435134 -2.295904
#> HOVAL -2.790347 -1.386199 -2.172682
#> 
#> Simulated p-values:
#>       Direct    Indirect Total   
#> INC   0.0031637 0.15125  0.021681
#> HOVAL 0.0052652 0.16569  0.029804

summary(impacts(COL.lag.Bayes, evalues=ev), short=TRUE, zstats=TRUE)

#> Impact measures (lag, evalues):
#>           Direct   Indirect      Total
#> INC   -1.0970636 -0.6983134 -1.7953770
#> HOVAL -0.2805878 -0.1786024 -0.4591902
#> ========================================================
#> MCMC sample impact results:
#> ========================================================
#> Simulated standard errors
#>          Direct  Indirect     Total
#> INC   0.3751709 0.5472321 0.8243543
#> HOVAL 0.1015676 0.1464146 0.2238607
#> 
#> Simulated z-values:
#>          Direct  Indirect     Total
#> INC   -2.951362 -1.434984 -2.295779
#> HOVAL -2.790341 -1.385952 -2.172475
#> 
#> Simulated p-values:
#>       Direct    Indirect Total   
#> INC   0.0031638 0.15129  0.021689
#> HOVAL 0.0052653 0.16576  0.029820

set.seed(1)
COL.D0.Bayes <- spBreg_lag(CRIME ~ INC + HOVAL, data=COL.OLD,
 listw=listw, Durbin=TRUE)
summary(COL.D0.Bayes)

#> 
#> Iterations = 501:2500
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean       SD Naive SE Time-series SE
#> (Intercept)  45.6552 13.92684 0.311414       0.311414
#> INC          -0.9357  0.37241 0.008327       0.008397
#> HOVAL        -0.2916  0.09815 0.002195       0.002038
#> lag.INC      -0.6098  0.61022 0.013645       0.013645
#> lag.HOVAL     0.2458  0.18971 0.004242       0.004734
#> rho           0.3883  0.16462 0.003681       0.003681
#> sige        108.6633 25.45638 0.569222       0.655387
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%     25%      50%      75%    97.5%
#> (Intercept) 19.42270 36.0252  45.3814  55.2111  73.4654
#> INC         -1.66848 -1.1944  -0.9352  -0.6836  -0.1840
#> HOVAL       -0.48299 -0.3576  -0.2930  -0.2232  -0.1015
#> lag.INC     -1.85700 -1.0306  -0.5826  -0.1992   0.5816
#> lag.HOVAL   -0.12660  0.1229   0.2465   0.3739   0.6106
#> rho          0.05948  0.2706   0.3962   0.5040   0.6979
#> sige        70.62697 90.7192 104.5505 123.4942 169.7947
#> 

summary(impacts(COL.D0.Bayes, tr=trMatc), short=TRUE, zstats=TRUE)

#> Impact measures (mixed, trace):
#>           Direct   Indirect       Total
#> INC   -1.0389664 -1.4874152 -2.52638157
#> HOVAL -0.2775359  0.2026609 -0.07487497
#> ========================================================
#> MCMC sample impact results:
#> ========================================================
#> Simulated standard errors
#>          Direct  Indirect     Total
#> INC   0.3927953 1.3104789 1.4692286
#> HOVAL 0.1014637 0.3249945 0.3616705
#> 
#> Simulated z-values:
#>          Direct   Indirect      Total
#> INC   -2.700710 -1.2936313 -1.8758840
#> HOVAL -2.749323  0.6152303 -0.2184584
#> 
#> Simulated p-values:
#>       Direct    Indirect Total   
#> INC   0.0069192 0.19579  0.060671
#> HOVAL 0.0059718 0.53840  0.827072

set.seed(1)
COL.D1.Bayes <- spBreg_lag(CRIME ~ DISCBD + INC + HOVAL, data=COL.OLD,
 listw=listw, Durbin= ~ INC)
summary(COL.D1.Bayes)

#> 
#> Iterations = 501:2500
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD Naive SE Time-series SE
#> (Intercept)  56.7873 13.5871 0.303817       0.303817
#> DISCBD       -4.7577  2.0863 0.046652       0.044748
#> INC          -0.9431  0.3481 0.007784       0.007784
#> HOVAL        -0.1795  0.1008 0.002255       0.002255
#> lag.INC       0.4036  0.6071 0.013575       0.012161
#> rho           0.1821  0.1832 0.004095       0.004095
#> sige        102.7442 23.7839 0.531824       0.608744
#> 
#> 2. Quantiles for each variable:
#> 
#>                2.5%       25%     50%      75%     97.5%
#> (Intercept) 30.4765 47.703086 56.7720  65.3671  84.50212
#> DISCBD      -8.8651 -6.222039 -4.6800  -3.3613  -0.61213
#> INC         -1.6560 -1.156154 -0.9413  -0.7174  -0.28960
#> HOVAL       -0.3748 -0.245735 -0.1811  -0.1107   0.01867
#> lag.INC     -0.8123  0.003337  0.4135   0.8014   1.58038
#> rho         -0.2147  0.062531  0.1826   0.3027   0.53984
#> sige        67.4492 85.860315 98.7754 116.5171 159.56314
#> 

summary(impacts(COL.D1.Bayes, tr=trMatc), short=TRUE, zstats=TRUE)

#> Impact measures (mixed, trace):
#>            Direct    Indirect      Total
#> DISCBD -4.7952447 -1.02140655 -5.8166512
#> INC    -0.9330808  0.27348153 -0.6595992
#> HOVAL  -0.1808675 -0.03852551 -0.2193930
#> ========================================================
#> MCMC sample impact results:
#> ========================================================
#> Simulated standard errors
#>           Direct   Indirect     Total
#> DISCBD 2.1324430 1.76669611 3.2740683
#> INC    0.3521918 0.77882425 0.8652785
#> HOVAL  0.1030000 0.07360172 0.1512414
#> 
#> Simulated z-values:
#>           Direct   Indirect      Total
#> DISCBD -2.272544 -0.7459585 -1.8826587
#> INC    -2.669828  0.3187837 -0.7997599
#> HOVAL  -1.775459 -0.6886619 -1.5442793
#> 
#> Simulated p-values:
#>        Direct   Indirect Total   
#> DISCBD 0.023054 0.45569  0.059747
#> INC    0.007589 0.74989  0.423850
#> HOVAL  0.075822 0.49104  0.122521

data(elect80, package="spData")
lw <- nb2listw(e80_queen, zero.policy=TRUE)
el_ml <- lagsarlm(log(pc_turnout) ~ log(pc_college) + log(pc_homeownership)
 + log(pc_income), data=elect80, listw=lw, zero.policy=TRUE, method="LU")
summary(el_ml)

#> 
#> Call:
#> lagsarlm(formula = log(pc_turnout) ~ log(pc_college) + log(pc_homeownership) + 
#>     log(pc_income), data = elect80, listw = lw, method = "LU", 
#>     zero.policy = TRUE)
#> 
#> Residuals:
#>        Min         1Q     Median         3Q        Max 
#> -1.1132095 -0.0690070 -0.0037755  0.0670525  0.8084858 
#> 
#> Type: lag 
#> Regions with no neighbours included:
#>  1183 1189 1832 2945 
#> Coefficients: (numerical Hessian approximate standard errors) 
#>                        Estimate Std. Error z value  Pr(>|z|)
#> (Intercept)            0.637925   0.041787 15.2660 < 2.2e-16
#> log(pc_college)        0.226366   0.015759 14.3645 < 2.2e-16
#> log(pc_homeownership)  0.481409   0.014524 33.1464 < 2.2e-16
#> log(pc_income)        -0.104942   0.016550 -6.3408 2.285e-10
#> 
#> Rho: 0.57742, LR test value: 1085.5, p-value: < 2.22e-16
#> Approximate (numerical Hessian) standard error: 0.014966
#>     z-value: 38.581, p-value: < 2.22e-16
#> Wald statistic: 1488.5, p-value: < 2.22e-16
#> 
#> Log likelihood: 2132.772 for lag model
#> ML residual variance (sigma squared): 0.013815, (sigma: 0.11754)
#> Number of observations: 3107 
#> Number of parameters estimated: 6 
#> AIC: -4253.5, (AIC for lm: -3170)
#> 

set.seed(1)
el_B <- spBreg_lag(log(pc_turnout) ~ log(pc_college) + log(pc_homeownership)
 + log(pc_income), data=elect80, listw=lw, zero.policy=TRUE)
summary(el_B)

#> 
#> Iterations = 501:2500
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                           Mean        SD  Naive SE Time-series SE
#> (Intercept)            0.63880 0.0412774 9.230e-04      8.895e-04
#> log(pc_college)        0.22707 0.0162196 3.627e-04      3.627e-04
#> log(pc_homeownership)  0.48157 0.0149993 3.354e-04      3.354e-04
#> log(pc_income)        -0.10539 0.0164751 3.684e-04      3.684e-04
#> rho                    0.57617 0.0152953 3.420e-04      3.420e-04
#> sige                   0.01385 0.0003612 8.077e-06      8.077e-06
#> 
#> 2. Quantiles for each variable:
#> 
#>                           2.5%     25%      50%      75%    97.5%
#> (Intercept)            0.55795  0.6113  0.63883  0.66694  0.71933
#> log(pc_college)        0.19707  0.2163  0.22726  0.23783  0.25725
#> log(pc_homeownership)  0.45325  0.4714  0.48133  0.49111  0.51015
#> log(pc_income)        -0.13776 -0.1165 -0.10535 -0.09452 -0.07317
#> rho                    0.54577  0.5658  0.57629  0.58679  0.60480
#> sige                   0.01317  0.0136  0.01384  0.01410  0.01455
#> 

el_ml$timings

#>           user.self elapsed
#> set_up        0.015   0.014
#> LU_set_up     0.003   0.003
#> LU_opt        0.243   0.247
#> coefs         0.002   0.002
#> LU_fdHess     0.335   0.340

attr(el_B, "timings")

#>                   user.self sys.self elapsed user.child sys.child
#> set_up                0.011    0.000   0.012          0         0
#> SE_classic_set_up     3.255    0.004   3.311          0         0
#> complete_setup        0.000    0.000   0.000          0         0
#> sampling              1.352    0.000   1.373          0         0
#> finalise              0.013    0.000   0.014          0         0