Spatial neighbour sparse representation

The function makes a "spatial neighbour" object representation (similar to the S-PLUS spatial statististics module representation of a "listw" spatial weights object. sn2listw() is the inverse function to listw2sn(), creating a "listw" object from a "spatial neighbour" object. The as.spam.listw method converts a "listw" object to a sparse matrix as defined in the spam package, using listw2sn().

listw2sn(listw)
sn2listw(sn)
as.spam.listw(listw)

Arguments

listw
a listw object from for example nb2listw

sn
a spatial.neighbour object

Value

listw2sn()returns a data frame with three columns, and with class spatial.neighbour:

from

region number id for the start of the link (S-PLUS row.id)

to

region number id for the end of the link (S-PLUS col.id)

weights

weight for this link

%\code{logSpwdet} returns log det(I - rho * W).
See also

Examples

if (require(rgdal, quietly=TRUE)) {
example(columbus)
col.listw <- nb2listw(col.gal.nb)
col.listw$neighbours[[1]]
col.listw$weights[[1]]
col.sn <- listw2sn(col.listw)
str(col.sn)

#> Error: <text>:9:0: unexpected end of input
#> 7: str(col.sn)
#> 8: 
#>   ^
# NOT RUN {
col.sp <- as.spam.listw(col.listw)
str(col.sp)
# }
}

#> Error: <text>:1:1: unexpected '}'
#> 1: }
#>     ^