Lagrange Multiplier diagnostics for spatial dependence in linear models

The function reports the estimates of tests chosen among five statistics for testing for spatial dependence in linear models. The statistics are the simple LM test for error dependence (LMerr), the simple LM test for a missing spatially lagged dependent variable (LMlag), variants of these robust to the presence of the other (RLMerr, RLMlag - RLMerr tests for error dependence in the possible presence of a missing lagged dependent variable, RLMlag the other way round), and a portmanteau test (SARMA, in fact LMerr + RLMlag). Note: from spdep 0.3-32, the value of the weights matrix trace term is returned correctly for both underlying symmetric and asymmetric neighbour lists, before 0.3-32, the value was wrong for listw objects based on asymmetric neighbour lists, such as k-nearest neighbours (thanks to Luc Anselin for finding the bug).

lm.LMtests(model, listw, zero.policy=NULL, test="LMerr", spChk=NULL, naSubset=TRUE)
# S3 method for LMtestlist
print(x, ...)
# S3 method for LMtestlist
summary(object, p.adjust.method="none", ...)
# S3 method for LMtestlist.summary
print(x, digits=max(3, getOption("digits") - 2), ...)
<!-- %tracew(listw) -->

Arguments

model
an object of class lm returned by lm, or optionally a vector of externally calculated residuals (run though na.omit if any NAs present) for use when only "LMerr" is chosen; weights and offsets should not be used in the lm object

listw
a listw object created for example by nb2listw, expected to be row-standardised (W-style)

zero.policy
default NULL, use global option value; if TRUE assign zero to the lagged value of zones without neighbours, if FALSE assign NA

test
a character vector of tests requested chosen from LMerr, LMlag, RLMerr, RLMlag, SARMA; test="all" computes all the tests.

spChk
should the data vector names be checked against the spatial objects for identity integrity, TRUE, or FALSE, default NULL to use get.spChkOption()

naSubset
default TRUE to subset listw object for omitted observations in model object (this is a change from earlier behaviour, when the model$na.action component was ignored, and the listw object had to be subsetted by hand)

x, object
object to be printed

p.adjust.method
a character string specifying the probability value adjustment (see p.adjust) for multiple tests, default "none"

digits
minimum number of significant digits to be used for most numbers

...
printing arguments to be passed through

Details

The two types of dependence are for spatial lag \(\rho\) and spatial error \(\lambda\):

$$ \mathbf{y} = \mathbf{X \beta} + \rho \mathbf{W_{(1)} y} + \mathbf{u}, $$ $$ \mathbf{u} = \lambda \mathbf{W_{(2)} u} + \mathbf{e} $$

where \(\mathbf{e}\) is a well-behaved, uncorrelated error term. Tests for a missing spatially lagged dependent variable test that \(\rho = 0\), tests for spatial autocorrelation of the error \(\mathbf{u}\) test whether \(\lambda = 0\). \(\mathbf{W}\) is a spatial weights matrix; for the tests used here they are identical.

Value

A list of class LMtestlist of htest objects, each with:

statistic

the value of the Lagrange Multiplier test.

parameter

number of degrees of freedom

p.value

the p-value of the test.

method

a character string giving the method used.

data.name

a character string giving the name(s) of the data.

References

Anselin, L. 1988 Spatial econometrics: methods and models. (Dordrecht: Kluwer); Anselin, L., Bera, A. K., Florax, R. and Yoon, M. J. 1996 Simple diagnostic tests for spatial dependence. Regional Science and Urban Economics, 26, 77--104.

Author

Roger Bivand Roger.Bivand@nhh.no and Andrew Bernat

See also

Examples

data(oldcol)
oldcrime.lm <- lm(CRIME ~ HOVAL + INC, data = COL.OLD)
summary(oldcrime.lm)

#> 
#> Call:
#> lm(formula = CRIME ~ HOVAL + INC, data = COL.OLD)
#> 
#> Residuals:
#>     Min      1Q  Median      3Q     Max 
#> -34.418  -6.388  -1.580   9.052  28.649 
#> 
#> Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)  68.6190     4.7355  14.490  < 2e-16 ***
#> HOVAL        -0.2739     0.1032  -2.654   0.0109 *  
#> INC          -1.5973     0.3341  -4.780 1.83e-05 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error: 11.43 on 46 degrees of freedom
#> Multiple R-squared:  0.5524,	Adjusted R-squared:  0.5329 
#> F-statistic: 28.39 on 2 and 46 DF,  p-value: 9.341e-09
#> 

res <- lm.LMtests(oldcrime.lm, nb2listw(COL.nb), test=c("LMerr", "LMlag",
  "RLMerr", "RLMlag", "SARMA"))
summary(res)

#> 	Lagrange multiplier diagnostics for spatial dependence
#> data:  
#> model: lm(formula = CRIME ~ HOVAL + INC, data = COL.OLD)
#> weights: nb2listw(COL.nb)
#>  
#>        statistic parameter  p.value   
#> LMerr   5.723131         1 0.016743 * 
#> LMlag   9.363684         1 0.002213 **
#> RLMerr  0.079495         1 0.777983   
#> RLMlag  3.720048         1 0.053763 . 
#> SARMA   9.443178         2 0.008901 **
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1

lm.LMtests(oldcrime.lm, nb2listw(COL.nb))

#> 
#> 	Lagrange multiplier diagnostics for spatial dependence
#> 
#> data:  
#> model: lm(formula = CRIME ~ HOVAL + INC, data = COL.OLD)
#> weights: nb2listw(COL.nb)
#> 
#> LMErr = 5.7231, df = 1, p-value = 0.01674
#> 

lm.LMtests(residuals(oldcrime.lm), nb2listw(COL.nb))

#> 
#> 	Lagrange multiplier diagnostics for spatial dependence
#> 
#> data:  
#> residuals: residuals(oldcrime.lm)
#> weights: nb2listw(COL.nb)
#> 
#> LMErr = 5.7231, df = 1, p-value = 0.01674
#>