Compute the Local Bivariate Moran's I Statistic

Given two continuous numeric variables, calculate the bivariate Local Moran's I.

localmoran_bv(x, y, listw, nsim = 199, scale = TRUE)

Arguments

x

a numeric vector of same length as y.

y

a numeric vector of same length as x.

listw

a listw object for example as created by nb2listw().

nsim

the number of simulations to run.

scale

default TRUE.

Value

a data.frame containing two columns Ib and p_sim containing the local bivariate Moran's I and simulated p-values respectively.

Details

The Bivariate Local Moran, like its global counterpart, evaluates the value of x at observation i with its spatial neighbors' value of y. The value of \(I_i^B\) is xi * Wyi. Or, in simpler words, the local bivariate Moran is the result of multiplying x by the spatial lag of y. Formally it is defined as

\( I_i^B= cx_i\Sigma_j{w_{ij}y_j} \)

References

Anselin, Luc, Ibnu Syabri, and Oleg Smirnov. 2002. “Visualizing Multivariate Spatial Correlation with Dynamically Linked Windows.” In New Tools for Spatial Data Analysis: Proceedings of the Specialist Meeting, edited by Luc Anselin and Sergio Rey. University of California, Santa Barbara: Center for Spatially Integrated Social Science (CSISS).

Author

Josiah Parry josiah.parry@gmail.com

Examples

# load columbus data
columbus <- st_read(system.file("shapes/columbus.shp", package="spData"))
#> Reading layer `columbus' from data source 
#>   `/home/rsb/lib/r_libs/spData/shapes/columbus.shp' using driver `ESRI Shapefile'
#> Simple feature collection with 49 features and 20 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 5.874907 ymin: 10.78863 xmax: 11.28742 ymax: 14.74245
#> CRS:           NA
nb <- poly2nb(columbus)
listw <- nb2listw(nb)
localmoran_bv(columbus$CRIME, columbus$INC, listw, nsim = 99)
#>               Ib p_sim
#> 1  -0.8578252210  0.15
#> 2   0.1802618568  0.42
#> 3   0.0118856298  0.49
#> 4  -0.0163525887  0.46
#> 5  -0.5037734606  0.03
#> 6  -0.0041841415  0.43
#> 7   1.5201688711  0.04
#> 8  -0.2016456990  0.00
#> 9   0.0297432891  0.37
#> 10 -0.0374203146  0.11
#> 11 -1.2368494456  0.01
#> 12 -1.2010054212  0.00
#> 13 -0.5377131737  0.04
#> 14 -0.9879712511  0.00
#> 15 -0.6057303757  0.00
#> 16 -0.7878039284  0.01
#> 17  0.1379508106  0.02
#> 18 -0.3253226893  0.07
#> 19 -0.3888822685  0.26
#> 20 -0.4544889957  0.13
#> 21 -0.0006014046  0.50
#> 22 -0.0068813684  0.36
#> 23 -0.8788967496  0.04
#> 24 -0.1256865340  0.03
#> 25 -0.9307459211  0.06
#> 26 -0.2252775788  0.00
#> 27  0.1413018326  0.37
#> 28 -0.7804482928  0.00
#> 29 -0.7654416689  0.07
#> 30 -0.9074829015  0.13
#> 31 -0.5510119267  0.20
#> 32 -1.9581476375  0.00
#> 33  0.0907015890  0.32
#> 34 -0.4297843082  0.11
#> 35  0.1181174820  0.13
#> 36 -0.9901007017  0.02
#> 37 -0.2195812498  0.07
#> 38 -0.3096268627  0.21
#> 39 -0.6090192229  0.20
#> 40 -1.2709838927  0.00
#> 41 -1.3940681427  0.01
#> 42 -0.4820987026  0.22
#> 43 -0.0064402410  0.44
#> 44  0.0761664412  0.46
#> 45 -0.0560342555  0.30
#> 46 -0.8249832451  0.18
#> 47 -0.8915318222  0.00
#> 48 -0.1282863305  0.24
#> 49  0.0094888641  0.40