Plot the object of skater class

This function displays the results of the skater function. The subgraphs are plotted with different colours.

# S3 method for skater
plot(x, coords, label.areas = NULL, 
   groups.colors, cex.circles = 1, cex.labels = 1, ...)

Arguments

x

An object of skater class.

coords

A matrix of two colums with coordinates of nodes.

label.areas

A vector of labels of nodes.

groups.colors

A vector with colors of groups ou sub-graphs.

cex.circles

The length of circles with represent the nodes.

cex.labels

The length of labels of nodes.

...

Further arguments passed to plotting funcitons.

Author

Elias T. Krainski and Renato M. Assuncao

See also

Examples

### see example in the skater function documentation