Spatial simultaneous autoregressive SAC model estimation

Maximum likelihood estimation of spatial simultaneous autoregressive “SAC/SARAR” models of the form:

$$y = \rho W1 y + X \beta + u, u = \lambda W2 u + \varepsilon$$

where \(\rho\) and \(\lambda\) are found by nlminb or optim() first, and \(\beta\) and other parameters by generalized least squares subsequently

sacsarlm(formula, data = list(), listw, listw2 = NULL, na.action, Durbin, type,
 method = "eigen", quiet = NULL, zero.policy = NULL, tol.solve = 1e-10,
 llprof=NULL, interval1=NULL, interval2=NULL, trs1=NULL, trs2=NULL,
 control = list())
spBreg_sac(formula, data = list(), listw, listw2=NULL, na.action,
    Durbin, type, zero.policy=NULL, control=list())

Arguments

formula
a symbolic description of the model to be fit. The details of model specification are given for lm()

data
an optional data frame containing the variables in the model. By default the variables are taken from the environment which the function is called

listw
a listw object created for example by nb2listw

listw2
a listw object created for example by nb2listw, if not given, set to the same spatial weights as the listw argument

na.action
a function (default options("na.action")), can also be na.omit or na.exclude with consequences for residuals and fitted values - in these cases the weights list will be subsetted to remove NAs in the data. It may be necessary to set zero.policy to TRUE because this subsetting may create no-neighbour observations. Note that only weights lists created without using the glist argument to nb2listw may be subsetted.

Durbin
default FALSE (SARAR model); if TRUE, full spatial general nested (Manski) model; if a formula object, the subset of explanatory variables to lag

type
(use the ‘Durbin=’ argument - retained for backwards compatibility only) default "sac", may be set to "sacmixed" for the spatial general nested (Manski) model to include the spatially lagged independent variables added to X using listw; when "sacmixed", the lagged intercept is dropped for spatial weights style "W", that is row-standardised weights, but otherwise included

method
"eigen" (default) - the Jacobian is computed as the product of (1 - rho*eigenvalue) using eigenw, and "spam" or "Matrix" for strictly symmetric weights lists of styles "B" and "C", or made symmetric by similarity (Ord, 1975, Appendix C) if possible for styles "W" and "S", using code from the spam or Matrix packages to calculate the determinant; "LU" provides an alternative sparse matrix decomposition approach. In addition, there are "Chebyshev" and Monte Carlo "MC" approximate log-determinant methods.

quiet
default NULL, use !verbose global option value; if FALSE, reports function values during optimization.

zero.policy
default NULL, use global option value; if TRUE assign zero to the lagged value of zones without neighbours, if FALSE (default) assign NA - causing sacsarlm() to terminate with an error

tol.solve
the tolerance for detecting linear dependencies in the columns of matrices to be inverted - passed to solve() (default=1.0e-10). This may be used if necessary to extract coefficient standard errors (for instance lowering to 1e-12), but errors in solve() may constitute indications of poorly scaled variables: if the variables have scales differing much from the autoregressive coefficient, the values in this matrix may be very different in scale, and inverting such a matrix is analytically possible by definition, but numerically unstable; rescaling the RHS variables alleviates this better than setting tol.solve to a very small value

llprof
default NULL, can either be an integer, to divide the feasible ranges into a grid of points, or a two-column matrix of spatial coefficient values, at which to evaluate the likelihood function

trs1, trs2
default NULL, if given, vectors for each weights object of powered spatial weights matrix traces output by trW; when given, used in some Jacobian methods

interval1, interval2
default is NULL, search intervals for each weights object for autoregressive parameters

control
list of extra control arguments - see section below

Details

Because numerical optimisation is used to find the values of lambda and rho, care needs to be shown. It has been found that the surface of the 2D likelihood function often forms a “banana trench” from (low rho, high lambda) through (high rho, high lambda) to (high rho, low lambda) values. In addition, sometimes the banana has optima towards both ends, one local, the other global, and conseqently the choice of the starting point for the final optimization becomes crucial. The default approach is not to use just (0, 0) as a starting point, nor the (rho, lambda) values from gstsls, which lie in a central part of the “trench”, but either four values at (low rho, high lambda), (0, 0), (high rho, high lambda), and (high rho, low lambda), and to use the best of these start points for the final optimization. Optionally, nine points can be used spanning the whole (lower, upper) space.

Control arguments

fdHess:

default NULL, then set to (method != "eigen") internally; use fdHess to compute an approximate Hessian using finite differences when using sparse matrix methods with fdHess from nlme; used to make a coefficient covariance matrix when the number of observations is large; may be turned off to save resources if need be

LAPACK:

default FALSE; logical value passed to qr in the SSE log likelihood function

Imult:

default 2; used for preparing the Cholesky decompositions for updating in the Jacobian function

cheb_q:

default 5; highest power of the approximating polynomial for the Chebyshev approximation

MC_p:

default 16; number of random variates

MC_m:

default 30; number of products of random variates matrix and spatial weights matrix

super:

default FALSE using a simplicial decomposition for the sparse Cholesky decomposition, if TRUE, use a supernodal decomposition

opt_method:

default “nlminb”, may be set to “L-BFGS-B” to use box-constrained optimisation in optim

opt_control:

default list(), a control list to pass to nlminb or optim

pars:

default NULL, for which five trial starting values spanning the lower/upper range are tried and the best selected, starting values of \(\rho\) and \(\lambda\)

npars

default integer 4L, four trial points; if not default value, nine trial points

pre_eig1, pre_eig2

default NULL; may be used to pass pre-computed vectors of eigenvalues

Extra Bayesian control arguments

ldet_method

default “SE_classic”; equivalent to the method argument in lagsarlm

interval1

default c(-1, 1); used unmodified or set internally by jacobianSetup

interval2

default c(-1, 1); used unmodified or set internally by jacobianSetup

ndraw

default 2500L; integer total number of draws

nomit

default 500L; integer total number of omitted burn-in draws

thin

default 1L; integer thinning proportion

detval1

default NULL; not yet in use, precomputed matrix of log determinants

detval2

default NULL; not yet in use, precomputed matrix of log determinants

prior

a list with the following components:

Tbeta

default NULL; values of the betas variance-covariance matrix, set to diag(k)*1e+12 if NULL

c_beta

default NULL; values of the betas set to 0 if NULL

lambda

default 0.5; value of the autoregressive coefficient

rho

default 0.5; value of the autoregressive coefficient

sige

default 1; value of the residual variance

nu

default 0; informative Gamma(nu,d0) prior on sige

d0

default 0; informative Gamma(nu,d0) prior on sige

a1

default 1.01; parameter for beta(a1,a2) prior on rho

a2

default 1.01; parameter for beta(a1,a2) prior on rho

cc1

default 0.2; initial tuning parameter for M-H sampling

cc2

default 0.2; initial tuning parameter for M-H sampling

Value

A list object of class sarlm

type

“sac” or “sacmixed”

dvars

vector of length 2, numbers of columns in X and WX; if Durbin is given as a formula, the formula as attribute “f”, the indices of the included Wx as “inds”, and indices of added zero Wx coefficients as “zero_fill”

rho

lag simultaneous autoregressive lag coefficient

lambda

error simultaneous autoregressive error coefficient

coefficients

GLS coefficient estimates

rest.se

asymptotic standard errors if ase=TRUE, otherwise approximate numeriacal Hessian-based values

ase

TRUE if method=eigen

LL

log likelihood value at computed optimum

s2

GLS residual variance

SSE

sum of squared GLS errors

parameters

number of parameters estimated

logLik_lm.model

Log likelihood of the non-spatial linear model

AIC_lm.model

AIC of the non-spatial linear model

% \item{lm.model}{the \code{lm} object returned when estimating for \eqn{\rho=0}{rho=0}}
method

the method used to calculate the Jacobian

call

the call used to create this object

residuals

GLS residuals

% \item{lm.target}{the \code{lm} object returned for the GLS fit}
tarX

model matrix of the GLS model

tary

response of the GLS model

y

response of the linear model for \(\rho=0\)

X

model matrix of the linear model for \(\rho=0\)

opt

object returned from numerical optimisation

pars

starting parameter values for final optimization, either given or found by trial point evaluation

mxs

if default input pars, optimal objective function values at trial points

fitted.values

Difference between residuals and response variable

se.fit

Not used yet

% \item{formula}{model formula}
rho.se

if ase=TRUE, the asymptotic standard error of \(\rho\), otherwise approximate numeriacal Hessian-based value

lambda.se

if ase=TRUE, the asymptotic standard error of \(\lambda\)

resvar

the asymptotic coefficient covariance matrix for (s2, rho, lambda, B)

zero.policy

zero.policy for this model

aliased

the aliased explanatory variables (if any)

LLNullLlm

Log-likelihood of the null linear model

fdHess

the numerical Hessian-based coefficient covariance matrix for (rho, lambda, B) if computed

resvar

asymptotic coefficient covariance matrix

optimHess

FALSE

timings

processing timings

na.action

(possibly) named vector of excluded or omitted observations if non-default na.action argument used

References

Anselin, L. 1988 Spatial econometrics: methods and models. (Dordrecht: Kluwer); LeSage J and RK Pace (2009) Introduction to Spatial Econometrics. CRC Press, Boca Raton.

Roger Bivand, Gianfranco Piras (2015). Comparing Implementations of Estimation Methods for Spatial Econometrics. Journal of Statistical Software, 63(18), 1-36. https://www.jstatsoft.org/v63/i18/.

Bivand, R. S., Hauke, J., and Kossowski, T. (2013). Computing the Jacobian in Gaussian spatial autoregressive models: An illustrated comparison of available methods. Geographical Analysis, 45(2), 150-179.

See also

Examples

data(oldcol)
listw <- nb2listw(COL.nb, style="W")
ev <- eigenw(listw)
COL.sacW.eig <- sacsarlm(CRIME ~ INC + HOVAL, data=COL.OLD, listw,
 control=list(pre_eig1=ev, pre_eig2=ev))
summary(COL.sacW.eig)

#> 
#> Call:sacsarlm(formula = CRIME ~ INC + HOVAL, data = COL.OLD, listw = listw, 
#>     control = list(pre_eig1 = ev, pre_eig2 = ev))
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -37.32081  -5.33662  -0.20219   6.59672  23.25604 
#> 
#> Type: sac 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) 47.783764   9.902659  4.8253 1.398e-06
#> INC         -1.025894   0.326326 -3.1438  0.001668
#> HOVAL       -0.281651   0.090033 -3.1283  0.001758
#> 
#> Rho: 0.36807
#> Asymptotic standard error: 0.19668
#>     z-value: 1.8714, p-value: 0.061285
#> Lambda: 0.16668
#> Asymptotic standard error: 0.29661
#>     z-value: 0.56196, p-value: 0.57415
#> 
#> LR test value: 10.285, p-value: 0.0058432
#> 
#> Log likelihood: -182.2348 for sac model
#> ML residual variance (sigma squared): 95.604, (sigma: 9.7777)
#> Number of observations: 49 
#> Number of parameters estimated: 6 
#> AIC: 376.47, (AIC for lm: 382.75)
#> 

W <- as(listw, "CsparseMatrix")
trMatc <- trW(W, type="mult")
set.seed(1)
summary(impacts(COL.sacW.eig, tr=trMatc, R=2000), zstats=TRUE, short=TRUE)

#> Impact measures (sac, trace):
#>           Direct   Indirect      Total
#> INC   -1.0632723 -0.5601502 -1.6234225
#> HOVAL -0.2919129 -0.1537848 -0.4456977
#> ========================================================
#> Simulation results (asymptotic variance matrix):
#> ========================================================
#> Simulated standard errors
#>          Direct  Indirect     Total
#> INC   0.3360326 0.8773349 1.0073913
#> HOVAL 0.1001824 0.3723400 0.4254181
#> 
#> Simulated z-values:
#>          Direct   Indirect     Total
#> INC   -3.225224 -0.8379554 -1.805602
#> HOVAL -2.992993 -0.6154364 -1.243475
#> 
#> Simulated p-values:
#>       Direct    Indirect Total   
#> INC   0.0012587 0.40206  0.070981
#> HOVAL 0.0027626 0.53827  0.213693

if (require(coda, quietly=TRUE)) {
set.seed(1)
COL.sacW.B0 <- spBreg_sac(CRIME ~ INC + HOVAL, data=COL.OLD, listw,
 Durbin=FALSE, control=list(ndraw=1500L, nomit=500L))
print(summary(COL.sacW.B0))
print(summary(impacts(COL.sacW.B0, tr=trMatc), zstats=TRUE, short=TRUE))
set.seed(1)
COL.sacW.B1 <- spBreg_sac(CRIME ~ INC + HOVAL, data=COL.OLD, listw,
 Durbin=TRUE, control=list(ndraw=1500L, nomit=500L))
print(summary(COL.sacW.B1))
print(summary(impacts(COL.sacW.B1, tr=trMatc), zstats=TRUE, short=TRUE))
}

#> 
#> Iterations = 501:1500
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 1000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean       SD Naive SE Time-series SE
#> (Intercept)  49.7487  9.65954 0.305462        0.80288
#> INC          -1.0133  0.34700 0.010973        0.01306
#> HOVAL        -0.2812  0.09521 0.003011        0.00334
#> rho           0.3037  0.20103 0.006357        0.01908
#> lambda        0.2036  0.28522 0.009019        0.02728
#> sige        104.0852 23.19791 0.733582        0.83380
#> 
#> 2. Quantiles for each variable:
#> 
#>                2.5%      25%      50%      75%     97.5%
#> (Intercept) 30.6943 43.55329  49.6265  55.9348  69.25088
#> INC         -1.6926 -1.23577  -1.0207  -0.7706  -0.33203
#> HOVAL       -0.4689 -0.34422  -0.2817  -0.2228  -0.08857
#> rho         -0.1016  0.19369   0.3362   0.4264   0.65502
#> lambda      -0.3415  0.01603   0.2258   0.3913   0.69422
#> sige        71.1594 87.59002 100.0702 116.7883 160.77222
#> 
#> Impact measures (sac, trace):
#>           Direct   Indirect      Total
#> INC   -1.0372824 -0.4180106 -1.4552930
#> HOVAL -0.2878461 -0.1159980 -0.4038441
#> ========================================================
#> MCMC sample impact results:
#> ========================================================
#> Simulated standard errors
#>           Direct  Indirect     Total
#> INC   0.35632326 0.4452514 0.6656137
#> HOVAL 0.09843362 0.1381719 0.1963475
#> 
#> Simulated z-values:
#>          Direct  Indirect     Total
#> INC   -2.950929 -1.152887 -2.350927
#> HOVAL -2.965772 -1.029062 -2.210973
#> 
#> Simulated p-values:
#>       Direct    Indirect Total   
#> INC   0.0031682 0.24896  0.018727
#> HOVAL 0.0030192 0.30345  0.027038
#> 
#> Iterations = 501:1500
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 1000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                Mean       SD Naive SE Time-series SE
#> (Intercept) 59.0609 24.07621 0.761357       2.885394
#> INC         -0.9854  0.36353 0.011496       0.016762
#> HOVAL       -0.2880  0.09725 0.003075       0.003075
#> lag.INC     -0.8411  0.75844 0.023984       0.056331
#> lag.HOVAL    0.1752  0.22457 0.007101       0.012878
#> rho          0.1946  0.30961 0.009791       0.042581
#> lambda       0.2158  0.30532 0.009655       0.035795
#> sige        99.1912 21.68244 0.685659       0.732837
#> 
#> 2. Quantiles for each variable:
#> 
#>                2.5%       25%     50%      75%     97.5%
#> (Intercept) 21.0234 40.176668 56.2825  73.1358 114.42123
#> INC         -1.6568 -1.231704 -0.9995  -0.7444  -0.25073
#> HOVAL       -0.4814 -0.351803 -0.2914  -0.2232  -0.09721
#> lag.INC     -2.4646 -1.343720 -0.7909  -0.3135   0.56286
#> lag.HOVAL   -0.2591  0.024796  0.1941   0.3181   0.60547
#> rho         -0.5053 -0.006441  0.2361   0.4455   0.65175
#> lambda      -0.4007 -0.012730  0.2565   0.4349   0.74335
#> sige        65.5174 83.825061 96.4492 110.2511 149.90304
#> 
#> Impact measures (sacmixed, trace):
#>           Direct   Indirect      Total
#> INC   -1.0336050 -1.2342060 -2.2678109
#> HOVAL -0.2824441  0.1423536 -0.1400905
#> ========================================================
#> MCMC sample impact results:
#> ========================================================
#> Simulated standard errors
#>           Direct  Indirect     Total
#> INC   0.35681532 0.9491189 1.0525755
#> HOVAL 0.09938295 0.3057824 0.3445356
#> 
#> Simulated z-values:
#>          Direct   Indirect      Total
#> INC   -2.915633 -1.4203263 -2.2691021
#> HOVAL -2.830873  0.5166975 -0.3579994
#> 
#> Simulated p-values:
#>       Direct    Indirect Total   
#> INC   0.0035497 0.15551  0.023262
#> HOVAL 0.0046421 0.60537  0.720344

COL.msacW.eig <- sacsarlm(CRIME ~ INC + HOVAL, data=COL.OLD, listw,
 type="sacmixed", control=list(pre_eig1=ev, pre_eig2=ev))
summary(COL.msacW.eig)

#> 
#> Call:sacsarlm(formula = CRIME ~ INC + HOVAL, data = COL.OLD, listw = listw, 
#>     type = "sacmixed", control = list(pre_eig1 = ev, pre_eig2 = ev))
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -37.8045  -6.5244  -0.2207   5.9944  22.8691 
#> 
#> Type: sacmixed 
#> Coefficients: (asymptotic standard errors) 
#>             Estimate Std. Error z value Pr(>|z|)
#> (Intercept) 50.92026   68.25721  0.7460 0.455664
#> INC         -0.95072    0.44033 -2.1591 0.030841
#> HOVAL       -0.28650    0.09994 -2.8667 0.004148
#> lag.INC     -0.69261    1.69113 -0.4096 0.682132
#> lag.HOVAL    0.20852    0.28702  0.7265 0.467546
#> 
#> Rho: 0.31557
#> Asymptotic standard error: 0.9458
#>     z-value: 0.33365, p-value: 0.73864
#> Lambda: 0.15415
#> Asymptotic standard error: 1.0643
#>     z-value: 0.14484, p-value: 0.88484
#> 
#> LR test value: 12.07, p-value: 0.016837
#> 
#> Log likelihood: -181.3422 for sacmixed model
#> ML residual variance (sigma squared): 93.149, (sigma: 9.6514)
#> Number of observations: 49 
#> Number of parameters estimated: 8 
#> AIC: 378.68, (AIC for lm: 382.75)
#> 

set.seed(1)
summary(impacts(COL.msacW.eig, tr=trMatc, R=2000), zstats=TRUE, short=TRUE)

#> Impact measures (sacmixed, trace):
#>           Direct   Indirect      Total
#> INC   -1.0317003 -1.3693141 -2.4010144
#> HOVAL -0.2768608  0.1629265 -0.1139344
#> ========================================================
#> Simulation results (asymptotic variance matrix):
#> ========================================================
#> Simulated standard errors
#>          Direct  Indirect     Total
#> INC   0.3782512 2.0621650 2.2059732
#> HOVAL 0.1136749 0.7877222 0.8509034
#> 
#> Simulated z-values:
#>          Direct   Indirect      Total
#> INC   -2.735748 -0.7339053 -1.1551517
#> HOVAL -2.384665  0.2179576 -0.1168013
#> 
#> Simulated p-values:
#>       Direct    Indirect Total  
#> INC   0.0062239 0.46301  0.24803
#> HOVAL 0.0170947 0.82746  0.90702

COL.msacW1.eig <- sacsarlm(CRIME ~ INC + HOVAL, data=COL.OLD, listw,
 Durbin=TRUE, control=list(pre_eig1=ev, pre_eig2=ev))
summary(COL.msacW1.eig)

#> 
#> Call:sacsarlm(formula = CRIME ~ INC + HOVAL, data = COL.OLD, listw = listw, 
#>     Durbin = TRUE, control = list(pre_eig1 = ev, pre_eig2 = ev))
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -37.8045  -6.5244  -0.2207   5.9944  22.8691 
#> 
#> Type: sacmixed 
#> Coefficients: (asymptotic standard errors) 
#>             Estimate Std. Error z value Pr(>|z|)
#> (Intercept) 50.92026   68.25721  0.7460 0.455664
#> INC         -0.95072    0.44033 -2.1591 0.030841
#> HOVAL       -0.28650    0.09994 -2.8667 0.004148
#> lag.INC     -0.69261    1.69113 -0.4096 0.682132
#> lag.HOVAL    0.20852    0.28702  0.7265 0.467546
#> 
#> Rho: 0.31557
#> Asymptotic standard error: 0.9458
#>     z-value: 0.33365, p-value: 0.73864
#> Lambda: 0.15415
#> Asymptotic standard error: 1.0643
#>     z-value: 0.14484, p-value: 0.88484
#> 
#> LR test value: 12.07, p-value: 0.016837
#> 
#> Log likelihood: -181.3422 for sacmixed model
#> ML residual variance (sigma squared): 93.149, (sigma: 9.6514)
#> Number of observations: 49 
#> Number of parameters estimated: 8 
#> AIC: 378.68, (AIC for lm: 382.75)
#> 

set.seed(1)
summary(impacts(COL.msacW1.eig, tr=trMatc, R=2000), zstats=TRUE, short=TRUE)

#> Impact measures (sacmixed, trace):
#>           Direct   Indirect      Total
#> INC   -1.0317003 -1.3693141 -2.4010144
#> HOVAL -0.2768608  0.1629265 -0.1139344
#> ========================================================
#> Simulation results (asymptotic variance matrix):
#> ========================================================
#> Simulated standard errors
#>          Direct  Indirect     Total
#> INC   0.3782512 2.0621650 2.2059732
#> HOVAL 0.1136749 0.7877222 0.8509034
#> 
#> Simulated z-values:
#>          Direct   Indirect      Total
#> INC   -2.735748 -0.7339053 -1.1551517
#> HOVAL -2.384665  0.2179576 -0.1168013
#> 
#> Simulated p-values:
#>       Direct    Indirect Total  
#> INC   0.0062239 0.46301  0.24803
#> HOVAL 0.0170947 0.82746  0.90702

COL.msacW2.eig <- sacsarlm(CRIME ~ DISCBD + INC + HOVAL, data=COL.OLD,
 listw, Durbin= ~ INC, control=list(pre_eig1=ev, pre_eig2=ev))
summary(COL.msacW2.eig)

#> 
#> Call:sacsarlm(formula = CRIME ~ DISCBD + INC + HOVAL, data = COL.OLD, 
#>     listw = listw, Durbin = ~INC, control = list(pre_eig1 = ev, 
#>         pre_eig2 = ev))
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -34.2794  -7.0786   1.0543   6.0019  17.8891 
#> 
#> Type: sacmixed 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value Pr(>|z|)
#> (Intercept) 74.064506  37.738937  1.9625 0.049699
#> DISCBD      -5.707678   3.248629 -1.7569 0.078926
#> INC         -0.900975   0.314996 -2.8603 0.004233
#> HOVAL       -0.203399   0.092982 -2.1875 0.028705
#> lag.INC      0.061568   0.926947  0.0664 0.947043
#> 
#> Rho: -0.077599
#> Asymptotic standard error: 0.57955
#>     z-value: -0.1339, p-value: 0.89348
#> Lambda: 0.29646
#> Asymptotic standard error: 0.51021
#>     z-value: 0.58104, p-value: 0.56121
#> 
#> LR test value: 1.4982, p-value: 0.68269
#> 
#> Log likelihood: -178.963 for sacmixed model
#> ML residual variance (sigma squared): 85.135, (sigma: 9.2269)
#> Number of observations: 49 
#> Number of parameters estimated: 8 
#> AIC: 373.93, (AIC for lm: 369.42)
#> 

summary(impacts(COL.msacW2.eig, tr=trMatc, R=2000), zstats=TRUE, short=TRUE)

#> Impact measures (sacmixed, trace):
#>            Direct   Indirect      Total
#> DISCBD -5.7150801 0.41841708 -5.2966631
#> INC    -0.9031729 0.12421194 -0.7789609
#> HOVAL  -0.2036631 0.01491075 -0.1887524
#> ========================================================
#> Simulation results (asymptotic variance matrix):
#> ========================================================
#> Simulated standard errors
#>           Direct  Indirect     Total
#> DISCBD 3.2369649 5.4951520 5.6486738
#> INC    0.3627111 2.0073024 2.1711143
#> HOVAL  0.1048754 0.4887553 0.5425235
#> 
#> Simulated z-values:
#>           Direct    Indirect      Total
#> DISCBD -1.793118  0.03441804 -0.9940614
#> INC    -2.422720  0.15135275 -0.2648119
#> HOVAL  -2.081309 -0.22460039 -0.6046793
#> 
#> Simulated p-values:
#>        Direct   Indirect Total  
#> DISCBD 0.072954 0.97254  0.32019
#> INC    0.015405 0.87970  0.79115
#> HOVAL  0.037406 0.82229  0.54539