Defunct Functions in Package spdep

These functions are defunct from release 1.2-1. The functions have been moved to the spatialreg package.

Usage

aple.plot()
localAple()
aple.mc()
aple()
lextrB()
lextrW()
lextrS()
griffith_sone()
subgraph_eigenw()
mom_calc()
mom_calc_int2()
stsls()
# S3 method for class 'stsls'
impacts()
GMerrorsar()
# S3 method for class 'gmsar'
summary()
gstsls() 
# S3 method for class 'gmsar'
impacts()
# S3 method for class 'gmsar'
Hausman.test()
lagmess()
ME()
SpatialFiltering()
LR.sarlm()
# S3 method for class 'sarlm'
logLik()
LR1.sarlm()
Wald1.sarlm()
# S3 method for class 'sarlm'
Hausman.test()
as.spam.listw()
as_dgRMatrix_listw()
as_dsTMatrix_listw()
as_dsCMatrix_I()
as_dsCMatrix_IrW()
Jacobian_W()
powerWeights()
# S3 method for class 'lagImpact'
plot()
# S3 method for class 'lagImpact'
print()
# S3 method for class 'lagImpact'
summary()
# S3 method for class 'lagImpact'
HPDinterval()
intImpacts()
can.be.simmed()
eigenw()
similar.listw()
do_ldet()
jacobianSetup()
cheb_setup()
mcdet_setup()
eigen_setup()
eigen_pre_setup()
spam_setup()
spam_update_setup()
Matrix_setup()
Matrix_J_setup()
LU_setup()
LU_prepermutate_setup()
moments_setup()
SE_classic_setup()
SE_whichMin_setup()
SE_interp_setup()
MCMCsamp()
# S3 method for class 'spautolm'
MCMCsamp()
# S3 method for class 'sarlm'
MCMCsamp()
spautolm()
# S3 method for class 'spautolm'
summary()
spBreg_sac()
# S3 method for class 'MCMC_sar_g'
impacts()
# S3 method for class 'MCMC_sem_g'
impacts()
# S3 method for class 'MCMC_sac_g'
impacts()
spBreg_err()
spBreg_lag()
# S3 method for class 'SLX'
predict()
lmSLX()
# S3 method for class 'SLX'
impacts()
create_WX()
# S3 method for class 'sarlm'
anova()
bptest.sarlm()
errorsarlm()
# S3 method for class 'sarlm'
impacts()
lagsarlm()
# S3 method for class 'sarlm'
predict()
# S3 method for class 'sarlm.pred'
print()
# S3 method for class 'sarlm.pred'
as.data.frame()
# S3 method for class 'sarlm'
residuals()
# S3 method for class 'sarlm'
deviance()
# S3 method for class 'sarlm'
coef()
# S3 method for class 'sarlm'
vcov()
# S3 method for class 'sarlm'
fitted()
sacsarlm()
# S3 method for class 'sarlm'
summary()
# S3 method for class 'sarlm'
print()
# S3 method for class 'summary.sarlm'
print()
trW()

Details

Model-fitting functions and functions supporting model fitting have been moved to the spatialreg package.

See also