summary method for class sarlm

Methods used for presenting the results of estimating spatial SAR models.

# S3 method for sarlm
summary(object, correlation = FALSE, Nagelkerke = FALSE, Hausman=FALSE, adj.se=FALSE, ...)
# S3 method for sarlm
print(x, ...)
# S3 method for summary.sarlm
print(x, digits = max(5, .Options$digits - 3),
        signif.stars = FALSE, ...)

Arguments

object
sarlm object from lagsarlm or errorsarlm

correlation
logical; if 'TRUE', the correlation matrix of the estimated parameters including sigma is returned and printed (default=FALSE)

Nagelkerke
if TRUE, the Nagelkerke pseudo R-squared is reported

Hausman
if TRUE, the results of the Hausman test for error models are reported

adj.se
if TRUE, adjust the coefficient standard errors for the number of fitted coefficients

x
sarlm object from lagsarlm or errorsarlm in print.sarlm, summary object from summary.sarlm for print.summary.sarlm

digits
the number of significant digits to use when printing

signif.stars
logical. If TRUE, "significance stars" are printed for each coefficient.

…
further arguments passed to or from other methods

Value

The summary function summary.sarlm returns the sarlm object augmented with a coefficient matrix with probability values for coefficient asymptotic standard errors for type="error" and for type="lag" or "mixed" when object\$ase=TRUE, or a coefficient matrix with probability values for likelihood ratio tests between the model as reported and models with independent variables dropped in turn.

References

Cliff, A. D., Ord, J. K. 1981 Spatial processes, Pion; Ord, J. K. 1975 Estimation methods for models of spatial interaction, Journal of the American Statistical Association, 70, 120-126; Anselin, L. 1988 Spatial econometrics: methods and models. (Dordrecht: Kluwer); Anselin, L. 1995 SpaceStat, a software program for the analysis of spatial data, version 1.80. Regional Research Institute, West Virginia University, Morgantown, WV (www.spacestat.com); Anselin L, Bera AK (1998) Spatial dependence in linear regression models with an introduction to spatial econometrics. In: Ullah A, Giles DEA (eds) Handbook of applied economic statistics. Marcel Dekker, New York, pp. 237-289; Nagelkerke NJD (1991) A note on a general definition of the coefficient of determination. Biometrika 78: 691-692.

See also

Examples

data(oldcol)
COL.mix.eig <- lagsarlm(CRIME ~ INC + HOVAL, data=COL.OLD,
 nb2listw(COL.nb), type="mixed", method="eigen")
summary(COL.mix.eig, correlation=TRUE, Nagelkerke=TRUE)

#> 
#> Call:
#> lagsarlm(formula = CRIME ~ INC + HOVAL, data = COL.OLD, listw = nb2listw(COL.nb), 
#>     type = "mixed", method = "eigen")
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -37.47829  -6.46731  -0.33835   6.05200  22.62969 
#> 
#> Type: mixed 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) 42.822415  12.667205  3.3806 0.0007233
#> INC         -0.914223   0.331094 -2.7612 0.0057586
#> HOVAL       -0.293738   0.089212 -3.2926 0.0009927
#> lag.INC     -0.520284   0.565129 -0.9206 0.3572355
#> lag.HOVAL    0.245640   0.178917  1.3729 0.1697756
#> 
#> Rho: 0.42634, LR test value: 5.3693, p-value: 0.020494
#> Asymptotic standard error: 0.15623
#>     z-value: 2.7288, p-value: 0.0063561
#> Wald statistic: 7.4465, p-value: 0.0063561
#> 
#> Log likelihood: -181.3935 for mixed model
#> ML residual variance (sigma squared): 91.791, (sigma: 9.5808)
#> Nagelkerke pseudo-R-squared: 0.6494 
#> Number of observations: 49 
#> Number of parameters estimated: 7 
#> AIC: 376.79, (AIC for lm: 380.16)
#> LM test for residual autocorrelation
#> test value: 0.28919, p-value: 0.59074
#> 
#>  Correlation of coefficients 
#>             sigma rho   (Intercept) INC   HOVAL lag.INC
#> rho         -0.18                                      
#> (Intercept)  0.16 -0.89                                
#> INC         -0.03  0.14 -0.19                          
#> HOVAL        0.02 -0.09  0.03       -0.45              
#> lag.INC     -0.09  0.49 -0.53       -0.36  0.05        
#> lag.HOVAL   -0.04  0.19 -0.36        0.19 -0.24 -0.41  
#> 

COL.mix.M <- lagsarlm(CRIME ~ INC + HOVAL, data=COL.OLD,
 nb2listw(COL.nb), type="mixed", method="Matrix")
summary(COL.mix.M, correlation=TRUE, Nagelkerke=TRUE)

#> 
#> Call:
#> lagsarlm(formula = CRIME ~ INC + HOVAL, data = COL.OLD, listw = nb2listw(COL.nb), 
#>     type = "mixed", method = "Matrix")
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -37.47829  -6.46731  -0.33835   6.05200  22.62969 
#> 
#> Type: mixed 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) 42.822413  12.667204  3.3806 0.0007233
#> INC         -0.914223   0.331094 -2.7612 0.0057586
#> HOVAL       -0.293738   0.089212 -3.2926 0.0009927
#> lag.INC     -0.520283   0.565129 -0.9206 0.3572355
#> lag.HOVAL    0.245640   0.178917  1.3729 0.1697756
#> 
#> Rho: 0.42634, LR test value: 5.3693, p-value: 0.020494
#> Asymptotic standard error: 0.15623
#>     z-value: 2.7288, p-value: 0.0063561
#> Wald statistic: 7.4465, p-value: 0.0063561
#> 
#> Log likelihood: -181.3935 for mixed model
#> ML residual variance (sigma squared): 91.791, (sigma: 9.5808)
#> Nagelkerke pseudo-R-squared: 0.6494 
#> Number of observations: 49 
#> Number of parameters estimated: 7 
#> AIC: 376.79, (AIC for lm: 380.16)
#> LM test for residual autocorrelation
#> test value: 0.28919, p-value: 0.59074
#> 
#>  Correlation of coefficients 
#>             sigma rho   (Intercept) INC   HOVAL lag.INC
#> rho         -0.18                                      
#> (Intercept)  0.16 -0.89                                
#> INC         -0.03  0.14 -0.19                          
#> HOVAL        0.02 -0.09  0.03       -0.45              
#> lag.INC     -0.09  0.49 -0.53       -0.36  0.05        
#> lag.HOVAL   -0.04  0.19 -0.36        0.19 -0.24 -0.41  
#> 

COL.errW.eig <- errorsarlm(CRIME ~ INC + HOVAL, data=COL.OLD,
  nb2listw(COL.nb, style="W"), method="eigen")
summary(COL.errW.eig, correlation=TRUE, Nagelkerke=TRUE, Hausman=TRUE)

#> 
#> Call:
#> errorsarlm(formula = CRIME ~ INC + HOVAL, data = COL.OLD, listw = nb2listw(COL.nb, 
#>     style = "W"), method = "eigen")
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -34.81174  -6.44031  -0.72142   7.61476  23.33626 
#> 
#> Type: error 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) 59.893219   5.366163 11.1613 < 2.2e-16
#> INC         -0.941312   0.330569 -2.8476 0.0044057
#> HOVAL       -0.302250   0.090476 -3.3407 0.0008358
#> 
#> Lambda: 0.56179, LR test value: 7.9935, p-value: 0.0046945
#> Asymptotic standard error: 0.13387
#>     z-value: 4.1966, p-value: 2.7098e-05
#> Wald statistic: 17.611, p-value: 2.7098e-05
#> 
#> Log likelihood: -183.3805 for error model
#> ML residual variance (sigma squared): 95.575, (sigma: 9.7762)
#> Nagelkerke pseudo-R-squared: 0.61978 
#> Number of observations: 49 
#> Number of parameters estimated: 5 
#> AIC: 376.76, (AIC for lm: 382.75)
#> Hausman test: 4.902, df: 3, p-value: 0.17911
#> 
#>  Correlation of coefficients 
#>             sigma lambda (Intercept) INC  
#> lambda      -0.24                         
#> (Intercept)  0.00  0.00                   
#> INC          0.00  0.00  -0.56            
#> HOVAL        0.00  0.00  -0.26       -0.45
#>