MCMC sample from fitted spatial regression

The MCMCsamp method uses rwmetrop, a random walk Metropolis algorithm, from LearnBayes to make MCMC samples from fitted maximum likelihood spatial regression models.

MCMCsamp(object, mcmc = 1L, verbose = NULL, ...)
# S3 method for spautolm
MCMCsamp(object, mcmc = 1L, verbose = NULL, ...,
 burnin = 0L, scale=1, listw, control = list())
# S3 method for sarlm
MCMCsamp(object, mcmc = 1L, verbose = NULL, ...,
    burnin=0L, scale=1, listw, listw2=NULL, control=list())

Arguments

object
A spatial regression model object fitted by maximum likelihood with spautolm

mcmc
The number of MCMC iterations after burnin

verbose
default NULL, use global option value; if TRUE, reports progress

…
Arguments passed through

burnin
The number of burn-in iterations for the sampler

scale
a positive scale parameter

listw, listw2
listw objects created for example by nb2listw; should be the same object(s) used for fitting the model

control
list of extra control arguments - see spautolm

Value

An object of class “mcmc” suited to coda, with attributes: “accept” acceptance rate; “type” input ML fitted model type “SAR”, “CAR”, “SMA”, “lag”, “mixed”, “error”, “sac”, “sacmixed”; “timings” run times

Note

If the acceptance rate is below 0.05, a warning will be issued; consider increasing mcmc.

References

Jim Albert (2007) Bayesian Computation with R, Springer, New York, pp. 104-105.

See also

Examples

run <- FALSE
if (require("sf", quietly=TRUE)) run <- TRUE

if (run) {
nydata <- st_read(system.file("shapes/NY8_bna_utm18.gpkg", package="spData")[1], quiet=TRUE)
suppressMessages(nyadjmat <- as.matrix(foreign::read.dbf(system.file(
 "misc/nyadjwts.dbf", package="spData")[1])[-1]))
suppressMessages(ID <- as.character(names(foreign::read.dbf(system.file(
 "misc/nyadjwts.dbf", package="spData")[1]))[-1]))
identical(substring(ID, 2, 10), substring(as.character(nydata$AREAKEY), 2, 10))
}

#> [1] TRUE

if (run) {
esar1f <- spautolm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, family="SAR", method="eigen")
summary(esar1f)
}

#> 
#> Call: 
#> spautolm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, family = "SAR", method = "eigen")
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.56754 -0.38239 -0.02643  0.33109  4.01219 
#> 
#> Coefficients: 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.618193   0.176784 -3.4969 0.0004707
#> PEXPOSURE    0.071014   0.042051  1.6888 0.0912635
#> PCTAGE65P    3.754200   0.624722  6.0094 1.862e-09
#> PCTOWNHOME  -0.419890   0.191329 -2.1946 0.0281930
#> 
#> Lambda: 0.040487 LR test value: 5.2438 p-value: 0.022026 
#> Numerical Hessian standard error of lambda: 0.017199 
#> 
#> Log likelihood: -276.1069 
#> ML residual variance (sigma squared): 0.41388, (sigma: 0.64333)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 564.21
#> 

if (run) {
res <- MCMCsamp(esar1f, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
}

#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.04282 0.01677 0.0002371      0.0009019
#> (Intercept) -0.63718 0.17466 0.0024700      0.0093910
#> PEXPOSURE    0.07097 0.04455 0.0006301      0.0024970
#> PCTAGE65P    3.77933 0.59473 0.0084107      0.0324052
#> PCTOWNHOME  -0.39580 0.19705 0.0027867      0.0107066
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%      25%      50%      75%    97.5%
#> lambda       0.01016  0.03078  0.04257  0.05538  0.07346
#> (Intercept) -0.98316 -0.75204 -0.63947 -0.52801 -0.28626
#> PEXPOSURE   -0.01615  0.04145  0.06955  0.10065  0.16295
#> PCTAGE65P    2.62044  3.38224  3.76076  4.18324  4.96386
#> PCTOWNHOME  -0.80735 -0.53110 -0.39439 -0.25430 -0.03386
#> 

if (run) {
esar1fw <- spautolm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, weights=POP8, family="SAR", method="eigen")
summary(esar1fw)
}

#> 
#> Call: 
#> spautolm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, weights = POP8, family = "SAR", method = "eigen")
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.48488 -0.26823  0.09489  0.46552  4.28343 
#> 
#> Coefficients: 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.797063   0.144054 -5.5331 3.146e-08
#> PEXPOSURE    0.080545   0.028334  2.8428  0.004473
#> PCTAGE65P    3.816731   0.576037  6.6258 3.453e-11
#> PCTOWNHOME  -0.380778   0.156507 -2.4330  0.014975
#> 
#> Lambda: 0.0095636 LR test value: 0.32665 p-value: 0.56764 
#> Numerical Hessian standard error of lambda: 0.016466 
#> 
#> Log likelihood: -251.6017 
#> ML residual variance (sigma squared): 1104.1, (sigma: 33.229)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 515.2
#> 

if (run) {
res <- MCMCsamp(esar1fw, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
}

#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.01377 0.01712 0.0002421       0.001026
#> (Intercept) -0.81191 0.15140 0.0021411       0.009379
#> PEXPOSURE    0.08204 0.03115 0.0004405       0.001885
#> PCTAGE65P    3.82425 0.62075 0.0087787       0.037782
#> PCTOWNHOME  -0.37040 0.16288 0.0023035       0.009650
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%       25%      50%      75%    97.5%
#> lambda      -0.01944  0.001427  0.01375  0.02547  0.04651
#> (Intercept) -1.12146 -0.913920 -0.80993 -0.71103 -0.52526
#> PEXPOSURE    0.02483  0.061976  0.07970  0.10285  0.14663
#> PCTAGE65P    2.51714  3.458128  3.82721  4.25531  4.99891
#> PCTOWNHOME  -0.68368 -0.481408 -0.37068 -0.25842 -0.04238
#> 

if (run) {
ecar1f <- spautolm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, family="CAR", method="eigen")
summary(ecar1f)
}

#> 
#> Call: 
#> spautolm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, family = "CAR", method = "eigen")
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.539732 -0.384311 -0.030646  0.335126  3.808848 
#> 
#> Coefficients: 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.648362   0.181129 -3.5796 0.0003442
#> PEXPOSURE    0.077899   0.043692  1.7829 0.0745986
#> PCTAGE65P    3.703830   0.627185  5.9055 3.516e-09
#> PCTOWNHOME  -0.382789   0.195564 -1.9574 0.0503053
#> 
#> Lambda: 0.084123 LR test value: 5.8009 p-value: 0.016018 
#> Numerical Hessian standard error of lambda: 0.030868 
#> 
#> Log likelihood: -275.8283 
#> ML residual variance (sigma squared): 0.40758, (sigma: 0.63842)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 563.66
#> 

if (run) {
res <- MCMCsamp(ecar1f, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
}

#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.08353 0.03054 0.0004319       0.001873
#> (Intercept) -0.64806 0.22033 0.0031159       0.014431
#> PEXPOSURE    0.07789 0.04818 0.0006814       0.002820
#> PCTAGE65P    3.69287 0.62479 0.0088358       0.039091
#> PCTOWNHOME  -0.38888 0.24264 0.0034314       0.015695
#> 
#> 2. Quantiles for each variable:
#> 
#>                  2.5%      25%      50%     75%   97.5%
#> lambda       0.019452  0.06245  0.08663  0.1065  0.1343
#> (Intercept) -1.116582 -0.79139 -0.63734 -0.4999 -0.2616
#> PEXPOSURE   -0.009699  0.04352  0.07719  0.1089  0.1755
#> PCTAGE65P    2.444603  3.27215  3.69336  4.1157  4.9037
#> PCTOWNHOME  -0.851388 -0.54402 -0.39697 -0.2351  0.1369
#> 

if (run) {
esar1fw <- spautolm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, weights=POP8, family="SAR", method="eigen")
summary(esar1fw)
}

#> 
#> Call: 
#> spautolm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, weights = POP8, family = "SAR", method = "eigen")
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.48488 -0.26823  0.09489  0.46552  4.28343 
#> 
#> Coefficients: 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.797063   0.144054 -5.5331 3.146e-08
#> PEXPOSURE    0.080545   0.028334  2.8428  0.004473
#> PCTAGE65P    3.816731   0.576037  6.6258 3.453e-11
#> PCTOWNHOME  -0.380778   0.156507 -2.4330  0.014975
#> 
#> Lambda: 0.0095636 LR test value: 0.32665 p-value: 0.56764 
#> Numerical Hessian standard error of lambda: 0.016466 
#> 
#> Log likelihood: -251.6017 
#> ML residual variance (sigma squared): 1104.1, (sigma: 33.229)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 515.2
#> 

if (run) {
res <- MCMCsamp(esar1fw, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
}

#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.01259 0.01782 0.0002520       0.001141
#> (Intercept) -0.80532 0.14617 0.0020671       0.008324
#> PEXPOSURE    0.08237 0.03081 0.0004357       0.001906
#> PCTAGE65P    3.80933 0.57546 0.0081383       0.032882
#> PCTOWNHOME  -0.36919 0.16378 0.0023162       0.009787
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%        25%      50%      75%    97.5%
#> lambda      -0.02568  0.0006598  0.01327  0.02496  0.04539
#> (Intercept) -1.08505 -0.9110284 -0.80614 -0.70236 -0.52086
#> PEXPOSURE    0.02818  0.0597828  0.08038  0.10351  0.14596
#> PCTAGE65P    2.63669  3.4194302  3.81151  4.21190  4.94661
#> PCTOWNHOME  -0.68432 -0.4807121 -0.37347 -0.26044 -0.03870
#> 

if (run) {
ecar1fw <- spautolm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, weights=POP8, family="CAR", method="eigen")
summary(ecar1fw)
}

#> 
#> Call: 
#> spautolm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, weights = POP8, family = "CAR", method = "eigen")
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.491042 -0.270906  0.081435  0.451556  4.198134 
#> 
#> Coefficients: 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.790154   0.144862 -5.4545 4.910e-08
#> PEXPOSURE    0.081922   0.028593  2.8651  0.004169
#> PCTAGE65P    3.825858   0.577720  6.6223 3.536e-11
#> PCTOWNHOME  -0.386820   0.157436 -2.4570  0.014010
#> 
#> Lambda: 0.022419 LR test value: 0.38785 p-value: 0.53343 
#> Numerical Hessian standard error of lambda: 0.038543 
#> 
#> Log likelihood: -251.5711 
#> ML residual variance (sigma squared): 1102.9, (sigma: 33.21)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 515.14
#> 

if (run) {
res <- MCMCsamp(ecar1fw, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
}

#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.03740 0.03851 0.0005446       0.002636
#> (Intercept) -0.83163 0.16388 0.0023176       0.009623
#> PEXPOSURE    0.08957 0.03642 0.0005151       0.002463
#> PCTAGE65P    3.78337 0.58847 0.0083223       0.033007
#> PCTOWNHOME  -0.36137 0.17291 0.0024454       0.010419
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%      25%      50%      75%    97.5%
#> lambda      -0.04388  0.01189  0.03822  0.06419  0.10769
#> (Intercept) -1.18212 -0.93085 -0.82464 -0.71936 -0.53673
#> PEXPOSURE    0.02479  0.06562  0.08729  0.10979  0.17830
#> PCTAGE65P    2.65196  3.40302  3.78923  4.15970  4.94874
#> PCTOWNHOME  -0.67284 -0.47748 -0.37590 -0.24927  0.01155
#> 

if (run) {
esar0 <- errorsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY)
summary(esar0)
}

#> 
#> Call:errorsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, 
#>     data = nydata, listw = listw_NY)
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.56754 -0.38239 -0.02643  0.33109  4.01219 
#> 
#> Type: error 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.618193   0.176784 -3.4969 0.0004707
#> PEXPOSURE    0.071014   0.042051  1.6888 0.0912635
#> PCTAGE65P    3.754200   0.624722  6.0094 1.862e-09
#> PCTOWNHOME  -0.419890   0.191329 -2.1946 0.0281930
#> 
#> Lambda: 0.040487, LR test value: 5.2438, p-value: 0.022026
#> Asymptotic standard error: 0.016214
#>     z-value: 2.4971, p-value: 0.01252
#> Wald statistic: 6.2356, p-value: 0.01252
#> 
#> Log likelihood: -276.1069 for error model
#> ML residual variance (sigma squared): 0.41388, (sigma: 0.64333)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 564.21, (AIC for lm: 567.46)
#> 

if (run) {
res <- MCMCsamp(esar0, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
}

#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.04321 0.01626 0.0002300      0.0009438
#> (Intercept) -0.61861 0.18662 0.0026392      0.0112293
#> PEXPOSURE    0.06892 0.04457 0.0006303      0.0026865
#> PCTAGE65P    3.68728 0.61256 0.0086629      0.0322578
#> PCTOWNHOME  -0.38822 0.20491 0.0028979      0.0120127
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%      25%      50%      75%     97.5%
#> lambda       0.01234  0.03192  0.04278  0.05444  0.074958
#> (Intercept) -0.99402 -0.74534 -0.62179 -0.49032 -0.261240
#> PEXPOSURE   -0.01256  0.03776  0.06683  0.09668  0.162983
#> PCTAGE65P    2.41938  3.27690  3.71353  4.11049  4.864336
#> PCTOWNHOME  -0.76388 -0.53112 -0.39057 -0.25129  0.009339
#> 

if (run) {
esar0w <- errorsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, weights=POP8)
summary(esar0)
}

#> 
#> Call:errorsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, 
#>     data = nydata, listw = listw_NY)
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.56754 -0.38239 -0.02643  0.33109  4.01219 
#> 
#> Type: error 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.618193   0.176784 -3.4969 0.0004707
#> PEXPOSURE    0.071014   0.042051  1.6888 0.0912635
#> PCTAGE65P    3.754200   0.624722  6.0094 1.862e-09
#> PCTOWNHOME  -0.419890   0.191329 -2.1946 0.0281930
#> 
#> Lambda: 0.040487, LR test value: 5.2438, p-value: 0.022026
#> Asymptotic standard error: 0.016214
#>     z-value: 2.4971, p-value: 0.01252
#> Wald statistic: 6.2356, p-value: 0.01252
#> 
#> Log likelihood: -276.1069 for error model
#> ML residual variance (sigma squared): 0.41388, (sigma: 0.64333)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 564.21, (AIC for lm: 567.46)
#> 

if (run) {
res <- MCMCsamp(esar0w, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
}

#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.01346 0.01546 0.0002186      0.0008649
#> (Intercept) -0.79989 0.14872 0.0021033      0.0091383
#> PEXPOSURE    0.08318 0.03005 0.0004250      0.0018445
#> PCTAGE65P    3.73680 0.58421 0.0082619      0.0329871
#> PCTOWNHOME  -0.36902 0.16294 0.0023044      0.0104440
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%       25%      50%      75%    97.5%
#> lambda      -0.01787  0.003276  0.01411  0.02357  0.04293
#> (Intercept) -1.09013 -0.900266 -0.79020 -0.69811 -0.51735
#> PEXPOSURE    0.02355  0.063539  0.08350  0.10321  0.14078
#> PCTAGE65P    2.56445  3.337293  3.74391  4.14984  4.83605
#> PCTOWNHOME  -0.68213 -0.479511 -0.37838 -0.26449 -0.02673
#> 

if (run) {
esar1 <- errorsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, etype="emixed")
summary(esar1)
}

#> 
#> Call:errorsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, 
#>     data = nydata, listw = listw_NY, etype = "emixed")
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.81562 -0.37641 -0.02224  0.33638  4.00054 
#> 
#> Type: error 
#> Coefficients: (asymptotic standard errors) 
#>                  Estimate Std. Error z value  Pr(>|z|)
#> (Intercept)     -1.118019   0.247425 -4.5186 6.225e-06
#> PEXPOSURE        0.218279   0.079245  2.7545  0.005879
#> PCTAGE65P        3.416477   0.645587  5.2920 1.210e-07
#> PCTOWNHOME       0.036593   0.249835  0.1465  0.883551
#> lag.(Intercept)  0.121515   0.057636  2.1083  0.035003
#> lag.PEXPOSURE   -0.035075   0.015943 -2.2000  0.027808
#> lag.PCTAGE65P    0.263096   0.220118  1.1953  0.231989
#> lag.PCTOWNHOME  -0.155680   0.059213 -2.6291  0.008560
#> 
#> Lambda: 0.022723, LR test value: 1.6846, p-value: 0.19432
#> Asymptotic standard error: 0.017169
#>     z-value: 1.3235, p-value: 0.18567
#> Wald statistic: 1.7516, p-value: 0.18567
#> 
#> Log likelihood: -269.5398 for error model
#> ML residual variance (sigma squared): 0.39759, (sigma: 0.63055)
#> Number of observations: 281 
#> Number of parameters estimated: 10 
#> AIC: 559.08, (AIC for lm: 558.76)
#> 

if (run) {
res <- MCMCsamp(esar1, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
}

#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                     Mean      SD  Naive SE Time-series SE
#> lambda           0.02987 0.01917 0.0002711       0.001526
#> (Intercept)     -1.11364 0.24665 0.0034882       0.018097
#> PEXPOSURE        0.22144 0.08566 0.0012114       0.006763
#> PCTAGE65P        3.40696 0.62628 0.0088569       0.046242
#> PCTOWNHOME       0.03632 0.23679 0.0033488       0.017021
#> lag.(Intercept)  0.11612 0.05829 0.0008243       0.004343
#> lag.PEXPOSURE   -0.03430 0.01704 0.0002410       0.001319
#> lag.PCTAGE65P    0.24112 0.22755 0.0032180       0.017189
#> lag.PCTOWNHOME  -0.14705 0.06347 0.0008975       0.005022
#> 
#> 2. Quantiles for each variable:
#> 
#>                      2.5%      25%      50%      75%     97.5%
#> lambda          -0.011839  0.01692  0.03112  0.04328  0.065742
#> (Intercept)     -1.628272 -1.27164 -1.12205 -0.93901 -0.663981
#> PEXPOSURE        0.061283  0.16044  0.21897  0.28335  0.395202
#> PCTAGE65P        2.258838  2.98669  3.40917  3.79310  4.693603
#> PCTOWNHOME      -0.417539 -0.12695  0.03076  0.20076  0.513815
#> lag.(Intercept) -0.004942  0.07969  0.11545  0.15287  0.229647
#> lag.PEXPOSURE   -0.071662 -0.04540 -0.03412 -0.02360 -0.002453
#> lag.PCTAGE65P   -0.232816  0.09088  0.24395  0.38533  0.690176
#> lag.PCTOWNHOME  -0.272333 -0.19060 -0.14538 -0.10622 -0.014337
#> 

if (run) {
lsar0 <- lagsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY)
summary(lsar0)
}

#> 
#> Call:lagsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY)
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.586752 -0.391580 -0.022469  0.338017  4.029430 
#> 
#> Type: lag 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.514495   0.156154 -3.2948  0.000985
#> PEXPOSURE    0.047627   0.034509  1.3801  0.167542
#> PCTAGE65P    3.648198   0.599046  6.0900 1.129e-09
#> PCTOWNHOME  -0.414601   0.169554 -2.4453  0.014475
#> 
#> Rho: 0.038893, LR test value: 6.9683, p-value: 0.0082967
#> Asymptotic standard error: 0.015053
#>     z-value: 2.5837, p-value: 0.0097755
#> Wald statistic: 6.6754, p-value: 0.0097755
#> 
#> Log likelihood: -275.2447 for lag model
#> ML residual variance (sigma squared): 0.41166, (sigma: 0.6416)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 562.49, (AIC for lm: 567.46)
#> LM test for residual autocorrelation
#> test value: 1.4633, p-value: 0.22641
#> 

if (run) {
res <- MCMCsamp(lsar0, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
}

#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> rho          0.03685 0.01446 0.0002044      0.0007793
#> (Intercept) -0.52320 0.15889 0.0022470      0.0090904
#> PEXPOSURE    0.05197 0.03569 0.0005047      0.0021615
#> PCTAGE65P    3.69330 0.61439 0.0086888      0.0359511
#> PCTOWNHOME  -0.43281 0.18261 0.0025825      0.0110830
#> 
#> 2. Quantiles for each variable:
#> 
#>                  2.5%      25%      50%      75%    97.5%
#> rho          0.007429  0.02711  0.03658  0.04701  0.06458
#> (Intercept) -0.845316 -0.62822 -0.52371 -0.41688 -0.21065
#> PEXPOSURE   -0.014143  0.02783  0.05137  0.07548  0.12190
#> PCTAGE65P    2.487654  3.28697  3.71710  4.06781  4.90802
#> PCTOWNHOME  -0.770760 -0.55995 -0.43919 -0.30454 -0.06648
#> 

if (run) {
lsar1 <- lagsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, type="mixed")
summary(lsar1)
}

#> 
#> Call:lagsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, type = "mixed")
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.799308 -0.390125 -0.021371  0.346128  3.965251 
#> 
#> Type: mixed 
#> Coefficients: (asymptotic standard errors) 
#>                  Estimate Std. Error z value  Pr(>|z|)
#> (Intercept)     -1.131233   0.249631 -4.5316 5.853e-06
#> PEXPOSURE        0.218364   0.079301  2.7536  0.005894
#> PCTAGE65P        3.361158   0.654123  5.1384 2.771e-07
#> PCTOWNHOME       0.071903   0.253967  0.2831  0.777085
#> lag.(Intercept)  0.132544   0.056175  2.3595  0.018300
#> lag.PEXPOSURE   -0.035239   0.015536 -2.2681  0.023322
#> lag.PCTAGE65P    0.161685   0.223690  0.7228  0.469798
#> lag.PCTOWNHOME  -0.140681   0.058529 -2.4036  0.016234
#> 
#> Rho: 0.026981, LR test value: 2.558, p-value: 0.10974
#> Asymptotic standard error: 0.016766
#>     z-value: 1.6093, p-value: 0.10755
#> Wald statistic: 2.5899, p-value: 0.10755
#> 
#> Log likelihood: -269.1031 for mixed model
#> ML residual variance (sigma squared): 0.39587, (sigma: 0.62918)
#> Number of observations: 281 
#> Number of parameters estimated: 10 
#> AIC: 558.21, (AIC for lm: 558.76)
#> LM test for residual autocorrelation
#> test value: 4.908, p-value: 0.026732
#> 

if (run) {
res <- MCMCsamp(lsar1, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
}

#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                     Mean      SD  Naive SE Time-series SE
#> rho              0.02842 0.01664 0.0002353       0.001361
#> (Intercept)     -1.14133 0.25378 0.0035890       0.019366
#> PEXPOSURE        0.22681 0.07778 0.0011000       0.005754
#> PCTAGE65P        3.33073 0.67471 0.0095419       0.052853
#> PCTOWNHOME       0.08334 0.25072 0.0035457       0.019503
#> lag.(Intercept)  0.13422 0.05742 0.0008120       0.004652
#> lag.PEXPOSURE   -0.03649 0.01485 0.0002100       0.001076
#> lag.PCTAGE65P    0.15563 0.22755 0.0032181       0.018921
#> lag.PCTOWNHOME  -0.13894 0.05988 0.0008469       0.004910
#> 
#> 2. Quantiles for each variable:
#> 
#>                      2.5%      25%      50%      75%     97.5%
#> rho             -0.005507  0.01703  0.02841  0.04005  0.059096
#> (Intercept)     -1.615696 -1.34037 -1.15079 -0.92353 -0.674610
#> PEXPOSURE        0.079840  0.17944  0.22357  0.27914  0.382500
#> PCTAGE65P        1.968482  2.86014  3.33212  3.78840  4.640799
#> PCTOWNHOME      -0.372715 -0.10111  0.08690  0.24758  0.589208
#> lag.(Intercept)  0.020615  0.09473  0.13684  0.17135  0.246722
#> lag.PEXPOSURE   -0.065608 -0.04878 -0.03552 -0.02716 -0.006812
#> lag.PCTAGE65P   -0.277453 -0.01477  0.14652  0.31023  0.624269
#> lag.PCTOWNHOME  -0.267898 -0.17783 -0.13766 -0.09604 -0.036953
#> 

if (run) {
ssar0 <- sacsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY)
summary(ssar0)
}

#> 
#> Call:sacsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY)
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.468382 -0.375687 -0.034996  0.314714  3.833950 
#> 
#> Type: sac 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.386572   0.123188 -3.1381  0.001701
#> PEXPOSURE    0.026684   0.024013  1.1112  0.266479
#> PCTAGE65P    3.089824   0.562851  5.4896 4.029e-08
#> PCTOWNHOME  -0.323052   0.137449 -2.3503  0.018756
#> 
#> Rho: 0.089451
#> Asymptotic standard error: 0.019427
#>     z-value: 4.6046, p-value: 4.1325e-06
#> Lambda: -0.08192
#> Asymptotic standard error: 0.033201
#>     z-value: -2.4674, p-value: 0.01361
#> 
#> LR test value: 10.114, p-value: 0.0063661
#> 
#> Log likelihood: -273.672 for sac model
#> ML residual variance (sigma squared): 0.3766, (sigma: 0.61368)
#> Number of observations: 281 
#> Number of parameters estimated: 7 
#> AIC: 561.34, (AIC for lm: 567.46)
#> 

if (run) {
res <- MCMCsamp(ssar0, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
}

#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> rho         -0.06763 0.06153 0.0008702        0.02099
#> lambda       0.09096 0.05870 0.0008302        0.02096
#> (Intercept) -0.84256 0.25498 0.0036060        0.03869
#> PEXPOSURE    0.14285 0.06811 0.0009633        0.01186
#> PCTAGE65P    3.29653 0.62517 0.0088413        0.05123
#> PCTOWNHOME  -0.18582 0.22428 0.0031718        0.02641
#> 
#> 2. Quantiles for each variable:
#> 
#>                  2.5%      25%     50%      75%    97.5%
#> rho         -0.140058 -0.10615 -0.0853 -0.05694  0.08943
#> lambda      -0.079002  0.09258  0.1126  0.12465  0.13941
#> (Intercept) -1.288107 -1.02061 -0.8538 -0.68276 -0.30895
#> PEXPOSURE    0.003802  0.09818  0.1496  0.19483  0.25802
#> PCTAGE65P    2.150245  2.82707  3.2788  3.73703  4.54430
#> PCTOWNHOME  -0.572328 -0.34388 -0.1971 -0.04495  0.28430
#> 

if (run) {
ssar1 <- sacsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, type="sacmixed")
summary(ssar1)
}

#> 
#> Call:sacsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, type = "sacmixed")
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.633958 -0.363826 -0.019927  0.348238  3.655509 
#> 
#> Type: sacmixed 
#> Coefficients: (asymptotic standard errors) 
#>                  Estimate Std. Error z value  Pr(>|z|)
#> (Intercept)     -1.133298   0.247495 -4.5791 4.670e-06
#> PEXPOSURE        0.206963   0.074480  2.7788  0.005456
#> PCTAGE65P        3.083983   0.671081  4.5955 4.316e-06
#> PCTOWNHOME       0.174800   0.256280  0.6821  0.495196
#> lag.(Intercept)  0.153427   0.050817  3.0192  0.002534
#> lag.PEXPOSURE   -0.033400   0.013817 -2.4173  0.015634
#> lag.PCTAGE65P   -0.079738   0.222144 -0.3589  0.719634
#> lag.PCTOWNHOME  -0.102502   0.056760 -1.8059  0.070940
#> 
#> Rho: 0.092495
#> Asymptotic standard error: 0.023829
#>     z-value: 3.8817, p-value: 0.00010375
#> Lambda: -0.091069
#> Asymptotic standard error: 0.038431
#>     z-value: -2.3697, p-value: 0.017804
#> 
#> LR test value: 22.379, p-value: 0.0010335
#> 
#> Log likelihood: -267.5392 for sacmixed model
#> ML residual variance (sigma squared): 0.35617, (sigma: 0.5968)
#> Number of observations: 281 
#> Number of parameters estimated: 11 
#> AIC: 557.08, (AIC for lm: 567.46)
#> 

if (run) {
res <- MCMCsamp(ssar1, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
}

#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                     Mean      SD  Naive SE Time-series SE
#> rho             -0.03171 0.06990 0.0009886       0.018115
#> lambda           0.04838 0.06501 0.0009194       0.015765
#> (Intercept)     -1.11526 0.23654 0.0033452       0.017499
#> PEXPOSURE        0.21879 0.06916 0.0009781       0.005318
#> PCTAGE65P        3.38160 0.60900 0.0086126       0.043879
#> PCTOWNHOME       0.03524 0.25162 0.0035584       0.020546
#> lag.(Intercept)  0.09898 0.08024 0.0011348       0.013796
#> lag.PEXPOSURE   -0.03229 0.01650 0.0002334       0.001616
#> lag.PCTAGE65P    0.28722 0.34079 0.0048195       0.047037
#> lag.PCTOWNHOME  -0.14379 0.06452 0.0009125       0.006701
#> 
#> 2. Quantiles for each variable:
#> 
#>                     2.5%       25%      50%      75%      97.5%
#> rho             -0.13903 -0.089325 -0.04183  0.02259  1.041e-01
#> lambda          -0.10206  0.006913  0.06920  0.10151  1.271e-01
#> (Intercept)     -1.58219 -1.272563 -1.09888 -0.97188 -6.426e-01
#> PEXPOSURE        0.08979  0.173908  0.21909  0.26444  3.615e-01
#> PCTAGE65P        2.08316  2.994408  3.41633  3.78876  4.547e+00
#> PCTOWNHOME      -0.45786 -0.134565  0.02518  0.20870  5.404e-01
#> lag.(Intercept) -0.06798  0.041130  0.10490  0.15913  2.310e-01
#> lag.PEXPOSURE   -0.06566 -0.042841 -0.03170 -0.02175  4.235e-05
#> lag.PCTAGE65P   -0.36118  0.035942  0.27981  0.52501  9.295e-01
#> lag.PCTOWNHOME  -0.27612 -0.189819 -0.14133 -0.09996 -1.677e-02
#>