MCMC sample from fitted spatial regression

The MCMCsamp method uses rwmetrop, a random walk Metropolis algorithm, from LearnBayes to make MCMC samples from fitted maximum likelihood spatial regression models.

MCMCsamp(object, mcmc = 1L, verbose = NULL, ...)
# S3 method for spautolm
MCMCsamp(object, mcmc = 1L, verbose = NULL, ...,
 burnin = 0L, scale=1, listw, control = list())
# S3 method for sarlm
MCMCsamp(object, mcmc = 1L, verbose = NULL, ...,
    burnin=0L, scale=1, listw, listw2=NULL, control=list())

Arguments

object
A spatial regression model object fitted by maximum likelihood with spautolm

mcmc
The number of MCMC iterations after burnin

verbose
default NULL, use global option value; if TRUE, reports progress

…
Arguments passed through

burnin
The number of burn-in iterations for the sampler

scale
a positive scale parameter

listw, listw2
listw objects created for example by nb2listw; should be the same object(s) used for fitting the model

control
list of extra control arguments - see spautolm

Value

An object of class “mcmc” suited to coda, with attributes: “accept” acceptance rate; “type” input ML fitted model type “SAR”, “CAR”, “SMA”, “lag”, “mixed”, “error”, “sac”, “sacmixed”; “timings” run times

Note

If the acceptance rate is below 0.05, a warning will be issued; consider increasing mcmc.

References

Jim Albert (2007) Bayesian Computation with R, Springer, New York, pp. 104-105.

See also

Examples

run <- FALSE
if (require("sf", quietly=TRUE)) run <- TRUE

if (run) {
nydata <- st_read(system.file("shapes/NY8_bna_utm18.gpkg", package="spData")[1], quiet=TRUE)
suppressMessages(nyadjmat <- as.matrix(foreign::read.dbf(system.file(
 "misc/nyadjwts.dbf", package="spData")[1])[-1]))
suppressMessages(ID <- as.character(names(foreign::read.dbf(system.file(
 "misc/nyadjwts.dbf", package="spData")[1]))[-1]))
identical(substring(ID, 2, 10), substring(as.character(nydata$AREAKEY), 2, 10))
}

#> [1] TRUE

if (run) {
esar1f <- spautolm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, family="SAR", method="eigen")
summary(esar1f)
}

#> 
#> Call: 
#> spautolm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, family = "SAR", method = "eigen")
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.56754 -0.38239 -0.02643  0.33109  4.01219 
#> 
#> Coefficients: 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.618193   0.176784 -3.4969 0.0004707
#> PEXPOSURE    0.071014   0.042051  1.6888 0.0912635
#> PCTAGE65P    3.754200   0.624722  6.0094 1.862e-09
#> PCTOWNHOME  -0.419890   0.191329 -2.1946 0.0281930
#> 
#> Lambda: 0.040487 LR test value: 5.2438 p-value: 0.022026 
#> Numerical Hessian standard error of lambda: 0.017199 
#> 
#> Log likelihood: -276.1069 
#> ML residual variance (sigma squared): 0.41388, (sigma: 0.64333)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 564.21
#> 

if (run) {
res <- MCMCsamp(esar1f, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
}

#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.04414 0.01722 0.0002435      0.0009719
#> (Intercept) -0.64198 0.19317 0.0027318      0.0112218
#> PEXPOSURE    0.07395 0.04546 0.0006429      0.0027379
#> PCTAGE65P    3.76061 0.66332 0.0093808      0.0391459
#> PCTOWNHOME  -0.38545 0.20954 0.0029634      0.0119281
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%      25%      50%      75%    97.5%
#> lambda       0.01236  0.03205  0.04401  0.05521  0.07877
#> (Intercept) -1.04205 -0.77055 -0.64438 -0.51201 -0.26715
#> PEXPOSURE   -0.01375  0.04302  0.07280  0.10412  0.16575
#> PCTAGE65P    2.53404  3.27849  3.76919  4.22391  5.09011
#> PCTOWNHOME  -0.77111 -0.53571 -0.39398 -0.24189  0.05586
#> 

if (run) {
esar1fw <- spautolm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, weights=POP8, family="SAR", method="eigen")
summary(esar1fw)
}

#> 
#> Call: 
#> spautolm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, weights = POP8, family = "SAR", method = "eigen")
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.48488 -0.26823  0.09489  0.46552  4.28343 
#> 
#> Coefficients: 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.797063   0.144054 -5.5331 3.146e-08
#> PEXPOSURE    0.080545   0.028334  2.8428  0.004473
#> PCTAGE65P    3.816731   0.576037  6.6258 3.453e-11
#> PCTOWNHOME  -0.380778   0.156507 -2.4330  0.014975
#> 
#> Lambda: 0.0095636 LR test value: 0.32665 p-value: 0.56764 
#> Numerical Hessian standard error of lambda: 0.016466 
#> 
#> Log likelihood: -251.6017 
#> ML residual variance (sigma squared): 1104.1, (sigma: 33.229)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 515.2
#> 

if (run) {
res <- MCMCsamp(esar1fw, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
}

#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.01432 0.01703 0.0002409       0.001034
#> (Intercept) -0.80185 0.15873 0.0022448       0.009546
#> PEXPOSURE    0.08564 0.03294 0.0004658       0.001975
#> PCTAGE65P    3.82593 0.61631 0.0087159       0.036749
#> PCTOWNHOME  -0.38907 0.17679 0.0025002       0.010976
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%       25%      50%      75%    97.5%
#> lambda      -0.02088  0.003317  0.01482  0.02671  0.04682
#> (Intercept) -1.11474 -0.905768 -0.80101 -0.69291 -0.50360
#> PEXPOSURE    0.01391  0.065037  0.08559  0.10871  0.14794
#> PCTAGE65P    2.62410  3.443854  3.84478  4.22642  5.07766
#> PCTOWNHOME  -0.74540 -0.506030 -0.38939 -0.26953 -0.04054
#> 

if (run) {
ecar1f <- spautolm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, family="CAR", method="eigen")
summary(ecar1f)
}

#> 
#> Call: 
#> spautolm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, family = "CAR", method = "eigen")
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.539732 -0.384311 -0.030646  0.335126  3.808848 
#> 
#> Coefficients: 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.648362   0.181129 -3.5796 0.0003442
#> PEXPOSURE    0.077899   0.043692  1.7829 0.0745986
#> PCTAGE65P    3.703830   0.627185  5.9055 3.516e-09
#> PCTOWNHOME  -0.382789   0.195564 -1.9574 0.0503053
#> 
#> Lambda: 0.084123 LR test value: 5.8009 p-value: 0.016018 
#> Numerical Hessian standard error of lambda: 0.030868 
#> 
#> Log likelihood: -275.8283 
#> ML residual variance (sigma squared): 0.40758, (sigma: 0.63842)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 563.66
#> 

if (run) {
res <- MCMCsamp(ecar1f, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
}

#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.08789 0.02738 0.0003872       0.001480
#> (Intercept) -0.66632 0.19546 0.0027643       0.011405
#> PEXPOSURE    0.08279 0.05101 0.0007214       0.003301
#> PCTAGE65P    3.67565 0.64034 0.0090558       0.036333
#> PCTOWNHOME  -0.36726 0.21576 0.0030514       0.012522
#> 
#> 2. Quantiles for each variable:
#> 
#>                  2.5%      25%      50%     75%    97.5%
#> lambda       0.031966  0.06829  0.08958  0.1084  0.13644
#> (Intercept) -1.080640 -0.79075 -0.65955 -0.5292 -0.29414
#> PEXPOSURE   -0.009009  0.04914  0.07981  0.1148  0.19190
#> PCTAGE65P    2.352635  3.27210  3.70343  4.1302  4.89761
#> PCTOWNHOME  -0.765768 -0.51592 -0.38228 -0.2177  0.09219
#> 

if (run) {
esar1fw <- spautolm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, weights=POP8, family="SAR", method="eigen")
summary(esar1fw)
}

#> 
#> Call: 
#> spautolm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, weights = POP8, family = "SAR", method = "eigen")
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.48488 -0.26823  0.09489  0.46552  4.28343 
#> 
#> Coefficients: 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.797063   0.144054 -5.5331 3.146e-08
#> PEXPOSURE    0.080545   0.028334  2.8428  0.004473
#> PCTAGE65P    3.816731   0.576037  6.6258 3.453e-11
#> PCTOWNHOME  -0.380778   0.156507 -2.4330  0.014975
#> 
#> Lambda: 0.0095636 LR test value: 0.32665 p-value: 0.56764 
#> Numerical Hessian standard error of lambda: 0.016466 
#> 
#> Log likelihood: -251.6017 
#> ML residual variance (sigma squared): 1104.1, (sigma: 33.229)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 515.2
#> 

if (run) {
res <- MCMCsamp(esar1fw, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
}

#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.01223 0.01628 0.0002302      0.0008918
#> (Intercept) -0.81008 0.14450 0.0020436      0.0071485
#> PEXPOSURE    0.08333 0.03024 0.0004277      0.0016623
#> PCTAGE65P    3.83015 0.54901 0.0077641      0.0289888
#> PCTOWNHOME  -0.37589 0.15894 0.0022478      0.0088513
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%       25%      50%      75%    97.5%
#> lambda      -0.01950  0.001449  0.01258  0.02253  0.04561
#> (Intercept) -1.10339 -0.904325 -0.80883 -0.70550 -0.54256
#> PEXPOSURE    0.02629  0.061589  0.08190  0.10328  0.14268
#> PCTAGE65P    2.69715  3.471721  3.85731  4.20233  4.83788
#> PCTOWNHOME  -0.67970 -0.487368 -0.37586 -0.27418 -0.05980
#> 

if (run) {
ecar1fw <- spautolm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, weights=POP8, family="CAR", method="eigen")
summary(ecar1fw)
}

#> 
#> Call: 
#> spautolm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, weights = POP8, family = "CAR", method = "eigen")
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.491042 -0.270906  0.081435  0.451556  4.198134 
#> 
#> Coefficients: 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.790154   0.144862 -5.4545 4.910e-08
#> PEXPOSURE    0.081922   0.028593  2.8651  0.004169
#> PCTAGE65P    3.825858   0.577720  6.6223 3.536e-11
#> PCTOWNHOME  -0.386820   0.157436 -2.4570  0.014010
#> 
#> Lambda: 0.022419 LR test value: 0.38785 p-value: 0.53343 
#> Numerical Hessian standard error of lambda: 0.038543 
#> 
#> Log likelihood: -251.5711 
#> ML residual variance (sigma squared): 1102.9, (sigma: 33.21)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 515.14
#> 

if (run) {
res <- MCMCsamp(ecar1fw, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
}

#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.03541 0.03918 0.0005540       0.002337
#> (Intercept) -0.82709 0.16312 0.0023068       0.009943
#> PEXPOSURE    0.08706 0.03268 0.0004622       0.001914
#> PCTAGE65P    3.77190 0.56610 0.0080058       0.030749
#> PCTOWNHOME  -0.36346 0.17131 0.0024227       0.010363
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%       25%      50%      75%     97.5%
#> lambda      -0.04274  0.009742  0.03520  0.06288  0.109157
#> (Intercept) -1.17422 -0.931777 -0.81463 -0.71695 -0.531921
#> PEXPOSURE    0.02690  0.063601  0.08551  0.10732  0.154246
#> PCTAGE65P    2.68839  3.371537  3.80605  4.16482  4.902716
#> PCTOWNHOME  -0.68261 -0.483173 -0.36581 -0.25363  0.007252
#> 

if (run) {
esar0 <- errorsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY)
summary(esar0)
}

#> 
#> Call:errorsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, 
#>     data = nydata, listw = listw_NY)
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.56754 -0.38239 -0.02643  0.33109  4.01219 
#> 
#> Type: error 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.618193   0.176784 -3.4969 0.0004707
#> PEXPOSURE    0.071014   0.042051  1.6888 0.0912635
#> PCTAGE65P    3.754200   0.624722  6.0094 1.862e-09
#> PCTOWNHOME  -0.419890   0.191329 -2.1946 0.0281930
#> 
#> Lambda: 0.040487, LR test value: 5.2438, p-value: 0.022026
#> Asymptotic standard error: 0.016214
#>     z-value: 2.4971, p-value: 0.01252
#> Wald statistic: 6.2356, p-value: 0.01252
#> 
#> Log likelihood: -276.1069 for error model
#> ML residual variance (sigma squared): 0.41388, (sigma: 0.64333)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 564.21, (AIC for lm: 567.46)
#> 

if (run) {
res <- MCMCsamp(esar0, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
}

#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.04495 0.01688 0.0002387       0.001025
#> (Intercept) -0.64223 0.18870 0.0026686       0.012154
#> PEXPOSURE    0.07735 0.04453 0.0006298       0.002607
#> PCTAGE65P    3.76603 0.66687 0.0094310       0.040179
#> PCTOWNHOME  -0.40697 0.21105 0.0029847       0.013534
#> 
#> 2. Quantiles for each variable:
#> 
#>                  2.5%      25%      50%      75%    97.5%
#> lambda       0.011357  0.03407  0.04546  0.05722  0.07653
#> (Intercept) -1.015897 -0.76397 -0.63316 -0.51969 -0.28163
#> PEXPOSURE   -0.008832  0.04832  0.07686  0.10675  0.16231
#> PCTAGE65P    2.544777  3.29569  3.78732  4.18747  5.07729
#> PCTOWNHOME  -0.827336 -0.53941 -0.41621 -0.28127  0.00612
#> 

if (run) {
esar0w <- errorsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, weights=POP8)
summary(esar0)
}

#> 
#> Call:errorsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, 
#>     data = nydata, listw = listw_NY)
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.56754 -0.38239 -0.02643  0.33109  4.01219 
#> 
#> Type: error 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.618193   0.176784 -3.4969 0.0004707
#> PEXPOSURE    0.071014   0.042051  1.6888 0.0912635
#> PCTAGE65P    3.754200   0.624722  6.0094 1.862e-09
#> PCTOWNHOME  -0.419890   0.191329 -2.1946 0.0281930
#> 
#> Lambda: 0.040487, LR test value: 5.2438, p-value: 0.022026
#> Asymptotic standard error: 0.016214
#>     z-value: 2.4971, p-value: 0.01252
#> Wald statistic: 6.2356, p-value: 0.01252
#> 
#> Log likelihood: -276.1069 for error model
#> ML residual variance (sigma squared): 0.41388, (sigma: 0.64333)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 564.21, (AIC for lm: 567.46)
#> 

if (run) {
res <- MCMCsamp(esar0w, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
}

#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.01371 0.01608 0.0002274      0.0008841
#> (Intercept) -0.79983 0.15625 0.0022098      0.0097961
#> PEXPOSURE    0.08457 0.03127 0.0004422      0.0019717
#> PCTAGE65P    3.77770 0.58678 0.0082983      0.0347447
#> PCTOWNHOME  -0.38039 0.16234 0.0022959      0.0098801
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%       25%      50%      75%    97.5%
#> lambda      -0.01760  0.002331  0.01400  0.02489  0.04370
#> (Intercept) -1.08795 -0.915372 -0.79948 -0.69809 -0.47440
#> PEXPOSURE    0.02192  0.064060  0.08316  0.10577  0.14508
#> PCTAGE65P    2.63576  3.380985  3.75396  4.19167  4.92910
#> PCTOWNHOME  -0.69578 -0.488224 -0.37782 -0.26641 -0.04992
#> 

if (run) {
esar1 <- errorsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, etype="emixed")
summary(esar1)
}

#> 
#> Call:errorsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, 
#>     data = nydata, listw = listw_NY, etype = "emixed")
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.81562 -0.37641 -0.02224  0.33638  4.00054 
#> 
#> Type: error 
#> Coefficients: (asymptotic standard errors) 
#>                  Estimate Std. Error z value  Pr(>|z|)
#> (Intercept)     -1.118019   0.247425 -4.5186 6.225e-06
#> PEXPOSURE        0.218279   0.079245  2.7545  0.005879
#> PCTAGE65P        3.416477   0.645587  5.2920 1.210e-07
#> PCTOWNHOME       0.036593   0.249835  0.1465  0.883551
#> lag.(Intercept)  0.121515   0.057636  2.1083  0.035003
#> lag.PEXPOSURE   -0.035075   0.015943 -2.2000  0.027808
#> lag.PCTAGE65P    0.263096   0.220118  1.1953  0.231989
#> lag.PCTOWNHOME  -0.155680   0.059213 -2.6291  0.008560
#> 
#> Lambda: 0.022723, LR test value: 1.6846, p-value: 0.19432
#> Asymptotic standard error: 0.017169
#>     z-value: 1.3235, p-value: 0.18567
#> Wald statistic: 1.7516, p-value: 0.18567
#> 
#> Log likelihood: -269.5398 for error model
#> ML residual variance (sigma squared): 0.39759, (sigma: 0.63055)
#> Number of observations: 281 
#> Number of parameters estimated: 10 
#> AIC: 559.08, (AIC for lm: 558.76)
#> 

if (run) {
res <- MCMCsamp(esar1, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
}

#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                     Mean      SD  Naive SE Time-series SE
#> lambda           0.02913 0.01930 0.0002729       0.001653
#> (Intercept)     -1.10960 0.24984 0.0035332       0.018843
#> PEXPOSURE        0.21468 0.07965 0.0011265       0.005503
#> PCTAGE65P        3.44774 0.69299 0.0098003       0.052067
#> PCTOWNHOME       0.02211 0.27043 0.0038244       0.020961
#> lag.(Intercept)  0.11750 0.06047 0.0008552       0.004552
#> lag.PEXPOSURE   -0.03465 0.01682 0.0002378       0.001290
#> lag.PCTAGE65P    0.25624 0.22310 0.0031551       0.016924
#> lag.PCTOWNHOME  -0.14959 0.06489 0.0009177       0.004977
#> 
#> 2. Quantiles for each variable:
#> 
#>                      2.5%      25%       50%      75%      97.5%
#> lambda          -0.013140  0.01783  0.031084  0.04151  6.345e-02
#> (Intercept)     -1.604602 -1.28705 -1.130482 -0.92777 -6.296e-01
#> PEXPOSURE        0.051199  0.15874  0.220576  0.26791  3.781e-01
#> PCTAGE65P        2.123696  2.98456  3.436155  3.93723  4.774e+00
#> PCTOWNHOME      -0.498821 -0.16627  0.007635  0.20700  5.730e-01
#> lag.(Intercept) -0.005492  0.07512  0.119716  0.15998  2.271e-01
#> lag.PEXPOSURE   -0.065512 -0.04614 -0.036261 -0.02284 -6.481e-06
#> lag.PCTAGE65P   -0.189743  0.10315  0.253714  0.41515  7.023e-01
#> lag.PCTOWNHOME  -0.276893 -0.19388 -0.148544 -0.10507 -1.604e-02
#> 

if (run) {
lsar0 <- lagsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY)
summary(lsar0)
}

#> 
#> Call:lagsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY)
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.586752 -0.391580 -0.022469  0.338017  4.029430 
#> 
#> Type: lag 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.514495   0.156154 -3.2948  0.000985
#> PEXPOSURE    0.047627   0.034509  1.3801  0.167542
#> PCTAGE65P    3.648198   0.599046  6.0900 1.129e-09
#> PCTOWNHOME  -0.414601   0.169554 -2.4453  0.014475
#> 
#> Rho: 0.038893, LR test value: 6.9683, p-value: 0.0082967
#> Asymptotic standard error: 0.015053
#>     z-value: 2.5837, p-value: 0.0097755
#> Wald statistic: 6.6754, p-value: 0.0097755
#> 
#> Log likelihood: -275.2447 for lag model
#> ML residual variance (sigma squared): 0.41166, (sigma: 0.6416)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 562.49, (AIC for lm: 567.46)
#> LM test for residual autocorrelation
#> test value: 1.4633, p-value: 0.22641
#> 

if (run) {
res <- MCMCsamp(lsar0, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
}

#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> rho          0.03747 0.01419 0.0002007      0.0007808
#> (Intercept) -0.50985 0.15675 0.0022168      0.0092052
#> PEXPOSURE    0.04836 0.03219 0.0004552      0.0018160
#> PCTAGE65P    3.66723 0.61332 0.0086737      0.0370054
#> PCTOWNHOME  -0.42636 0.17580 0.0024861      0.0106249
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%      25%      50%      75%    97.5%
#> rho          0.01003  0.02733  0.03720  0.04721  0.06646
#> (Intercept) -0.83125 -0.61157 -0.50667 -0.40135 -0.19840
#> PEXPOSURE   -0.01927  0.02931  0.04824  0.06955  0.11136
#> PCTAGE65P    2.39615  3.27229  3.68505  4.07689  4.82372
#> PCTOWNHOME  -0.77453 -0.54731 -0.42341 -0.30284 -0.09088
#> 

if (run) {
lsar1 <- lagsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, type="mixed")
summary(lsar1)
}

#> 
#> Call:lagsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, type = "mixed")
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.799308 -0.390125 -0.021371  0.346128  3.965251 
#> 
#> Type: mixed 
#> Coefficients: (asymptotic standard errors) 
#>                  Estimate Std. Error z value  Pr(>|z|)
#> (Intercept)     -1.131233   0.249631 -4.5316 5.853e-06
#> PEXPOSURE        0.218364   0.079301  2.7536  0.005894
#> PCTAGE65P        3.361158   0.654123  5.1384 2.771e-07
#> PCTOWNHOME       0.071903   0.253967  0.2831  0.777085
#> lag.(Intercept)  0.132544   0.056175  2.3595  0.018300
#> lag.PEXPOSURE   -0.035239   0.015536 -2.2681  0.023322
#> lag.PCTAGE65P    0.161685   0.223690  0.7228  0.469798
#> lag.PCTOWNHOME  -0.140681   0.058529 -2.4036  0.016234
#> 
#> Rho: 0.026981, LR test value: 2.558, p-value: 0.10974
#> Asymptotic standard error: 0.016766
#>     z-value: 1.6093, p-value: 0.10755
#> Wald statistic: 2.5899, p-value: 0.10755
#> 
#> Log likelihood: -269.1031 for mixed model
#> ML residual variance (sigma squared): 0.39587, (sigma: 0.62918)
#> Number of observations: 281 
#> Number of parameters estimated: 10 
#> AIC: 558.21, (AIC for lm: 558.76)
#> LM test for residual autocorrelation
#> test value: 4.908, p-value: 0.026732
#> 

if (run) {
res <- MCMCsamp(lsar1, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
}

#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                     Mean      SD  Naive SE Time-series SE
#> rho              0.02586 0.01569 0.0002219       0.001173
#> (Intercept)     -1.15961 0.23273 0.0032914       0.017887
#> PEXPOSURE        0.22594 0.07947 0.0011238       0.006369
#> PCTAGE65P        3.38019 0.58525 0.0082767       0.041335
#> PCTOWNHOME       0.08772 0.22773 0.0032206       0.016692
#> lag.(Intercept)  0.13847 0.05392 0.0007626       0.004351
#> lag.PEXPOSURE   -0.03686 0.01514 0.0002141       0.001184
#> lag.PCTAGE65P    0.16398 0.21947 0.0031038       0.017529
#> lag.PCTOWNHOME  -0.14582 0.05704 0.0008067       0.004633
#> 
#> 2. Quantiles for each variable:
#> 
#>                      2.5%      25%      50%      75%     97.5%
#> rho             -0.006199  0.01549  0.02618  0.03560  0.057101
#> (Intercept)     -1.612966 -1.30297 -1.16623 -1.01763 -0.669928
#> PEXPOSURE        0.057677  0.18064  0.22260  0.28077  0.370572
#> PCTAGE65P        2.226986  3.02743  3.35530  3.75766  4.620171
#> PCTOWNHOME      -0.436558 -0.04732  0.09193  0.24400  0.504711
#> lag.(Intercept)  0.026620  0.10289  0.13884  0.17500  0.244560
#> lag.PEXPOSURE   -0.065674 -0.04778 -0.03701 -0.02657 -0.003934
#> lag.PCTAGE65P   -0.293421  0.01169  0.16096  0.30975  0.586923
#> lag.PCTOWNHOME  -0.259616 -0.18193 -0.14556 -0.11002 -0.031721
#> 

if (run) {
ssar0 <- sacsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY)
summary(ssar0)
}

#> 
#> Call:sacsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY)
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.468382 -0.375687 -0.034996  0.314714  3.833950 
#> 
#> Type: sac 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.386572   0.123188 -3.1381  0.001701
#> PEXPOSURE    0.026684   0.024013  1.1112  0.266479
#> PCTAGE65P    3.089824   0.562851  5.4896 4.029e-08
#> PCTOWNHOME  -0.323052   0.137449 -2.3503  0.018756
#> 
#> Rho: 0.089451
#> Asymptotic standard error: 0.019427
#>     z-value: 4.6046, p-value: 4.1325e-06
#> Lambda: -0.08192
#> Asymptotic standard error: 0.033201
#>     z-value: -2.4674, p-value: 0.01361
#> 
#> LR test value: 10.114, p-value: 0.0063661
#> 
#> Log likelihood: -273.672 for sac model
#> ML residual variance (sigma squared): 0.3766, (sigma: 0.61368)
#> Number of observations: 281 
#> Number of parameters estimated: 7 
#> AIC: 561.34, (AIC for lm: 567.46)
#> 

if (run) {
res <- MCMCsamp(ssar0, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
}

#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD Naive SE Time-series SE
#> rho         -0.03296 0.07395 0.001046        0.01975
#> lambda       0.05813 0.07410 0.001048        0.02136
#> (Intercept) -0.77991 0.32753 0.004632        0.07375
#> PEXPOSURE    0.12934 0.10335 0.001462        0.02366
#> PCTAGE65P    3.36334 0.65856 0.009313        0.04790
#> PCTOWNHOME  -0.28705 0.20857 0.002950        0.01835
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%       25%      50%      75%   97.5%
#> rho         -0.13465 -0.093139 -0.05970  0.04001  0.1010
#> lambda      -0.10145 -0.002409  0.09389  0.11672  0.1355
#> (Intercept) -1.43102 -1.031506 -0.77677 -0.49551 -0.2679
#> PEXPOSURE   -0.02294  0.036799  0.11912  0.21034  0.3290
#> PCTAGE65P    2.10377  2.893842  3.35412  3.76731  4.6757
#> PCTOWNHOME  -0.68773 -0.424730 -0.29029 -0.15274  0.1660
#> 

if (run) {
ssar1 <- sacsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, type="sacmixed")
summary(ssar1)
}

#> 
#> Call:sacsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, type = "sacmixed")
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.633958 -0.363826 -0.019927  0.348238  3.655509 
#> 
#> Type: sacmixed 
#> Coefficients: (asymptotic standard errors) 
#>                  Estimate Std. Error z value  Pr(>|z|)
#> (Intercept)     -1.133298   0.247495 -4.5791 4.670e-06
#> PEXPOSURE        0.206963   0.074480  2.7788  0.005456
#> PCTAGE65P        3.083983   0.671081  4.5955 4.316e-06
#> PCTOWNHOME       0.174800   0.256280  0.6821  0.495196
#> lag.(Intercept)  0.153427   0.050817  3.0192  0.002534
#> lag.PEXPOSURE   -0.033400   0.013817 -2.4173  0.015634
#> lag.PCTAGE65P   -0.079738   0.222144 -0.3589  0.719634
#> lag.PCTOWNHOME  -0.102502   0.056760 -1.8059  0.070940
#> 
#> Rho: 0.092495
#> Asymptotic standard error: 0.023829
#>     z-value: 3.8817, p-value: 0.00010375
#> Lambda: -0.091069
#> Asymptotic standard error: 0.038431
#>     z-value: -2.3697, p-value: 0.017804
#> 
#> LR test value: 22.379, p-value: 0.0010335
#> 
#> Log likelihood: -267.5392 for sacmixed model
#> ML residual variance (sigma squared): 0.35617, (sigma: 0.5968)
#> Number of observations: 281 
#> Number of parameters estimated: 11 
#> AIC: 557.08, (AIC for lm: 567.46)
#> 

if (run) {
res <- MCMCsamp(ssar1, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
}

#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                      Mean      SD  Naive SE Time-series SE
#> rho             -0.025660 0.06503 0.0009197       0.014003
#> lambda           0.041280 0.06191 0.0008755       0.013819
#> (Intercept)     -1.074360 0.26571 0.0037577       0.022295
#> PEXPOSURE        0.212544 0.08470 0.0011979       0.007028
#> PCTAGE65P        3.314485 0.63801 0.0090228       0.050714
#> PCTOWNHOME       0.003484 0.25978 0.0036738       0.021865
#> lag.(Intercept)  0.112005 0.07102 0.0010044       0.007936
#> lag.PEXPOSURE   -0.036565 0.01668 0.0002359       0.001424
#> lag.PCTAGE65P    0.306235 0.31875 0.0045078       0.038162
#> lag.PCTOWNHOME  -0.151066 0.06645 0.0009397       0.006463
#> 
#> 2. Quantiles for each variable:
#> 
#>                     2.5%        25%       50%      75%      97.5%
#> rho             -0.12868 -0.0798728 -0.035985  0.01955  0.0987633
#> lambda          -0.09863  0.0006317  0.057962  0.09088  0.1193741
#> (Intercept)     -1.57995 -1.2504463 -1.082594 -0.88786 -0.5048974
#> PEXPOSURE        0.04331  0.1541882  0.214118  0.27257  0.3730577
#> PCTAGE65P        2.22775  2.8877944  3.249345  3.72756  4.6974025
#> PCTOWNHOME      -0.47022 -0.1810743 -0.005259  0.17530  0.5270675
#> lag.(Intercept) -0.01752  0.0588096  0.109781  0.16516  0.2450747
#> lag.PEXPOSURE   -0.06923 -0.0480602 -0.036216 -0.02617 -0.0008617
#> lag.PCTAGE65P   -0.29408  0.0917982  0.290299  0.53337  0.9974834
#> lag.PCTOWNHOME  -0.29153 -0.1907786 -0.150445 -0.10345 -0.0285370
#>