MCMC sample from fitted spatial regression

The MCMCsamp method uses rwmetrop, a random walk Metropolis algorithm, from LearnBayes to make MCMC samples from fitted maximum likelihood spatial regression models.

MCMCsamp(object, mcmc = 1L, verbose = NULL, ...)
# S3 method for spautolm
MCMCsamp(object, mcmc = 1L, verbose = NULL, ...,
 burnin = 0L, scale=1, listw, control = list())
# S3 method for sarlm
MCMCsamp(object, mcmc = 1L, verbose = NULL, ...,
    burnin=0L, scale=1, listw, listw2=NULL, control=list())

Arguments

object
A spatial regression model object fitted by maximum likelihood with spautolm

mcmc
The number of MCMC iterations after burnin

verbose
default NULL, use global option value; if TRUE, reports progress

…
Arguments passed through

burnin
The number of burn-in iterations for the sampler

scale
a positive scale parameter

listw, listw2
listw objects created for example by nb2listw; should be the same object(s) used for fitting the model

control
list of extra control arguments - see spautolm

Value

An object of class “mcmc” suited to coda, with attributes: “accept” acceptance rate; “type” input ML fitted model type “SAR”, “CAR”, “SMA”, “lag”, “mixed”, “error”, “sac”, “sacmixed”; “timings” run times

Note

If the acceptance rate is below 0.05, a warning will be issued; consider increasing mcmc.

References

Jim Albert (2007) Bayesian Computation with R, Springer, New York, pp. 104-105.

See also

Examples

nydata <- st_read(system.file("shapes/NY8_bna_utm18.gpkg", package="spData")[1], quiet=TRUE)
suppressMessages(nyadjmat <- as.matrix(foreign::read.dbf(system.file(
 "misc/nyadjwts.dbf", package="spData")[1])[-1]))
suppressMessages(ID <- as.character(names(foreign::read.dbf(system.file(
 "misc/nyadjwts.dbf", package="spData")[1]))[-1]))
identical(substring(ID, 2, 10), substring(as.character(nydata$AREAKEY), 2, 10))

#> [1] TRUE

esar1f <- spautolm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, family="SAR", method="eigen")
summary(esar1f)

#> 
#> Call: 
#> spautolm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, family = "SAR", method = "eigen")
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.56754 -0.38239 -0.02643  0.33109  4.01219 
#> 
#> Coefficients: 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.618193   0.176784 -3.4969 0.0004707
#> PEXPOSURE    0.071014   0.042051  1.6888 0.0912635
#> PCTAGE65P    3.754200   0.624722  6.0094 1.862e-09
#> PCTOWNHOME  -0.419890   0.191329 -2.1946 0.0281930
#> 
#> Lambda: 0.040487 LR test value: 5.2438 p-value: 0.022026 
#> Numerical Hessian standard error of lambda: 0.017199 
#> 
#> Log likelihood: -276.1069 
#> ML residual variance (sigma squared): 0.41388, (sigma: 0.64333)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 564.21
#> 

res <- MCMCsamp(esar1f, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)

#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.04429 0.01691 0.0002391      0.0009689
#> (Intercept) -0.62975 0.19309 0.0027307      0.0117292
#> PEXPOSURE    0.07230 0.04628 0.0006545      0.0027868
#> PCTAGE65P    3.74094 0.62754 0.0088748      0.0356924
#> PCTOWNHOME  -0.39726 0.20235 0.0028617      0.0123184
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%      25%      50%      75%    97.5%
#> lambda       0.01059  0.03220  0.04461  0.05573  0.07621
#> (Intercept) -1.04023 -0.75276 -0.61931 -0.51131 -0.23580
#> PEXPOSURE   -0.01696  0.04113  0.06945  0.09939  0.17155
#> PCTAGE65P    2.41718  3.34443  3.77069  4.13446  4.99866
#> PCTOWNHOME  -0.78005 -0.53332 -0.40694 -0.27064  0.01811
#> 

esar1fw <- spautolm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, weights=POP8, family="SAR", method="eigen")
summary(esar1fw)

#> 
#> Call: 
#> spautolm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, weights = POP8, family = "SAR", method = "eigen")
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.48488 -0.26823  0.09489  0.46552  4.28343 
#> 
#> Coefficients: 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.797063   0.144054 -5.5331 3.146e-08
#> PEXPOSURE    0.080545   0.028334  2.8428  0.004473
#> PCTAGE65P    3.816731   0.576037  6.6258 3.453e-11
#> PCTOWNHOME  -0.380778   0.156507 -2.4330  0.014975
#> 
#> Lambda: 0.0095636 LR test value: 0.32665 p-value: 0.56764 
#> Numerical Hessian standard error of lambda: 0.016466 
#> 
#> Log likelihood: -251.6017 
#> ML residual variance (sigma squared): 1104.1, (sigma: 33.229)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 515.2
#> 

res <- MCMCsamp(esar1fw, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)

#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.01105 0.01638 0.0002316      0.0009186
#> (Intercept) -0.78457 0.14454 0.0020441      0.0077326
#> PEXPOSURE    0.08026 0.02867 0.0004054      0.0015588
#> PCTAGE65P    3.75868 0.59779 0.0084541      0.0332261
#> PCTOWNHOME  -0.38700 0.15685 0.0022182      0.0082559
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%        25%      50%      75%    97.5%
#> lambda      -0.02146  0.0004359  0.01188  0.02153  0.04376
#> (Intercept) -1.07522 -0.8822177 -0.77820 -0.68170 -0.51232
#> PEXPOSURE    0.02650  0.0606119  0.07958  0.09942  0.13993
#> PCTAGE65P    2.59748  3.3227380  3.76137  4.16785  4.90894
#> PCTOWNHOME  -0.68665 -0.5001917 -0.39285 -0.28251 -0.07474
#> 

ecar1f <- spautolm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, family="CAR", method="eigen")
summary(ecar1f)

#> 
#> Call: 
#> spautolm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, family = "CAR", method = "eigen")
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.539732 -0.384311 -0.030646  0.335126  3.808848 
#> 
#> Coefficients: 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.648362   0.181129 -3.5796 0.0003442
#> PEXPOSURE    0.077899   0.043692  1.7829 0.0745986
#> PCTAGE65P    3.703830   0.627185  5.9055 3.516e-09
#> PCTOWNHOME  -0.382789   0.195564 -1.9574 0.0503053
#> 
#> Lambda: 0.084123 LR test value: 5.8009 p-value: 0.016018 
#> Numerical Hessian standard error of lambda: 0.030868 
#> 
#> Log likelihood: -275.8283 
#> ML residual variance (sigma squared): 0.40758, (sigma: 0.63842)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 563.66
#> 

res <- MCMCsamp(ecar1f, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)

#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.08760 0.02864 0.0004051       0.001590
#> (Intercept) -0.65226 0.20677 0.0029242       0.012460
#> PEXPOSURE    0.08446 0.04896 0.0006924       0.003029
#> PCTAGE65P    3.62157 0.64895 0.0091776       0.041144
#> PCTOWNHOME  -0.37230 0.22785 0.0032223       0.013376
#> 
#> 2. Quantiles for each variable:
#> 
#>                  2.5%      25%      50%     75%    97.5%
#> lambda       0.028033  0.06826  0.08922  0.1085  0.13671
#> (Intercept) -1.069635 -0.78403 -0.65554 -0.5078 -0.25039
#> PEXPOSURE   -0.007364  0.04987  0.08260  0.1172  0.18543
#> PCTAGE65P    2.265603  3.20707  3.62840  4.0753  4.83742
#> PCTOWNHOME  -0.815589 -0.53130 -0.36539 -0.2281  0.07847
#> 

esar1fw <- spautolm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, weights=POP8, family="SAR", method="eigen")
summary(esar1fw)

#> 
#> Call: 
#> spautolm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, weights = POP8, family = "SAR", method = "eigen")
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.48488 -0.26823  0.09489  0.46552  4.28343 
#> 
#> Coefficients: 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.797063   0.144054 -5.5331 3.146e-08
#> PEXPOSURE    0.080545   0.028334  2.8428  0.004473
#> PCTAGE65P    3.816731   0.576037  6.6258 3.453e-11
#> PCTOWNHOME  -0.380778   0.156507 -2.4330  0.014975
#> 
#> Lambda: 0.0095636 LR test value: 0.32665 p-value: 0.56764 
#> Numerical Hessian standard error of lambda: 0.016466 
#> 
#> Log likelihood: -251.6017 
#> ML residual variance (sigma squared): 1104.1, (sigma: 33.229)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 515.2
#> 

res <- MCMCsamp(esar1fw, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)

#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.01237 0.01688 0.0002387       0.001030
#> (Intercept) -0.82916 0.16119 0.0022796       0.011108
#> PEXPOSURE    0.08571 0.03127 0.0004422       0.001873
#> PCTAGE65P    3.81056 0.60911 0.0086142       0.036006
#> PCTOWNHOME  -0.34950 0.16804 0.0023764       0.010258
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%       25%      50%      75%     97.5%
#> lambda      -0.02048  0.001295  0.01219  0.02409  0.044644
#> (Intercept) -1.14674 -0.932404 -0.82553 -0.72091 -0.505857
#> PEXPOSURE    0.02749  0.064070  0.08599  0.10469  0.152937
#> PCTAGE65P    2.56424  3.416681  3.82234  4.23224  4.965195
#> PCTOWNHOME  -0.68099 -0.461585 -0.35395 -0.23798 -0.007876
#> 

ecar1fw <- spautolm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, weights=POP8, family="CAR", method="eigen")
summary(ecar1fw)

#> 
#> Call: 
#> spautolm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, weights = POP8, family = "CAR", method = "eigen")
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.491042 -0.270906  0.081435  0.451556  4.198134 
#> 
#> Coefficients: 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.790154   0.144862 -5.4545 4.910e-08
#> PEXPOSURE    0.081922   0.028593  2.8651  0.004169
#> PCTAGE65P    3.825858   0.577720  6.6223 3.536e-11
#> PCTOWNHOME  -0.386820   0.157436 -2.4570  0.014010
#> 
#> Lambda: 0.022419 LR test value: 0.38785 p-value: 0.53343 
#> Numerical Hessian standard error of lambda: 0.038543 
#> 
#> Log likelihood: -251.5711 
#> ML residual variance (sigma squared): 1102.9, (sigma: 33.21)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 515.14
#> 

res <- MCMCsamp(ecar1fw, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)

#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.03542 0.03787 0.0005355       0.002168
#> (Intercept) -0.81507 0.16062 0.0022715       0.009554
#> PEXPOSURE    0.08695 0.03206 0.0004534       0.001879
#> PCTAGE65P    3.76438 0.58703 0.0083019       0.032643
#> PCTOWNHOME  -0.37172 0.16913 0.0023918       0.009840
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%      25%      50%      75%   97.5%
#> lambda      -0.04292  0.01031  0.03565  0.06337  0.1031
#> (Intercept) -1.12996 -0.92547 -0.80796 -0.70018 -0.5262
#> PEXPOSURE    0.02923  0.06506  0.08589  0.10794  0.1528
#> PCTAGE65P    2.62210  3.36051  3.77349  4.17907  4.9388
#> PCTOWNHOME  -0.69948 -0.49013 -0.37220 -0.25684 -0.0556
#> 

esar0 <- errorsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY)
summary(esar0)

#> 
#> Call:errorsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, 
#>     data = nydata, listw = listw_NY)
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.56754 -0.38239 -0.02643  0.33109  4.01219 
#> 
#> Type: error 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.618193   0.176784 -3.4969 0.0004707
#> PEXPOSURE    0.071014   0.042051  1.6888 0.0912635
#> PCTAGE65P    3.754200   0.624722  6.0094 1.862e-09
#> PCTOWNHOME  -0.419890   0.191329 -2.1946 0.0281930
#> 
#> Lambda: 0.040487, LR test value: 5.2438, p-value: 0.022026
#> Asymptotic standard error: 0.016214
#>     z-value: 2.4971, p-value: 0.01252
#> Wald statistic: 6.2356, p-value: 0.01252
#> 
#> Log likelihood: -276.1069 for error model
#> ML residual variance (sigma squared): 0.41388, (sigma: 0.64333)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 564.21, (AIC for lm: 567.46)
#> 

res <- MCMCsamp(esar0, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)

#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.04419 0.01697 0.0002399       0.001078
#> (Intercept) -0.64099 0.18642 0.0026364       0.011304
#> PEXPOSURE    0.07325 0.04515 0.0006386       0.002729
#> PCTAGE65P    3.79120 0.63486 0.0089783       0.035664
#> PCTOWNHOME  -0.40094 0.20137 0.0028479       0.011945
#> 
#> 2. Quantiles for each variable:
#> 
#>                  2.5%      25%      50%      75%    97.5%
#> lambda       0.009440  0.03336  0.04506  0.05603  0.07546
#> (Intercept) -0.996575 -0.77405 -0.64310 -0.51143 -0.27833
#> PEXPOSURE   -0.009371  0.04240  0.07255  0.10534  0.16520
#> PCTAGE65P    2.568115  3.35875  3.80270  4.20366  5.10342
#> PCTOWNHOME  -0.785386 -0.54117 -0.40394 -0.26812  0.02865
#> 

esar0w <- errorsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, weights=POP8)
summary(esar0)

#> 
#> Call:errorsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, 
#>     data = nydata, listw = listw_NY)
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.56754 -0.38239 -0.02643  0.33109  4.01219 
#> 
#> Type: error 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.618193   0.176784 -3.4969 0.0004707
#> PEXPOSURE    0.071014   0.042051  1.6888 0.0912635
#> PCTAGE65P    3.754200   0.624722  6.0094 1.862e-09
#> PCTOWNHOME  -0.419890   0.191329 -2.1946 0.0281930
#> 
#> Lambda: 0.040487, LR test value: 5.2438, p-value: 0.022026
#> Asymptotic standard error: 0.016214
#>     z-value: 2.4971, p-value: 0.01252
#> Wald statistic: 6.2356, p-value: 0.01252
#> 
#> Log likelihood: -276.1069 for error model
#> ML residual variance (sigma squared): 0.41388, (sigma: 0.64333)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 564.21, (AIC for lm: 567.46)
#> 

res <- MCMCsamp(esar0w, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)

#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.01448 0.01679 0.0002375      0.0009561
#> (Intercept) -0.80541 0.15177 0.0021464      0.0091791
#> PEXPOSURE    0.08178 0.03175 0.0004491      0.0020138
#> PCTAGE65P    3.80780 0.60355 0.0085356      0.0356224
#> PCTOWNHOME  -0.37314 0.15762 0.0022292      0.0091866
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%       25%      50%      75%    97.5%
#> lambda      -0.01812  0.003413  0.01468  0.02599  0.04589
#> (Intercept) -1.10475 -0.903970 -0.80590 -0.70560 -0.49326
#> PEXPOSURE    0.01853  0.059548  0.08189  0.10210  0.14611
#> PCTAGE65P    2.64051  3.400986  3.81346  4.20398  4.98465
#> PCTOWNHOME  -0.68348 -0.479277 -0.36672 -0.26772 -0.07536
#> 

esar1 <- errorsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, etype="emixed")
summary(esar1)

#> 
#> Call:errorsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, 
#>     data = nydata, listw = listw_NY, etype = "emixed")
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.81562 -0.37641 -0.02224  0.33638  4.00054 
#> 
#> Type: error 
#> Coefficients: (asymptotic standard errors) 
#>                  Estimate Std. Error z value  Pr(>|z|)
#> (Intercept)     -1.118019   0.247425 -4.5186 6.225e-06
#> PEXPOSURE        0.218279   0.079245  2.7545  0.005879
#> PCTAGE65P        3.416477   0.645587  5.2920 1.210e-07
#> PCTOWNHOME       0.036593   0.249835  0.1465  0.883551
#> lag.(Intercept)  0.121515   0.057636  2.1083  0.035003
#> lag.PEXPOSURE   -0.035075   0.015943 -2.2000  0.027808
#> lag.PCTAGE65P    0.263096   0.220118  1.1953  0.231989
#> lag.PCTOWNHOME  -0.155680   0.059213 -2.6291  0.008560
#> 
#> Lambda: 0.022723, LR test value: 1.6846, p-value: 0.19432
#> Asymptotic standard error: 0.017169
#>     z-value: 1.3235, p-value: 0.18567
#> Wald statistic: 1.7516, p-value: 0.18567
#> 
#> Log likelihood: -269.5398 for error model
#> ML residual variance (sigma squared): 0.39759, (sigma: 0.63055)
#> Number of observations: 281 
#> Number of parameters estimated: 10 
#> AIC: 559.08, (AIC for lm: 558.76)
#> 

res <- MCMCsamp(esar1, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)

#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                     Mean      SD  Naive SE Time-series SE
#> lambda           0.02919 0.01724 0.0002437       0.001259
#> (Intercept)     -1.11495 0.25196 0.0035633       0.020446
#> PEXPOSURE        0.21609 0.08304 0.0011744       0.006769
#> PCTAGE65P        3.48058 0.59625 0.0084323       0.041798
#> PCTOWNHOME       0.03078 0.25793 0.0036476       0.020568
#> lag.(Intercept)  0.12371 0.06193 0.0008758       0.005190
#> lag.PEXPOSURE   -0.03461 0.01682 0.0002379       0.001360
#> lag.PCTAGE65P    0.22839 0.24200 0.0034224       0.020680
#> lag.PCTOWNHOME  -0.15228 0.06247 0.0008834       0.005118
#> 
#> 2. Quantiles for each variable:
#> 
#>                       2.5%      25%      50%      75%     97.5%
#> lambda          -0.0047772  0.01697  0.03003  0.04087  0.060582
#> (Intercept)     -1.5858692 -1.29223 -1.11093 -0.95116 -0.613975
#> PEXPOSURE        0.0521332  0.16147  0.21194  0.26958  0.386386
#> PCTAGE65P        2.1815917  3.11859  3.50938  3.90931  4.555704
#> PCTOWNHOME      -0.4671286 -0.15470  0.03101  0.19426  0.573397
#> lag.(Intercept)  0.0005569  0.08611  0.11926  0.16670  0.243600
#> lag.PEXPOSURE   -0.0687185 -0.04577 -0.03437 -0.02356 -0.003167
#> lag.PCTAGE65P   -0.2602480  0.09090  0.23601  0.37639  0.722080
#> lag.PCTOWNHOME  -0.2682098 -0.19677 -0.15204 -0.10520 -0.038668
#> 

lsar0 <- lagsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY)
summary(lsar0)

#> 
#> Call:lagsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY)
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.586752 -0.391580 -0.022469  0.338017  4.029430 
#> 
#> Type: lag 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.514495   0.156154 -3.2948  0.000985
#> PEXPOSURE    0.047627   0.034509  1.3801  0.167542
#> PCTAGE65P    3.648198   0.599046  6.0900 1.129e-09
#> PCTOWNHOME  -0.414601   0.169554 -2.4453  0.014475
#> 
#> Rho: 0.038893, LR test value: 6.9683, p-value: 0.0082967
#> Asymptotic standard error: 0.015053
#>     z-value: 2.5837, p-value: 0.0097755
#> Wald statistic: 6.6754, p-value: 0.0097755
#> 
#> Log likelihood: -275.2447 for lag model
#> ML residual variance (sigma squared): 0.41166, (sigma: 0.6416)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 562.49, (AIC for lm: 567.46)
#> LM test for residual autocorrelation
#> test value: 1.4633, p-value: 0.22641
#> 

res <- MCMCsamp(lsar0, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)

#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> rho          0.03786 0.01360 0.0001923       0.000721
#> (Intercept) -0.50551 0.15203 0.0021500       0.008676
#> PEXPOSURE    0.04683 0.03384 0.0004786       0.002011
#> PCTAGE65P    3.64851 0.57471 0.0081277       0.032169
#> PCTOWNHOME  -0.42391 0.16886 0.0023880       0.009651
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%      25%      50%      75%    97.5%
#> rho          0.01113  0.02918  0.03790  0.04650  0.06347
#> (Intercept) -0.79479 -0.61485 -0.50512 -0.40505 -0.20793
#> PEXPOSURE   -0.02273  0.02354  0.04856  0.06967  0.11117
#> PCTAGE65P    2.50073  3.27416  3.65774  4.04029  4.70560
#> PCTOWNHOME  -0.74781 -0.53847 -0.42351 -0.31428 -0.08230
#> 

lsar1 <- lagsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, type="mixed")
summary(lsar1)

#> 
#> Call:lagsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, type = "mixed")
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.799308 -0.390125 -0.021371  0.346128  3.965251 
#> 
#> Type: mixed 
#> Coefficients: (asymptotic standard errors) 
#>                  Estimate Std. Error z value  Pr(>|z|)
#> (Intercept)     -1.131233   0.249631 -4.5316 5.853e-06
#> PEXPOSURE        0.218364   0.079301  2.7536  0.005894
#> PCTAGE65P        3.361158   0.654123  5.1384 2.771e-07
#> PCTOWNHOME       0.071903   0.253967  0.2831  0.777085
#> lag.(Intercept)  0.132544   0.056175  2.3595  0.018300
#> lag.PEXPOSURE   -0.035239   0.015536 -2.2681  0.023322
#> lag.PCTAGE65P    0.161685   0.223690  0.7228  0.469798
#> lag.PCTOWNHOME  -0.140681   0.058529 -2.4036  0.016234
#> 
#> Rho: 0.026981, LR test value: 2.558, p-value: 0.10974
#> Asymptotic standard error: 0.016766
#>     z-value: 1.6093, p-value: 0.10755
#> Wald statistic: 2.5899, p-value: 0.10755
#> 
#> Log likelihood: -269.1031 for mixed model
#> ML residual variance (sigma squared): 0.39587, (sigma: 0.62918)
#> Number of observations: 281 
#> Number of parameters estimated: 10 
#> AIC: 558.21, (AIC for lm: 558.76)
#> LM test for residual autocorrelation
#> test value: 4.908, p-value: 0.026732
#> 

res <- MCMCsamp(lsar1, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)

#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                     Mean      SD  Naive SE Time-series SE
#> rho              0.02560 0.01527 0.0002160       0.001083
#> (Intercept)     -1.14317 0.22347 0.0031604       0.016320
#> PEXPOSURE        0.21636 0.07524 0.0010640       0.005739
#> PCTAGE65P        3.34703 0.61555 0.0087052       0.047015
#> PCTOWNHOME       0.08351 0.24779 0.0035043       0.019739
#> lag.(Intercept)  0.13217 0.05430 0.0007680       0.004333
#> lag.PEXPOSURE   -0.03393 0.01509 0.0002134       0.001167
#> lag.PCTAGE65P    0.15921 0.22678 0.0032072       0.017972
#> lag.PCTOWNHOME  -0.14006 0.05448 0.0007705       0.004214
#> 
#> 2. Quantiles for each variable:
#> 
#>                      2.5%      25%      50%      75%     97.5%
#> rho             -0.002664  0.01494  0.02582  0.03562  0.055310
#> (Intercept)     -1.621930 -1.29916 -1.14895 -0.99153 -0.738536
#> PEXPOSURE        0.062140  0.16474  0.22033  0.26921  0.350867
#> PCTAGE65P        2.069235  2.90420  3.36557  3.77704  4.505405
#> PCTOWNHOME      -0.394840 -0.08070  0.08761  0.25772  0.622804
#> lag.(Intercept)  0.027041  0.09921  0.13596  0.16899  0.232320
#> lag.PEXPOSURE   -0.061436 -0.04500 -0.03422 -0.02365 -0.003468
#> lag.PCTAGE65P   -0.286543  0.01158  0.14290  0.32435  0.563685
#> lag.PCTOWNHOME  -0.245727 -0.17663 -0.14268 -0.10378 -0.030811
#> 

ssar0 <- sacsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY)
summary(ssar0)

#> 
#> Call:sacsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY)
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.468382 -0.375687 -0.034996  0.314714  3.833950 
#> 
#> Type: sac 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.386572   0.123188 -3.1381  0.001701
#> PEXPOSURE    0.026684   0.024013  1.1112  0.266479
#> PCTAGE65P    3.089824   0.562851  5.4896 4.029e-08
#> PCTOWNHOME  -0.323052   0.137449 -2.3503  0.018756
#> 
#> Rho: 0.089451
#> Asymptotic standard error: 0.019427
#>     z-value: 4.6046, p-value: 4.1325e-06
#> Lambda: -0.08192
#> Asymptotic standard error: 0.033201
#>     z-value: -2.4674, p-value: 0.01361
#> 
#> LR test value: 10.114, p-value: 0.0063661
#> 
#> Log likelihood: -273.672 for sac model
#> ML residual variance (sigma squared): 0.3766, (sigma: 0.61368)
#> Number of observations: 281 
#> Number of parameters estimated: 7 
#> AIC: 561.34, (AIC for lm: 567.46)
#> 

res <- MCMCsamp(ssar0, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)

#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                  Mean      SD Naive SE Time-series SE
#> rho         -0.008426 0.07758 0.001097        0.02277
#> lambda       0.035056 0.07997 0.001131        0.02024
#> (Intercept) -0.660294 0.29377 0.004155        0.04947
#> PEXPOSURE    0.088792 0.06908 0.000977        0.01023
#> PCTAGE65P    3.360986 0.65211 0.009222        0.04525
#> PCTOWNHOME  -0.291329 0.21916 0.003099        0.02010
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%      25%       50%      75%   97.5%
#> rho         -0.13010 -0.08160 -0.008581  0.06788  0.1052
#> lambda      -0.10450 -0.04059  0.056054  0.11176  0.1329
#> (Intercept) -1.32733 -0.85710 -0.642072 -0.42623 -0.1788
#> PEXPOSURE   -0.01472  0.03130  0.075126  0.14270  0.2213
#> PCTAGE65P    2.09662  2.91291  3.355083  3.77601  4.6007
#> PCTOWNHOME  -0.71343 -0.43204 -0.297317 -0.16763  0.1952
#> 

ssar1 <- sacsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, type="sacmixed")
summary(ssar1)

#> 
#> Call:sacsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, type = "sacmixed")
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.633958 -0.363826 -0.019927  0.348238  3.655509 
#> 
#> Type: sacmixed 
#> Coefficients: (asymptotic standard errors) 
#>                  Estimate Std. Error z value  Pr(>|z|)
#> (Intercept)     -1.133298   0.247495 -4.5791 4.670e-06
#> PEXPOSURE        0.206963   0.074480  2.7788  0.005456
#> PCTAGE65P        3.083983   0.671081  4.5955 4.316e-06
#> PCTOWNHOME       0.174800   0.256280  0.6821  0.495196
#> lag.(Intercept)  0.153427   0.050817  3.0192  0.002534
#> lag.PEXPOSURE   -0.033400   0.013817 -2.4173  0.015634
#> lag.PCTAGE65P   -0.079738   0.222144 -0.3589  0.719634
#> lag.PCTOWNHOME  -0.102502   0.056760 -1.8059  0.070940
#> 
#> Rho: 0.092495
#> Asymptotic standard error: 0.023829
#>     z-value: 3.8817, p-value: 0.00010375
#> Lambda: -0.091069
#> Asymptotic standard error: 0.038431
#>     z-value: -2.3697, p-value: 0.017804
#> 
#> LR test value: 22.379, p-value: 0.0010335
#> 
#> Log likelihood: -267.5392 for sacmixed model
#> ML residual variance (sigma squared): 0.35617, (sigma: 0.5968)
#> Number of observations: 281 
#> Number of parameters estimated: 11 
#> AIC: 557.08, (AIC for lm: 567.46)
#> 

res <- MCMCsamp(ssar1, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)

#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                     Mean      SD  Naive SE Time-series SE
#> rho             -0.06557 0.06756 0.0009554       0.017791
#> lambda           0.07677 0.06056 0.0008564       0.017566
#> (Intercept)     -1.11599 0.21993 0.0031102       0.016899
#> PEXPOSURE        0.22925 0.07556 0.0010686       0.006399
#> PCTAGE65P        3.25183 0.67178 0.0095004       0.070504
#> PCTOWNHOME       0.01836 0.25882 0.0036602       0.022713
#> lag.(Intercept)  0.05327 0.10269 0.0014523       0.021517
#> lag.PEXPOSURE   -0.02390 0.02386 0.0003375       0.004835
#> lag.PCTAGE65P    0.27734 0.32872 0.0046488       0.046247
#> lag.PCTOWNHOME  -0.11522 0.08866 0.0012538       0.015617
#> 
#> 2. Quantiles for each variable:
#> 
#>                     2.5%      25%      50%       75%    97.5%
#> rho             -0.15242 -0.12127 -0.07608 -0.024144  0.08929
#> lambda          -0.08581  0.05118  0.09400  0.123168  0.13909
#> (Intercept)     -1.58991 -1.23960 -1.12162 -0.970776 -0.70638
#> PEXPOSURE        0.07616  0.18406  0.23263  0.276078  0.38636
#> PCTAGE65P        1.99200  2.80501  3.22411  3.739504  4.53456
#> PCTOWNHOME      -0.52010 -0.15218  0.03849  0.211835  0.51930
#> lag.(Intercept) -0.11413 -0.03243  0.05570  0.140118  0.24530
#> lag.PEXPOSURE   -0.06179 -0.04157 -0.02858 -0.004586  0.02806
#> lag.PCTAGE65P   -0.25813  0.03119  0.24677  0.496821  1.00365
#> lag.PCTOWNHOME  -0.32429 -0.16882 -0.11640 -0.042202  0.03730
#>