Approximate profile-likelihood estimator (APLE)

The Approximate profile-likelihood estimator (APLE) of the simultaneous autoregressive model's spatial dependence parameter was introduced in Li et al. (2007). It employs a correction term using the eigenvalues of the spatial weights matrix, and consequently should not be used for large numbers of observations. It also requires that the variable has a mean of zero, and it is assumed that it has been detrended. The spatial weights object is assumed to be row-standardised, that is using default style="W" in nb2listw.

aple(x, listw, override_similarity_check=FALSE, useTrace=TRUE)

Arguments

x
a zero-mean detrended continuous variable

listw
a listw object from for example nb2listw

override\_similarity\_check
default FALSE, if TRUE - typically for row-standardised weights with asymmetric underlying general weights - similarity is not checked

useTrace
default TRUE, use trace of sparse matrix W %*% W (Li et al. (2010)), if FALSE, use crossproduct of eigenvalues of W as in Li et al. (2007)

Details

This implementation has been checked with Hongfei Li's own implementation using her data; her help was very valuable.

Value

A scalar APLE value.

References

Li, H, Calder, C. A. and Cressie N. A. C. (2007) Beyond Moran's I: testing for spatial dependence based on the spatial autoregressive model. Geographical Analysis 39, 357-375; Li, H, Calder, C. A. and Cressie N. A. C. (2012) One-step estimation of spatial dependence parameters: Properties and extensions of the APLE statistic, Journal of Multivariate Analysis 105, 68-84.

See also

Examples

wheat <- st_read(system.file("shapes/wheat.shp", package="spData")[1], quiet=TRUE)
nbr1 <- poly2nb(wheat, queen=FALSE)
nbrl <- nblag(nbr1, 2)
nbr12 <- nblag_cumul(nbrl)
cms0 <- with(as.data.frame(wheat), tapply(yield, c, median))
cms1 <- c(model.matrix(~ factor(c) -1, data=wheat) %*% cms0)
wheat$yield_detrend <- wheat$yield - cms1
isTRUE(all.equal(c(with(as.data.frame(wheat),
 tapply(yield_detrend, c, median))), rep(0.0, 25),
 check.attributes=FALSE))

#> [1] TRUE

moran.test(wheat$yield_detrend, nb2listw(nbr12, style="W"))

#> 
#> 	Moran I test under randomisation
#> 
#> data:  wheat$yield_detrend  
#> weights: nb2listw(nbr12, style = "W")    
#> 
#> Moran I statistic standard deviate = 8.6225, p-value < 2.2e-16
#> alternative hypothesis: greater
#> sample estimates:
#> Moran I statistic       Expectation          Variance 
#>       0.133391679      -0.002004008       0.000246570 
#> 

aple(as.vector(scale(wheat$yield_detrend, scale=FALSE)), nb2listw(nbr12, style="W"))

#> Registered S3 methods overwritten by 'spatialreg':
#>   method                   from 
#>   residuals.stsls          spdep
#>   deviance.stsls           spdep
#>   coef.stsls               spdep
#>   print.stsls              spdep
#>   summary.stsls            spdep
#>   print.summary.stsls      spdep
#>   residuals.gmsar          spdep
#>   deviance.gmsar           spdep
#>   coef.gmsar               spdep
#>   fitted.gmsar             spdep
#>   print.gmsar              spdep
#>   summary.gmsar            spdep
#>   print.summary.gmsar      spdep
#>   print.lagmess            spdep
#>   summary.lagmess          spdep
#>   print.summary.lagmess    spdep
#>   residuals.lagmess        spdep
#>   deviance.lagmess         spdep
#>   coef.lagmess             spdep
#>   fitted.lagmess           spdep
#>   logLik.lagmess           spdep
#>   fitted.SFResult          spdep
#>   print.SFResult           spdep
#>   fitted.ME_res            spdep
#>   print.ME_res             spdep
#>   print.lagImpact          spdep
#>   plot.lagImpact           spdep
#>   summary.lagImpact        spdep
#>   HPDinterval.lagImpact    spdep
#>   print.summary.lagImpact  spdep
#>   print.sarlm              spdep
#>   summary.sarlm            spdep
#>   residuals.sarlm          spdep
#>   deviance.sarlm           spdep
#>   coef.sarlm               spdep
#>   vcov.sarlm               spdep
#>   fitted.sarlm             spdep
#>   logLik.sarlm             spdep
#>   anova.sarlm              spdep
#>   predict.sarlm            spdep
#>   print.summary.sarlm      spdep
#>   print.sarlm.pred         spdep
#>   as.data.frame.sarlm.pred spdep
#>   residuals.spautolm       spdep
#>   deviance.spautolm        spdep
#>   coef.spautolm            spdep
#>   fitted.spautolm          spdep
#>   print.spautolm           spdep
#>   summary.spautolm         spdep
#>   logLik.spautolm          spdep
#>   print.summary.spautolm   spdep
#>   print.WXImpact           spdep
#>   summary.WXImpact         spdep
#>   print.summary.WXImpact   spdep
#>   predict.SLX              spdep

#> [1] 0.6320073

if (FALSE) {
errorsarlm(yield_detrend ~ 1, wheat, nb2listw(nbr12, style="W"))
}