bhicv {spdep}	R Documentation

Data set with 4 life condition indices of Belo Horizonte region

Description

The data are collected inthe Atlas of condition indices published by the Joao Pinheiro Foundation and UNDP.

Format

A shape polygon object with seven variables:

id

The identificator

Name

Name of city

Population

The population of city

HLCI

Health Life Condition Index

ELCI

Education Life Condition Index

CLCI

Children Life Condition Index

ELCI

Economic Life Condition Index

Examples

### see example in 'skater' function help

[Package spdep version 0.6-6 Index]