geary {spdep}	R Documentation

Compute Geary's C

Description

A simple function to compute Geary's C, called by geary.test and geary.mc;

C = \frac{(n-1)}{2∑_{i=1}^{n}∑_{j=1}^{n}w_{ij}} \frac{∑_{i=1}^{n}∑_{j=1}^{n}w_{ij}(x_i-x_j)^2}{∑_{i=1}^{n}(x_i - \bar{x})^2}

geary.intern is an internal function used to vary the similarity criterion.

Usage

geary(x, listw, n, n1, S0, zero.policy=NULL)

Arguments

x
a numeric vector the same length as the neighbours list in listw

listw
a listw object created for example by nb2listw

n
number of zones

n1
n - 1

S0
global sum of weights

zero.policy
default NULL, use global option value; if TRUE assign zero to the lagged value of zones without neighbours, if FALSE assign NA

Value

a list with

C
Geary's C

K
sample kurtosis of x

Author(s)

Roger Bivand Roger.Bivand@nhh.no

References

Cliff, A. D., Ord, J. K. 1981 Spatial processes, Pion, p. 17.

See Also

geary.test, geary.mc, sp.mantel.mc

Examples

data(oldcol)
col.W <- nb2listw(COL.nb, style="W")
str(geary(COL.OLD$CRIME, col.W, length(COL.nb), length(COL.nb)-1,
 Szero(col.W)))

[Package spdep version 0.6-6 Index]