Impacts in spatial lag models

The calculation of impacts for spatial lag and spatial Durbin models is needed in order to interpret the regression coefficients correctly, because of the spillovers between the terms in these data generation processes (unlike the spatial error model). Methods for “SLX” and Bayesian fitted models are also provided, the former do not need MC simulations, while the latter pass through MCMC draws.

# S3 method for sarlm
impacts(obj, …, tr, R = NULL, listw = NULL, evalues=NULL,
 useHESS = NULL, tol = 1e-06, empirical = FALSE, Q=NULL)
# S3 method for stsls
impacts(obj, …, tr, R = NULL, listw = NULL, evalues=NULL,
 tol = 1e-06, empirical = FALSE, Q=NULL)
# S3 method for gmsar
impacts(obj, …, n = NULL, tr = NULL, R = NULL,
 listw = NULL, evalues=NULL, tol = 1e-06, empirical = FALSE, Q=NULL)
# S3 method for lagmess
impacts(obj, ..., R=NULL, listw=NULL, tol=1e-6,
 empirical=FALSE)
# S3 method for SLX
impacts(obj, ...)
# S3 method for MCMC_sar_g
impacts(obj, ..., tr=NULL, listw=NULL, evalues=NULL, Q=NULL)
# S3 method for MCMC_sem_g
impacts(obj, ..., tr=NULL, listw=NULL, evalues=NULL, Q=NULL)
# S3 method for MCMC_sac_g
impacts(obj, ..., tr=NULL, listw=NULL, evalues=NULL, Q=NULL)
# S3 method for lagImpact
plot(x, ..., choice="direct", trace=FALSE, density=TRUE)
# S3 method for lagImpact
print(x, ..., reportQ=NULL)
# S3 method for lagImpact
summary(object, ..., zstats=FALSE, short=FALSE, reportQ=NULL)
# S3 method for WXImpact
print(x, ...)
# S3 method for WXImpact
summary(object, ..., adjust_k=(attr(object, "type") == "SDEM"))
# S3 method for lagImpact
HPDinterval(obj, prob = 0.95, ..., choice="direct")
intImpacts(rho, beta, P, n, mu, Sigma, irho, drop2beta, bnames, interval,
 type, tr, R, listw, evalues, tol, empirical, Q, icept, iicept, p, mess=FALSE,
 samples=NULL, zero_fill = NULL, dvars = NULL)

Arguments

obj
A spatial regression object created by lagsarlm, lagmess or by lmSLX; in HPDinterval.lagImpact, a lagImpact object

…
Arguments passed through to methods in the coda package

tr
A vector of traces of powers of the spatial weights matrix created using trW, for approximate impact measures; if not given, listw must be given for exact measures (for small to moderate spatial weights matrices); the traces must be for the same spatial weights as were used in fitting the spatial regression, and must be row-standardised

listw
If tr is not given, a spatial weights object as created by nb2listw; they must be the same spatial weights as were used in fitting the spatial regression, but do not have to be row-standardised

evalues
vector of eigenvalues of spatial weights matrix for impacts calculations

n
defaults to length(obj$residuals); in the method for gmsar objects it may be used in panel settings to compute the impacts for cross-sectional weights only, suggested by Angela Parenti

R
If given, simulations are used to compute distributions for the impact measures, returned as mcmc objects; the objects are used for convenience but are not output by an MCMC process

useHESS
Use the Hessian approximation (if available) even if the asymptotic coefficient covariance matrix is available; used for comparing methods

tol
Argument passed to mvrnorm: tolerance (relative to largest variance) for numerical lack of positive-definiteness in the coefficient covariance matrix

empirical
Argument passed to mvrnorm (default FALSE): if true, the coefficients and their covariance matrix specify the empirical not population mean and covariance matrix

Q
default NULL, else an integer number of cumulative power series impacts to calculate if tr is given

reportQ
default NULL; if TRUE and Q given as an argument to impacts, report impact components

x, object
lagImpact objects created by impacts methods

zstats
default FALSE, if TRUE, also return z-values and p-values for the impacts based on the simulations

short
default FALSE, if TRUE passed to the print summary method to omit printing of the mcmc summaries

choice
One of three impacts: direct, indirect, or total

trace
Argument passed to plot.mcmc: plot trace plots

density
Argument passed to plot.mcmc: plot density plots

prob
Argument passed to HPDinterval.mcmc: a numeric scalar in the interval (0,1) giving the target probability content of the intervals

adjust_k
default TRUE if SDEM else FALSE, adjust internal OLS SDEM standard errors by dividing by n rather than (n-k) (default changed and bug fixed after 0.7-8; standard errors now ML in SDEM summary and impacts summary and identical - for SLX use FALSE)

rho, beta, P, mu, Sigma, irho, drop2beta, bnames, interval, type, icept, iicept, p, mess, samples, zero_fill, dvars
internal arguments shared inside impacts methods

Details

If called without R being set, the method returns the direct, indirect and total impacts for the variables in the model, for the variables themselves in tha spatial lag model case, for the variables and their spatial lags in the spatial Durbin (mixed) model case. The spatial lag impact measures are computed using eq. 2.46 (LeSage and Pace, 2009, p. 38), either using the exact dense matrix (when listw is given), or traces of powers of the weights matrix (when tr is given). When the traces are created by powering sparse matrices, the exact and the trace methods should give very similar results, unless the number of powers used is very small, or the spatial coefficient is close to its bounds.

If R is given, simulations will be used to create distributions for the impact measures, provided that the fitted model object contains a coefficient covariance matrix. The simulations are made using mvrnorm with the coefficients and their covariance matrix from the fitted model.

The simulations are stored as mcmc objects as defined in the coda package; the objects are used for convenience but are not output by an MCMC process. The simulated values of the coefficients are checked to see that the spatial coefficient remains within its valid interval --- draws outside the interval are discarded.

If a model is fitted with the “Durbin=” set to a formula subsetting the explanatory variables, the impacts object returned reports Durbin impacts for variables included in the formula and lag impacts for the other variables.

When Q and tr are given, addition impact component results are provided for each step in the traces of powers of the weights matrix up to and including the Q'th power. This increases computing time because the output object is substantially increased in size in proportion to the size of Q.

The method for gmsar objects is only for those of type SARAR output by gstsls, and assume that the spatial error coefficient is fixed, and thus omitted from the coefficients and covariance matrix used for simulation.

Value

An object of class lagImpact.

If no simulation is carried out, the object returned is a list with:

direct

numeric vector

indirect

numeric vector

total

numeric vector

and a matching Qres list attribute if Q was given. If simulation is carried out, the object returned is a list with:
res

a list with three components as for the non-simulation case, with a matching Qres list attribute if Q was given

sres

a list with three mcmc matrices, for the direct, indirect and total impacts with a matching Qmcmc list attribute if Q was given

References

LeSage J and RK Pace (2009) Introduction to Spatial Econometrics. CRC Press, Boca Raton, pp. 33--42, 114--115; LeSage J and MM Fischer (2008) Spatial growth regressions: model specification, estimation and interpretation. Spatial Economic Analysis 3 (3), pp. 275--304.

Roger Bivand, Gianfranco Piras (2015). Comparing Implementations of Estimation Methods for Spatial Econometrics. Journal of Statistical Software, 63(18), 1-36. https://www.jstatsoft.org/v63/i18/.

See also

Examples

columbus <- st_read(system.file("shapes/columbus.shp", package="spData")[1], quiet=TRUE)
col.gal.nb <- read.gal(system.file("weights/columbus.gal", package="spData")[1])
listw <- nb2listw(col.gal.nb)
ev <- eigenw(listw)
lobj <- lagsarlm(CRIME ~ INC + HOVAL, columbus, listw,
 control=list(pre_eig=ev))
summary(lobj)

#> 
#> Call:lagsarlm(formula = CRIME ~ INC + HOVAL, data = columbus, listw = listw, 
#>     control = list(pre_eig = ev))
#> 
#> Residuals:
#>         Min          1Q      Median          3Q         Max 
#> -37.4497094  -5.4565567   0.0016388   6.7159553  24.7107975 
#> 
#> Type: lag 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) 46.851430   7.314754  6.4051 1.503e-10
#> INC         -1.073533   0.310872 -3.4533 0.0005538
#> HOVAL       -0.269997   0.090128 -2.9957 0.0027381
#> 
#> Rho: 0.40389, LR test value: 8.4179, p-value: 0.0037154
#> Asymptotic standard error: 0.12071
#>     z-value: 3.3459, p-value: 0.00082027
#> Wald statistic: 11.195, p-value: 0.00082027
#> 
#> Log likelihood: -183.1683 for lag model
#> ML residual variance (sigma squared): 99.164, (sigma: 9.9581)
#> Number of observations: 49 
#> Number of parameters estimated: 5 
#> AIC: 376.34, (AIC for lm: 382.75)
#> LM test for residual autocorrelation
#> test value: 0.19184, p-value: 0.66139
#> 

mobj <- lagsarlm(CRIME ~ INC + HOVAL, columbus, listw, Durbin=TRUE,
 control=list(pre_eig=ev))
summary(mobj)

#> 
#> Call:lagsarlm(formula = CRIME ~ INC + HOVAL, data = columbus, listw = listw, 
#>     Durbin = TRUE, control = list(pre_eig = ev))
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -37.15904  -6.62594  -0.39823   6.57561  23.62757 
#> 
#> Type: mixed 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) 45.592896  13.128680  3.4728 0.0005151
#> INC         -0.939088   0.338229 -2.7765 0.0054950
#> HOVAL       -0.299605   0.090843 -3.2980 0.0009736
#> lag.INC     -0.618375   0.577052 -1.0716 0.2838954
#> lag.HOVAL    0.266615   0.183971  1.4492 0.1472760
#> 
#> Rho: 0.38251, LR test value: 4.1648, p-value: 0.041272
#> Asymptotic standard error: 0.16237
#>     z-value: 2.3557, p-value: 0.018488
#> Wald statistic: 5.5493, p-value: 0.018488
#> 
#> Log likelihood: -182.0161 for mixed model
#> ML residual variance (sigma squared): 95.051, (sigma: 9.7494)
#> Number of observations: 49 
#> Number of parameters estimated: 7 
#> AIC: 378.03, (AIC for lm: 380.2)
#> LM test for residual autocorrelation
#> test value: 0.101, p-value: 0.75063
#> 

mobj1 <- lagsarlm(CRIME ~ INC + HOVAL, columbus, listw, Durbin= ~ INC,
 control=list(pre_eig=ev))
summary(mobj1)

#> 
#> Call:lagsarlm(formula = CRIME ~ INC + HOVAL, data = columbus, listw = listw, 
#>     Durbin = ~INC, control = list(pre_eig = ev))
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -37.62163  -6.13108   0.11211   6.70682  24.76777 
#> 
#> Type: mixed 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) 51.951208  12.577338  4.1305 3.619e-05
#> INC         -1.038812   0.337656 -3.0765  0.002094
#> HOVAL       -0.269345   0.090406 -2.9793  0.002889
#> lag.INC     -0.254653   0.544298 -0.4679  0.639888
#> 
#> Rho: 0.35028, LR test value: 3.4351, p-value: 0.063823
#> Asymptotic standard error: 0.1617
#>     z-value: 2.1662, p-value: 0.030293
#> Wald statistic: 4.6926, p-value: 0.030293
#> 
#> Log likelihood: -183.065 for mixed model
#> ML residual variance (sigma squared): 99.846, (sigma: 9.9923)
#> Number of observations: 49 
#> Number of parameters estimated: 6 
#> AIC: 378.13, (AIC for lm: 379.57)
#> LM test for residual autocorrelation
#> test value: 2.5646, p-value: 0.10928
#> 

W <- as(listw, "CsparseMatrix")
trMatc <- trW(W, type="mult")
trMC <- trW(W, type="MC")
set.seed(1)
impacts(lobj, listw=listw)

#> Impact measures (lag, exact):
#>           Direct   Indirect      Total
#> INC   -1.1225155 -0.6783818 -1.8008973
#> HOVAL -0.2823163 -0.1706152 -0.4529315

impacts(lobj, tr=trMatc)

#> Impact measures (lag, trace):
#>           Direct   Indirect      Total
#> INC   -1.1225155 -0.6783818 -1.8008973
#> HOVAL -0.2823163 -0.1706152 -0.4529315

impacts(lobj, tr=trMC)

#> Impact measures (lag, trace):
#>           Direct   Indirect      Total
#> INC   -1.1211859 -0.6797114 -1.8008973
#> HOVAL -0.2819819 -0.1709496 -0.4529315

impacts(lobj, evalues=ev)

#> Impact measures (lag, evalues):
#>           Direct   Indirect      Total
#> INC   -1.1225155 -0.6783818 -1.8008973
#> HOVAL -0.2823163 -0.1706152 -0.4529315

lobj1 <- stsls(CRIME ~ INC + HOVAL, columbus, listw)
loobj1 <- impacts(lobj1, R=200, tr=trMatc)
summary(loobj1, zstats=TRUE, short=TRUE)

#> Impact measures (lag, trace):
#>           Direct   Indirect      Total
#> INC   -1.0687585 -0.7790438 -1.8478023
#> HOVAL -0.2858263 -0.2083456 -0.4941719
#> ========================================================
#> Simulation results (asymptotic IV variance matrix):
#> ========================================================
#> Simulated standard errors
#>          Direct  Indirect     Total
#> INC   0.3755271 0.7998518 0.9041439
#> HOVAL 0.1072510 0.3733382 0.4372305
#> 
#> Simulated z-values:
#>          Direct   Indirect     Total
#> INC   -2.886403 -1.1699705 -2.233854
#> HOVAL -2.757889 -0.8259039 -1.381715
#> 
#> Simulated p-values:
#>       Direct    Indirect Total   
#> INC   0.0038967 0.24201  0.025493
#> HOVAL 0.0058176 0.40886  0.167059

loobj2 <- impacts(lobj1, R=200, evalues=ev)
summary(loobj2, zstats=TRUE, short=TRUE)

#> Impact measures (lag, evalues):
#>           Direct   Indirect      Total
#> INC   -1.0687585 -0.7790438 -1.8478023
#> HOVAL -0.2858263 -0.2083456 -0.4941719
#> ========================================================
#> Simulation results (asymptotic IV variance matrix):
#> ========================================================
#> Simulated standard errors
#>          Direct  Indirect     Total
#> INC   0.4106540 0.5926178 0.7088727
#> HOVAL 0.1327871 0.4825350 0.5812652
#> 
#> Simulated z-values:
#>          Direct   Indirect      Total
#> INC   -2.561322 -1.3695611 -2.6287421
#> HOVAL -2.194698 -0.5619779 -0.9678914
#> 
#> Simulated p-values:
#>       Direct   Indirect Total    
#> INC   0.010427 0.17082  0.0085701
#> HOVAL 0.028185 0.57413  0.3330986

library(coda)
HPDinterval(loobj1)

#>            lower       upper
#> INC   -1.7286485 -0.26439087
#> HOVAL -0.4840639 -0.09027924
#> attr(,"Probability")
#> [1] 0.95

lobj1r <- stsls(CRIME ~ INC + HOVAL, columbus, listw, robust=TRUE)
loobj1r <- impacts(lobj1r, tr=trMatc, R=200)
summary(loobj1r, zstats=TRUE, short=TRUE)

#> Impact measures (lag, trace):
#>           Direct   Indirect      Total
#> INC   -1.0687585 -0.7790438 -1.8478023
#> HOVAL -0.2858263 -0.2083456 -0.4941719
#> ========================================================
#> Simulation results (HC0 IV variance matrix):
#> ========================================================
#> Simulated standard errors
#>          Direct  Indirect     Total
#> INC   0.5274394 1.0931855 1.3983255
#> HOVAL 0.1958572 0.2426985 0.4008787
#> 
#> Simulated z-values:
#>         Direct   Indirect     Total
#> INC   -2.09808 -0.8223885 -1.434311
#> HOVAL -1.46588 -0.9639231 -1.299759
#> 
#> Simulated p-values:
#>       Direct   Indirect Total  
#> INC   0.035898 0.41086  0.15148
#> HOVAL 0.142681 0.33508  0.19368

lobjIQ5 <- impacts(lobj, tr=trMatc, R=200, Q=5)
summary(lobjIQ5, zstats=TRUE, short=TRUE)

#> Impact measures (lag, trace):
#>           Direct   Indirect      Total
#> INC   -1.1225155 -0.6783818 -1.8008973
#> HOVAL -0.2823163 -0.1706152 -0.4529315
#> ========================================================
#> Simulation results (asymptotic variance matrix):
#> ========================================================
#> Simulated standard errors
#>          Direct  Indirect     Total
#> INC   0.3462030 0.4013775 0.6174125
#> HOVAL 0.1055706 0.1320129 0.2136489
#> 
#> Simulated z-values:
#>          Direct  Indirect     Total
#> INC   -3.228212 -1.674543 -2.898777
#> HOVAL -2.667191 -1.371279 -2.165252
#> 
#> Simulated p-values:
#>       Direct    Indirect Total    
#> INC   0.0012457 0.094024 0.0037462
#> HOVAL 0.0076488 0.170288 0.0303684

summary(lobjIQ5, zstats=TRUE, short=TRUE, reportQ=TRUE)

#> Impact measures (lag, trace):
#>           Direct   Indirect      Total
#> INC   -1.1225155 -0.6783818 -1.8008973
#> HOVAL -0.2823163 -0.1706152 -0.4529315
#> =================================
#> Impact components
#> $direct
#>             INC         HOVAL
#> Q1 -1.073533437 -0.2699971234
#> Q2  0.000000000  0.0000000000
#> Q3 -0.038985418 -0.0098049584
#> Q4 -0.005269655 -0.0013253352
#> Q5 -0.003276080 -0.0008239446
#> 
#> $indirect
#>            INC        HOVAL
#> Q1  0.00000000  0.000000000
#> Q2 -0.43358911 -0.109049060
#> Q3 -0.13613676 -0.034238835
#> Q4 -0.06546039 -0.016463500
#> Q5 -0.02529106 -0.006360783
#> 
#> $total
#>            INC        HOVAL
#> Q1 -1.07353344 -0.269997123
#> Q2 -0.43358911 -0.109049060
#> Q3 -0.17512218 -0.044043793
#> Q4 -0.07073005 -0.017788835
#> Q5 -0.02856714 -0.007184727
#> 
#> ========================================================
#> Simulation results (asymptotic variance matrix):
#> ========================================================
#> Simulated standard errors
#>          Direct  Indirect     Total
#> INC   0.3462030 0.4013775 0.6174125
#> HOVAL 0.1055706 0.1320129 0.2136489
#> 
#> Simulated z-values:
#>          Direct  Indirect     Total
#> INC   -3.228212 -1.674543 -2.898777
#> HOVAL -2.667191 -1.371279 -2.165252
#> 
#> Simulated p-values:
#>       Direct    Indirect Total    
#> INC   0.0012457 0.094024 0.0037462
#> HOVAL 0.0076488 0.170288 0.0303684
#> ========================================================
#> Simulated impact components z-values:
#> $Direct
#>           INC      HOVAL
#> Q1 -3.1454607 -2.6874102
#> Q2        NaN        NaN
#> Q3 -1.5907172 -1.3149445
#> Q4 -1.0828785 -0.9578991
#> Q5 -0.7854446 -0.7335789
#> 
#> $Indirect
#>           INC      HOVAL
#> Q1        NaN        NaN
#> Q2 -2.5512196 -1.9340091
#> Q3 -1.5907172 -1.3149445
#> Q4 -1.0828785 -0.9578991
#> Q5 -0.7854446 -0.7335789
#> 
#> $Total
#>           INC      HOVAL
#> Q1 -3.1454607 -2.6874102
#> Q2 -2.5512196 -1.9340091
#> Q3 -1.5907172 -1.3149445
#> Q4 -1.0828785 -0.9578991
#> Q5 -0.7854446 -0.7335789
#> 
#> 
#> Simulated impact components p-values:
#> $Direct
#>    INC       HOVAL    
#> Q1 0.0016583 0.0072008
#> Q2 NA        NA       
#> Q3 0.1116732 0.1885286
#> Q4 0.2788624 0.3381136
#> Q5 0.4321930 0.4632054
#> 
#> $Indirect
#>    INC      HOVAL   
#> Q1 NA       NA      
#> Q2 0.010735 0.053112
#> Q3 0.111673 0.188529
#> Q4 0.278862 0.338114
#> Q5 0.432193 0.463205
#> 
#> $Total
#>    INC       HOVAL    
#> Q1 0.0016583 0.0072008
#> Q2 0.0107347 0.0531120
#> Q3 0.1116732 0.1885286
#> Q4 0.2788624 0.3381136
#> Q5 0.4321930 0.4632054
#> 

impacts(mobj, listw=listw)

#> Impact measures (mixed, exact):
#>           Direct   Indirect       Total
#> INC   -1.0418080 -1.4804246 -2.52223255
#> HOVAL -0.2836325  0.2302055 -0.05342697

impacts(mobj, tr=trMatc)

#> Impact measures (mixed, trace):
#>           Direct   Indirect       Total
#> INC   -1.0418080 -1.4804246 -2.52223255
#> HOVAL -0.2836325  0.2302055 -0.05342697

impacts(mobj, tr=trMC)

#> Impact measures (mixed, trace):
#>          Direct   Indirect       Total
#> INC   -1.039187 -1.4830455 -2.52223255
#> HOVAL -0.284040  0.2306131 -0.05342697

impacts(mobj1, tr=trMatc)

#> Impact measures (mixed, trace):
#>           Direct   Indirect      Total
#> INC   -1.0968247 -0.8939687 -1.9907934
#> HOVAL -0.2781941 -0.1363596 -0.4145537

impacts(mobj1, listw=listw)

#> Impact measures (mixed, exact):
#>           Direct   Indirect      Total
#> INC   -1.0968247 -0.8939687 -1.9907934
#> HOVAL -0.2781941 -0.1363596 -0.4145537

cat(try(impacts(mobj, evalues=ev), silent=TRUE), "\n")

#> Error in intImpacts(rho = rho, beta = beta, P = P, n = n, mu = mu, Sigma = Sigma,  : 
#>   eigenvalue mixed impacts not available
#>  

summary(impacts(mobj, tr=trMatc, R=200), short=TRUE, zstats=TRUE)

#> Impact measures (mixed, trace):
#>           Direct   Indirect       Total
#> INC   -1.0418080 -1.4804246 -2.52223255
#> HOVAL -0.2836325  0.2302055 -0.05342697
#> ========================================================
#> Simulation results (asymptotic variance matrix):
#> ========================================================
#> Simulated standard errors
#>          Direct  Indirect     Total
#> INC   0.3138781 1.0464126 1.1155472
#> HOVAL 0.1010350 0.4021256 0.4401912
#> 
#> Simulated z-values:
#>          Direct   Indirect       Total
#> INC   -3.263683 -1.4485963 -2.27711376
#> HOVAL -2.824144  0.6175382 -0.08407608
#> 
#> Simulated p-values:
#>       Direct    Indirect Total   
#> INC   0.0010997 0.14745  0.022779
#> HOVAL 0.0047407 0.53688  0.932996

summary(impacts(mobj1, tr=trMatc, R=200), short=TRUE, zstats=TRUE)

#> Impact measures (mixed, trace):
#>           Direct   Indirect      Total
#> INC   -1.0968247 -0.8939687 -1.9907934
#> HOVAL -0.2781941 -0.1363596 -0.4145537
#> ========================================================
#> Simulation results (asymptotic variance matrix):
#> ========================================================
#> Simulated standard errors
#>           Direct  Indirect     Total
#> INC   0.32590243 0.5584296 0.6593439
#> HOVAL 0.09715763 0.1264950 0.1933227
#> 
#> Simulated z-values:
#>          Direct  Indirect     Total
#> INC   -3.391265 -1.640173 -3.065384
#> HOVAL -2.891161 -1.329411 -2.322863
#> 
#> Simulated p-values:
#>       Direct     Indirect Total    
#> INC   0.00069571 0.10097  0.0021739
#> HOVAL 0.00383821 0.18371  0.0201865

xobj <- lmSLX(CRIME ~ INC + HOVAL, columbus, listw)
summary(impacts(xobj))

#> Impact measures (SLX, estimable, n-k):
#>           Direct   Indirect       Total
#> INC   -1.1081273 -1.3834468 -2.49157410
#> HOVAL -0.2949095  0.2261538 -0.06875574
#> ========================================================
#> Standard errors:
#>          Direct  Indirect     Total
#> INC   0.3749956 0.5591789 0.4928197
#> HOVAL 0.1013524 0.2026169 0.2049626
#> ========================================================
#> Z-values:
#>          Direct  Indirect      Total
#> INC   -2.955041 -2.474068 -5.0557523
#> HOVAL -2.909744  1.116164 -0.3354551
#> 
#> p-values:
#>       Direct    Indirect Total    
#> INC   0.0031263 0.013358 4.287e-07
#> HOVAL 0.0036172 0.264352 0.73728  
#> 

eobj <- errorsarlm(CRIME ~ INC + HOVAL, columbus, listw, etype="emixed")
summary(impacts(eobj), adjust_k=TRUE)

#> Impact measures (SDEM, estimable, n):
#>           Direct   Indirect      Total
#> INC   -1.0695301 -1.1967736 -2.2663036
#> HOVAL -0.2803441  0.1467585 -0.1335856
#> ========================================================
#> Standard errors:
#>           Direct  Indirect     Total
#> INC   0.32471853 0.5689676 0.6209448
#> HOVAL 0.09180929 0.2008722 0.2323518
#> ========================================================
#> Z-values:
#>          Direct   Indirect      Total
#> INC   -3.293714 -2.1034125 -3.6497667
#> HOVAL -3.053548  0.7306064 -0.5749284
#> 
#> p-values:
#>       Direct     Indirect Total     
#> INC   0.00098873 0.03543  0.00026248
#> HOVAL 0.00226152 0.46502  0.56533972
#> 

mobj1 <- lagsarlm(CRIME ~ INC + HOVAL, columbus, listw, type="mixed",
method="Matrix", control=list(fdHess=TRUE))
summary(mobj1)

#> 
#> Call:lagsarlm(formula = CRIME ~ INC + HOVAL, data = columbus, listw = listw, 
#>     type = "mixed", method = "Matrix", control = list(fdHess = TRUE))
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -37.15904  -6.62594  -0.39823   6.57561  23.62757 
#> 
#> Type: mixed 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) 45.592896  14.240041  3.2017 0.0013660
#> INC         -0.939088   0.343069 -2.7373 0.0061942
#> HOVAL       -0.299605   0.090393 -3.3145 0.0009182
#> lag.INC     -0.618375   0.607708 -1.0176 0.3088903
#> lag.HOVAL    0.266615   0.181524  1.4688 0.1418998
#> 
#> Rho: 0.38251, LR test value: 4.1648, p-value: 0.041272
#> Asymptotic standard error: 0.17375
#>     z-value: 2.2015, p-value: 0.027702
#> Wald statistic: 4.8465, p-value: 0.027702
#> 
#> Log likelihood: -182.0161 for mixed model
#> ML residual variance (sigma squared): 95.051, (sigma: 9.7494)
#> Number of observations: 49 
#> Number of parameters estimated: 7 
#> AIC: 378.03, (AIC for lm: 380.2)
#> LM test for residual autocorrelation
#> test value: 0.101, p-value: 0.75063
#> 

set.seed(1)
summary(impacts(mobj1, tr=trMatc, R=1000), zstats=TRUE, short=TRUE)

#> Impact measures (mixed, trace):
#>           Direct   Indirect       Total
#> INC   -1.0418080 -1.4804246 -2.52223255
#> HOVAL -0.2836325  0.2302055 -0.05342697
#> ========================================================
#> Simulation results (asymptotic variance matrix):
#> ========================================================
#> Simulated standard errors
#>           Direct  Indirect     Total
#> INC   0.34686331 0.8324743 0.8959688
#> HOVAL 0.09085583 0.3131242 0.3483785
#> 
#> Simulated z-values:
#>          Direct   Indirect      Total
#> INC   -3.023902 -1.8284275 -2.8695190
#> HOVAL -3.120537  0.7910739 -0.1028036
#> 
#> Simulated p-values:
#>       Direct    Indirect Total   
#> INC   0.0024954 0.067485 0.004111
#> HOVAL 0.0018052 0.428901 0.918119

summary(impacts(mobj, tr=trMatc, R=1000), zstats=TRUE, short=TRUE)

#> Impact measures (mixed, trace):
#>           Direct   Indirect       Total
#> INC   -1.0418080 -1.4804246 -2.52223255
#> HOVAL -0.2836325  0.2302055 -0.05342697
#> ========================================================
#> Simulation results (asymptotic variance matrix):
#> ========================================================
#> Simulated standard errors
#>          Direct  Indirect     Total
#> INC   0.3123750 0.8291512 0.8821167
#> HOVAL 0.0922758 0.2961092 0.3267773
#> 
#> Simulated z-values:
#>          Direct   Indirect      Total
#> INC   -3.330304 -1.8060414 -2.8769263
#> HOVAL -3.099811  0.8140556 -0.1376722
#> 
#> Simulated p-values:
#>       Direct     Indirect Total    
#> INC   0.00086751 0.070912 0.0040157
#> HOVAL 0.00193644 0.415613 0.8904995

mobj2 <- lagsarlm(CRIME ~ INC + HOVAL, columbus, listw, type="mixed",
method="Matrix", control=list(fdHess=TRUE, optimHess=TRUE))
summary(impacts(mobj2, tr=trMatc, R=1000), zstats=TRUE, short=TRUE)

#> Impact measures (mixed, trace):
#>           Direct   Indirect       Total
#> INC   -1.0418080 -1.4804246 -2.52223255
#> HOVAL -0.2836325  0.2302055 -0.05342697
#> ========================================================
#> Simulation results (asymptotic variance matrix):
#> ========================================================
#> Simulated standard errors
#>           Direct  Indirect     Total
#> INC   0.32853332 0.8044059 0.8615603
#> HOVAL 0.09291975 0.3169361 0.3571304
#> 
#> Simulated z-values:
#>          Direct   Indirect      Total
#> INC   -3.231204 -1.8690203 -2.9771674
#> HOVAL -3.026494  0.7323897 -0.1374858
#> 
#> Simulated p-values:
#>       Direct    Indirect Total    
#> INC   0.0012327 0.06162  0.0029093
#> HOVAL 0.0024741 0.46393  0.8906468

mobj3 <- lagsarlm(CRIME ~ INC + HOVAL, columbus, listw, type="mixed",
method="spam", control=list(fdHess=TRUE))
summary(impacts(mobj3, tr=trMatc, R=1000), zstats=TRUE, short=TRUE)

#> Impact measures (mixed, trace):
#>           Direct   Indirect       Total
#> INC   -1.0418080 -1.4804246 -2.52223255
#> HOVAL -0.2836325  0.2302055 -0.05342697
#> ========================================================
#> Simulation results (asymptotic variance matrix):
#> ========================================================
#> Simulated standard errors
#>           Direct  Indirect     Total
#> INC   0.32866009 0.8664620 0.9215792
#> HOVAL 0.09837849 0.3110899 0.3492674
#> 
#> Simulated z-values:
#>          Direct   Indirect     Total
#> INC   -3.190433 -1.7325351 -2.766712
#> HOVAL -2.890221  0.7694814 -0.128720
#> 
#> Simulated p-values:
#>       Direct    Indirect Total    
#> INC   0.0014206 0.083178 0.0056625
#> HOVAL 0.0038497 0.441608 0.8975792

data(boston, package="spData")
Wb <- as(nb2listw(boston.soi), "CsparseMatrix")
trMatb <- trW(Wb, type="mult")
gp2mMi <- lagsarlm(log(CMEDV) ~ CRIM + ZN + INDUS + CHAS + I(NOX^2) +
I(RM^2) +  AGE + log(DIS) + log(RAD) + TAX + PTRATIO + B + log(LSTAT),
data=boston.c, nb2listw(boston.soi), type="mixed", method="Matrix",
control=list(fdHess=TRUE), trs=trMatb)
summary(gp2mMi)

#> 
#> Call:lagsarlm(formula = log(CMEDV) ~ CRIM + ZN + INDUS + CHAS + I(NOX^2) + 
#>     I(RM^2) + AGE + log(DIS) + log(RAD) + TAX + PTRATIO + B + 
#>     log(LSTAT), data = boston.c, listw = nb2listw(boston.soi), 
#>     type = "mixed", method = "Matrix", trs = trMatb, control = list(fdHess = TRUE))
#> 
#> Residuals:
#>        Min         1Q     Median         3Q        Max 
#> -0.6316833 -0.0629790 -0.0090776  0.0682421  0.6991072 
#> 
#> Type: mixed 
#> Coefficients: (asymptotic standard errors) 
#>                   Estimate  Std. Error  z value  Pr(>|z|)
#> (Intercept)     1.89816218  0.24400182   7.7793 7.327e-15
#> CRIM           -0.00571021  0.00093857  -6.0839 1.173e-09
#> ZN              0.00069091  0.00051874   1.3319 0.1828931
#> INDUS          -0.00111343  0.00307380  -0.3622 0.7171778
#> CHAS1          -0.04163225  0.02738839  -1.5201 0.1284937
#> I(NOX^2)       -0.01034950  0.19358633  -0.0535 0.9573639
#> I(RM^2)         0.00794979  0.00102109   7.7856 6.883e-15
#> AGE            -0.00128789  0.00048929  -2.6322 0.0084838
#> log(DIS)       -0.12404108  0.09510145  -1.3043 0.1921304
#> log(RAD)        0.05863502  0.02257529   2.5973 0.0093957
#> TAX            -0.00049084  0.00012146  -4.0412 5.317e-05
#> PTRATIO        -0.01319853  0.00595331  -2.2170 0.0266227
#> B               0.00056383  0.00011084   5.0867 3.643e-07
#> log(LSTAT)     -0.24724454  0.02265149 -10.9152 < 2.2e-16
#> lag.CRIM       -0.00464215  0.00173900  -2.6694 0.0075978
#> lag.ZN         -0.00037937  0.00070703  -0.5366 0.5915659
#> lag.INDUS       0.00025064  0.00385911   0.0649 0.9482165
#> lag.CHAS1       0.12518252  0.04083949   3.0652 0.0021750
#> lag.I(NOX^2)   -0.38640401  0.22253428  -1.7364 0.0824967
#> lag.I(RM^2)    -0.00451252  0.00148919  -3.0302 0.0024440
#> lag.AGE         0.00149678  0.00068470   2.1860 0.0288128
#> lag.log(DIS)   -0.00453785  0.10056961  -0.0451 0.9640105
#> lag.log(RAD)   -0.00940702  0.03127787  -0.3008 0.7636001
#> lag.TAX         0.00041083  0.00017859   2.3004 0.0214237
#> lag.PTRATIO     0.00060355  0.00789994   0.0764 0.9391011
#> lag.B          -0.00050781  0.00014107  -3.5996 0.0003187
#> lag.log(LSTAT)  0.09846781  0.03574182   2.7550 0.0058697
#> 
#> Rho: 0.59578, LR test value: 181.68, p-value: < 2.22e-16
#> Asymptotic standard error: 0.037474
#>     z-value: 15.899, p-value: < 2.22e-16
#> Wald statistic: 252.76, p-value: < 2.22e-16
#> 
#> Log likelihood: 300.6131 for mixed model
#> ML residual variance (sigma squared): 0.016011, (sigma: 0.12654)
#> Number of observations: 506 
#> Number of parameters estimated: 29 
#> AIC: -543.23, (AIC for lm: -363.55)
#> LM test for residual autocorrelation
#> test value: 29.772, p-value: 4.8604e-08
#> 

summary(impacts(gp2mMi, tr=trMatb, R=1000), zstats=TRUE, short=TRUE)

#> Impact measures (mixed, trace):
#>                   Direct      Indirect         Total
#> CRIM       -0.0074555753 -1.815485e-02 -0.0256104225
#> ZN          0.0006979073  7.279849e-05  0.0007707058
#> INDUS      -0.0012029822 -9.314672e-04 -0.0021344494
#> CHAS1      -0.0198526436  2.265454e-01  0.2066927455
#> I(NOX^2)   -0.0955268264 -8.859909e-01 -0.9815177559
#> I(RM^2)     0.0079983430  5.050177e-04  0.0085033608
#> AGE        -0.0011296134  1.646365e-03  0.0005167515
#> log(DIS)   -0.1410601687 -1.770278e-01 -0.3180879407
#> log(RAD)    0.0641735548  5.761026e-02  0.1217838138
#> TAX        -0.0004651543  2.672120e-04 -0.0001979424
#> PTRATIO    -0.0147737151 -1.638465e-02 -0.0311583673
#> B           0.0005265343 -3.879424e-04  0.0001385920
#> log(LSTAT) -0.2578403213 -1.102144e-01 -0.3680547068
#> ========================================================
#> Simulation results (asymptotic variance matrix):
#> ========================================================
#> Simulated standard errors
#>                  Direct     Indirect        Total
#> CRIM       0.0010076798 0.0036288635 0.0040404924
#> ZN         0.0004892696 0.0012374985 0.0012596482
#> INDUS      0.0028820758 0.0060784278 0.0057782835
#> CHAS1      0.0271174438 0.0816949094 0.0902306757
#> I(NOX^2)   0.1798525308 0.2954129372 0.2426994803
#> I(RM^2)    0.0011004204 0.0031441374 0.0036256801
#> AGE        0.0004759027 0.0012824325 0.0013506324
#> log(DIS)   0.0884514493 0.1106832532 0.0757292416
#> log(RAD)   0.0197638019 0.0513249132 0.0502885291
#> TAX        0.0001125321 0.0003293654 0.0003385003
#> PTRATIO    0.0054150585 0.0127889718 0.0122254626
#> B          0.0001047613 0.0002376555 0.0002352788
#> log(LSTAT) 0.0224736606 0.0610705732 0.0660452868
#> 
#> Simulated z-values:
#>                 Direct   Indirect      Total
#> CRIM        -7.4241338 -5.0475149 -6.3848388
#> ZN           1.4320070  0.1005269  0.6549761
#> INDUS       -0.4184563 -0.1088975 -0.3232704
#> CHAS1       -0.7260662  2.7907466  2.3085356
#> I(NOX^2)    -0.5230875 -3.0572802 -4.1089447
#> I(RM^2)      7.3364331  0.1789518  2.3818456
#> AGE         -2.4211451  1.2544530  0.3380060
#> log(DIS)    -1.5778892 -1.6651491 -4.2766929
#> log(RAD)     3.2376717  1.1761160  2.4727856
#> TAX         -4.1072661  0.7673289 -0.6188112
#> PTRATIO     -2.7323833 -1.2913715 -2.5611570
#> B            5.0195551 -1.6302856  0.5882763
#> log(LSTAT) -11.4429268 -1.7773204 -5.5372071
#> 
#> Simulated p-values:
#>            Direct     Indirect   Total     
#> CRIM       1.1346e-13 4.4759e-07 1.7158e-10
#> ZN         0.1521418  0.9199260  0.512483  
#> INDUS      0.6756135  0.9132838  0.746490  
#> CHAS1      0.4677982  0.0052587  0.020969  
#> I(NOX^2)   0.6009133  0.0022336  3.9747e-05
#> I(RM^2)    2.1938e-13 0.8579755  0.017226  
#> AGE        0.0154717  0.2096774  0.735359  
#> log(DIS)   0.1145911  0.0958830  1.8969e-05
#> log(RAD)   0.0012051  0.2395485  0.013406  
#> TAX        4.0037e-05 0.4428860  0.536041  
#> PTRATIO    0.0062878  0.1965749  0.010432  
#> B          5.1791e-07 0.1030411  0.556347  
#> log(LSTAT) < 2.22e-16 0.0755155  3.0733e-08

data(house, package="spData")
lw <- nb2listw(LO_nb)
form <- formula(log(price) ~ age + I(age^2) + I(age^3) + log(lotsize) +
   rooms + log(TLA) + beds + syear)
lobj <- lagsarlm(form, house, lw, method="Matrix",
 control=list(fdHess=TRUE), trs=trMat)
summary(lobj)

#> 
#> Call:lagsarlm(formula = form, data = house, listw = lw, method = "Matrix", 
#>     trs = trMat, control = list(fdHess = TRUE))
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -2.488174 -0.128732  0.030689  0.174778  2.845118 
#> 
#> Type: lag 
#> Coefficients: (numerical Hessian approximate standard errors) 
#>                Estimate Std. Error  z value  Pr(>|z|)
#> (Intercept)   0.2583272  0.0656170   3.9369 8.254e-05
#> age           1.3084686  0.0563647  23.2143 < 2.2e-16
#> I(age^2)     -2.3213256  0.1024831 -22.6508 < 2.2e-16
#> I(age^3)      0.6548946  0.0551842  11.8674 < 2.2e-16
#> log(lotsize)  0.0729753  0.0030642  23.8153 < 2.2e-16
#> rooms        -0.0025340  0.0030414  -0.8332 0.4047438
#> log(TLA)      0.5778330  0.0101653  56.8437 < 2.2e-16
#> beds          0.0156215  0.0045220   3.4545 0.0005512
#> syear1994     0.0444752  0.0073880   6.0199 1.745e-09
#> syear1995     0.0860740  0.0072149  11.9300 < 2.2e-16
#> syear1996     0.1059371  0.0069873  15.1613 < 2.2e-16
#> syear1997     0.1473471  0.0069302  21.2615 < 2.2e-16
#> syear1998     0.2007216  0.0071253  28.1703 < 2.2e-16
#> 
#> Rho: 0.52281, LR test value: 12594, p-value: < 2.22e-16
#> Approximate (numerical Hessian) standard error: 0.0038416
#>     z-value: 136.09, p-value: < 2.22e-16
#> Wald statistic: 18521, p-value: < 2.22e-16
#> 
#> Log likelihood: -7670.362 for lag model
#> ML residual variance (sigma squared): 0.094786, (sigma: 0.30787)
#> Number of observations: 25357 
#> Number of parameters estimated: 15 
#> AIC: 15371, (AIC for lm: 27962)
#> 

loobj <- impacts(lobj, tr=trMat, R=1000)
summary(loobj, zstats=TRUE, short=TRUE)

#> Impact measures (lag, trace):
#>                    Direct     Indirect        Total
#> age           1.495027201  1.247025340  2.742052541
#> I(age^2)     -2.652295154 -2.212320461 -4.864615615
#> I(age^3)      0.748268069  0.624141983  1.372410052
#> log(lotsize)  0.083379995  0.069548545  0.152928539
#> rooms        -0.002895346 -0.002415053 -0.005310398
#> log(TLA)      0.660219207  0.550699064  1.210918271
#> beds          0.017848752  0.014887921  0.032736673
#> syear1994     0.050816402  0.042386748  0.093203150
#> syear1995     0.098346272  0.082032148  0.180378420
#> syear1996     0.121041417  0.100962520  0.222003938
#> syear1997     0.168355573  0.140427991  0.308783564
#> syear1998     0.229340074  0.191296107  0.420636180
#> ========================================================
#> Simulation results (mixed Hessian approximation variance matrix):
#> ========================================================
#> Simulated standard errors
#>                   Direct    Indirect       Total
#> age          0.064296438 0.055323249 0.118497939
#> I(age^2)     0.115360929 0.098123688 0.211488373
#> I(age^3)     0.062144038 0.052041435 0.113886372
#> log(lotsize) 0.003366268 0.002819239 0.006117351
#> rooms        0.003373582 0.002816658 0.006190009
#> log(TLA)     0.011357951 0.011054895 0.021274151
#> beds         0.005234585 0.004380382 0.009612636
#> syear1994    0.008477181 0.007110487 0.015577465
#> syear1995    0.008361287 0.007043902 0.015367150
#> syear1996    0.008041709 0.006838250 0.014820783
#> syear1997    0.007818221 0.006751750 0.014453900
#> syear1998    0.008347923 0.007398409 0.015549411
#> 
#> Simulated z-values:
#>                  Direct    Indirect       Total
#> age           23.246302  22.5444222  23.1386730
#> I(age^2)     -22.993254 -22.5570316 -23.0078949
#> I(age^3)      12.045514  12.0023762  12.0574372
#> log(lotsize)  24.789519  24.6983542  25.0236969
#> rooms         -0.868709  -0.8686937  -0.8687345
#> log(TLA)      58.086508  49.7994169  56.8892760
#> beds           3.453218   3.4438856   3.4498027
#> syear1994      5.977552   5.9470557   5.9675471
#> syear1995     11.730998  11.6199189  11.7091203
#> syear1996     15.001064  14.7211639  14.9318155
#> syear1997     21.514961  20.7895766  21.3488924
#> syear1998     27.484866  25.8792188  27.0689717
#> 
#> Simulated p-values:
#>              Direct     Indirect   Total     
#> age          < 2.22e-16 < 2.22e-16 < 2.22e-16
#> I(age^2)     < 2.22e-16 < 2.22e-16 < 2.22e-16
#> I(age^3)     < 2.22e-16 < 2.22e-16 < 2.22e-16
#> log(lotsize) < 2.22e-16 < 2.22e-16 < 2.22e-16
#> rooms        0.38500632 0.38501467 0.384992  
#> log(TLA)     < 2.22e-16 < 2.22e-16 < 2.22e-16
#> beds         0.00055394 0.00057342 0.000561  
#> syear1994    2.2652e-09 2.7301e-09 2.4085e-09
#> syear1995    < 2.22e-16 < 2.22e-16 < 2.22e-16
#> syear1996    < 2.22e-16 < 2.22e-16 < 2.22e-16
#> syear1997    < 2.22e-16 < 2.22e-16 < 2.22e-16
#> syear1998    < 2.22e-16 < 2.22e-16 < 2.22e-16

lobj1 <- stsls(form, house, lw)
loobj1 <- impacts(lobj1, tr=trMat, R=1000)
summary(loobj1, zstats=TRUE, short=TRUE)

#> Impact measures (lag, trace):
#>                    Direct     Indirect       Total
#> age           1.492873357  1.265395569  2.75826893
#> I(age^2)     -2.642549749 -2.239889088 -4.88243884
#> I(age^3)      0.744861040  0.631362235  1.37622327
#> log(lotsize)  0.082511049  0.069938361  0.15244941
#> rooms        -0.003035731 -0.002573159 -0.00560889
#> log(TLA)      0.658778386  0.558396495  1.21717488
#> beds          0.018256951  0.015475033  0.03373198
#> syear1994     0.050971955  0.043205062  0.09417702
#> syear1995     0.098646802  0.083615416  0.18226222
#> syear1996     0.121389057  0.102892301  0.22428136
#> syear1997     0.168907859  0.143170387  0.31207825
#> syear1998     0.230052473  0.194998041  0.42505051
#> ========================================================
#> Simulation results (asymptotic IV variance matrix):
#> ========================================================
#> Simulated standard errors
#>                   Direct    Indirect       Total
#> age          0.063940403 0.058147878 0.118645008
#> I(age^2)     0.115356332 0.102073661 0.211300675
#> I(age^3)     0.061737586 0.052713683 0.113534373
#> log(lotsize) 0.003638760 0.002922621 0.006362103
#> rooms        0.003597558 0.003054142 0.006651031
#> log(TLA)     0.011573499 0.014387138 0.023064479
#> beds         0.005239154 0.004488962 0.009721066
#> syear1994    0.008450587 0.007249922 0.015667840
#> syear1995    0.008414578 0.007392580 0.015690713
#> syear1996    0.007853922 0.007120803 0.014791393
#> syear1997    0.008034115 0.007755663 0.015469672
#> syear1998    0.008226236 0.008551863 0.016246736
#> 
#> Simulated z-values:
#>                   Direct    Indirect       Total
#> age           23.2989820  21.7222194  23.2023860
#> I(age^2)     -22.8521297 -21.8956086 -23.0529449
#> I(age^3)      12.0096399  11.9245663  12.0671295
#> log(lotsize)  22.6998261  23.9565936  23.9881786
#> rooms         -0.8739561  -0.8746701  -0.8743718
#> log(TLA)      56.9078117  38.8154572  52.7679744
#> beds           3.5109410   3.4768651   3.4977518
#> syear1994      6.0577454   5.9875499   6.0378954
#> syear1995     11.6940386  11.2870056  11.5890516
#> syear1996     15.4457534  14.4463751  15.1560791
#> syear1997     21.0088714  18.4557137  20.1635809
#> syear1998     27.9328908  22.7855472  26.1370308
#> 
#> Simulated p-values:
#>              Direct     Indirect   Total     
#> age          < 2.22e-16 < 2.22e-16 < 2.22e-16
#> I(age^2)     < 2.22e-16 < 2.22e-16 < 2.22e-16
#> I(age^3)     < 2.22e-16 < 2.22e-16 < 2.22e-16
#> log(lotsize) < 2.22e-16 < 2.22e-16 < 2.22e-16
#> rooms        0.38214217 0.38175344 0.3819158 
#> log(TLA)     < 2.22e-16 < 2.22e-16 < 2.22e-16
#> beds         0.00044652 0.00050731 0.0004692 
#> syear1994    1.3804e-09 2.1303e-09 1.5614e-09
#> syear1995    < 2.22e-16 < 2.22e-16 < 2.22e-16
#> syear1996    < 2.22e-16 < 2.22e-16 < 2.22e-16
#> syear1997    < 2.22e-16 < 2.22e-16 < 2.22e-16
#> syear1998    < 2.22e-16 < 2.22e-16 < 2.22e-16

mobj <- lagsarlm(form, house, lw, type="mixed",
 method="Matrix", control=list(fdHess=TRUE), trs=trMat)
summary(mobj)

#> 
#> Call:lagsarlm(formula = form, data = house, listw = lw, type = "mixed", 
#>     method = "Matrix", trs = trMat, control = list(fdHess = TRUE))
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -2.421437 -0.126022  0.029935  0.167534  2.895359 
#> 
#> Type: mixed 
#> Coefficients: (numerical Hessian approximate standard errors) 
#>                     Estimate  Std. Error  z value  Pr(>|z|)
#> (Intercept)       0.67918314  0.08721235   7.7877 6.883e-15
#> age               0.97905382  0.08453611  11.5815 < 2.2e-16
#> I(age^2)         -1.87856579  0.13822485 -13.5907 < 2.2e-16
#> I(age^3)          0.55454470  0.06800288   8.1547 4.441e-16
#> log(lotsize)      0.10944213  0.00522190  20.9583 < 2.2e-16
#> rooms             0.00136728  0.00302395   0.4521  0.651162
#> log(TLA)          0.62148335  0.01060003  58.6303 < 2.2e-16
#> beds              0.01294189  0.00448716   2.8842  0.003924
#> syear1994         0.04031417  0.00725573   5.5562 2.757e-08
#> syear1995         0.08394251  0.00708737  11.8440 < 2.2e-16
#> syear1996         0.10400813  0.00686112  15.1591 < 2.2e-16
#> syear1997         0.14696274  0.00680970  21.5814 < 2.2e-16
#> syear1998         0.20021074  0.00699781  28.6105 < 2.2e-16
#> lag.age          -0.04653535  0.09647336  -0.4824  0.629547
#> lag.I(age^2)      0.27074756  0.16312167   1.6598  0.096957
#> lag.I(age^3)     -0.32908972  0.08295047  -3.9673 7.269e-05
#> lag.log(lotsize) -0.05072331  0.00577489  -8.7834 < 2.2e-16
#> lag.rooms         0.00050653  0.00441843   0.1146  0.908731
#> lag.log(TLA)     -0.07588857  0.01502932  -5.0494 4.433e-07
#> lag.beds         -0.04306553  0.00657416  -6.5507 5.726e-11
#> lag.syear1994    -0.02428053  0.01111700  -2.1841  0.028956
#> lag.syear1995    -0.04415580  0.01084657  -4.0709 4.682e-05
#> lag.syear1996    -0.05712325  0.01050600  -5.4372 5.412e-08
#> lag.syear1997    -0.09260271  0.01042633  -8.8816 < 2.2e-16
#> lag.syear1998    -0.09696080  0.01078352  -8.9916 < 2.2e-16
#> 
#> Rho: 0.53828, LR test value: 9593.4, p-value: < 2.22e-16
#> Approximate (numerical Hessian) standard error: 0.0046668
#>     z-value: 115.34, p-value: < 2.22e-16
#> Wald statistic: 13304, p-value: < 2.22e-16
#> 
#> Log likelihood: -7307.507 for mixed model
#> ML residual variance (sigma squared): 0.091319, (sigma: 0.30219)
#> Number of observations: 25357 
#> Number of parameters estimated: 27 
#> AIC: 14669, (AIC for lm: 24260)
#> 

moobj <- impacts(mobj, tr=trMat, R=1000)
summary(moobj, zstats=TRUE, short=TRUE)

#> Impact measures (mixed, trace):
#>                    Direct     Indirect        Total
#> age           1.116551228  0.903101121  2.019652349
#> I(age^2)     -2.090461358 -1.391758011 -3.482219369
#> I(age^3)      0.545790447 -0.057499114  0.488291333
#> log(lotsize)  0.111785027  0.015388439  0.127173466
#> rooms         0.001722848  0.002335444  0.004058292
#> log(TLA)      0.695500044  0.486151389  1.181651434
#> beds          0.002611230 -0.067853136 -0.065241906
#> syear1994     0.039575758 -0.004850020  0.034725738
#> syear1995     0.084236861  0.001933350  0.086170211
#> syear1996     0.103682464 -0.002139009  0.101543455
#> syear1997     0.143100555 -0.025367367  0.117733187
#> syear1998     0.203303767  0.020315386  0.223619153
#> ========================================================
#> Simulation results (mixed Hessian approximation variance matrix):
#> ========================================================
#> Simulated standard errors
#>                   Direct    Indirect       Total
#> age          0.077912086 0.131881976 0.143882369
#> I(age^2)     0.129796126 0.237613423 0.273074699
#> I(age^3)     0.066227363 0.129395850 0.154590639
#> log(lotsize) 0.004782001 0.007064361 0.007246254
#> rooms        0.003469062 0.008355429 0.010599350
#> log(TLA)     0.011915239 0.026025735 0.032701944
#> beds         0.005156452 0.013021177 0.016492973
#> syear1994    0.008933529 0.022310364 0.028858981
#> syear1995    0.008774074 0.021322232 0.027731068
#> syear1996    0.008265616 0.020924180 0.026923516
#> syear1997    0.008255624 0.020592579 0.026557065
#> syear1998    0.008734415 0.021710987 0.028265858
#> 
#> Simulated z-values:
#>                   Direct    Indirect       Total
#> age           14.3647778  6.85083025  14.0579477
#> I(age^2)     -16.1327047 -5.86383428 -12.7704555
#> I(age^3)       8.2576961 -0.44044912   3.1689703
#> log(lotsize)  23.3692705  2.17781122  17.5451661
#> rooms          0.5358941  0.31678906   0.4251165
#> log(TLA)      58.3752150 18.66943702  36.1275297
#> beds           0.4257952 -5.24641974  -4.0089174
#> syear1994      4.4062433 -0.22894696   1.1869931
#> syear1995      9.5881162  0.08576743   3.0996136
#> syear1996     12.5708803 -0.08723417   3.7915092
#> syear1997     17.3303952 -1.20514282   4.4529101
#> syear1998     23.2902952  0.94925557   7.9260418
#> 
#> Simulated p-values:
#>              Direct     Indirect   Total     
#> age          < 2.22e-16 7.3423e-12 < 2.22e-16
#> I(age^2)     < 2.22e-16 4.5230e-09 < 2.22e-16
#> I(age^3)     2.2204e-16 0.65961    0.00152980
#> log(lotsize) < 2.22e-16 0.02942    < 2.22e-16
#> rooms        0.59203    0.75140    0.67075179
#> log(TLA)     < 2.22e-16 < 2.22e-16 < 2.22e-16
#> beds         0.67026    1.5508e-07 6.0998e-05
#> syear1994    1.0518e-05 0.81891    0.23523035
#> syear1995    < 2.22e-16 0.93165    0.00193773
#> syear1996    < 2.22e-16 0.93049    0.00014973
#> syear1997    < 2.22e-16 0.22815    8.4714e-06
#> syear1998    < 2.22e-16 0.34249    2.2204e-15