include.self {spdep}	R Documentation

Include self in neighbours list

Description

The function includes the region itself in its own list of neighbours, and sets attribute "self.included" to TRUE.

Usage

include.self(nb)

Arguments

nb
input neighbours list of class nb

Value

The function returns an object of class nb with a list of integer vectors containing neighbour region number ids; attribute "self.included" is set to TRUE.

Author(s)

Roger Bivand Roger.Bivand@nhh.no

See Also

summary.nb

Examples

example(columbus)
coords <- coordinates(columbus)
summary(col.gal.nb, coords)
summary(include.self(col.gal.nb), coords)

[Package spdep version 0.6-6 Index]