Permutation test for same colour join count statistics

A permutation test for same colour join count statistics calculated by using nsim random permutations of fx for the given spatial weighting scheme, to establish the ranks of the observed statistics (for each colour) in relation to the nsim simulated values.

joincount.mc(fx, listw, nsim, zero.policy=FALSE, alternative="greater",
 spChk=NULL)

Arguments

fx
a factor of the same length as the neighbours and weights objects in listw

listw
a listw object created for example by nb2listw

nsim
number of permutations

zero.policy
if TRUE assign zero to the lagged value of zones without neighbours, if FALSE assign NA

alternative
a character string specifying the alternative hypothesis, must be one of "greater" (default), or "less".

spChk
should the data vector names be checked against the spatial objects for identity integrity, TRUE, or FALSE, default NULL to use get.spChkOption()

Value

A list with class jclist of lists with class htest and mc.sim for each of the k colours containing the following components:

statistic

the value of the observed statistic.

parameter

the rank of the observed statistic.

method

a character string giving the method used.

data.name

a character string giving the name(s) of the data.

p.value

the pseudo p-value of the test.

alternative

a character string describing the alternative hypothesis.

estimate

the mean and variance of the simulated distribution.

res

nsim simulated values of statistic, the final element is the observed statistic

References

Cliff, A. D., Ord, J. K. 1981 Spatial processes, Pion, p. 63-5.

See also

Examples

data(oldcol)
HICRIME <- cut(COL.OLD$CRIME, breaks=c(0,35,80), labels=c("low","high"))
names(HICRIME) <- rownames(COL.OLD)
joincount.mc(HICRIME, nb2listw(COL.nb, style="B"), nsim=99)

#> 
#> 	Monte-Carlo simulation of join-count statistic
#> 
#> data:  HICRIME 
#> weights: nb2listw(COL.nb, style = "B") 
#> number of simulations + 1: 100 
#> 
#> Join-count statistic for low = 34, rank of observed statistic = 83.5,
#> p-value = 0.165
#> alternative hypothesis: greater
#> sample estimates:
#>     mean of simulation variance of simulation 
#>               29.77778               16.80726 
#> 
#> 
#> 	Monte-Carlo simulation of join-count statistic
#> 
#> data:  HICRIME 
#> weights: nb2listw(COL.nb, style = "B") 
#> number of simulations + 1: 100 
#> 
#> Join-count statistic for high = 54, rank of observed statistic = 100,
#> p-value = 0.01
#> alternative hypothesis: greater
#> sample estimates:
#>     mean of simulation variance of simulation 
#>               28.06061               18.03711 
#> 

joincount.test(HICRIME, nb2listw(COL.nb, style="B"))

#> 
#> 	Join count test under nonfree sampling
#> 
#> data:  HICRIME 
#> weights: nb2listw(COL.nb, style = "B") 
#> 
#> Std. deviate for low = 1.0141, p-value = 0.1553
#> alternative hypothesis: greater
#> sample estimates:
#> Same colour statistic           Expectation              Variance 
#>              34.00000              29.59184              18.89550 
#> 
#> 
#> 	Join count test under nonfree sampling
#> 
#> data:  HICRIME 
#> weights: nb2listw(COL.nb, style = "B") 
#> 
#> Std. deviate for high = 6.3307, p-value = 1.22e-10
#> alternative hypothesis: greater
#> sample estimates:
#> Same colour statistic           Expectation              Variance 
#>              54.00000              27.22449              17.88838 
#>