lee {spdep}	R Documentation

Compute Lee's statistic

Description

A simple function to compute Lee's L statistic for bivariate spatial data;

L(x,y) = (n sum_i (sum_j w_ij (x_i - xbar)) (sum_j w_ij (y_j - ybar))) / (S2 sqrt(sum_i (x_i - xbar)^2)) sqrt(sum_i (x_i - xbar)^2))

Usage

lee(x, y, listw, n, S2, zero.policy=NULL, NAOK=FALSE)

Arguments

x
a numeric vector the same length as the neighbours list in listw

y
a numeric vector the same length as the neighbours list in listw

listw
a listw object created for example by nb2listw

n
number of zones

S2
Sum of squared sum of weights by rows.

zero.policy
default NULL, use global option value; if TRUE assign zero to the lagged value of zones without neighbours, if FALSE assign NA

NAOK
if 'TRUE' then any 'NA' or 'NaN' or 'Inf' values in x are passed on to the foreign function. If 'FALSE', the presence of 'NA' or 'NaN' or 'Inf' values is regarded as an error.

Value

a list of

L
Lee's L statistic

local L
Lee's local L statistic

Author(s)

Roger Bivand and Virgiio Gómez-Rubio Virgilio.Gomez@uclm.es

References

Lee (2001). Developing a bivariate spatial association measure: An integration of Pearson's r and Moran's I. J Geograph Syst 3: 369-385

See Also

lee.mc

Examples

data(boston)
lw<-nb2listw(boston.soi)

x<-boston.c$CMEDV
y<-boston.c$CRIM
z<-boston.c$RAD

Lxy<-lee(x, y, lw, length(x), zero.policy=TRUE)
Lxz<-lee(x, z, lw, length(x), zero.policy=TRUE)

[Package spdep version 0.6-6 Index]