listw2sn {spdep}	R Documentation

Spatial neighbour sparse representation

Description

The function makes a "spatial neighbour" object representation (similar to the S-PLUS spatial statististics module representation of a "listw" spatial weights object. sn2listw() is the inverse function to listw2sn(), creating a "listw" object from a "spatial neighbour" object. The as.spam.listw method converts a "listw" object to a sparse matrix as defined in the spam package, using listw2sn().

Usage

listw2sn(listw)
sn2listw(sn)
as.spam.listw(listw)

Arguments

listw
a listw object from for example nb2listw

sn
a spatial.neighbour object

Value

listw2sn()returns a data frame with three columns, and with class spatial.neighbour:

from
region number id for the start of the link (S-PLUS row.id)

to
region number id for the end of the link (S-PLUS col.id)

weights
weight for this link

Author(s)

Roger Bivand Roger.Bivand@nhh.no

See Also

nb2listw

Examples

example(columbus)
col.listw <- nb2listw(col.gal.nb)
col.listw$neighbours[[1]]
col.listw$weights[[1]]
col.sn <- listw2sn(col.listw)
str(col.sn)
## Not run: 
col.sp <- as.spam.listw(col.listw)
str(col.sp)

## End(Not run)

[Package spdep version 0.6-6 Index]