lm.LMtests {spdep}	R Documentation

Lagrange Multiplier diagnostics for spatial dependence in linear models

Description

The function reports the estimates of tests chosen among five statistics for testing for spatial dependence in linear models. The statistics are the simple LM test for error dependence (LMerr), the simple LM test for a missing spatially lagged dependent variable (LMlag), variants of these robust to the presence of the other (RLMerr, RLMlag - RLMerr tests for error dependence in the possible presence of a missing lagged dependent variable, RLMlag the other way round), and a portmanteau test (SARMA, in fact LMerr + RLMlag). Note: from spdep 0.3-32, the value of the weights matrix trace term is returned correctly for both underlying symmetric and asymmetric neighbour lists, before 0.3-32, the value was wrong for listw objects based on asymmetric neighbour lists, such as k-nearest neighbours (thanks to Luc Anselin for finding the bug).

Usage

lm.LMtests(model, listw, zero.policy=NULL, test="LMerr", spChk=NULL, naSubset=TRUE)
## S3 method for class 'LMtestlist'
print(x, ...)
## S3 method for class 'LMtestlist'
summary(object, p.adjust.method="none", ...)
## S3 method for class 'LMtestlist.summary'
print(x, digits=max(3, getOption("digits") - 2), ...)

Arguments

model
an object of class lm returned by lm, or optionally a vector of externally calculated residuals (run though na.omit if any NAs present) for use when only "LMerr" is chosen; weights and offsets should not be used in the lm object

listw
a listw object created for example by nb2listw, expected to be row-standardised (W-style)

zero.policy
default NULL, use global option value; if TRUE assign zero to the lagged value of zones without neighbours, if FALSE assign NA

test
a character vector of tests requested chosen from LMerr, LMlag, RLMerr, RLMlag, SARMA; test="all" computes all the tests.

spChk
should the data vector names be checked against the spatial objects for identity integrity, TRUE, or FALSE, default NULL to use get.spChkOption()

naSubset
default TRUE to subset listw object for omitted observations in model object (this is a change from earlier behaviour, when the model$na.action component was ignored, and the listw object had to be subsetted by hand)

x, object
object to be printed

p.adjust.method
a character string specifying the probability value adjustment (see p.adjust) for multiple tests, default "none"

digits
minimum number of significant digits to be used for most numbers

...
printing arguments to be passed through

Details

The two types of dependence are for spatial lag rho and spatial error lambda:

y = X beta + rho W1 y + u

u = lambda W2 u + e

where e is a well-behaved, uncorrelated error term. Tests for a missing spatially lagged dependent variable test that rho = 0, tests for spatial autocorrelation of the error u test whether lambda = 0. W is a spatial weights matrix; for the tests used here they are identical.

Value

A list of class LMtestlist of htest objects, each with:

statistic
the value of the Lagrange Multiplier test.

parameter
number of degrees of freedom

p.value
the p-value of the test.

method
a character string giving the method used.

data.name
a character string giving the name(s) of the data.

Author(s)

Roger Bivand Roger.Bivand@nhh.no and Andrew Bernat

References

Anselin, L. 1988 Spatial econometrics: methods and models. (Dordrecht: Kluwer); Anselin, L., Bera, A. K., Florax, R. and Yoon, M. J. 1996 Simple diagnostic tests for spatial dependence. Regional Science and Urban Economics, 26, 77–104.

See Also

lm

Examples

data(oldcol)
oldcrime.lm <- lm(CRIME ~ HOVAL + INC, data = COL.OLD)
summary(oldcrime.lm)
res <- lm.LMtests(oldcrime.lm, nb2listw(COL.nb), test=c("LMerr", "LMlag",
  "RLMerr", "RLMlag", "SARMA"))
summary(res)
lm.LMtests(oldcrime.lm, nb2listw(COL.nb))
lm.LMtests(residuals(oldcrime.lm), nb2listw(COL.nb))

[Package spdep version 0.6-6 Index]