Exact local Moran's Ii tests

localmoran.exact provides exact local Moran's Ii tests under the null hypothesis, while localmoran.exact.alt provides exact local Moran's Ii tests under the alternative hypothesis. In this case, the model may be a fitted model derived from a model fitted by errorsarlm, with the covariance matrix can be passed through the Omega= argument.

localmoran.exact(model, select, nb, glist = NULL, style = "W", 
 zero.policy = NULL, alternative = "greater", spChk = NULL, 
 resfun = weighted.residuals, save.Vi = FALSE, useTP=FALSE, truncErr=1e-6, 
 zeroTreat=0.1)
localmoran.exact.alt(model, select, nb, glist = NULL, style = "W",
 zero.policy = NULL, alternative = "greater", spChk = NULL,
 resfun = weighted.residuals, Omega = NULL, save.Vi = FALSE,
 save.M = FALSE, useTP=FALSE, truncErr=1e-6, zeroTreat=0.1)
# S3 method for localmoranex
print(x, …)
# S3 method for localmoranex
as.data.frame(x, row.names=NULL, optional=FALSE, …)

Arguments

model
an object of class lm returned by lm (assuming no global spatial autocorrelation), or an object of class sarlm returned by a spatial simultaneous autoregressive model fit (assuming global spatial autocorrelation represented by the model spatial coefficient); weights may be specified in the lm fit, but offsets should not be used

select
an integer vector of the id. numbers of zones to be tested; if missing, all zones

nb
a list of neighbours of class nb

glist
a list of general weights corresponding to neighbours

style
can take values W, B, C, and S

zero.policy
default NULL, use global option value; if TRUE assign zero to the lagged value of zones without neighbours, if FALSE assign NA

alternative
a character string specifying the alternative hypothesis, must be one of greater (default), less or two.sided.

spChk
should the data vector names be checked against the spatial objects for identity integrity, TRUE, or FALSE, default NULL to use get.spChkOption()

resfun
default: weighted.residuals; the function to be used to extract residuals from the lm object, may be residuals, weighted.residuals, rstandard, or rstudent

Omega
A SAR process matrix may be passed in to test an alternative hypothesis, for example Omega <- invIrW(listw, rho=0.1); Omega <- tcrossprod(Omega), chol() is taken internally

save.Vi
if TRUE, return the star-shaped weights lists for each zone tested

save.M
if TRUE, save a list of left and right M products

useTP
default FALSE, if TRUE, use truncation point in integration rather than upper=Inf, see Tiefelsdorf (2000), eq. 6.7, p.69

truncErr
when useTP=TRUE, pass truncation error to truncation point function

zeroTreat
when useTP=TRUE, pass zero adjustment to truncation point function

x
object to be printed

row.names
ignored argument to as.data.frame.localmoranex; row names assigned from localmoranex object

optional
ignored argument to as.data.frame.localmoranex; row names assigned from localmoranex object

…
arguments to be passed through

Value

A list with class localmoranex containing "select" lists, each with class moranex with the following components:

statistic

the value of the exact standard deviate of global Moran's I.

p.value

the p-value of the test.

estimate

the value of the observed local Moran's Ii.

method

a character string giving the method used.

alternative

a character string describing the alternative hypothesis.

gamma

eigenvalues (two extreme values for null, vector for alternative)

oType

usually set to "E", but set to "N" if the integration leads to an out of domain value for qnorm, when the Normal assumption is substituted. This only occurs when the output p-value would be very close to zero

data.name

a character string giving the name(s) of the data.

df

degrees of freedom

i

zone tested

Vi

zone tested

When the alternative is being tested, a list of left and right M products in attribute M.
References

Bivand RS, Müller W, Reder M (2009) Power calculations for global and local Moran’s I. Comput Stat Data Anal 53:2859–2872; Bivand RS, Wong DWS 2018 Comparing implementations of global and local indicators of spatial association. TEST, 27(3), 716--748 https://doi.org/10.1007/s11749-018-0599-x

See also

Examples

if (require(rgdal, quietly=TRUE)) {
eire <- readOGR(system.file("shapes/eire.shp", package="spData")[1])
row.names(eire) <- as.character(eire$names)
proj4string(eire) <- CRS("+proj=utm +zone=30 +ellps=airy +units=km")
eire.nb <- poly2nb(eire)
e.lm <- lm(OWNCONS ~ ROADACC, data=eire)
localmoran.sad(e.lm, nb=eire.nb)
localmoran.exact(e.lm, nb=eire.nb)
localmoran.exact(e.lm, nb=eire.nb, useTP=TRUE)
e.errorsar <- errorsarlm(OWNCONS ~ ROADACC, data=eire,
 listw=nb2listw(eire.nb))
lm.target <- lm(e.errorsar$tary ~ e.errorsar$tarX - 1)
localmoran.exact.alt(lm.target, nb=eire.nb)
Omega <- invIrW(nb2listw(eire.nb), rho=0.6)
Omega1 <- tcrossprod(Omega)
localmoran.exact.alt(lm.target, nb=eire.nb, Omega=Omega1)
localmoran.exact.alt(lm.target, nb=eire.nb, Omega=Omega1, useTP=TRUE)
}

#> OGR data source with driver: ESRI Shapefile 
#> Source: "/Library/Frameworks/R.framework/Versions/3.4/Resources/library/spData/shapes/eire.shp", layer: "eire"
#> with 26 features
#> It has 10 fields

#>              Local Morans I    Exact SD Pr. (exact)
#> 1 Carlow         0.17958462  0.23106345  0.40863276
#> 2 Cavan         -0.24752628 -1.23468995  0.89152704
#> 3 Clare         -0.27901334 -1.23568183  0.89171157
#> 4 Cork           0.37808655  0.37129179  0.35521011
#> 5 Donegal       -1.01894688 -1.84902980  0.96777324
#> 6 Dublin        -0.18171297 -0.98014903  0.83649372
#> 7 Galway         1.02193390  1.60349632  0.05441256
#> 8 Kerry         -0.94967914 -2.12070821  0.98302682
#> 9 Kildare        0.07005053  0.10764406  0.45713902
#> 10 Kilkenny      0.43022231  0.72290590  0.23486885
#> 11 Laoghis      -0.12239133 -0.87106429  0.80814047
#> 12 Leitrim      -0.24203970 -1.37942813  0.91611860
#> 13 Limerick     -0.03214546 -0.68260211  0.75257085
#> 14 Longford      0.38307454  0.47071586  0.31892183
#> 15 Louth         0.21301552  0.02198203  0.49123115
#> 16 Mayo          0.93971200  1.30000026  0.09680044
#> 17 Meath         0.12484415  0.15994587  0.43646186
#> 18 Monaghan     -0.16109919 -1.05555896  0.85441512
#> 19 Offaly       -0.00632492 -0.29479749  0.61592571
#> 20 Roscommon     1.02089429  1.58863472  0.05607144
#> 21 Sligo        -2.01629233 -3.26455056  0.99945181
#> 22 Tipperary    -0.10810709 -0.98217470  0.83699310
#> 23 Waterford     0.44099279  0.66779690  0.25213162
#> 24 Westmeath    -0.06329661 -0.70910521  0.76087041
#> 25 Wexford       0.30764883  0.37770018  0.35282667
#> 26 Wicklow      -0.01696406 -0.54778213  0.70807925