plot.mst {spdep}	R Documentation

Plot the Minimum Spanning Tree

Description

This function plots a MST, the nodes are circles and the edges are segments.

Usage

## S3 method for class 'mst'
plot(x, coords, label.areas = NULL, 
   cex.circles = 1, cex.labels = 1, ...)

Arguments

x
Object of mst class.

coords
A two column matrix with the coordinates of nodes.

label.areas
A vector with the labels of nodes

cex.circles
The length of circles to plot.

cex.labels
The length of nodes labels ploted.

...
Further arguments passed to plotting funcitons.

Author(s)

Elias T. Krainski and Renato M. Assuncao

See Also

See Also as skater and mstree

Examples

### see example in mstree function documentation

[Package spdep version 0.6-6 Index]