residuals.sarlm {spdep}	R Documentation

Access functions for spatial simultaneous autoregressive linear model objects

Description

Access functions for residuals, deviance, coefficients and fitted values from spatial simultaneous autoregressive linear model objects

Usage

## S3 method for class 'sarlm'
residuals(object, ...)
## S3 method for class 'sarlm'
deviance(object, ...)
## S3 method for class 'sarlm'
coef(object, ...)
## S3 method for class 'sarlm'
vcov(object, ...)
## S3 method for class 'sarlm'
fitted(object, ...)

Arguments

object
sarlm object returned by lagsarlm or errorsarlm

...
further arguments passed through

Value

Revelant vectors of numerical values.

Note

fitted.sarlm() returns the difference between residuals() for the same object and the response variable; predict.sarlm() returns a decomposition into trend and signal for the fit.

Author(s)

Roger Bivand, Roger.Bivand@nhh.no

See Also

errorsarlm, lagsarlm, predict.sarlm

[Package spdep version 0.6-6 Index]