Access functions for spatial simultaneous autoregressive linear model objects

Access functions for residuals, deviance, coefficients and fitted values from spatial simultaneous autoregressive linear model objects

# S3 method for sarlm
residuals(object, ...)
# S3 method for sarlm
deviance(object, ...)
# S3 method for sarlm
coef(object, ...)
# S3 method for sarlm
vcov(object, ...)
# S3 method for sarlm
fitted(object, ...)

Arguments

object
sarlm object returned by lagsarlm or errorsarlm

...
further arguments passed through

Value

Revelant vectors of numerical values.

Note

fitted.sarlm() returns the difference between residuals() for the same object and the response variable; predict.sarlm() returns a decomposition into trend and signal for the fit.

See also