Options for parallel support

Provides support for the use of parallel computation in the parallel package.

set.mcOption(value)
get.mcOption()
set.coresOption(value)
get.coresOption()
set.ClusterOption(cl)
get.ClusterOption()

Arguments

value

valid replacement value

cl

a cluster object created by makeCluster in parallel

Details

Options in the spdep package are held in an environment local to the package namespace and not exported. Option values are set and retrieved with pairs of access functions, get and set. The mc option is set by default to FALSE on Windows systems, as they cannot fork the R session; by default it is TRUE on other systems, but may be set FALSE. If mc is FALSE, the Cluster option is used: if mc is FALSE and the Cluster option is NULL no parallel computing is done, or the Cluster option is passed a “cluster” object created by the parallel or snow package for access without being passed as an argument. The cores option is set to NULL by default, and can be used to store the number of cores to use as an integer. If cores is NULL, facilities from the parallel package will not be used.

Value

The option access functions return their current settings, the assignment functions usually return the previous value of the option.

Note

An extended example is shown in the documentation of aple.mc, including treatment of seeding of RNG for multicore/cluster.

Author

Roger Bivand Roger.Bivand@nhh.no

Examples

ls(envir=spdep:::.spdepOptions)
#> [1] "cluster"                "cores"                  "listw_is_CsparseMatrix"
#> [4] "mc"                     "rlecuyerSeed"           "spChkID"               
#> [7] "verbose"                "zeroPolicy"            
library(parallel)
nc <- detectCores(logical=FALSE)
nc
#> [1] 6
# set nc to 1L here
if (nc > 1L) nc <- 1L
#nc <- ifelse(nc > 2L, 2L, nc)
coresOpt <- get.coresOption()
coresOpt
#> NULL
if (!is.na(nc)) {
 invisible(set.coresOption(nc))
 print(exists("aple.mc"))
 if(.Platform$OS.type == "windows") {
# forking not permitted on Windows - start cluster
  print(get.mcOption())
  cl <- makeCluster(get.coresOption())
  print(clusterEvalQ(cl, exists("aple.mc")))
  set.ClusterOption(cl)
  clusterEvalQ(get.ClusterOption(), library(spdep))
  print(clusterEvalQ(cl, exists("aple.mc")))
  clusterEvalQ(get.ClusterOption(), detach(package:spdep))
  set.ClusterOption(NULL)
  print(clusterEvalQ(cl, exists("aple.mc")))
  stopCluster(cl)
 } else {
  mcOpt <- get.mcOption()
  print(mcOpt)
  print(mclapply(1:get.coresOption(), function(i) exists("aple.mc"),
   mc.cores=get.coresOption()))
  invisible(set.mcOption(FALSE))
  cl <- makeCluster(nc)
  print(clusterEvalQ(cl, exists("aple.mc")))
  set.ClusterOption(cl)
  clusterEvalQ(get.ClusterOption(), library(spdep))
  print(clusterEvalQ(cl, exists("aple.mc")))
  clusterEvalQ(get.ClusterOption(), detach(package:spdep))
  set.ClusterOption(NULL)
  print(clusterEvalQ(cl, exists("aple.mc")))
  stopCluster(cl)
  invisible(set.mcOption(mcOpt))
 }
 invisible(set.coresOption(coresOpt))
}
#> [1] TRUE
#> [1] TRUE
#> [[1]]
#> [1] TRUE
#> 
#> [[1]]
#> [1] FALSE
#> 
#> [[1]]
#> [1] FALSE
#> 
#> [[1]]
#> [1] FALSE
#>