summary.sarlm {spdep}	R Documentation

summary method for class sarlm

Description

Methods used for presenting the results of estimating spatial SAR models.

Usage

## S3 method for class 'sarlm'
summary(object, correlation = FALSE, Nagelkerke = FALSE, Hausman=FALSE, adj.se=FALSE, ...)
## S3 method for class 'sarlm'
print(x, ...)
## S3 method for class 'summary.sarlm'
print(x, digits = max(5, .Options$digits - 3),
	signif.stars = FALSE, ...)

Arguments

object
sarlm object from lagsarlm or errorsarlm

correlation
logical; if 'TRUE', the correlation matrix of the estimated parameters including sigma is returned and printed (default=FALSE)

Nagelkerke
if TRUE, the Nagelkerke pseudo R-squared is reported

Hausman
if TRUE, the results of the Hausman test for error models are reported

adj.se
if TRUE, adjust the coefficient standard errors for the number of fitted coefficients

x
sarlm object from lagsarlm or errorsarlm in print.sarlm, summary object from summary.sarlm for print.summary.sarlm

digits
the number of significant digits to use when printing

signif.stars
logical. If TRUE, "significance stars" are printed for each coefficient.

...
further arguments passed to or from other methods

Value

The summary function summary.sarlm returns the sarlm object augmented with a coefficient matrix with probability values for coefficient asymptotic standard errors for type="error" and for type="lag" or "mixed" when object\$ase=TRUE, or a coefficient matrix with probability values for likelihood ratio tests between the model as reported and models with independent variables dropped in turn.

Author(s)

Roger Bivand Roger.Bivand@nhh.no

References

Cliff, A. D., Ord, J. K. 1981 Spatial processes, Pion; Ord, J. K. 1975 Estimation methods for models of spatial interaction, Journal of the American Statistical Association, 70, 120-126; Anselin, L. 1988 Spatial econometrics: methods and models. (Dordrecht: Kluwer); Anselin, L. 1995 SpaceStat, a software program for the analysis of spatial data, version 1.80. Regional Research Institute, West Virginia University, Morgantown, WV (www.spacestat.com); Anselin L, Bera AK (1998) Spatial dependence in linear regression models with an introduction to spatial econometrics. In: Ullah A, Giles DEA (eds) Handbook of applied economic statistics. Marcel Dekker, New York, pp. 237-289; Nagelkerke NJD (1991) A note on a general definition of the coefficient of determination. Biometrika 78: 691-692.

See Also

errorsarlm, lagsarlm, summary.lm

Examples

data(oldcol)
COL.mix.eig <- lagsarlm(CRIME ~ INC + HOVAL, data=COL.OLD,
 nb2listw(COL.nb), type="mixed", method="eigen")
summary(COL.mix.eig, correlation=TRUE, Nagelkerke=TRUE)
COL.mix.M <- lagsarlm(CRIME ~ INC + HOVAL, data=COL.OLD,
 nb2listw(COL.nb), type="mixed", method="Matrix")
summary(COL.mix.M, correlation=TRUE, Nagelkerke=TRUE)
COL.errW.eig <- errorsarlm(CRIME ~ INC + HOVAL, data=COL.OLD,
  nb2listw(COL.nb, style="W"), method="eigen")
summary(COL.errW.eig, correlation=TRUE, Nagelkerke=TRUE, Hausman=TRUE)

[Package spdep version 0.6-6 Index]