write.nb.gal {spdep}	R Documentation

Write a neighbours list as a GAL lattice file

Description

Write a neighbours list as a GAL lattice file, may also use newer GeoDa header format

Usage

write.nb.gal(nb, file, oldstyle=TRUE, shpfile=NULL, ind=NULL)

Arguments

nb
an object of class nb with a list of integer vectors containing neighbour region number ids.

file
name of file with GAL lattice data

oldstyle
if TRUE, first line of file contains only number of spatial units, if FALSE, uses newer GeoDa style

shpfile
Shapefile name taken from GAL file for this dataset

ind
region id indicator variable name

Author(s)

Roger Bivand Roger.Bivand@nhh.no

See Also

read.gal

Examples

example(columbus)
GALfile <- tempfile("GAL")
write.nb.gal(col.gal.nb, GALfile)
col.queen <- read.gal(GALfile)
summary(diffnb(col.queen, col.gal.nb))

[Package spdep version 0.6-6 Index]